% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_counties.R
\name{add_counties}
\alias{add_counties}
\title{Add county name, FIPs, and geometry to data frame with zip codes}
\usage{
add_counties(
  df,
  statecol,
  zipcol,
  add_fips = FALSE,
  add_geo = FALSE,
  as_sf = FALSE
)
}
\arguments{
\item{df}{data frame}

\item{statecol}{bare column name column containing state abbreviations}

\item{zipcol}{bare column name containing zip codes}

\item{add_fips}{add county FIPS code column, default is \code{FALSE}}

\item{add_geo}{add county geometry column, default is \code{FALSE}}

\item{as_sf}{convert tibble to an \code{{sf}} object, default is \code{FALSE}}
}
\description{
Add county name, FIPs, and geometry to data frame with zip codes
}
\examples{
# Example data frame containing
# state abbreviation and zip code

(ex <- dplyr::tibble(state = "GA", zip = "31605"))

# Add county and latitude/longitude
ex |> add_counties(state, zip)

# Add county FIPS
ex |> add_counties(state, zip, add_fips = TRUE)

# Add county `geometry` column,
# based on county FIPS column
ex |> add_counties(state, zip, add_fips = TRUE, add_geo = TRUE)

# Convert to an `sf` object
ex |> add_counties(state, zip, add_fips = TRUE, add_geo = TRUE, as_sf = TRUE)

}
