% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-affiliation.R
\name{affiliations}
\alias{affiliations}
\title{Provider-Facility Affiliations}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}
}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  last = NULL,
  fac_type = NULL,
  ccn = NULL,
  parent_ccn = NULL,
  count = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} Individual National Provider Identifier}

\item{pac}{\verb{<chr>} Individual PECOS Associate Control ID}

\item{first, last}{\verb{<chr>} Individual provider's name}

\item{fac_type}{\verb{<enum>} facility type:
\itemize{
\item \code{esrd} = Dialysis facility
\item \code{hha} = Home health agency
\item \code{hospice} = Hospice
\item \code{hospital} = Hospital
\item \code{irf} = Inpatient rehabilitation facility
\item \code{ltch} = Long-term care hospital
\item \code{nurse} = Nursing home
\item \code{snf} = Skilled nursing facility
}}

\item{ccn}{\verb{<chr>} CCN of \code{fac_type} column's
facility \strong{or} of a \strong{unit} within the hospital where the individual
provider provides services.}

\item{parent_ccn}{\verb{<int>} CCN of the \strong{primary} hospital containing the
unit where the individual provider provides services.}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
affiliations(ccn = 331302)
affiliations(parent_ccn = 331302)
\dontshow{\}) # examplesIf}
}
