% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_days.R
\name{age_days}
\alias{age_days}
\title{Calculate Number of Days Between Two Dates}
\usage{
age_days(df, start_date = start_date, end_date = end_date)
}
\arguments{
\item{df}{data frame containing two date columns}

\item{start_date}{column containing date(s) prior to end_date column}

\item{end_date}{column containing date(s) after end_date column}
}
\value{
data frame
}
\description{
Note: This calculation includes the end date in the sum (see example)
}
\examples{
dates <- data.frame(x = seq.Date(as.Date("2021-01-01"),
                    by = "month", length.out = 3),
                    y = seq.Date(as.Date("2022-01-01"),
                    by = "month", length.out = 3))

age_days(dates, x, y)
}
