% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc_enrollment.R
\name{asc_ifed_enrollment}
\alias{asc_ifed_enrollment}
\title{Search the Ambulatory Surgical Centers and Independent Free Standing
Emergency Departments Enrolled in Medicare As Hospital Providers}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
asc_ifed_enrollment(
  npi = NULL,
  enroll_id = NULL,
  enroll_state = NULL,
  enroll_type = NULL,
  org_name = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{NPI of the Provider}

\item{enroll_id}{Unique number assigned by PECOS to the provider}

\item{enroll_state}{State where the provider is enrolled}

\item{enroll_type}{Type of enrollment - ASC to Hospital OR IFED}

\item{org_name}{Organizational name of the enrolled provider}

\item{address}{description}

\item{city}{City where the enrolled provider is located}

\item{state}{State where the provider is located}

\item{zipcode}{Zip code of the provider's location}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on Medicare Enrolled Ambulatory Surgical Center
(ASC) Providers and Independent Free Standing Emergency Departments
(IFEDs) during the COVID-19 public health emergency.
}
\details{
The Ambulatory Surgical Centers (ASCs) and Independent Free
Standing Emergency Departments (IFEDs) Enrolled in Medicare As Hospital
Providers dataset includes all Medicare enrolled Ambulatory Surgical
Centers (ASC) that converted to Hospitals during the COVID-19 public
health emergency. Additionally, this list includes Independent Free
Standing Emergency Departments (IFEDs) that have been temporarily
certified as Hospitals during the COVID-19 public health emergency.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{ASCs and IFEDs Enrolled in Medicare As Hospital Providers}
}
}
}
\note{
Update Frequency: \strong{Weekly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
asc_ifed_enrollment(city = "Atlanta")
\dontshow{\}) # examplesIf}
}
