% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beneficiaries.R
\name{beneficiaries}
\alias{beneficiaries}
\title{Beneficiary Enrollment in Medicare}
\usage{
beneficiaries(
  year = NULL,
  period = NULL,
  level = NULL,
  state = NULL,
  county = NULL,
  fips = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{Calendar year of Medicare enrollment; current years can be
checked with:
\itemize{
\item \code{bene_years("year")}: Years available for all 12 months
\item \code{bene_years("month")}: Years available for individual months
}}

\item{period}{Time frame of Medicare enrollment; options are:
\itemize{
\item \code{Year}
\item \code{Month}
\item Any individual month name
}}

\item{level}{Geographic level of data; options are:
\itemize{
\item \code{National}
\item \code{State}
\item \code{County}
}}

\item{state}{Full state name or abbreviation of beneficiary residence}

\item{county}{County of beneficiary residence}

\item{fips}{FIPS code of beneficiary residence}

\item{tidy}{Tidy output; default is \code{TRUE}}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

\code{\link[=beneficiaries]{beneficiaries()}} allows the user access current data on enrolled Medicare
beneficiaries.
}
\section{Medicare Monthly Enrollment}{

Current monthly information on the number of Medicare beneficiaries with
hospital/medical coverage and prescription drug coverage, available for
several geographic areas including national, state and county.

The hospital/medical coverage data can be broken down further by health care
delivery (Original Medicare versus Medicare Advantage and Other Health Plans)
and the prescription drug coverage data can be examined by those enrolled in
stand-alone Prescription Drug Plans and those enrolled in Medicare Advantage
Prescription Drug plans.

The dataset includes enrollee counts on a \emph{rolling 12 month basis} and also
provides information on yearly trends.
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/summary-statistics-on-beneficiary-enrollment/medicare-and-medicaid-reports/medicare-monthly-enrollment}{Medicare Monthly Enrollment}
}

\emph{Update Frequency:} \strong{Monthly}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
beneficiaries(year = 2022, period = "Year", level = "County", county = "Autauga")
beneficiaries(year = 2022, period = "July", state = "Georgia")
beneficiaries(level = "State", fips = "10")
\dontshow{\}) # examplesIf}
}
