% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_crosswalk.R
\name{betos_classification}
\alias{betos_classification}
\title{Restructured BETOS Classification System}
\usage{
betos_classification(
  hcpcs_code = NULL,
  category = NULL,
  subcategory = NULL,
  family = NULL,
  procedure = NULL,
  tidy = TRUE
)
}
\arguments{
\item{hcpcs_code}{< \emph{character} > HCPCS or CPT code}

\item{category}{< \emph{character} > RBCS Category Description}

\item{subcategory}{< \emph{character} > RBCS Subcategory Description}

\item{family}{< \emph{character} > RBCS Family Description}

\item{procedure}{< \emph{character} > Whether the HCPCS code is a Major (\code{"M"}),
Other (\code{"O"}), or non-procedure code (\code{"N"}).}

\item{tidy}{< \emph{boolean} > Tidy output; default is \code{TRUE}}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{specialty_code} \tab Code that corresponds to the Medicare specialty \cr
   \code{specialty_description} \tab Description of the Medicare provider/Supplier Type \cr
   \code{taxonomy_code} \tab Provider's taxonomy code \cr
   \code{taxonomy_description} \tab Description of the taxonomy code \cr
}
}
\description{
\code{betos_classification()} allows the user to group HCPCS codes into clinically
meaningful categories based on the original Berenson-Eggers Type of Service
(BETOS) classification. Users may use the RBCS to analyze trends and perform
other types of health services analytic work.
\subsection{BETOS}{

The Restructured BETOS Classification System (RBCS) is a taxonomy that allows
researchers to group healthcare service codes for Medicare Part B services
into clinically meaningful categories and subcategories. It is based on the
original Berenson-Eggers Type of Service (BETOS) classification created in
the 1980s, and includes notable updates such as Part B non-physician services.
The RBCS will undergo annual updates by a technical expert panel of
researchers and clinicians.

The general framework for grouping service codes into the new RBCS taxonomy
largely follows the same structure of BETOS. Like BETOS, the RBCS groups
HCPCS codes into categories, subcategories, and families – with categories
as the most aggregate level and families as the more granular level.

All Medicare Part B service codes, including non-physician services, are
assigned to a 6-character RBCS taxonomy code.

Links:
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/provider-service-classifications/restructured-betos-classification-system}{Restructured BETOS Classification System}
\item \href{https://data.cms.gov/resources/restructured-betos-classification-system-data-dictionary}{Restructured BETOS Classification System Data Dictionary}
}

\emph{Update Frequency:} \strong{Annually}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
betos_classification(hcpcs_code = "0001U")
betos_classification(category = "Test")
betos_classification(subcategory = "General Laboratory")
betos_classification(family = "Immunoassay")
betos_classification(procedure = "M")
\dontshow{\}) # examplesIf}
}
