% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_geography.R
\name{by_geography}
\alias{by_geography}
\title{Medicare Part B Utilization Statistics by \strong{Geography and Service}}
\usage{
by_geography(
  year,
  state = NULL,
  hcpcs_code = NULL,
  pos = NULL,
  level = NULL,
  fips = NULL,
  hcpcs_desc = NULL,
  drug = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{int (\emph{required}); Year in YYYY format. Run helper function
\code{by_geography_years()} to return a vector of the years currently available.}

\item{state}{State where the provider is located, as reported
in NPPES. Values include the 50 United States, District of Columbia,
U.S. territories, Armed Forces areas, Unknown and Foreign Country. Data
aggregated at the National level are identified by the word 'National'.}

\item{hcpcs_code}{HCPCS code used to identify the specific medical service
furnished by the provider.}

\item{pos}{Identifies whether the place of service submitted on claims
is a facility (\code{F}) or non-facility (\code{O}). Non-facility
is generally an office setting.}

\item{level}{Level of geography by which the data will be aggregated. \code{State}
indicates the data is aggregated to a single state identified in the Rendering Provider
State column. \code{National} indicates the data is aggregated across all states for a
given HCPCS Code Level.}

\item{fips}{FIPS code of the referring provider state. This variable is
blank when reported at the national level.}

\item{hcpcs_desc}{Description of the HCPCS code for the specific medical
service furnished by the provider.}

\item{drug}{Flag that identifies whether the HCPCS code for the specific
service furnished by the provider is a HCPCS listed on the Medicare Part
B Drug Average Sales Price (ASP) File.}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{by_geography()} allows you to search for information on services and
procedures provided to Original Medicare (fee-for-service) Part B
beneficiaries by physicians and other healthcare professionals, aggregated by
geography and service.

The \strong{Geography and Service} subset contains information on utilization,
allowed amount, Medicare payment, and submitted charges organized nationally
and state-wide by HCPCS code and place of service.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service}{Physician & Other Practitioners: by Geography and Service API}
}

\emph{Update Frequency:} \strong{Annually}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
by_geography(hcpcs_code = "0002A", year = 2020)
\dontshow{\}) # examplesIf}
}
