% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physician_by_provider.R
\name{by_provider}
\alias{by_provider}
\title{Yearly Utilization Data on Medicare Part B Providers}
\usage{
by_provider(
  year,
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  organization_name = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{integer (\emph{required}); Year in \code{YYYY} format. Run helper function
\code{by_provider_years()} to return a vector of the years
currently available.}

\item{npi}{National Provider Identifier for the rendering provider on the claim.}

\item{first_name}{Individual provider's first name.}

\item{last_name}{Individual provider's last name.}

\item{organization_name}{Organization name.}

\item{credential}{Individual provider's credentials.}

\item{gender}{Individual provider's gender.}

\item{entype}{Provider entity type.}

\item{city}{City where provider is located.}

\item{state}{State where provider is located.}

\item{zip}{Provider’s zip code.}

\item{fips}{Provider's state FIPS code.}

\item{ruca}{Rural-Urban Commuting Area Code (RUCA); a Census tract-based
classification scheme that utilizes the standard Bureau of Census
Urbanized Area and Urban Cluster definitions in combination with work
commuting information to characterize all of the nation's Census tracts
regarding their rural and urban status and relationships. The Referring
Provider ZIP code was cross walked to the United States Department of
Agriculture (USDA) 2010 Rural-Urban Commuting Area Codes.}

\item{country}{Country where provider is located.}

\item{specialty}{Provider specialty code reported on the largest number of
claims submitted.}

\item{par}{Identifies whether the provider participates in Medicare
and/or accepts assignment of Medicare allowed amounts. The value will
be \code{Y} for any provider that had at least one claim identifying the
provider as participating in Medicare or accepting assignment of
Medicare allowed amounts within HCPCS code and place of service. A
non-participating provider may elect to accept Medicare allowed amounts
for some services and not accept Medicare allowed amounts for other
services.}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{by_provider()} allows you to access data such as
services and procedures performed; charges submitted and payment received;
and beneficiary demographic and health characteristics for providers treating
Original Medicare (fee-for-service) Part B beneficiaries, aggregated by year.
\subsection{Links:}{
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider}{Medicare Physician & Other Practitioners: by Provider API}
}

\emph{Update Frequency:} \strong{Annually}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
physician_by_provider(year = 2020,
                      npi = 1003000423)

# Use the years helper function to retrieve results for every year:
physician_by_provider_years() |>
map(\(x) physician_by_provider(year = x, npi = 1043477615)) |>
list_rbind()
\dontshow{\}) # examplesIf}
}
