% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_multiple.R
\name{cc_multiple_years}
\alias{cc_multiple_years}
\title{Check the current years available for the Multiple Chronic Conditions API}
\usage{
cc_multiple_years()
}
\value{
integer vector of years available
}
\description{
Check the current years available for the Multiple Chronic Conditions API
}
\examples{
cc_multiple_years()
}
