% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_specific.R
\name{cc_specific_years}
\alias{cc_specific_years}
\title{Check the current years available for the Specific Chronic Conditions API}
\usage{
cc_specific_years()
}
\value{
integer vector of years available
}
\description{
Check the current years available for the Specific Chronic Conditions API
}
\examples{
cc_specific_years()
}
