% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{change}
\alias{change}
\title{Calculate lagged values by columns}
\usage{
change(df, cols, digits = 3)
}
\arguments{
\item{df}{data frame}

\item{cols}{numeric columns}

\item{digits}{Number of digits to round to}
}
\value{
A \code{tibble}
}
\description{
Calculate lagged values by columns
}
\examples{
dplyr::tibble(year = 2015:2020,
charges = sample(1000:2000, size = 6),
payment = sample(1000:2000, size = 6)) |>
change(c(charges, payment))
}
