% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{change_year}
\alias{change_year}
\title{Calculate lagged values by column}
\usage{
change_year(df, col, by = year)
}
\arguments{
\item{df}{data frame}

\item{col}{column of numeric values to calculate lag}

\item{by}{column to calculate lag by}
}
\value{
A \code{tibble}
}
\description{
Calculate lagged values by column
}
\examples{
dplyr::tibble(year = 2015:2020,
              pay = 1000:1005) |>
change_year(pay, year)
}
