% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-clia.R
\name{clia}
\alias{clia}
\title{Clinical Laboratories}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories}{API: Provider of Services File - Clinical Laboratories}
\item \href{https://www.cms.gov/medicare/quality/clinical-laboratory-improvement-amendments}{CMS: CLIA}
\item \href{https://www.cdc.gov/clia/php/about/index.html}{CDC: CLIA}
\item \href{https://qcor.cms.gov/main.jsp}{CMS: QCOR}
\item \href{https://www.cdc.gov/labs/clia-certificates/index.html}{CLIA: Certificates}
\item \href{https://www.cms.gov/files/document/clia-certificate-fee-schedule-updated-06/7/2024.pdf}{CLIA: Certificate Fee Schedule}
\item \href{https://www.cms.gov/files/document/clia-cert-quick-start-guide.pdf}{CLIA: Certification Guide}
}
}
\usage{
clia(
  facility_name = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  certificate = NULL,
  accreditation = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  chow = NULL,
  compliant = NULL,
  active = NULL,
  multi = NULL,
  campus = NULL,
  non_profit = NULL,
  eligible = NULL,
  temporary = NULL,
  poc = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{facility_name}{\verb{<chr>} Provider/Laboratory name}

\item{facility_ccn}{\verb{<chr>} 10-digit CMS Certification Number}

\item{parent_ccn}{\verb{<chr>} 6-digit CMS Certification Number}

\item{certificate}{\verb{<enum>} CLIA certificate type (see Details):
\itemize{
\item \code{"waiver"} = Waiver
\item \code{"ppm"} = Provider-Performed Microscopy (PPM)
\item \code{"registration"} = Registration
\item \code{"compliance"} = Compliance
\item \code{"accreditation"} = Accreditation
}}

\item{accreditation}{\verb{<enum>} CLIA accrediting organization (see Details):
\itemize{
\item \code{"a2la"} = A2LA
\item \code{"aabb"} = AABB
\item \code{"aoa"} = AOA
\item \code{"ashi"} = ASHI-HLA
\item \code{"cap"} = CAP
\item \code{"cola"} = COLA
\item \code{"jcaho"} = JCAHO
}}

\item{city, state, zip}{\verb{<chr>} Lab city, state, zip}

\item{chow}{\verb{<int>} Number of times there has been a Change of Ownership.}

\item{compliant}{\verb{<lgl>} Provider compliance status at time of certification
survey.}

\item{active}{\verb{<lgl>} Return only active labs}

\item{multi}{\verb{<lgl>} Indicates lab has applied for a single site CLIA to
cover multiple testing locations.}

\item{campus}{\verb{<lgl>} Indicates single site CLIA is for hospital with several
labs on a single hospital campus.}

\item{non_profit}{\verb{<lgl>} Indicates single site CLIA is for multiple sites
with non-profit, federal, state or local government status and engaged in
limited public health testing.}

\item{eligible}{\verb{<lgl>} Indicates lab is eligible to participate in
Medicare/Medicaid.}

\item{temporary}{\verb{<lgl>} Indicates single site CLIA is for lab with multiple
temporary testing sites.}

\item{poc}{\verb{<lgl>} Indicates provider is in compliance with Plan of
Correction.}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Clinical laboratories including demographics and the type of testing
services the facility provides.
}
\section{CLIA}{


CMS regulates all laboratory testing (except research) performed on humans
in the U.S. through the \strong{Clinical Laboratory Improvement Amendments}.
In total, \strong{CLIA} covers approximately 320,000 laboratory entities.

Although all clinical laboratories must be properly certified to receive
Medicare or Medicaid payments, CLIA has no direct Medicare or Medicaid
program responsibilities.
\subsection{Certification}{

There are five CLIA certificate types, all of which are effective for a period of
two years. They are as follows, in order of increasing complexity:
\enumerate{
\item \strong{Waiver}: Laboratory can only perform waived tests, those that are simple
to perform. Although routine inspections are not conducted, 2\% are visited
annually to ensure quality. Does not waive the lab from all CLIA requirements.
\item \strong{Provider-Performed Microscopy (PPM)}: In addition to waived tests,
a physician, mid-level practitioner or dentist can perform limited tests of
moderate complexity requiring microscopic examination. No routine inspections
are conducted.
\item \strong{Registration}: Laboratory has applied for a \strong{Compliance} or
\strong{Accreditation} certificate, enabling it to conduct moderate/high complexity
testing until the survey is completed. Includes PPM and waived testing.
\item \strong{Compliance}: Allows the laboratory to conduct moderate/high complexity
testing and is issued after an inspection finds the lab to be in compliance
with all applicable CLIA requirements. Includes PPM and waived testing.
\item \strong{Accreditation}: Exactly the same as \strong{Compliance}, except that the
laboratory is accredited by one of the CMS-approved organizations:
\itemize{
\item \href{https://a2la.org/}{A2LA: American Association for Laboratory Accreditation}
\item \href{https://www.aabb.org/}{AABB: Association for the Advancement of Blood & Biotherapies}
\item \href{https://osteopathic.org/}{AOA: American Osteopathic Association}
\item \href{https://www.ashi-hla.org/}{ASHI-HLA: American Society for Histocompatibility & Immunogenetics}
\item \href{https://www.cap.org/}{CAP: College of American Pathologists}
\item \href{https://www.cola.org/}{COLA: Commission on Office Laboratory Accreditation}
\item \href{https://www.jointcommission.org/}{JCAHO: The Joint Commission}
}
}
}
}

\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
clia(count = TRUE)

clia(compliant = FALSE,
     active = TRUE,
     count = TRUE)

clia(facility_ccn = provider:::cdc_labs$ccn) |> str()

clia(certificate = c("accreditation", "registration"),
     city = "Valdosta",
     state = "GA") |>
     str()

clia(accreditation = c("cap", "cola", "jcaho")) |> str()
\dontshow{\}) # examplesIf}
}
