% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clia.R
\name{clia_labs}
\alias{clia_labs}
\title{Clinical Laboratories}
\usage{
clia_labs(
  facility_name = NULL,
  clia_number = NULL,
  ccn = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  tidy = TRUE
)
}
\arguments{
\item{facility_name}{Hospital’s CMS Certification Number (CCN)}

\item{clia_number}{description}

\item{ccn}{description}

\item{city}{City of the hospital’s practice location address}

\item{state}{State of the hospital’s practice location address}

\item{zip}{State of the hospital’s practice location address}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{clia()} allows you to search for information on clinical laboratories
including demographics and the type of testing services the facility provides.
\subsection{Clinical Laboratory Improvement Amendments (CLIA)}{

CMS regulates all laboratory testing (except research) performed on humans
in the U.S. through the Clinical Laboratory Improvement Amendments (CLIA).
In total, CLIA covers approximately 320,000 laboratory entities. The Division
of Clinical Laboratory Improvement & Quality, within the Quality, Safety &
Oversight Group, under the Center for Clinical Standards and Quality (CCSQ)
has the responsibility for implementing the CLIA Program. Although all
clinical laboratories must be properly certified to receive Medicare or
Medicaid payments, CLIA has no direct Medicare or Medicaid program
responsibilities.
\subsection{Links:}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories}{Provider of Services File - Clinical Laboratories}
}

\emph{Update Frequency:} \strong{Quarterly}
}

}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clia_labs(city = "Valdosta")
clia_labs(clia_number = "00205114A3")
clia_labs(ccn = "11D0265516")
\dontshow{\}) # examplesIf}
}
