% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{compare_hcpcs}
\alias{compare_conditions}
\title{Compare Yearly Provider Data To State And National Averages}
\usage{
compare_hcpcs(serv_tbl)

compare_conditions(prov_tbl, pivot = TRUE)
}
\arguments{
\item{serv_tbl}{< \emph{tbl_df} > // \strong{required} \link[tibble:tibble-package]{tibble} returned from \code{\link[=by_service]{by_service()}}}

\item{prov_tbl}{< \emph{tbl_df} > // \strong{required} \link[tibble:tibble-package]{tibble} returned from \code{\link[=by_provider]{by_provider()}}}

\item{pivot}{< \emph{boolean} > // \strong{default:} \code{TRUE} Pivot output}
}
\value{
A \link[tibble:tibble-package]{tibble} containing:
\itemize{
\item \code{compare_hcpcs()}
\item \code{compare_conditions()}
}
}
\description{
\itemize{
\item \code{compare_hcpcs()} allows the user to compare a provider's yearly HCPCS
utilization data to state and national averages
\item \code{compare_conditions()} allows the user to compare the average yearly
prevalence of chronic conditions among a provider's patient mix to state and
national averages
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
compare_hcpcs(by_service(year = 2018, npi = 1023076643))
compare_conditions(by_provider(year = 2018, npi = 1023076643))

compare_hcpcs(map_dfr(pop_years(), ~by_service(year = .x, npi = 1023076643)))
compare_conditions(map_dfr(pop_years(), ~by_provider(year = .x, npi = 1023076643)))
\dontshow{\}) # examplesIf}
}
