% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_conditions}
\alias{compare_conditions}
\title{Compare Yearly Chronic Condition Prevalence Data}
\usage{
compare_conditions(df)
}
\arguments{
\item{df}{data frame returned by \code{by_provider()}}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the results.
}
\description{
\code{compare_conditions()} allows you to compare yearly chronic condition
prevalence by provider, state and national averages
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
by_provider_years() |>
map(\(x) by_provider(year = x, npi = 1023076643)) |>
list_rbind() |>
compare_conditions()
\dontshow{\}) # examplesIf}
}
