% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_geography}
\alias{compare_geography}
\title{Compare}
\usage{
compare_geography(df)
}
\arguments{
\item{df}{data frame returned by \code{by_service()}}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the results.
}
\description{
\code{clia()} allows you to search for information on clinical laboratories
including demographics and the type of testing services the facility provides.
\subsection{Clinical Laboratory Improvement Amendments (CLIA)}{

CMS regulates all laboratory testing (except research) performed on humans
in the U.S. through the Clinical Laboratory Improvement Amendments (CLIA).
In total, CLIA covers approximately 320,000 laboratory entities. The Division
of Clinical Laboratory Improvement & Quality, within the Quality, Safety &
Oversight Group, under the Center for Clinical Standards and Quality (CCSQ)
has the responsibility for implementing the CLIA Program. Although all
clinical laboratories must be properly certified to receive Medicare or
Medicaid payments, CLIA has no direct Medicare or Medicaid program
responsibilities.
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
by_service_years() |>
purrr::map(\(x) by_service(year = x, npi = 1023076643)) |>
purrr::list_rbind() |>
compare_geography()
mutate(level = "individual")
\dontshow{\}) # examplesIf}
}
