% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_hcpcs}
\alias{compare_hcpcs}
\title{Compare Provider Performance}
\usage{
compare_hcpcs(df)
}
\arguments{
\item{df}{< \emph{tbl_df} > // \strong{required}

\link[tibble:tibble-package]{tibble} returned from \code{utilization(type = "Service")}}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
\code{compare_hcpcs()} allows the user to compare a provider's yearly HCPCS
utilization data to state and national averages
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
compare_hcpcs(utilization(year = 2018,
                          type = "Service",
                          npi = 1023076643))

map_dfr(util_years(), ~utilization(year = .x,
                                   npi = 1023076643,
                                   type = "Service")) |>
compare_hcpcs()
\dontshow{\}) # examplesIf}
}
