% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{conditions}
\alias{conditions}
\title{Chronic Conditions Prevalence}
\usage{
conditions(
  year,
  condition = NULL,
  sublevel = NULL,
  set = c("Multiple", "Specific"),
  level = NULL,
  fips = NULL,
  age = NULL,
  demo = NULL,
  subdemo = NULL,
  mcc = NULL,
  tidy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{year}{< \emph{integer} > Calendar year of Medicare enrollment, in \code{YYYY}
format. Run \code{\link[=cc_years]{cc_years()}} to return a vector of currently available years.}

\item{condition}{< \emph{character} > Chronic condition for which the prevalence
and utilization is compiled (see details for list of conditions)}

\item{sublevel}{< \emph{character} > Beneficiary's state or county}

\item{set}{< \emph{character} > // \strong{required} \code{"Multiple"} or \code{"Specific"}}

\item{level}{< \emph{character} > Geographic level of aggregation: \code{"National"}, \code{"State"}, or \code{"County"}}

\item{fips}{< \emph{character} > Beneficiary's state or county FIPS code}

\item{age}{< \emph{character} > Age level of aggregation: \code{"All"}, \code{"<65"}, or \code{"65+"}}

\item{demo, subdemo}{< \emph{character} > Demographic/subdemographic level of aggregation:\tabular{ll}{
   \strong{\code{demo}} \tab \strong{\code{subdemo}} \cr
   \code{'All'} \tab \code{'All'} \cr
   \code{'Sex'} \tab \code{'Male'} \cr
   '' \tab \code{'Female'} \cr
   \code{'Race'} \tab \code{'White'} \cr
   '' \tab \code{'Black'} \cr
   '' \tab \code{'Island'} \cr
   '' \tab \code{'Hispanic'} \cr
   '' \tab \code{'Native'} \cr
   \code{'Dual Status'} \tab \code{'Dual'} (Medicare & Medicaid) \cr
   '' \tab \code{'Nondual'} (Medicare only) \cr
}}

\item{mcc}{< \emph{character} > Number of chronic conditions: \code{"0-1"}, \code{"2-3"}, \code{"4-5"}, \code{"6+"}}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}
}
\value{
A \link[tibble:tibble-package]{tibble} with the following columns:
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=conditions]{conditions()}} allows the user access to data concerning chronic conditions
among Original Medicare (or fee-for-service) beneficiaries.
}
\section{Specific Conditions}{


The \emph{Specific} Chronic Conditions(\code{set = "Specific"}) dataset provides
information on prevalence, use and spending organized by geography and 21
distinct chronic conditions:
\enumerate{
\item Alcohol Abuse
\item Alzheimer's Disease/Dementia
\item Arthritis
\item Asthma
\item Atrial Fibrillation
\item Autism Spectrum Disorders
\item Cancer
\item Chronic Kidney Disease
\item COPD
\item Depression
\item Diabetes
\item Drug Abuse/Substance Abuse
\item Heart Failure
\item Hepatitis (Chronic Viral B & C)
\item HIV/AIDS
\item Hyperlipidemia
\item Hypertension
\item Ischemic Heart Disease
\item Osteoporosis
\item Schizophrenia and Other Psychotic Disorders
\item Stroke
}
}

\section{Multiple Conditions}{


The \emph{Multiple} Chronic Conditions(\code{set = "Multiple"}) dataset provides
information on prevalence, use and spending organized by geography and the
count of chronic conditions from the set of select 21 chronic conditions.
The count of conditions is binned into four categories:
\enumerate{
\item \strong{0-1}
\item \strong{2-3}
\item \strong{4-5}
\item \strong{6 or More}
}
}

\section{Prevalence}{


Prevalence estimates are calculated by taking the beneficiaries within the
MCC category divided by the total number of beneficiaries in the
fee-for-service population, expressed as a percentage.
}

\section{Hospital Readmission Rates}{


Hospital readmissions are expressed as a percentage of all admissions. A
30-day readmission is defined as an admission to an acute care hospital for
any cause within 30 days of discharge from an acute care hospital. Except
when the patient died during the stay, each inpatient stay is classified as
an index admission, a readmission, or both.
}

\section{Emergency Department Visits}{


Emergency department visits are presented as the number of visits per 1,000
beneficiaries. ED visits include visits where the beneficiary was released
from the outpatient setting and where the beneficiary was admitted to an
inpatient setting.
}

\section{Payment Per Capita}{


Medicare spending includes total Medicare payments for all covered services
in Parts A and B and is presented per beneficiary (i.e. per capita).
Standardized payments are presented to allow for comparisons across
geographic areas in health care use among beneficiaries.
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/medicare-chronic-conditions/multiple-chronic-conditions}{Medicare Multiple Chronic Conditions}
\item \href{https://data.cms.gov/medicare-chronic-conditions/specific-chronic-conditions}{Medicare Specific Chronic Conditions}
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
conditions(year = 2018,
           set = "Specific",
           sublevel = "CA",
           demo = "All")

conditions(year = 2018,
           set = "Specific",
           sublevel = "CA",
           subdemo = "Female",
           age = "All")

conditions(year = 2018,
           set = "Multiple",
           sublevel = "California",
           subdemo = "Female")

conditions(year = 2007,
           set = "Specific",
           level = "National",
           demo = "Race",
           condition = "Alzheimer's Disease/Dementia")

conditions(year = 2007,
           set = "Multiple",
           level = "National",
           demo = "Race")
\dontshow{\}) # examplesIf}
}
