% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_today.R
\name{days_today}
\alias{days_today}
\title{Calculate Number of Days Between a Date and Today}
\usage{
days_today(df, start, colname = "age")
}
\arguments{
\item{df}{data frame containing date columns}

\item{start}{date column}

\item{colname}{desired column name of output}
}
\value{
data frame with a column containing the number of days
calculated.
}
\description{
Calculate Number of Days Between a Date and Today
}
\examples{
ex <- data.frame(x = c("02/05/92",
                       "01-04-2020",
                       "1996/05/01",
                       "2020-05-01",
                       "02-04-96"))

days_today(df = ex, date = x)

ex |> days_today(x)
}
