% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-facility.R
\name{dialysis}
\alias{dialysis}
\title{Dialysis Facilities}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/23ew-n7w9}{API: Dialysis Facility - Listing by Facility}
}
}
\usage{
dialysis(
  ccn = NULL,
  fac_name = NULL,
  org_name = NULL,
  rating = NULL,
  network = NULL,
  status = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  count = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} Facility CMS Certification Number}

\item{fac_name}{\verb{<chr>} Facility name}

\item{org_name}{\verb{<chr>} Name of the chain organization the facility is owned/managed by}

\item{rating}{\verb{<int>} Facility's Quality of Care star rating; (1-5)}

\item{network}{\verb{<int>} Numeric code for the network the facility participates in; (1-18)}

\item{status}{\verb{<enum>} \code{Non-profit} or \code{Profit}}

\item{address, city, state, zip, county}{\verb{<chr>} Facility's city, state, zip, county}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of all dialysis facilities registered with Medicare that
includes addresses and phone numbers, as well as services and quality of
care provided.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
dialysis(count = TRUE)
dialysis(org_name = "DaVita", state = "GA")
\dontshow{\}) # examplesIf}
}
