% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_long}
\alias{display_long}
\title{Pivot data frame to long format for easy printing}
\usage{
display_long(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Pivot data frame to long format for easy printing
}
\keyword{internal}
