% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{distributions_natl}
\alias{distributions_natl}
\title{Doctors and Clinicians (National Downloadable File) Distributions}
\usage{
distributions_natl()
}
\value{
description
}
\description{
Doctors and Clinicians (National Downloadable File) Distributions
}
\keyword{internal}
