% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{distros_dac}
\alias{distros_dac}
\title{Doctors and Clinicians (National Downloadable File) Distributions}
\usage{
distros_dac()
}
\value{
description
}
\description{
Doctors and Clinicians (National Downloadable File) Distributions
}
\keyword{internal}
