% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctors_and_clinicians.R
\name{doctors_and_clinicians}
\alias{doctors_and_clinicians}
\title{Search the CMS Doctors and Clinicians National Downloadable File API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
doctors_and_clinicians(
  npi = NULL,
  pac_id = NULL,
  enroll_id = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  gender = NULL,
  school = NULL,
  grad_year = NULL,
  specialty = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  ind_assn = NULL,
  group_assn = NULL,
  offset = 0L,
  tidy = TRUE
)
}
\arguments{
\item{npi}{Unique clinician ID assigned by NPPES}

\item{pac_id}{Unique individual clinician ID assigned by PECOS}

\item{enroll_id}{Unique ID for the clinician enrollment that is the source
for the data in the observation}

\item{first_name}{Individual clinician first name}

\item{middle_name}{Individual clinician middle name}

\item{last_name}{Individual clinician last name}

\item{gender}{Individual clinician gender}

\item{school}{Individual clinician’s medical school}

\item{grad_year}{Individual clinician’s medical school graduation year}

\item{specialty}{Primary medical specialty reported by the individual
clinician in the selected enrollment}

\item{city}{Group or individual's city}

\item{state}{Group or individual's state}

\item{zipcode}{Group or individual's ZIP code (9 digits when available)}

\item{ind_assn}{Indicator for whether clinician accepts Medicare approved
amount as payment in full. \code{Y} = Clinician accepts Medicare approved amount
as payment in full. \code{M} = Clinician may accept Medicare Assignment.}

\item{group_assn}{Indicator for whether group accepts Medicare approved
amount as payment in full. \code{Y} = Clinician accepts Medicare approved amount
as payment in full. \code{M} = Clinician may accept Medicare Assignment.}

\item{offset}{offset; API pagination}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Dataset of providers' facility affiliations
}
\details{
The Doctors and Clinicians national downloadable file is organized
such that each line is unique at the clinician/enrollment
record/group/address level. Clinicians with multiple Medicare enrollment
records and/or single enrollments linking to multiple practice locations
are listed on multiple lines.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/mj5m-pzi6}{Doctors and Clinicians National Downloadable File}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
doctors_and_clinicians(npi = 1407263999)
doctors_and_clinicians(enroll_id = "I20081002000549")
doctors_and_clinicians(first_name = "John")
doctors_and_clinicians(school = "NEW YORK UNIVERSITY SCHOOL OF MEDICINE")
doctors_and_clinicians(grad_year = 2003)
\dontshow{\}) # examplesIf}
}
