% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-enrolled.R
\name{enrolled}
\alias{enrolled}
\title{Provider Enrollment in Medicare}
\usage{
enrolled(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  last = NULL,
  prov_type = NULL,
  prov_desc = NULL,
  state = NULL,
  org_name = NULL,
  multi = NULL,
  count = FALSE,
  chains = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{first, last}{\verb{<chr>} Individual provider's name}

\item{prov_type, prov_desc}{\verb{<chr>} Enrollment specialty code/description}

\item{state}{\verb{<chr>} Enrollment state, full or abbreviation}

\item{org_name}{\verb{<chr>} Organizational provider's name}

\item{multi}{\verb{<lgl>} Provider has multiple NPIs}

\item{count}{\verb{<lgl>} Return the total row count}

\item{chains}{\verb{<lgl>} Add search chains}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Enrollment data on individual and organizational providers that are
actively approved to bill Medicare.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
enrolled(count = TRUE)

enrolled(count = TRUE, org_name = not_blank())

enrolled(org_name = starts("AB"), state = "GA")
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{API: Medicare Provider Supplier Enrollment}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Provider Enrollment Data Dictionary}
}
}
