% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-enrollment.R
\name{enrollment}
\alias{enrollment}
\alias{fqhc_enroll}
\alias{hospice_enroll}
\alias{rhc_enroll}
\alias{snf_enroll}
\title{Facility Enrollment}
\usage{
fqhc_enroll(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  org_name = NULL,
  org_dba = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  multi = NULL,
  status = NULL,
  org_type = NULL,
  count = FALSE,
  set = FALSE
)

hospice_enroll(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  org_name = NULL,
  org_dba = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  multi = NULL,
  status = NULL,
  org_type = NULL,
  count = FALSE,
  set = FALSE
)

rhc_enroll(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  org_name = NULL,
  org_dba = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  multi = NULL,
  status = NULL,
  org_type = NULL,
  count = FALSE,
  set = FALSE
)

snf_enroll(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  org_name = NULL,
  org_dba = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  multi = NULL,
  status = NULL,
  org_type = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{ccn}{\verb{<int>} CMS Certification Number}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{org_name}{\verb{<chr>} Legal business name}

\item{org_dba}{\verb{<chr>} Doing-business-as name}

\item{city, state, zip}{\verb{<chr>} Location city, state, zip}

\item{multi}{\verb{<lgl>} Does hospital have more than one NPI?}

\item{status}{\verb{<enum>} Organization status
\itemize{
\item \code{P} = Proprietary
\item \code{N} = Non-Profit
\item \code{D} = Unknown
}}

\item{org_type}{\verb{<enum>} Organization structure type
\itemize{
\item \code{corp} = Corporation
\item \code{other} = Other
\item \code{llc} = LLC
\item \code{part} = Partnership
\item \code{sole} = Sole Proprietor
}}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Providers with pending Medicare enrollment applications.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
fqhc_enroll(count = TRUE)
hospice_enroll(count = TRUE)
rhc_enroll(count = TRUE)
snf_enroll(count = TRUE)

fqhc_enroll(state = c("GA", "FL")) |> str()

hospice_enroll(state = c("GA", "FL")) |> str()

rhc_enroll(state = c("GA", "FL")) |> str()

snf_enroll(state = c("GA", "FL")) |> str()
\dontshow{\}) # examplesIf}
}
