% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-esrd.R
\name{esrd}
\alias{esrd}
\title{Dialysis Facilities}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/23ew-n7w9}{API: Dialysis Facility - Listing by Facility}
}
}
\usage{
esrd(
  ccn = NULL,
  facility_name = NULL,
  chain_name = NULL,
  rating = NULL,
  network = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  status = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} Facility CMS Certification Number}

\item{facility_name}{\verb{<chr>} Facility name}

\item{chain_name}{\verb{<chr>} Name of the chain organization the facility is owned/managed by}

\item{rating}{\verb{<int>} Facility's Quality of Care star rating (1 - 5)}

\item{network}{\verb{<int>} Numeric code for the network the facility participates in (1 - 18)}

\item{address, city, state, zip, county}{\verb{<chr>} Facility's city, state, zip, county}

\item{status}{\verb{<enum>} \code{Non-profit} or \code{profit}}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
A list of all dialysis facilities registered with Medicare that
includes addresses and phone numbers, as well as services and quality of
care provided.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
esrd(count = TRUE)

esrd()

esrd(rating = 1)

esrd(network = 15:18)
\dontshow{\}) # examplesIf}
}
