% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-facility.R
\name{facility}
\alias{facility}
\title{Facilities}
\source{
Medicare
}
\usage{
facility(
  fac_type = NULL,
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  org_name = NULL,
  org_dba = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  multi = NULL,
  status = NULL,
  org_type = NULL,
  count = FALSE
)
}
\arguments{
\item{fac_type}{\verb{<enum>} Facility type; if NULL (default), will search all:
\itemize{
\item \code{HHA} = Home Health Agency
\item \code{RHC} = Rural Health Clinic
\item \code{FQHC} = Federally Qualified Health Clinic
\item \code{SNF} = Skilled Nursing Facility
\item \code{Hospice} = Hospice
}}

\item{npi}{\verb{<int>} National Provider Identifier}

\item{ccn}{\verb{<int>} CMS Certification Number}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{org_name}{\verb{<chr>} Legal business name}

\item{org_dba}{\verb{<chr>} Doing-business-as name}

\item{city, state, zip}{\verb{<chr>} Location city, state, zip}

\item{multi}{\verb{<lgl>} Does facility have more than one NPI?}

\item{status}{\verb{<enum>} Facility organization status
\itemize{
\item \code{P} = Proprietary
\item \code{N} = Non-Profit
\item \code{D} = Unknown
}}

\item{org_type}{\verb{<enum>} Facility organization structure type
\itemize{
\item \code{corp} = Corporation
\item \code{other} = Other
\item \code{llc} = LLC
\item \code{part} = Partnership
\item \code{sole} = Sole Proprietor
}}

\item{count}{\verb{<lgl>} Return the total row count}
}
\description{
Facilities enrolled in Medicare.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
facility(count = TRUE)

facility(state = c("GA", "FL"), count = TRUE)

facility(city = "Valdosta", state = "GA")
\dontshow{\}) # examplesIf}
}
