% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_affiliations.R
\name{facility_affiliations}
\alias{facility_affiliations}
\title{Search the CMS Physician Facility Affiliations API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
facility_affiliations(
  npi = NULL,
  pac_id = NULL,
  last_name = NULL,
  first_name = NULL,
  middle_name = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  offset = 0,
  clean_names = TRUE
)
}
\arguments{
\item{npi}{Unique clinician ID assigned by NPPES}

\item{pac_id}{Unique individual clinician ID assigned by PECOS}

\item{last_name}{Individual clinician last name}

\item{first_name}{Individual clinician first name}

\item{middle_name}{Individual clinician middle name}

\item{facility_type}{Facilities can fall into the following type categories:
Hospital, Long-term Care Hospital, Nursing Home, Inpatient
Rehabilitation Facility, Home Health Agency, Hospice, and Dialysis
Facility}

\item{facility_ccn}{alphanumeric; Medicare CCN (CMS Certification Number) of
facility type or unit within hospital where an individual clinician
provides service. Effective 2007, the CCN replaced the terms
\emph{Medicare Provider Number}, \emph{Medicare Identification Number} and
\emph{OSCAR Number}. The CCN is used to verify Medicare/Medicaid certification
for survey and certification, assessment-related activities and
communications.}

\item{parent_ccn}{numeric; Medicare CCN of the primary hospital where an
individual clinician provides service, should the clinician provide
services in a unit within the hospital.}

\item{offset}{offset; API pagination}

\item{clean_names}{Convert column names to snakecase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Dataset of providers' facility affiliations
}
\details{
This file lists clinicians' facility affiliations, which have been
expanded to include Long-term Care Hospitals (LTCHs), Skilled Nursing
Facilities (SNFs), Inpatient Rehabilitation Facility (IRFs), Home Health
Agencies, Hospices, and/or Dialysis Facilities, in addition to hospitals.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{CMS Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
facility_affiliations(npi = 1003019563)
facility_affiliations(facility_ccn = "060004")
facility_affiliations(parent_ccn = 670055)
\dontrun{
facility_affiliations(first_name = "John")
facility_affiliations(facility_type = "Home Health Agency")
}
}
