% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_zip.R
\name{format_zipcode}
\alias{format_zipcode}
\title{Format US ZIP codes}
\usage{
format_zipcode(zip)
}
\arguments{
\item{zip}{Nine-digit US ZIP code}
}
\value{
ZIP code, hyphenated for ZIP+4 or 5-digit ZIP.
}
\description{
Format US ZIP codes
}
\examples{
format_zipcode(123456789)

format_zipcode(12345)

}
