% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{geomean}
\alias{geomean}
\title{Calculate geometric mean (average rate of return)
For use in conjunction with \code{\link[=ror]{ror()}}}
\usage{
geomean(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Calculate geometric mean (average rate of return)
For use in conjunction with \code{\link[=ror]{ror()}}
}
\keyword{internal}
