% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{get_pin}
\alias{get_pin}
\title{Get a pinned dataset from a \link[pins:pins-package]{pins} board}
\usage{
get_pin(pin, ...)
}
\arguments{
\item{pin}{\verb{<chr>} string name of pinned dataset}

\item{...}{additional arguments passed to \code{mount_board()}}
}
\value{
\verb{<tibble>} or \verb{<data.frame>}
}
\description{
Get a pinned dataset from a \link[pins:pins-package]{pins} board
}
\keyword{internal}
