% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gt.R
\name{gt_check_xmark}
\alias{gt_check_xmark}
\title{check or x mark icon}
\usage{
gt_check_xmark(gt_tbl, cols)
}
\arguments{
\item{gt_tbl}{gt_tbl object}

\item{cols}{columns in data frame}
}
\description{
check or x mark icon
}
\keyword{internal}
