% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gt.R
\name{gt_prov}
\alias{gt_prov}
\title{gt theme}
\usage{
gt_prov(
  df,
  divider = NULL,
  title = NULL,
  subtitle = NULL,
  source = NULL,
  checkmark = NULL,
  qmark = NULL,
  dollars = NULL,
  pct = NULL,
  pctchg = NULL,
  clean = TRUE
)
}
\arguments{
\item{df}{data frame}

\item{divider}{description}

\item{title}{description}

\item{subtitle}{description}

\item{source}{description}

\item{checkmark}{description}

\item{qmark}{description}

\item{dollars}{description}

\item{pct}{description}

\item{pctchg}{description}
}
\description{
gt theme
}
\keyword{internal}
