% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gt.R
\name{gt_qmark}
\alias{gt_qmark}
\title{check, x, question mark}
\usage{
gt_qmark(gt_tbl, cols)
}
\arguments{
\item{gt_tbl}{gt_tbl object}

\item{cols}{columns in data frame}
}
\description{
check, x, question mark
}
\keyword{internal}
