% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-hospital.R
\name{hospital2}
\alias{hospital2}
\title{Hospital General Information}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Hospital General Information}
\item \href{https://data.cms.gov/provider-data/dataset/xubh-q36u#data-dictionary}{API: Data Dictionary}
}
}
\usage{
hospital2(
  ccn = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  hosp_type = NULL,
  own_type = NULL,
  rating = NULL,
  count = FALSE
)
}
\arguments{
\item{county}{\verb{<chr>} Location county}

\item{hosp_type}{\verb{<enum>} Provider type:
\itemize{
\item \code{acute} = Acute Care
\item \code{cah} = Critical Access Hospital
\item \code{child} = Childrens' Hospital
\item \code{dod} = Acute Care - Department of Defense
\item \code{ltc} = Long-term
\item \code{psych} = Psychiatric
\item \code{reh} = Rural Emergency Hospital
\item \code{vha} = Acute Care - Veterans Administration
}}

\item{own_type}{\verb{<enum>} Ownership type:
\itemize{
\item \code{private} = Voluntary non-profit - Private
\item \code{other} = Voluntary non-profit - Other
\item \code{church} = Voluntary non-profit - Church
\item \code{district} = Government - Hospital District or Authority
\item \code{local} = Government - Local
\item \code{federal} = Government - Federal
\item \code{state} = Government - State
\item \code{dod} = Department of Defense
\item \code{profit} = Proprietary
\item \code{physician} = Physician
\item \code{tribal} = Tribal
\item \code{vha} = Veterans Health Administration
}}

\item{rating}{\verb{<int>} Hospital rating; 1-5 or "Not Available"}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of all hospitals that have been registered with Medicare.
The list includes addresses, phone numbers, hospital type, and overall
hospital rating.
}
\keyword{internal}
