% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_enrollment.R
\name{hospital_enrollment}
\alias{hospital_enrollment}
\title{Hospitals Enrolled in Medicare}
\usage{
hospital_enrollment(
  npi = NULL,
  facility_ccn = NULL,
  enroll_id_org = NULL,
  enroll_state = NULL,
  specialty_code = NULL,
  pac_id_org = NULL,
  organization_name = NULL,
  doing_business_as = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{Hospital’s National Provider Identifier}

\item{facility_ccn}{Hospital’s CMS Certification Number (CCN)}

\item{enroll_id_org}{Hospital’s enrollment ID}

\item{enroll_state}{Hospital’s enrollment state}

\item{specialty_code}{Enrollment specialty type code}

\item{pac_id_org}{Hospital’s PAC ID}

\item{organization_name}{Hospital’s legal business name}

\item{doing_business_as}{Hospital’s doing-business-as name}

\item{city}{City of the hospital’s practice location address}

\item{state}{State of the hospital’s practice location address}

\item{zip}{Zip code of the hospital’s practice location address}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{hospital_enrollment()} allows you to search for information on all hospitals
currently enrolled in Medicare. Data returned includes the hospital's
sub-group types, legal business name, doing-business-as name, organization type and address.
\subsection{Links:}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{Hospital Enrollments}
}

Update Frequency: \strong{Monthly}
}
}
\examples{
hospital_enrollment(pac_id_org = 6103733050)

}
\seealso{
\code{\link[=doctors_and_clinicians]{doctors_and_clinicians()}}, \code{\link[=provider_enrollment]{provider_enrollment()}}, \code{\link[=facility_affiliations]{facility_affiliations()}}
}
