% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-hospitals.R
\name{hospitals}
\alias{hospitals}
\alias{hospitals2}
\title{Hospitals Enrolled in Medicare}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{API: Hospital Enrollments}
}

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Hospital General Information}
\item \href{https://data.cms.gov/provider-data/dataset/xubh-q36u#data-dictionary}{API: Data Dictionary}
}
}
\usage{
hospitals(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  org_name = NULL,
  org_dba = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  multi = NULL,
  status = NULL,
  org_type = NULL,
  prov_type = NULL,
  loc_type = NULL,
  subgroup = subgroups(),
  count = FALSE
)

hospitals2(
  ccn = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  hosp_type = NULL,
  own_type = NULL,
  rating = NULL,
  count = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{ccn}{\verb{<int>} CMS Certification Number}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{org_name}{\verb{<chr>} Legal business name}

\item{org_dba}{\verb{<chr>} Doing-business-as name}

\item{city, state, zip}{\verb{<chr>} Location city, state, zip}

\item{multi}{\verb{<lgl>} Does hospital have more than one NPI?}

\item{status}{\verb{<enum>} Organization status
\itemize{
\item \code{P} = Proprietary
\item \code{N} = Non-Profit
}}

\item{org_type}{\verb{<enum>} Organization structure type
\itemize{
\item \code{corp} = Corporation
\item \code{other} = Other
\item \code{llc} = LLC
\item \code{part} = Partnership
\item \code{sole} = Sole Proprietor
}}

\item{prov_type}{\verb{<enum>} Provider type:
\itemize{
\item \code{hospital} = Medicare Part A Hospital
\item \code{reh} = Rural Emergency Hospital
\item \code{cah} = Critical Access Hospital
}}

\item{loc_type}{\verb{<enum>} Practice location type
\itemize{
\item \code{main} = Main/Primary Hospital Location
\item \code{psych} = Hospital Psychiatric Unit
\item \code{rehab} = Hospital Rehabilitation Unit
\item \code{swing} = Hospital Swing-Bed Unit
\item \code{ext} = Opt Extension Site
\item \code{other} = Other Hospital Practice Location
}}

\item{subgroup}{\verb{<subgroups>} Hospital’s subgroup/unit. See \code{\link[=subgroups]{subgroups()}}.}

\item{count}{\verb{<lgl>} Return the total row count}

\item{county}{\verb{<chr>} Location county}

\item{hosp_type}{\verb{<enum>} Provider type:
\itemize{
\item \code{acute} = Acute Care
\item \code{cah} = Critical Access Hospital
\item \code{child} = Childrens' Hospital
\item \code{dod} = Acute Care - Department of Defense
\item \code{ltc} = Long-term
\item \code{psych} = Psychiatric
\item \code{reh} = Rural Emergency Hospital
\item \code{vha} = Acute Care - Veterans Administration
}}

\item{own_type}{\verb{<enum>} Ownership type:
\itemize{
\item \code{private} = Voluntary non-profit - Private
\item \code{other} = Voluntary non-profit - Other
\item \code{church} = Voluntary non-profit - Church
\item \code{district} = Government - Hospital District or Authority
\item \code{local} = Government - Local
\item \code{federal} = Government - Federal
\item \code{state} = Government - State
\item \code{dod} = Department of Defense
\item \code{profit} = Proprietary
\item \code{physician} = Physician
\item \code{tribal} = Tribal
\item \code{vha} = Veterans Health Administration
}}

\item{rating}{\verb{<int>} Hospital rating; 1-5 or "Not Available"}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.

A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Hospitals currently enrolled in Medicare. Data includes the
hospital's sub-group types, legal business name, doing-business-as name,
organization type and address.

A list of all hospitals that have been registered with Medicare.
The list includes addresses, phone numbers, hospital type, and overall
hospital rating.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
hospitals(count = TRUE)

hospitals(prov_type = "cah", state = "GA")

hospitals(
  city = "Atlanta",
  state = "GA",
  subgroup = subgroups(acute = FALSE)
)
\dontshow{\}) # examplesIf}
}
