% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov-hospitals2.R
\name{hospitals2}
\alias{hospitals2}
\title{Hospitals 2}
\usage{
hospitals2(
  ccn = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  hosp_type = NULL,
  ownership = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{ccn}{\verb{<int>} CMS Certification Number}

\item{org_name}{\verb{<chr>} Legal business name}

\item{city, state, zip}{\verb{<chr>} Location city, state, zip}

\item{county}{\verb{<chr>}}

\item{hosp_type}{\verb{<chr>}}

\item{ownership}{\verb{<chr>}}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\description{
Hospitals 2
}
\keyword{internal}
