% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hrsa.R
\name{hrsa}
\alias{hrsa}
\alias{hrsa_items}
\alias{hrsa_layers}
\alias{hrsa_fields}
\alias{hrsa_select}
\title{HRSA Facilities}
\source{
\href{https://data.hrsa.gov/tools/web-services/registration#serviceInfo}{API: HRSA}
}
\usage{
hrsa_items()

hrsa_layers()

hrsa_fields(facility)

hrsa_select(facility, ...)
}
\arguments{
\item{facility}{Facility type}

\item{...}{arguments passed on to \code{arcgislayers::arc_select()}}
}
\value{
A list of endpoints.
}
\description{
Health Resources & Services Administration
}
\details{
Query modifiers are a small DSL for use in constructing query conditions.
}
\examples{
hrsa_items()
hrsa_layers()
hrsa_fields("snf_all")
}
