% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-in_outpatient.R
\name{inpatient}
\alias{inpatient}
\title{Inpatient Hospitals}
\usage{
inpatient(
  year = NULL,
  ccn = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  patients = NULL,
  discharges = NULL,
  charge = NULL,
  payment = NULL,
  avg_age = NULL,
  avg_risk = NULL,
  dual = NULL,
  ndual = NULL,
  count = FALSE
)
}
\arguments{
\item{year}{\verb{<int>} Year data was reported}

\item{ccn}{\verb{<int>} 10-digit national provider identifier}

\item{org_name}{\verb{<chr>} Individual/Organizational provider's name}

\item{city, state, zip}{\verb{<chr>} The provider's city and state, as reported in NPPES.}

\item{patients}{\verb{<int>} Total Medicare beneficiaries receiving services from
the provider}

\item{discharges}{description}

\item{charge}{\verb{<int>} The total charges that the provider submitted for all
services}

\item{payment}{\verb{<dbl>} Total amount that Medicare paid after deductible and
coinsurance amounts have been deducted for all the provider's line item
services.}

\item{avg_age}{\verb{<dbl>} Average age of beneficiaries. Beneficiary age is
calculated at the end of the calendar year or at the time of death}

\item{avg_risk}{\verb{<dbl>} Average Hierarchical Condition Category (HCC) risk
score of beneficiaries}

\item{dual}{\verb{<int>} Number of Medicare beneficiaries qualified to receive
Medicare and Medicaid benefits. Beneficiaries are classified as Medicare
and Medicaid entitlement if in any month in the given calendar year they
were receiving full or partial Medicaid benefits.}

\item{ndual}{\verb{<int>} Number of Medicare beneficiaries qualified to receive
Medicare only benefits. Beneficiaries are classified as Medicare only
entitlement if they received zero months of any Medicaid benefits (full or
partial) in the given calendar year.}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Inpatient Hospitals
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
inpatient(count = TRUE)
inpatient(state = "GA", city = "Valdosta")
\dontshow{\}) # examplesIf}
}
