% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laboratories.R
\name{laboratories}
\alias{laboratories}
\title{Clinical Laboratories}
\usage{
laboratories(
  name = NULL,
  ccn = NULL,
  cert = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  active = FALSE
)
}
\arguments{
\item{name}{\verb{<chr>} Provider or clinical laboratory's name}

\item{ccn}{\verb{<chr>} 10-character CLIA number}

\item{cert}{\verb{<chr>} CLIA certificate type:
\itemize{
\item \code{"waiver"}
\item \code{"ppm"}
\item \code{"registration"}
\item \code{"compliance"}
\item \code{"accreditation"}
}}

\item{city}{\verb{<chr>} City}

\item{state}{\verb{<chr>} State}

\item{zip}{\verb{<chr>} Zip code}

\item{active}{\verb{<lgl>} Return only active providers}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Clinical laboratories including demographics and the type of testing
services the facility provides.
}
\section{CLIA}{


CMS regulates all laboratory testing (except research) performed on humans
in the U.S. through the \strong{Clinical Laboratory Improvement Amendments (CLIA)}.
In total, CLIA covers approximately 320,000 laboratory entities.

The \emph{Division of Clinical Laboratory Improvement & Quality}, within the
\emph{Quality, Safety & Oversight Group}, under the \emph{Center for Clinical Standards and Quality}
(CCSQ) has the responsibility for implementing the CLIA Program.

Although all clinical laboratories must be properly certified to receive
Medicare or Medicaid payments, CLIA has no direct Medicare or Medicaid
program responsibilities.
}

\section{Certification}{


The five CLIA certificate types, all of which are effective for a period of
two years, are as follows, in order of increasing complexity:
\enumerate{
\item \strong{Waiver}: Issued to a laboratory to perform only waived tests; does
not waive the lab from all CLIA requirements. Waived tests are laboratory
tests that are simple to perform. Routine inspections are not conducted
for waiver labs, although 2\% are visited each year to ensure quality
laboratory testing.
\item \strong{Provider-Performed Microscopy Procedures (PPM)}: Issued to a
laboratory in which a physician, mid-level practitioner or dentist
performs limited tests that require microscopic examination. PPM tests are
considered moderate complexity. Waived tests can also be performed under this
certificate type. There are no routine inspections conducted for PPM labs.
\item \strong{Registration}: Initially issued to a laboratory that has applied for
a Certificate of Compliance or Accreditation, enabling the lab to conduct
moderate/high complexity testing until the survey is performed and the
laboratory is found to be in CLIA compliance. Includes PPM and waived
testing.
\item \strong{Compliance}: Allows the laboratory to conduct moderate/high complexity
testing and is issued after an inspection finds the lab to be in
compliance with all applicable CLIA requirements. Includes PPM and
waived testing.
\item \strong{Accreditation}: Exactly the same as the Certificate of Compliance,
except that the laboratory must be accredited by one of the following
CMS-approved accreditation organizations:
\itemize{
\item \href{https://a2la.org/}{A2LA}: American Association for Laboratory Accreditation
\item \href{https://www.aabb.org/}{AABB}: Association for the Advancement of Blood & Biotherapies
\item \href{https://osteopathic.org/}{AOA}: American Osteopathic Association
\item \href{https://www.ashi-hla.org/}{ASHI-HLA}: American Society for Histocompatibility & Immunogenetics
\item \href{https://www.cap.org/}{CAP}: College of American Pathologists
\item \href{https://www.cola.org/}{COLA}: Commission on Office Laboratory Accreditation
\item \href{https://www.jointcommission.org/}{JCAHO}: The Joint Commission
}
}
}

\examples{
provider:::cdc_labs

laboratories(ccn = provider:::cdc_labs$ccn)

laboratories(
   cert = c("ppm", "accreditation"),
   city = "Valdosta",
   state = "GA",
   active = TRUE
 )
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories}{Provider of Services File - Clinical Laboratories API}
\item \href{https://www.cms.gov/medicare/quality/clinical-laboratory-improvement-amendments}{CMS CLIA}
\item \href{https://www.cdc.gov/clia/php/about/index.html}{CDC CLIA}
\item \href{https://qcor.cms.gov/main.jsp}{CMS QCOR}
\item \href{https://www.cdc.gov/labs/clia-certificates/index.html}{CLIA Certificates}
\item \href{https://www.cms.gov/files/document/clia-certificate-fee-schedule-updated-06/7/2024.pdf}{CLIA Certificate Fee Schedule}
\item \href{https://www.cms.gov/files/document/clia-cert-quick-start-guide.pdf}{CLIA Certification Guide}
}
}
