% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{list_pins}
\alias{list_pins}
\title{List pins from a \link[pins:pins-package]{pins} board}
\usage{
list_pins(...)
}
\arguments{
\item{...}{arguments to pass to \code{\link[=mount_board]{mount_board()}}}
}
\value{
\verb{<list>} of \link[pins:pins-package]{pins}
}
\description{
List pins from a \link[pins:pins-package]{pins} board
}
\keyword{internal}
