% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-ambulatory.R
\name{long_term}
\alias{long_term}
\title{Long-Term Care Hospitals}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}
}
\usage{
long_term(
  ccn = NULL,
  name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  own_type = NULL,
  beds = NULL,
  count = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} desc}

\item{name}{\verb{<chr>} desc}

\item{city, state, zip, county}{\verb{<chr>} desc}

\item{own_type}{description}

\item{beds}{\verb{<int>} desc}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
long_term(count = TRUE)
long_term(state = "GA")
\dontshow{\}) # examplesIf}
}
