% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc.R
\name{medline}
\alias{medline}
\title{Medline Plus API}
\usage{
medline(ndc)
}
\arguments{
\item{ndc}{< \emph{character} > // \strong{required} 10- to 11-digit National Drug Code}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:
}
\description{
Medline Plus API
}
\keyword{internal}
