% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{meta.funs}
\alias{meta.funs}
\title{Look up Dataset's API title}
\usage{
meta.funs(fn)
}
\arguments{
\item{call}{function name}
}
\value{
A list containing the function name and the dataset's API title
}
\description{
Look up Dataset's API title
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
meta.funs('affiliations')
meta.funs('clinicians')
meta.funs('beneficiaries')
\dontshow{\}) # examplesIf}
}
\keyword{internal}
