% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{meta.json}
\alias{meta.json}
\title{Return Metadata about function's dataset}
\usage{
meta.json(fn = NULL, title = NULL, first = TRUE)
}
\arguments{
\item{fn}{function using the dataset}

\item{title}{dataset's API title}

\item{first}{only need the first row}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the updated ids.
}
\description{
Return Metadata about function's dataset
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
meta.json(title = 'Medicare Fee-For-Service  Public Provider Enrollment')
meta.json(title = 'Provider of Services File - Clinical Laboratories')
meta.json(fn = 'providers')
meta.json(fn = 'laboratories')
\dontshow{\}) # examplesIf}
}
\keyword{internal}
