% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{meta.open}
\alias{meta.open}
\title{Return Metadata about function's dataset}
\usage{
meta.open(search = NULL, tidy = TRUE, year = NULL)
}
\arguments{
\item{search}{dataset title}

\item{tidy}{TRUE}

\item{year}{dataset year}
}
\value{
A list of metadata describing each API's dataset
}
\description{
Return Metadata about function's dataset
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
meta.open('General Payment Data')
meta.open(meta_funs('open_payments'), year = 2022)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
