% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{meta.store}
\alias{meta.store}
\title{Return Metadata about function's dataset}
\usage{
meta.store(fn = NULL, title = NULL)
}
\arguments{
\item{fn}{function using the dataset}

\item{title}{dataset's API title}
}
\value{
A list of metadata describing each API's dataset
}
\description{
Return Metadata about function's dataset
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
meta.store(title = 'Facility Affiliation Data')
meta.store(title = 'National Downloadable File')
meta.store(fn = 'affiliations')
meta.store(fn = 'clinicians')
\dontshow{\}) # examplesIf}
}
\keyword{internal}
