% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_payment.R
\name{mips_2021}
\alias{mips_2021}
\title{2021 Quality Payment Performance}
\usage{
mips_2021(
  facility_name = NULL,
  pac_id_org = NULL,
  npi = NULL,
  pac_id_ind = NULL,
  first_name = NULL,
  last_name = NULL,
  offset = 0L,
  tidy = TRUE
)
}
\arguments{
\item{facility_name}{Organization name}

\item{pac_id_org}{Unique organization ID assigned by PECOS}

\item{npi}{Unique clinician ID assigned by NPPES}

\item{pac_id_ind}{Unique individual clinician ID assigned by PECOS}

\item{first_name}{Individual clinician first name}

\item{last_name}{Individual clinician last name}

\item{offset}{offset; API pagination}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Performance information for Merit-Based Incentive Payment System (MIPS)
submitted by groups.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mips_2021(pac_id_org = 4789842956)
mips_2021(npi = 1316172182)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
