% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_information.R
\name{missing_information}
\alias{missing_information}
\title{Search the CMS Public Reporting of Missing Digital Contact Information API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
missing_information(npi = NULL, tidy = TRUE)
}
\arguments{
\item{npi}{The provider’s National Provider Identifier}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on providers missing digital contact information
in NPPES.
}
\details{
In the May 2020 CMS Interoperability and Patient Access final rule,
CMS finalized the policy to publicly report the names and NPIs of those
providers who do not have digital contact information included in the
NPPES system (85 FR 25584). This data includes the NPI and provider
name of providers and clinicians without digital contact
information in NPPES.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-compliance/public-reporting-of-missing-digital-contact-information}{CMS Public Reporting of Missing Digital Contact Information API}
}
}
}
\note{
Update Frequency: \strong{Quarterly}
}
\examples{
missing_information(npi = 1134122013)
missing_information(npi = 1386128379)
}
