% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{modifier}
\alias{modifier}
\alias{excludes}
\alias{contains}
\alias{starts}
\alias{ends}
\alias{not}
\alias{not_blank}
\alias{is_blank}
\alias{greater}
\alias{less}
\alias{between}
\title{A variety of different query operators}
\source{
\href{https://jsonapi.org/format/#query-parameters}{JSON-API: Query Parameters}
}
\usage{
excludes(x)

contains(x)

starts(x)

ends(x)

not(x)

not_blank()

is_blank()

greater(x, equal = FALSE)

less(x, equal = FALSE)

between(x, y)
}
\arguments{
\item{x, y}{input}

\item{equal}{\verb{<lgl>} append \code{=} to \code{less()} or \code{greater()}}
}
\value{
An S7 \verb{<Modifier>} object.
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query
conditions, in the
\href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API}
format.
}
\examples{
excludes(c("AL", "AK", "AZ"))

ends("bar")

starts("foo")

contains("baz")

not("zzz")

less(1000)

less(0.125, equal = TRUE)

greater(1000)

greater(0.125, equal = TRUE)

between(0.125, 2)

not_blank()

is_blank()

}
