% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc.R
\name{ndc_lookup}
\alias{ndc_lookup}
\title{National Drug Code (NDC) Lookup}
\usage{
ndc_lookup(ndc)
}
\arguments{
\item{ndc}{< \emph{character} > // \strong{required} 10- to 11-digit National Drug Code}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{ndc} \tab National Drug Code \cr
   \code{rxcui} \tab RxNorm Identifier \cr
   \code{atc} \tab ATC Identifier \cr
   \code{status} \tab Status (NDC) \cr
   \code{brand_name} \tab Brand Name (RxCUI) \cr
   \code{drug_name} \tab Drug Name (RxCUI) \cr
   \code{atc_first} \tab ATC I: Anatomical, Pharmacological \cr
   \code{atc_first} \tab ATC II: Pharmacological, Therapeutic \cr
   \code{atc_first} \tab ATC III: Chemical, Pharmacological, Therapeutic \cr
   \code{atc_first} \tab ATC IV: Chemical, Pharmacological, Therapeutic \cr
   \code{subject} \tab Medline Plus: Subject \cr
   \code{summary} \tab Medline Plus: Summary \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{NDC}{

A National Drug Code (NDC) is a unique 10- or 11-digit, 3-segment numeric
identifier assigned to each medication listed under Section 510 of the US
Federal Food, Drug, and Cosmetic Act.
}

\section{RxCUI}{

An RxCUI is a machine-readable code or identifier that points to the common
meaning shared by the various source names grouped and assigned to a particular concept.

RxNorm provides normalized names for clinical drugs and links its names to
many of the drug vocabularies commonly used in pharmacy management and drug
interaction software, including those of First Databank, Micromedex, Multum,
and Gold Standard Drug Database. By providing links between these vocabularies,
RxNorm can mediate messages between systems not using the same software and vocabulary.
}

\section{ATC}{

ATC classifies drugs at five different levels. Each level is a different
grouping of drugs. Groupings of active substances include the organ or system
on which the drug acts as well as therapeutic, pharmacological, and chemical
properties of the drug.

The Anatomical Therapeutic Chemical (ATC) classification was developed as a
modification and extension of the EphMRA classification system. In the ATC
classification system, the active substances are classified in a hierarchy
with five different levels.

The system has fourteen main anatomical/pharmacological groups or
\strong{1st Levels}.

Each ATC main group is divided into \strong{2nd Levels} which could be either
pharmacological or therapeutic groups.

The \strong{3rd and 4th Levels} are chemical, pharmacological or therapeutic
subgroups and the \strong{5th Level} is the chemical substance.

The 2nd, 3rd and 4th levels are often used to identify pharmacological
subgroups when that is considered more appropriate than therapeutic or
chemical subgroups.
}

\section{Links}{

\itemize{
\item \href{https://www.whocc.no/atc/structure_and_principles/}{ATCs}
\item \href{https://www.nlm.nih.gov/research/umls/rxnorm/overview.html}{RxNorm}
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ndc_lookup("0002-1433-80")

medline("0002-1433-80")

rxnorm("0002-1433-80")
\dontshow{\}) # examplesIf}
}
