% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{noresults_cli}
\alias{noresults_cli}
\title{noresults_cli ------------------------------------------------------------}
\usage{
noresults_cli(apiname, url)
}
\arguments{
\item{apiname}{Name of API}

\item{url}{API url landing page}
}
\description{
noresults_cli ------------------------------------------------------------
}
