% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_api_input.R
\docType{data}
\name{nppes_api_input}
\alias{nppes_api_input}
\title{Reference Table for NPPES API Search Parameters}
\format{
A data frame with 16 rows and 3 variables:
\describe{
\item{field}{NPPES search field}
\item{description}{Description of the search field}
\item{notes}{Notes detailing search parameters including constraints}
}
}
\source{
\url{https://npiregistry.cms.hhs.gov/registry/help-api}
}
\usage{
nppes_api_input
}
\description{
Reference table of search parameters and constraints
for the NPPES NPI Public Registry, per the API documentation.
}
\keyword{datasets}
