% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_api_output.R
\docType{data}
\name{nppes_api_output}
\alias{nppes_api_output}
\title{Reference Table for NPPES API Output Parameters}
\format{
A data frame with 90 rows and 3 variables:
\describe{
\item{group}{NPPES output group}
\item{field}{NPPES output field}
\item{description}{Description of the output field}
}
}
\source{
\url{https://npiregistry.cms.hhs.gov/registry/Json-Conversion-Field-Map}
}
\usage{
nppes_api_output
}
\description{
Reference table of output parameters
for the NPPES NPI Public Registry,
per the API documentation.
}
\keyword{datasets}
