% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_npi.R
\name{nppes_npi_multi}
\alias{nppes_npi_multi}
\title{Search the NPPES National Provider Identifier Registry API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
nppes_npi_multi(df)
}
\arguments{
\item{df}{data frame, tibble}
}
\description{
\code{nppes_npi_multi()} allows you to search the NPPES NPI
Registry's public API by many of the parameters defined in the
API's documentation.
}
\details{
The NPPES NPI Registry Public Search is a free directory of all
active National Provider Identifier (NPI) records. Healthcare providers
acquire their unique 10-digit NPIs to identify themselves in a standard
way throughout their industry. After CMS supplies an NPI, they publish
the parts of the NPI record that have public relevance, including the
provider’s name, taxonomy and practice address. It enables you to search
for providers in the NPPES (National Plan and Provider Enumeration
System.) All information produced by the NPI Registry is provided in
accordance with the NPPES Data Dissemination Notice. There is no charge
to use the NPI Registry.
\subsection{Links}{
\itemize{
\item \href{https://npiregistry.cms.hhs.gov/api-page}{NPPES NPI Registry API Documentation}
\item \href{https://npiregistry.cms.hhs.gov/demo-api}{NPPES NPI Registry API Demo}
}
}
}
\note{
Update Frequency: \strong{Weekly}
}
