% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-nursing.R
\name{nursing}
\alias{nursing}
\title{Nursing Homes}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/4pq5-n9py}{API: Nursing Home Provider Information}
}
}
\usage{
nursing(
  ccn = NULL,
  name = NULL,
  org_dba = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  beds = NULL,
  rating = NULL,
  count = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} CMS Certification Number}

\item{name}{\verb{<chr>} Provider Name}

\item{org_dba}{\verb{<chr>} Legal Business Name}

\item{org_name}{\verb{<chr>} Unique name identifying a group of nursing homes that
share at least one individual or organizational owner, officer, or entity
with operational/managerial control}

\item{city, state, zip, county}{\verb{<chr>} desc}

\item{beds}{\verb{<int>} Number of Federally Certified Beds}

\item{rating}{\verb{<int>} Overall Rating (1-5)}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
General information on currently active nursing homes, including
number of certified beds, quality measure scores, staffing and other
information used in the Five-Star Rating System. Data are presented as one
row per nursing home.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
nursing(count = TRUE)
nursing(state = "GA")
\dontshow{\}) # examplesIf}
}
