% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_payments.R
\name{open_payments}
\alias{open_payments}
\title{Search CMS' Open Payments API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
open_payments(
  covered_recipient_type = NULL,
  covered_recipient_npi = NULL,
  covered_recipient_first_name = NULL,
  covered_recipient_middle_name = NULL,
  covered_recipient_last_name = NULL,
  recipient_city = NULL,
  recipient_state = NULL,
  recipient_zip_code = NULL,
  year = 2021,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{covered_recipient_type}{"Covered Recipient Physician"}

\item{covered_recipient_npi}{10-digit National Provider Identifier (NPI).}

\item{covered_recipient_first_name}{First name}

\item{covered_recipient_middle_name}{Middle name}

\item{covered_recipient_last_name}{Last name}

\item{recipient_city}{City}

\item{recipient_state}{State}

\item{recipient_zip_code}{Zip code}

\item{year}{Reporting year, 2015-2021, default is 2021}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=open_payments]{open_payments()}} allows you to search CMS' NCCI Procedure
to Procedure Edits (PTP) API.
}
\details{
The Open Payments program is a national disclosure program that promotes a more transparent and accountable health care system. Open Payments houses a publicly accessible database of payments that reporting entities, including drug and medical device companies, make to covered recipients like physicians. Please note that CMS does not comment on what relationships may be beneficial or potential conflicts of interest. CMS publishes the data attested to by reporting entities. The data is open to individual interpretation.
\subsection{Links}{
\itemize{
\item \href{https://www.cms.gov/OpenPayments}{What is the Open Payments Program?}
\item \href{https://www.cms.gov/OpenPaymentsData.cms.gov}{OpenPaymentsData.cms.gov}
}
}
}
\note{
Update Frequency: \strong{Quarterly}
}
\examples{
\dontrun{
open_payments(covered_recipient_npi = "1043218118")
}
}
