% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_payments.R
\name{open_payments_years}
\alias{open_payments_years}
\title{Check the current years available for the Open Payments API}
\usage{
open_payments_years()
}
\value{
integer vector of years available
}
\description{
Check the current years available for the Open Payments API
}
\examples{
open_payments_years()
}
