% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-opted_out.R
\name{opt_out}
\alias{opt_out}
\title{Opt-Out Providers}
\usage{
opt_out(
  npi = NULL,
  first = NULL,
  last = NULL,
  specialty = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  order_refer = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{first, last}{\verb{<chr>} Provider's name}

\item{specialty}{\verb{<chr>} Provider's specialty}

\item{address}{\verb{<chr>} Provider's address}

\item{city}{\verb{<chr>} Provider's city}

\item{state}{\verb{<chr>} Provider's state abbreviation}

\item{zip}{\verb{<chr>} Provider's zip code}

\item{order_refer}{\verb{<lgl>} Indicates order and refer eligibility}

\item{count}{\verb{<lgl>} Return the dataset's total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Information on providers who have decided not to participate in Medicare.
}
\section{Opting Out}{

Providers who do not wish to enroll in the Medicare program may "opt-out",
meaning neither they nor the beneficiary can bill Medicare for services rendered.

Instead, a private contract between provider and beneficiary is signed,
neither party is reimbursed by Medicare and the beneficiary pays
the provider out-of-pocket.

To opt out, a provider must:
\itemize{
\item Be of an \strong{eligible specialty} type
\item Submit an \strong{opt-out affidavit} to Medicare
\item Enter into a \strong{private contract} with their Medicare patients
}
}

\section{Opt-Out Periods}{

Opt-out periods last for two years and cannot be terminated early unless the
provider is opting out for the very first time and terminates the opt-out no
later than 90 days after the opt-out period's effective date. Opt-out
statuses are effective for two years and automatically renew.

Providers may \strong{NOT} opt-out if they intend to be a Medicare Advantage
(Part C) provider or furnish services covered by traditional Medicare
fee-for-service (Part B).
}

\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
opt_out(count = TRUE)
opt_out(npi = 1043522824)
opt_out(state = "AK")
opt_out(specialty = "Psychiatry", order_refer = FALSE)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits}{Medicare Opt Out Affidavits API}
}
}
