% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-opted_out.R
\name{opted_out}
\alias{opted_out}
\title{Opt-Out Providers}
\usage{
opted_out(
  npi = NULL,
  first = NULL,
  last = NULL,
  specialty = NULL,
  start_year = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  order_refer = NULL,
  count = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{first, last}{\verb{<chr>} Provider's name}

\item{specialty}{\verb{<chr>} Provider's specialty}

\item{start_year}{\verb{<int>} Opt-out effective date year}

\item{address, city, state, zip}{\verb{<chr>} Provider's address, city, state, zip}

\item{order_refer}{\verb{<lgl>} Indicates order and refer eligibility}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on providers who have decided not to participate in Medicare.
}
\section{Opting Out}{

Providers who do not wish to enroll in the Medicare program may "opt-out",
meaning neither they nor the beneficiary can bill Medicare for services rendered.

Instead, a private contract between provider and beneficiary is signed,
neither party is reimbursed by Medicare and the beneficiary pays
the provider out-of-pocket.

To opt out, a provider must:
\itemize{
\item Be of an \strong{eligible specialty} type
\item Submit an \strong{opt-out affidavit} to Medicare
\item Enter into a \strong{private contract} with their Medicare patients
}
}

\section{Opt-Out Periods}{

Opt-out periods last for two years and cannot be terminated early unless the
provider is opting out for the very first time and terminates the opt-out no
later than 90 days after the opt-out period's effective date. Opt-out
statuses are effective for two years and automatically renew.

Providers may \strong{NOT} opt-out if they intend to be a Medicare Advantage
(Part C) provider or furnish services covered by traditional Medicare
fee-for-service (Part B).
}

\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
opted_out(count = TRUE)

opted_out(state = "GA", specialty = contains("Psych"))
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits}{Medicare Opt Out Affidavits API}
}
}
