% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_refer.R
\name{order_refer}
\alias{order_refer}
\title{Order and Referral Eligibility}
\usage{
order_refer(
  npi = NULL,
  first = NULL,
  last = NULL,
  part_b = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  hospice = NULL
)
}
\arguments{
\item{npi}{\verb{<int>} 10-digit Individual National Provider Identifier}

\item{first, last}{\verb{<chr>} Individual provider's first/last name}

\item{part_b, dme, hha, pmd, hospice}{\verb{<lgl>} Whether a provider is eligible to
order and refer to:
\itemize{
\item \code{partb}: Medicare Part B
\item \code{dme}: Durable Medical Equipment
\item \code{hha}: Home Health Agency
\item \code{pmd}: Power Mobility Devices
\item \code{hospice}: Hospice
}}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab National Provider Identifier \cr
   \code{first} \tab Order and Referring Provider's First Name \cr
   \code{last} \tab Order and Referring Provider's Last Name \cr
   \code{eligible} \tab Services An Eligible Provider Can Order/Refer To \cr
}
}
\description{
\code{order_refer()} returns a provider's eligibility to order and refer within
Medicare to:
}
\details{
\itemize{
\item \strong{Part B}: Clinical Laboratory Services, Imaging Services
\item \strong{DME}: Durable Medical Equipment, Prosthetics, Orthotics, & Supplies (DMEPOS)
\item \strong{Part A}: Home Health Services
}

To be eligible, a provider must:
\itemize{
\item have an \emph{Individual} NPI
\item be enrolled in Medicare in either an \emph{Approved} or \emph{Opt-Out} status
\item be of an \emph{Eligible Specialty} type
}

\strong{Ordering Providers} can order non-physician services for patients.

\strong{Referring (or Certifying) Providers} can request items or services that
Medicare may reimburse on behalf of its beneficiaries.

\strong{Opt-Out Providers}: Providers who have opted out of Medicare may still
order and refer. They can also enroll solely to order and refer.
}
\examples{
order_refer()

order_refer(npi = 100)

order_refer(npi = 1003026055)

order_refer(first = "Jennifer", last = "Smith")

order_refer(
  part_b = TRUE,
  dme = TRUE,
  hha = FALSE,
  pmd = TRUE,
  hospice = FALSE
 )

}
\references{
links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{Medicare Order and Referring API}
\item \href{https://www.cms.gov/medicare/enrollment-renewal/providers-suppliers/chain-ownership-system-pecos/ordering-certifying}{CMS.gov: Ordering & Certifying}
}
}
