% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-in_outpatient.R
\name{outpatient}
\alias{outpatient}
\title{Outpatient Hospitals}
\usage{
outpatient(
  year = NULL,
  ccn = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  patients = NULL,
  services = NULL,
  charge = NULL,
  allowed = NULL,
  payment = NULL,
  outliers = NULL,
  count = FALSE
)
}
\arguments{
\item{year}{\verb{<int>} Year data was reported}

\item{ccn}{\verb{<int>} 10-digit national provider identifier}

\item{org_name}{\verb{<chr>} Individual/Organizational provider's name}

\item{city, state, zip}{\verb{<chr>} The provider's city and state, as reported in NPPES.}

\item{patients}{\verb{<int>} Total Medicare beneficiaries receiving services from
the provider}

\item{services}{\verb{<int>} Total provider services}

\item{charge}{\verb{<int>} The total charges that the provider submitted for all
services}

\item{allowed}{\verb{<dbl>} The Medicare allowed amount for all provider services.
This figure is the sum of the amount Medicare pays, the deductible and
coinsurance amounts that the beneficiary is responsible for paying, and any
amounts that a third party is responsible for paying.}

\item{payment}{\verb{<dbl>} Total amount that Medicare paid after deductible and
coinsurance amounts have been deducted for all the provider's line item
services.}

\item{outliers}{description}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Outpatient Hospitals
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
outpatient(count = TRUE)
outpatient(state = "GA", city = "Valdosta")
\dontshow{\}) # examplesIf}
}
