% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-owner.R
\name{owner}
\alias{owner}
\title{Owners}
\source{
Medicare
}
\usage{
owner(
  fac_type = NULL,
  org_enid = NULL,
  org_pac = NULL,
  org_name = NULL,
  pac = NULL,
  owner = NULL,
  dba = NULL,
  percent = NULL,
  role = NULL,
  entity = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  title = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE
)
}
\arguments{
\item{fac_type}{\verb{<enum>} Facility type
\itemize{
\item \code{hha} = Home Health Agency
\item \code{rhc} = Rural Health Clinic
\item \code{fqhc} = Federally Qualified Health Clinic
\item \code{snf} = Skilled Nursing Facility
\item \code{hospice} = Hospice
\item \code{hospital} = Hospital
}}

\item{org_enid}{\verb{<chr>} National Provider Identifier}

\item{org_pac}{\verb{<chr>} Provider's name}

\item{org_name}{\verb{<chr>} Provider's name}

\item{pac}{\verb{<chr>} Provider's name}

\item{owner}{\verb{<chr>} Provider's name}

\item{dba}{\verb{<chr>} Provider's name}

\item{percent}{\verb{<chr>} Provider's name}

\item{role}{\verb{<chr>} Provider's name}

\item{entity}{\verb{<chr>} Provider's name}

\item{first, middle, last}{\verb{<chr>} Provider's name}

\item{title}{\verb{<chr>} Provider's name}

\item{address, city, state, zip}{\verb{<chr>} Provider's name}

\item{count}{\verb{<lgl>} Return the total row count}
}
\description{
Owners of facilities enrolled in Medicare.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
owner(count = TRUE)

owner(state = c("GA", "FL"), count = TRUE)

owner(city = "Valdosta", state = "GA")
\dontshow{\}) # examplesIf}
}
