% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-owners.R
\name{owners}
\alias{owners}
\alias{fqhc_owner}
\alias{hha_owner}
\alias{hospital_owner}
\alias{hospice_owner}
\alias{rhc_owner}
\alias{snf_owner}
\title{Facility Owners}
\usage{
fqhc_owner(
  org_enid = NULL,
  org_pac = NULL,
  org_name = NULL,
  pac = NULL,
  owner = NULL,
  dba = NULL,
  percent = NULL,
  role = NULL,
  entity = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  title = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE,
  set = FALSE
)

hha_owner(
  org_enid = NULL,
  org_pac = NULL,
  org_name = NULL,
  pac = NULL,
  owner = NULL,
  dba = NULL,
  percent = NULL,
  role = NULL,
  entity = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  title = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE,
  set = FALSE
)

hospital_owner(
  org_enid = NULL,
  org_pac = NULL,
  org_name = NULL,
  pac = NULL,
  owner = NULL,
  dba = NULL,
  percent = NULL,
  role = NULL,
  entity = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  title = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE,
  set = FALSE
)

hospice_owner(
  org_enid = NULL,
  org_pac = NULL,
  org_name = NULL,
  pac = NULL,
  owner = NULL,
  dba = NULL,
  percent = NULL,
  role = NULL,
  entity = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  title = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE,
  set = FALSE
)

rhc_owner(
  org_enid = NULL,
  org_pac = NULL,
  org_name = NULL,
  pac = NULL,
  owner = NULL,
  dba = NULL,
  percent = NULL,
  role = NULL,
  entity = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  title = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE,
  set = FALSE
)

snf_owner(
  org_enid = NULL,
  org_pac = NULL,
  org_name = NULL,
  pac = NULL,
  owner = NULL,
  dba = NULL,
  percent = NULL,
  role = NULL,
  entity = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  title = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{org_enid}{\verb{<chr>} National Provider Identifier}

\item{org_pac}{\verb{<chr>} Provider's name}

\item{org_name}{\verb{<chr>} Provider's name}

\item{pac}{\verb{<chr>} Provider's name}

\item{owner}{\verb{<chr>} Provider's name}

\item{dba}{\verb{<chr>} Provider's name}

\item{percent}{\verb{<chr>} Provider's name}

\item{role}{\verb{<chr>} Provider's name}

\item{entity}{\verb{<chr>} Provider's name}

\item{first, middle, last}{\verb{<chr>} Provider's name}

\item{title}{\verb{<chr>} Provider's name}

\item{address, city, state, zip}{\verb{<chr>} Provider's name}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\description{
Providers with pending Medicare enrollment applications.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
fqhc_owner(count = TRUE)
hospital_owner(count = TRUE)
hospice_owner(count = TRUE)
rhc_owner(count = TRUE)
snf_owner(count = TRUE)
hha_owner(count = TRUE)

fqhc_owner(state = c("GA", "FL")) |> str()

hospital_owner(state = c("GA", "FL")) |> str()

hospice_owner(state = c("GA", "FL")) |> str()

rhc_owner(state = c("GA", "FL")) |> str()

snf_owner(state = c("GA", "FL")) |> str()

hha_owner(state = c("GA", "FL")) |> str()
\dontshow{\}) # examplesIf}
}
