% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-pending.R
\name{pending}
\alias{pending}
\title{Pending Medicare Enrollments}
\source{
Medicare Pending Initial Logging and Tracking:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-physicians}{API: Physicians}
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-non-physicians}{API: Non-Physicians}
}
}
\usage{
pending(prov_type = NULL, npi = NULL, first = NULL, last = NULL, count = FALSE)
}
\arguments{
\item{prov_type}{\verb{<chr>} \code{"Physician"} or \code{"Non-Physician"}}

\item{npi}{\verb{<int>} National Provider Identifier}

\item{first, last}{\verb{<chr>} Provider's name}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Providers with pending Medicare enrollment applications.

The \emph{Pending Initial Logging and Tracking (L & T)} dataset provides a list of
pending applications for both Physicians and Non-Physicians that have not
been processed by CMS contractors.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
pending(count = TRUE)

pending(first = starts("V"))
\dontshow{\}) # examplesIf}
}
