% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending_applications.R
\name{pending_applications}
\alias{pending_applications}
\title{Search the Medicare Pending Initial Logging and Tracking API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
pending_applications(
  npi = NULL,
  last_name = NULL,
  first_name = NULL,
  type = c("physician", "non-physician"),
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{npi}{National Provider Identifier (NPI) number}

\item{last_name}{Last name of provider}

\item{first_name}{First name of provider}

\item{type}{physician or non-physician}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of enrollment applications pending CMS contractor
review for physicians & non-physicians.
}
\details{
The Pending Initial Logging and Tracking (L & T) dataset provides
a list of pending applications that have not been processed by CMS
contractors.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-physicians}{Medicare Pending Initial Logging and Tracking Physicians API}
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-non-physicians}{Medicare Pending Initial Logging and Tracking Non-Physicians API}
}
}
}
\note{
Update Frequency: \strong{Weekly}
}
\examples{
pending_applications(npi = 1487003984, type = "physician")
pending_applications(npi = 1487003984, type = "non-physician")
pending_applications(last_name = "Abbott", type = "non-physician")
pending_applications(first_name = "John", type = "physician")
}
