% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physician_by_geography.R
\name{physician_by_geography_years}
\alias{physician_by_geography_years}
\title{Check the current years available for the Physician & Other Practitioners by Geography and Service API}
\usage{
physician_by_geography_years()
}
\value{
integer vector of years available
}
\description{
Check the current years available for the Physician & Other Practitioners by Geography and Service API
}
\examples{
physician_by_geography_years()
}
