% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physician_by_provider.R
\name{physician_by_provider}
\alias{physician_by_provider}
\title{Search the Medicare Physician & Other Practitioners API
by \strong{Provider}}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
physician_by_provider(
  npi = NULL,
  last_org = NULL,
  first = NULL,
  cred = NULL,
  gender = NULL,
  type = NULL,
  city = NULL,
  state = NULL,
  fips = NULL,
  zip = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par_ind = NULL,
  year = 2020,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{npi}{National Provider Identifier (NPI) for the rendering provider
on the claim. The provider NPI is the numeric identifier registered in
NPPES.}

\item{last_org}{Last name/Organization name of the provider. When the
provider is registered in NPPES as an individual (entity type code \code{I}),
this is the provider’s last name. When the provider is registered as an
organization (entity type code \code{O}), this is the organization name.}

\item{first}{An individual provider's (entity type code \code{I}) first name.
An organization's (entity type code \code{O}) will be blank.}

\item{cred}{An individual provider's (entity type code \code{I}) credentials.
An organization's will be blank.}

\item{gender}{An individual provider's gender.
An organization's will be blank.}

\item{type}{Type of entity reported in NPPES. An entity code of \code{I}
identifies providers registered as individuals while an entity type
code of \code{O} identifies providers registered as organizations.}

\item{city}{The city where the provider is located, as reported in NPPES.}

\item{state}{The state where the provider is located, as reported in NPPES.}

\item{fips}{FIPS code for the rendering provider's state.}

\item{zip}{The provider’s zip code, as reported in NPPES.}

\item{ruca}{Rural-Urban Commuting Area Code (RUCA); a Census tract-based
classification scheme that utilizes the standard Bureau of Census
Urbanized Area and Urban Cluster definitions in combination with work
commuting information to characterize all of the nation's Census tracts
regarding their rural and urban status and relationships. The Referring
Provider ZIP code was cross walked to the United States Department of
Agriculture (USDA) 2010 Rural-Urban Commuting Area Codes.}

\item{country}{The country where the provider is located, as reported
in NPPES.}

\item{specialty}{Derived from the provider specialty code reported on the
claim. For providers that reported more than one specialty code on their
claims, this is the specialty code associated with the largest number
of services.}

\item{par_ind}{Identifies whether the provider participates in Medicare
and/or accepts assignment of Medicare allowed amounts. The value will
be \code{Y} for any provider that had at least one claim identifying the
provider as participating in Medicare or accepting assignment of
Medicare allowed amounts within HCPCS code and place of service. A
non-participating provider may elect to accept Medicare allowed amounts
for some services and not accept Medicare allowed amounts for other
services.}

\item{year}{Year in YYYY format, between 2013-2020; default is 2020}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on services and procedures provided to
Original Medicare (fee-for-service) Part B (Medical Insurance)
beneficiaries by physicians and other healthcare professionals;
aggregated by provider.
}
\details{
The \strong{Provider} dataset provides information on use, payments,
submitted charges and beneficiary demographic and health characteristics
organized by National Provider Identifier (NPI).
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider}{Medicare Physician & Other Practitioners: by Provider API}
}
}
}
\note{
Update Frequency: \strong{Annually}
}
\examples{
physician_by_provider(npi = 1003000423,
                      year = 2020)

physician_by_provider(type = "I",
                      city = "Hershey",
                      state = "PA",
                      fips = 42,
                      ruca =1,
                      gender = "F",
                      cred = "M.D.",
                      specialty = "Anesthesiology")
\dontrun{
purrr::map_dfr(as.character(2013:2020), ~physician_by_provider(npi = 1003000126, year = .x))
}
}
