% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physician_by_provider.R
\name{physician_by_provider_years}
\alias{physician_by_provider_years}
\title{Check the current years available for the Physician & Other Practitioners by Provider API}
\usage{
physician_by_provider_years()
}
\value{
integer vector of years available
}
\description{
Check the current years available for the Physician & Other Practitioners by Provider API
}
\examples{
physician_by_provider_years()
}
