% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physician_by_service.R
\name{physician_by_service}
\alias{physician_by_service}
\title{Search the Medicare Physician & Other Practitioners API
by \strong{Provider and Service}}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
physician_by_service(
  npi = NULL,
  last_org = NULL,
  first = NULL,
  cred = NULL,
  gender = NULL,
  type = NULL,
  city = NULL,
  state_abb = NULL,
  fips = NULL,
  zip = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par_ind = NULL,
  hcpcs_code = NULL,
  drug_ind = NULL,
  pos = NULL,
  year = 2020,
  clean_names = TRUE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{first}{Provider's first name}

\item{year}{Year in YYYY format, between 2013-2020; default is 2020}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{last}{Provider's last name}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on services and procedures provided to
Original Medicare (fee-for-service) Part B (Medical Insurance)
beneficiaries by physicians and other healthcare professionals;
aggregated by provider and service.
}
\details{
The \strong{Provider and Service} dataset provides information on
services and procedures provided to Medicare (fee-for-service) Part B
beneficiaries by physicians and other healthcare professionals.

The data is based on information gathered from CMS administrative claims
data for Part B beneficiaries available from the CMS Chronic Conditions
Data Warehouse.

The spending and utilization data in the Physician and Other
Practitioners by Provider and Service Dataset are aggregated
to the following:
\enumerate{
\item the NPI for the performing provider,
\item the Healthcare Common Procedure Coding System (HCPCS) code, and
\item the place of service (either facility or non-facility).
}

There can be multiple records for a given NPI based on the number of
distinct HCPCS codes that were billed and where the services were
provided. Data have been aggregated based on the place of service
because separate fee schedules apply depending on whether the place
of service submitted on the claim is facility or non-facility.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}{Medicare Physician & Other Practitioners: by Provider and Service API}
}
}
}
\note{
Update Frequency: \strong{Annually}
}
\examples{
\dontrun{
# Search by NPI
physician_by_service(npi = 1003000126)

# Search by Last Name for 2019
physician_by_service(last_org = "Enkeshafi", year = 2019)

# Multiple NPIs
npis <- c(1003026055,
          1316405939,
          1720392988,
          1518184605,
          1922056829,
          1083879860)

npis |> purrr::map_dfr(physician_by_service)

# Retrieve All Provider Data, 2013-2020
dates <- as.character(seq(from = 2013, to = 2020))
purrr::map_dfr(dates, ~physician_by_service(npi = 1003000126, year = .x))
}
}
