% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-PDC.R
\name{polish2}
\alias{polish2}
\title{Polish generic
Defines data cleaning methods for results}
\usage{
polish2(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
Polish generic
Defines data cleaning methods for results
}
\keyword{internal}
