% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescribers.R
\name{prescribers}
\alias{prescribers}
\alias{prescribers_}
\title{Prescriber Utilization & Demographics by Year}
\usage{
prescribers(
  year,
  type = "provider",
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  tidy = TRUE,
  na.rm = TRUE,
  ...
)

prescribers_(year = rx_years(), ...)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Year data was reported, in \code{YYYY}
format. Run \code{\link[=rx_years]{rx_years()}} to return a vector of the years currently available.}

\item{type}{< \emph{character} > // \strong{required} dataset to query, \code{"provider"}, \code{"drug"}, \code{"geography"}}

\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{first, last, organization}{< \emph{character} > Individual/Organizational
prescriber's name}

\item{credential}{< \emph{character} > Individual prescriber's credentials}

\item{gender}{< \emph{character} > Individual prescriber's gender; \code{"F"} (Female),
\code{"M"} (Male)}

\item{entype}{< \emph{character} > Prescriber entity type; \code{"I"} (Individual),
\code{"O"} (Organization)}

\item{city}{< \emph{character} > City where prescriber is located}

\item{state}{< \emph{character} > State where prescriber is located}

\item{zip}{< \emph{character} > Prescriber’s zip code}

\item{fips}{< \emph{character} > Prescriber's state's FIPS code}

\item{ruca}{< \emph{character} > Prescriber’s RUCA code}

\item{country}{< \emph{character} > Country where prescriber is located}

\item{specialty}{< \emph{character} > Prescriber specialty code reported on the
largest number of claims submitted}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}

\item{...}{Pass arguments to \code{\link[=prescribers]{prescribers()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=prescribers]{prescribers()}} allows access to information on prescription drugs provided
to Medicare beneficiaries enrolled in Part D (Prescription Drug Coverage),
by physicians and other health care providers; aggregated by provider, drug
and geography.
}
\section{By Provider}{

\strong{type =}\code{"provider"}:

The \strong{Provider} dataset allows the user access to data such as
services and procedures performed; charges submitted and payment received;
and beneficiary demographic and health characteristics for providers
treating Original Medicare (fee-for-service) Part B beneficiaries,
aggregated by year.
}

\section{By Provider and Drug}{

\strong{type =}\code{"service"}:

The \strong{Provider and Drug} dataset is aggregated by:
\enumerate{
\item Rendering provider's NPI
\item Healthcare Common Procedure Coding System (HCPCS) code
\item Place of Service (Facility or Non-facility)
}

There can be multiple records for a given NPI based on the number of
distinct HCPCS codes that were billed and where the services were
provided. Data have been aggregated based on the place of service
because separate fee schedules apply depending on whether the place
of service submitted on the claim is facility or non-facility.
}

\section{By Geography and Drug}{

\strong{type =}\code{"geography"}:

The \strong{Geography and Drug} dataset contains information on utilization,
allowed amount, Medicare payment, and submitted charges organized nationally
and state-wide by HCPCS code and place of service.
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider}{Medicare Part D Prescribers: by Provider}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider-and-drug}{Medicare Part D Prescribers: by Provider and Drug}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-geography-and-drug}{Medicare Part D Prescribers: by Geography and Drug}
}

\emph{Update Frequency:} \strong{Annually}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
prescribers(year = 2020, npi = 1003000423) |> dplyr::glimpse()
prescribers(year = 2021, npi = 1003000126) |> dplyr::glimpse()
prescribers(year = 2019, npi = 1003000126)

# Use the years helper function to retrieve results for every year:
rx_years() |>
map(\(x) prescribers(year = x, npi = 1043477615)) |>
list_rbind()

# Parallelized version
prescribers_(npi = 1043477615)
\dontshow{\}) # examplesIf}
}
