% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_mcr_ffs.R
\name{prov_mcr_ffs}
\alias{prov_mcr_ffs}
\title{Search Medicare Fee-For-Service Public Provider Enrollment API}
\source{
https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment
}
\usage{
prov_mcr_ffs(npi)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
The Medicare Fee-For-Service Public Provider Enrollment
dataset includes information on providers who are actively approved to
bill Medicare or have completed the 855O at the time the data was pulled
from the Provider Enrollment and Chain Ownership System (PECOS). These
files are populated from PECOS and contain basic enrollment and provider
information, reassignment of benefits information and practice location
city, state and zip. These files are not intended to be used as real-time
reporting as the data changes from day to day and the files are updated
only on a quarterly basis.
}
\details{
Information on a point in time snapshot of enrollment
level data for providers actively enrolled in Medicare.
}
\examples{
prov_mcr_ffs(1003026055)
}
