% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_mcr_ordref.R
\name{prov_mcr_ordref}
\alias{prov_mcr_ordref}
\title{Search Medicare Order and Referring API}
\source{
https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring
}
\usage{
prov_mcr_ordref(npi)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
The Order and Referring dataset provides
information on all physicians and non-physician practitioners,
by their National Provider Identifier (NPI), who are of a
type/specialty that is legally eligible to order and refer in
the Medicare program and who have current enrollment records in Medicare.
}
\details{
All physicians and non-physician practitioners who are legally
eligible to order and refer in the Medicare program and
who have current enrollment records in Medicare.
}
\examples{
prov_mcr_ordref(1003026055)
}
