% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_mcr_phys_pract.R
\name{prov_mcr_phys_pract}
\alias{prov_mcr_phys_pract}
\title{Search the Medicare Physician & Other
Practitioners - by Provider and Service API}
\source{
https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service
}
\usage{
prov_mcr_phys_pract(npi, year = "2020")
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{year}{year between 2013-2020, in YYYY format}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
The Medicare Physician & Other Practitioners by
Provider and Service dataset provides information on use,
payments, and submitted charges organized by National Provider
Identifier (NPI), Healthcare Common Procedure Coding System
(HCPCS) code, and place of service.
}
\details{
Information on services and procedures provided
to Original Medicare (fee-for-service) Part B
(Medical Insurance) beneficiaries by physicians
and other healthcare professionals; aggregated by
provider and service.
}
\examples{
prov_mcr_phys_pract(npi = 1003000126, year = "2020")
}
