% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_mcr_reval_date.R
\name{prov_mcr_reval_date}
\alias{prov_mcr_reval_date}
\title{Search the Medicare Revalidation Due Date API}
\usage{
prov_mcr_reval_date(npi)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
The Revalidation Due Date List dataset contains
revalidation due dates for Medicare providers who are due to
revalidate in the following six months. If a provider's due
date does not fall within the ensuing six months, the due
date is marked 'TBD'. In addition the dataset also includes
subfiles with reassignment information for a given provider
as well as due date listings for clinics and group practices
and their providers.
}
\details{
Information on revalidation due dates for Medicare providers.
}
\examples{
prov_mcr_reval_date(1184699621)
}
