% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_mcr_taxcross.R
\name{prov_mcr_taxcross}
\alias{prov_mcr_taxcross}
\title{Search Medicare Provider and Supplier Taxonomy Crosswalk API}
\source{
https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-supplier-taxonomy-crosswalk
}
\usage{
prov_mcr_taxcross(taxonomy_code)
}
\arguments{
\item{taxonomy_code}{unique 10-character alphanumeric code that
designates a provider’s classification and specialization.}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
The Medicare Provider and Supplier Taxonomy
Crosswalk dataset lists the providers and suppliers eligible
to enroll in Medicare programs with the proper healthcare
provider taxonomy code. This data includes the Medicare
speciality codes, if available, provider/supplier type
description, taxonomy code, and the taxonomy description.
This dataset is derived from information gathered from the
National Plan and Provider Enumerator System (NPPES) and the
Provider Enrollment, Chain and Ownership System (PECOS).
}
\details{
A list of the type of providers and suppliers with the proper
taxonomy code eligible for medicare programs.
}
\examples{
prov_mcr_taxcross("2086S0102X")
}
