% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_nppes_missing.R
\name{prov_nppes_missing}
\alias{prov_nppes_missing}
\title{Search the CMS Public Reporting of Missing Digital Contact Information API}
\usage{
prov_nppes_missing(npi)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
In the May 2020 CMS Interoperability and Patient Access
final rule, CMS finalized the policy to publicly report the names
and NPIs of those providers who do not have digital contact
information included in the NPPES system (85 FR 25584). This
data includes the NPI and provider name of providers and
clinicians without digital contact information in NPPES.
}
\details{
Information on providers missing digital contact information in NPPES.
}
\examples{
prov_nppes_missing(1760485387)
}
