% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_nppes_unpack.R
\name{prov_nppes_unpack}
\alias{prov_nppes_unpack}
\title{Unpack NPI NPPES Search Data}
\usage{
prov_nppes_unpack(df)
}
\arguments{
\item{df}{data.frame or tbl_df, response from prov_npi_nppes()}
}
\value{
A tibble containing the NPI(s) searched for on the
NPPES, the date-time of the search, and the unnested and
tidied list-columns of results split into groups.
}
\description{
Unpack NPI NPPES Search Data
}
\examples{
\dontrun{prov_nppes_unpack(nppes_npi_response)}
}
