% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_nppes_unpack_city.R
\name{prov_nppes_unpack_city}
\alias{prov_nppes_unpack_city}
\title{Unpack NPPES Search Data}
\usage{
prov_nppes_unpack_city(df)
}
\arguments{
\item{df}{data.frame or tbl_df, response from prov_city_nppes()}
}
\value{
A tibble containing the NPI(s) searched for on the
NPPES, the date-time of the search, and the unnested and
tidied list-columns of results split into groups.
}
\description{
Unpack NPPES Search Data
}
\examples{
\dontrun{prov_nppes_unpack_city(nppes_response)}
}
