% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_opt_out.R
\name{prov_opt_out}
\alias{prov_opt_out}
\title{Search Medicare Opt Out Affidavits API}
\source{
https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits
}
\usage{
prov_opt_out(npi)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
The Opt Out Affidavits dataset provides
information on providers who have decided not to
participate in Medicare. It contains the provider's NPI,
specialty, address, and effective dates.
}
\details{
A list of practitioners who are currently opted out of Medicare.
}
\examples{
prov_opt_out(1114974490)
}
