% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-package.R
\docType{package}
\name{provider-package}
\alias{provider}
\alias{provider-package}
\title{provider: Access Public Healthcare Provider Data}
\description{
Provides various features to facilitate easy access to publicly available healthcare provider data. Includes functions to tidy the data being accessed, whether for reporting purposes or updating internal records.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrewallenbruce.github.io/provider/}
  \item \url{https://github.com/andrewallenbruce/provider}
  \item Report bugs at \url{https://github.com/andrewallenbruce/provider/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrew.bruce@northstarrevenueintegrity.com} [copyright holder]

}
\keyword{internal}
