% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physician_by_geography.R,
%   R/physician_by_provider.R, R/physician_by_service.R
\name{physician_by_geography}
\alias{physician_by_geography}
\alias{physician_by_provider}
\alias{physician_by_service}
\title{Search the Medicare Physician & Other Practitioners API
by \strong{Geography and Service}}
\source{
Centers for Medicare & Medicaid Services

Centers for Medicare & Medicaid Services
}
\usage{
physician_by_geography(
  year,
  level = NULL,
  sublevel = NULL,
  fips = NULL,
  hcpcs_code = NULL,
  hcpcs_desc = NULL,
  hcpcs_drug = NULL,
  place_of_srvc = NULL,
  tidy = TRUE
)

physician_by_provider(
  year,
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  organization_name = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  tidy = TRUE
)

physician_by_service(
  year,
  npi = NULL,
  last_name = NULL,
  first_name = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  hcpcs_code = NULL,
  hcpcs_drug = NULL,
  place_of_srvc = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{int (required); Year in YYYY format. Run helper function
\code{provider:::physician_by_service_years()} to return a vector of the years
currently available.}

\item{level}{Identifies the level of geography by which the data in the
row has been aggregated. A value of 'State' indicates the data in the
row is aggregated to a single state identified in the Rendering Provider
State column for a given HCPCS Code Level. A value of 'National'
indicates the data in the row is aggregated across all states for a
given HCPCS Code Level.}

\item{sublevel}{The state name where the provider is located, as reported
in NPPES. The values include the 50 United States, District of Columbia,
U.S. territories, Armed Forces areas, Unknown and Foreign Country. Data
aggregated at the National level are identified by the word 'National'.}

\item{fips}{FIPS code for the rendering provider's state.}

\item{hcpcs_code}{HCPCS code used to identify the specific medical service
furnished by the provider. HCPCS codes include two levels. Level I codes
are the Current Procedural Terminology (CPT) codes that are maintained
by the American Medical Association and Level II codes are created by
CMS to identify products, supplies and services not covered by the CPT
codes (such as ambulance services).}

\item{hcpcs_desc}{Description of the HCPCS code for the specific medical
service furnished by the provider. HCPCS descriptions associated with
CPT codes are consumer friendly descriptions provided by the AMA. CPT
Consumer Friendly Descriptors are lay synonyms for CPT descriptors that
are intended to help healthcare consumers who are not medical
professionals understand clinical procedures on bills and patient
portals. CPT Consumer Friendly Descriptors should not be used for
clinical coding or documentation. All other descriptions are CMS Level
II descriptions provided in long form. Due to variable length
restrictions, the CMS Level II descriptions have been truncated to
256 bytes. As a result, the same HCPCS description can be associated
with more than one HCPCS code.}

\item{hcpcs_drug}{Identifies whether the HCPCS code for the specific service
furnished by the provider is a HCPCS listed on the Medicare Part B Drug
Average Sales Price (ASP) File. Please visit the ASP drug pricing page
for additional information.}

\item{place_of_srvc}{Identifies whether the place of service submitted on the claims
is a facility (value of \code{F}) or non-facility (value of \code{O}). Non-facility
is generally an office setting; however other entities are included
in non-facility.}

\item{tidy}{Tidy output; default is \code{TRUE}.}

\item{npi}{National Provider Identifier (NPI) for the rendering provider
on the claim. The provider NPI is the numeric identifier registered in
NPPES.}

\item{first_name}{An individual provider's (entity type code = \code{I}) first name.
An organization's (entity type code = \code{O}) will be blank.}

\item{last_name}{Last name/Organization name of the provider. When the
provider is registered in NPPES as an individual (entity type code = \code{I}),
this is the provider’s last name. When the provider is registered as an
organization (entity type code = \code{O}), this is the organization name.}

\item{organization_name}{Organization name.}

\item{credential}{An individual provider's (entity type code=’I’) credentials. An organization's will be blank.}

\item{gender}{An individual provider's gender. An organization's will be blank.}

\item{entype}{Type of entity reported in NPPES. An entity code of ‘I’
identifies providers registered as individuals while an entity type
code of ‘O’ identifies providers registered as organizations.}

\item{city}{The city where the provider is located, as reported in NPPES.}

\item{state}{The state where the provider is located, as reported
in NPPES.}

\item{zipcode}{The provider’s zip code, as reported in NPPES.}

\item{ruca}{Rural-Urban Commuting Area Code (RUCA); a Census tract-based
classification scheme that utilizes the standard Bureau of Census
Urbanized Area and Urban Cluster definitions in combination with work
commuting information to characterize all of the nation's Census tracts
regarding their rural and urban status and relationships. The Referring
Provider ZIP code was cross walked to the United States Department of
Agriculture (USDA) 2010 Rural-Urban Commuting Area Codes.}

\item{country}{The country where the provider is located, as reported
in NPPES.}

\item{specialty}{Derived from the provider specialty code reported on the
claim. For providers that reported more than one specialty code on their
claims, this is the specialty code associated with the largest number
of services.}

\item{par}{Identifies whether the provider participates in Medicare
and/or accepts assignment of Medicare allowed amounts. The value will
be \code{Y} for any provider that had at least one claim identifying the
provider as participating in Medicare or accepting assignment of
Medicare allowed amounts within HCPCS code and place of service. A
non-participating provider may elect to accept Medicare allowed amounts
for some services and not accept Medicare allowed amounts for other
services.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.

A \link[tibble:tibble-package]{tibble} containing the search results.

A \link[tibble:tibble-package]{tibble} containing 29 columns:
\item{year}{year}
\item{npi}{year}
\item{entype}{Entity/Enumeration Type of the Provider}
\item{first_name}{year}
\item{middle_name}{year}
\item{last_name}{year}
\item{credential}{year}
\item{gender}{year}
\item{specialty}{year}
\item{street}{ the provider’s street address}
\item{city}{The city where the provider is located, as reported in NPPES.}
\item{state}{State Abbreviation of the Provider}
\item{fips}{State FIPS Code of the Provider}
\item{zipcode}{Zip Code of the Provider}
\item{ruca}{RUCA Code of the Provider}
\item{country}{Country Code of the Provider}
\item{par}{Medicare Participation Indicator}
\item{hcpcs_cd}{HCPCS Code}
\item{hcpcs_desc}{HCPCS Description}
\item{hcpcs_drug}{HCPCS Drug Indicator}
\item{pos}{Place of Service}
\item{tot_benes}{Number of Medicare Beneficiaries}
\item{tot_srvcs}{Number of Services}
\item{tot_day}{Number of Distinct Medicare Beneficiary/Per Day Services}
\item{avg_charge}{Average Submitted Charge Amount}
\item{avg_allowed}{Average Medicare Allowed Amount}
\item{avg_payment}{Average Medicare Payment Amount}
\item{avg_std_pymt}{Average Medicare Standardized Payment Amount}
}
\description{
Information on services and procedures provided to
Original Medicare (fee-for-service) Part B (Medical Insurance)
beneficiaries by physicians and other healthcare professionals;
aggregated by geography and service.

\code{physician_by_provider()} allows you to access data such as
services and procedures performed; charges submitted and payment received;
and beneficiary demographic and health characteristics for providers treating
Original Medicare (fee-for-service) Part B beneficiaries, aggregated by year.

Information on services and procedures provided to
Original Medicare (fee-for-service) Part B (Medical Insurance)
beneficiaries by physicians and other healthcare professionals;
aggregated by provider and service.
}
\details{
The \strong{Geography and Service} dataset contains information on
utilization, payment (allowed amount and Medicare payment), and
submitted charges organized by HCPCS and place of service in the
national table and organized by provider state, HCPCS and place of
service in the state table. The national and state tables also include
a HCPCS drug indicator to identify whether the HCPCS product/service is
a drug as defined from the Medicare Part B Drug ASP list.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service}{Medicare Physician & Other Practitioners: by Geography and Service API}
}
}

The \strong{Provider and Service} dataset provides information on
services and procedures provided to Medicare (fee-for-service) Part B
beneficiaries by physicians and other healthcare professionals.

The data is based on information gathered from CMS administrative claims
data for Part B beneficiaries available from the CMS Chronic Conditions
Data Warehouse.

The spending and utilization data in the Physician and Other
Practitioners by Provider and Service Dataset are aggregated
to the following:
\enumerate{
\item the NPI for the performing provider,
\item the Healthcare Common Procedure Coding System (HCPCS) code, and
\item the place of service (either facility or non-facility).
}

There can be multiple records for a given NPI based on the number of
distinct HCPCS codes that were billed and where the services were
provided. Data have been aggregated based on the place of service
because separate fee schedules apply depending on whether the place
of service submitted on the claim is facility or non-facility.
}
\note{
Update Frequency: \strong{Annually}

Update Frequency: \strong{Annually}
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider}{Medicare Physician & Other Practitioners: by Provider API}
}
}

\section{Update Frequency}{
 *\emph{Annually}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
physician_by_geography(hcpcs_code = "0002A", year = 2020)
service <- purrr::map_dfr(as.character(2013:2020),
           ~physician_by_service(npi = 1003000126, year = .x))

procedures <- service |>
              dplyr::distinct(hcpcs_cd) |>
              tibble::deframe()

arg_cross <- purrr::cross_df(list(
             x = as.character(2013:2020),
             y = procedures))

# National Level
purrr::map2_dfr(arg_cross$x, arg_cross$y,
~physician_by_geography(geo_level = "National", year = .x, hcpcs_code = .y))

# State Level
purrr::map2_dfr(arg_cross$x, arg_cross$y,
~physician_by_geography(geo_level = "Georgia", year = .x, hcpcs_code = .y))
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
physician_by_provider(year = 2020,
                      npi = 1003000423)

# Use the years helper function to retrieve results for every year:
physician_by_provider_years() |>
map(\(x) physician_by_provider(year = x, npi = 1043477615)) |>
list_rbind()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
physician_by_service(npi = 1003000126)
physician_by_service(year = 2019, last_name = "Enkeshafi")
c(1003026055, 1316405939, 1720392988) |>
  purrr::map(physician_by_service, year = 2020) |>
  purrr::list_rbind()
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}{Medicare Physician & Other Practitioners: by Provider and Service API}
}
}
\seealso{
\code{\link[=physician_by_provider_years]{physician_by_provider_years()}}, \code{\link[=physician_by_service]{physician_by_service()}},
\code{\link[=physician_by_geography]{physician_by_geography()}}, \code{\link[=beneficiary_enrollment]{beneficiary_enrollment()}}, \code{\link[=cc_multiple]{cc_multiple()}},
\code{\link[=cc_specific]{cc_specific()}}
}
