% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{provider_cli}
\alias{provider_cli}
\title{provider_cli ------------------------------------------------------------}
\usage{
provider_cli(apiname, resp, size)
}
\arguments{
\item{apiname}{Name of API}

\item{resp}{httr2 response object}

\item{size}{size of responses downloaded}
}
\description{
provider_cli ------------------------------------------------------------
}
