% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{provider_cli_2}
\alias{provider_cli_2}
\title{provider_cli_2 ------------------------------------------------------------}
\usage{
provider_cli_2(apiname, resp, size)
}
\arguments{
\item{apiname}{Name of API}

\item{resp}{httr2 response object}

\item{size}{size of responses downloaded}
}
\description{
provider_cli_2 ------------------------------------------------------------
}
