% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_clia.R
\name{provider_clia}
\alias{provider_clia}
\title{Search the Medicare Provider of Services File - Clinical Laboratories API}
\usage{
provider_clia(
  name = NULL,
  city = NULL,
  state = NULL,
  year = "2022",
  clean_names = TRUE,
  full = FALSE
)
}
\arguments{
\item{name}{Name of the provider certified to participate in the Medicare
and/or Medicaid programs.}

\item{city}{City in which the provider is physically located.}

\item{state}{Two-character state abbreviation.}

\item{year}{Year between 2010-2022, in YYYY format}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{full}{If true, downloads the first 1000 rows of data;
default is \code{FALSE}.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the search results with the
following columns:
\describe{
\item{prvdr_ctgry_sbtyp_cd}{Identifies the subtype of provider within the primary category. (01 = CLIA88 Lab)}
\item{prvdr_ctgry_cd}{Identifies the type of provider participating in the Medicare/Medicaid program. (22 = CLIA Lab)}
\item{chow_cnt}{Number of times this provider has undergone a change of ownership.}
\item{chow_dt}{Effective date of the most recent change of ownership for this provider.}
\item{city_name}{City in which the provider is physically located.}
}
}
\description{
\code{\link[=provider_clia]{provider_clia()}} allows access to a dataset that
provides information on clinical laboratories including
demographics and the type of testing services the facility
provides.
}
\section{Medicare Provider of Services File - Clinical Laboratories API}{
The Provider of Services (POS) Clinical Laboratories (CLIA) data
provides information on CLIA demographics and types of testing
services the facility provides. In this file you will find provider
number, name, address and characteristics of the participating
institution providers. This data is gathered as part of the CMS
Provider Certification process and is updated each time a provider
is recertified.
\subsection{Data Update Frequency}{

Quarterly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories}{Medicare Provider of Services File - Clinical Laboratories API}
\item \href{https://data.cms.gov/sites/default/files/2020-12/POS_CLIA_LAYOUT_SEP20.pdf}{Data Dictionary}
}
}
}

\examples{
\dontrun{
provider_clia(name = "carbon hill", year = "2022")

provider_clia(city = "alabaster")

# Download First 1,000 Rows of Dataset =====================================
provider_clia(full = TRUE)
}
}
