% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-common-params.R
\name{provider_common_params}
\alias{provider_common_params}
\title{Common parameters}
\arguments{
\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Common parameter definitions
}
\keyword{internal}
