% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_luhn.R
\name{provider_luhn}
\alias{provider_luhn}
\title{Checks NPI number for compliance with the Luhn algorithm}
\usage{
provider_luhn(npi = NULL)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI).}
}
\value{
boolean, \code{TRUE} or \code{FALSE}
}
\description{
\code{provider_luhn()} checks NPIs against the Luhn algorithm for
compliance with the CMS requirements stated in the linked PDF below.
}
\section{Requirements for NPI Check Digit}{
The National Provider Identifier (NPI) check digit is calculated using
the Luhn formula for computing the modulus 10 “double-add-double” check
digit. This algorithm is recognized as an ISO standard and is the specified
check digit algorithm to be used for the card issuer identifier on a
standard health identification card.

When an NPI is used as a card issuer identifier on a standard health
identification card, it is preceded by the prefix \code{80840}, in which \code{80}
indicates health applications and \code{840} indicates the United States.

The prefix is required only when the NPI is used as a card issuer
identifier. However, in order that any NPI could be used as a card issuer
identifier on a standard health identification card, the check digit will
always be calculated as if the prefix is present. This is accomplished by
adding the constant \code{24} in step 2 of the check digit calculation (as shown
in the second example below) when the NPI is used without the prefix.
\subsection{Example NPI Check Digit Calculation (Card Issuer Identifier)}{

Assume that the NPI is \code{123456789}.

If used as a card issuer identifier on a standard health identification
card, the full number would be \code{80840123456789}. Using the Luhn formula
on the identifier portion, the check digit is calculated as follows:

Card issuer identifier without check digit: \code{80840123456789}

\emph{Step 1}: Double the value of alternate digits,
beginning with the rightmost digit:

\verb{0 8 2 6 10 14 18}

\emph{Step 2}: Add the individual digits of products
of doubling, plus unaffected digits:

\code{8 + 0 + 8 + 8 + 0 + 2 + 2 + 6 + 4 + 1 + 0 + 6 + 1 + 4 + 8 + 1 + 8 = 67}

\emph{Step 3}: Subtract from next higher number ending in zero.

\verb{70 – 67 = 3}

The check digit equals 3, thus the card issuer identifier
with check digit is \strong{80840}123456789\strong{3}.
}

\subsection{Example NPI Check Digit Calculation (without Prefix)}{

Assume that the NPI is \code{123456789}.

Using the Luhn formula on the identifier portion, the check
digit is calculated as follows:

NPI without check digit: \code{123456789}

\emph{Step 1}: Double the value of alternate digits,
beginning with the rightmost digit:

\verb{2 6 10 14 18}

\emph{Step 2}: Add constant \code{24}, to account for the \code{80840}
prefix that would be present on a card issuer identifier,
plus the individual digits of products of doubling,
plus unaffected digits:

\code{24 + 2 + 2 + 6 + 4 + 1 + 0 + 6 + 1 + 4 + 8 + 1 + 8 = 67}

\emph{Step 3}: Subtract from next higher number ending in zero.

\verb{70 – 67 = 3}

The check digit equals \strong{3}, thus the NPI with check digit is 123456789\strong{3}.
}
}

\examples{
# Valid NPI:
provider_luhn(npi = 1528060837)

# Quoted NPIs are valid:
provider_luhn(npi = "1528060837")

# Invalid NPI (per Luhn algorithm):
provider_luhn(npi = 1234567891)

\dontrun{
# NPIs with less than 10 digits throw an error:
provider_luhn(npi = 123456789)

# Inputting letters will throw an error, quoted or not:
provider_luhn(npi = abcdefghij)
provider_luhn(npi = "abcdefghij")
}
}
\references{
\url{https://www.cms.gov/Regulations-and-Guidance/Administrative-Simplification/NationalProvIdentStand/Downloads/NPIcheckdigit.pdf}
\url{https://en.wikipedia.org/wiki/Luhn_algorithm}
\url{https://npiregistry.cms.hhs.gov/api-page}
}
