% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_moar.R
\name{provider_moar}
\alias{provider_moar}
\title{Search the Medicare Order and Referring API}
\usage{
provider_moar(
  npi = NULL,
  last = NULL,
  first = NULL,
  clean_names = TRUE,
  full = FALSE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{last}{Provider's last name}

\item{first}{Provider's first name}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{full}{If true, downloads the first 1000 rows of data;
default is \code{FALSE}.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the search results.
}
\description{
All physicians and non-physician practitioners who are legally
eligible to order and refer in the Medicare program and who have current
enrollment records in Medicare.
}
\section{Medicare Order and Referring API}{
The Order and Referring dataset provides information on all physicians and
non-physician practitioners, by their National Provider Identifier (NPI),
who are of a type/specialty that is legally eligible to order and refer in
the Medicare program and who have current enrollment records in Medicare.
\subsection{Data Update Frequency}{

Weekly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{Medicare Order and Referring API}
}
}
}

\examples{
\dontrun{
provider_moar(npi = 1003026055)

provider_moar(last = "phadke", first = "radhika")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |> purrr::map_dfr(provider_moar)

# Data frame of NPIs
npi_df <- data.frame(npi = c(1003026055,
                             1316405939,
                             1720392988,
                             1518184605,
                             1922056829,
                             1083879860))
npi_df |>
tibble::deframe() |>
purrr::map_dfr(provider_moar)

# Download First 1,000 Rows of Dataset =====================================
provider_moar(full = TRUE)
}
}
