% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_mpop.R
\name{provider_mpop}
\alias{provider_mpop}
\title{Search the Medicare Physician & Other Practitioners
(by Provider and Service) API}
\usage{
provider_mpop(
  npi = NULL,
  last = NULL,
  first = NULL,
  year = "2020",
  clean_names = TRUE,
  full = FALSE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{last}{Provider's last name}

\item{first}{Provider's first name}

\item{year}{Year between 2013-2020, in YYYY format}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{full}{If true, downloads the first 1000 rows of data;
default is \code{FALSE}.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the search results.
}
\description{
Information on services and procedures provided to
Original Medicare (fee-for-service) Part B (Medical Insurance)
beneficiaries by physicians and other healthcare professionals;
aggregated by provider and service.
}
\section{Medicare Physician & Other Practitioners Dataset}{
The Medicare Physician & Other Practitioners by
Provider and Service dataset provides information on use,
payments, and submitted charges organized by National Provider
Identifier (NPI), Healthcare Common Procedure Coding System
(HCPCS) code, and place of service.
}

\examples{
\dontrun{
# Search by NPI ===========================================================
provider_mpop(npi = 1003000126, year = "2020")

# Search by First Name ====================================================
provider_mpop(first = "Enkeshafi", year = "2019")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |> purrr::map_dfr(provider_mpop)

# Returns the First 1,000 Rows in the Dataset =============================
provider_mpop(full = TRUE)
}

}
\references{
\url{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}
}
