% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_mpop_serv.R
\name{provider_mpop_serv}
\alias{provider_mpop_serv}
\title{Search the Medicare Physician & Other Practitioners
by Provider and Service API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
provider_mpop_serv(
  npi = NULL,
  last = NULL,
  first = NULL,
  year = "2020",
  clean_names = TRUE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{last}{Provider's last name}

\item{first}{Provider's first name}

\item{year}{Year in YYYY format, between 2013-2020; default is 2020}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=provider_mpop]{provider_mpop()}} allows you to access data from Medicare's
Physician & Other Practitioners by Provider and Service API.
}
\details{
The \strong{Provider and Service} dataset provides information on
services and procedures provided to Medicare (fee-for-service) Part B
beneficiaries by physicians and other healthcare professionals.

The data is based on information gathered from CMS administrative claims
data for Part B beneficiaries available from the CMS Chronic Conditions
Data Warehouse.

The spending and utilization data in the Physician and Other
Practitioners by Provider and Service Dataset are aggregated
to the following:
\enumerate{
\item the NPI for the performing provider,
\item the Healthcare Common Procedure Coding System (HCPCS) code, and
\item the place of service (either facility or non-facility).
}

There can be multiple records for a given NPI based on the number of
distinct HCPCS codes that were billed and where the services were
provided. Data have been aggregated based on the place of service
because separate fee schedules apply depending on whether the place
of service submitted on the claim is facility or non-facility.
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}{Medicare Physician & Other Practitioners: by Provider and Service API}
}
}
\note{
Data Update Frequency: Annually
}
\examples{
\dontrun{
# Search by NPI
provider_mpop_serv(npi = 1003000126)

# Search by First Name for 2019
provider_mpop_serv(first = "Enkeshafi", year = "2019")

# Multiple NPIs
npis <- c(1003026055,
          1316405939,
          1720392988,
          1518184605,
          1922056829,
          1083879860)

npis |> purrr::map_dfr(provider_mpop_serv)

# Retrieve All Provider Data, 2013-2020
dates <- as.character(seq(from = 2013, to = 2020))
purrr::map_dfr(dates, ~provider_mpop_serv(npi = 1003000126, year = .x))
}
}
