% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_mppe.R
\name{provider_mppe}
\alias{provider_mppe}
\title{Search the Medicare Fee-For-Service Public Provider Enrollment API}
\usage{
provider_mppe(
  npi = NULL,
  last = NULL,
  first = NULL,
  type_desc = NULL,
  clean_names = TRUE,
  full = FALSE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI).}

\item{last}{Provider's last name}

\item{first}{Provider's first name}

\item{type_desc}{description of Provider specialty (taxonomy).}

\item{clean_names}{Clean column names with {janitor}'s
\code{\link[=clean_names]{clean_names()}} function; default is \code{TRUE}.}

\item{full}{If true, downloads the first 1000 rows of data;
default is \code{FALSE}.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the search results.
}
\description{
Information on a point in time snapshot of enrollment
level data for providers actively enrolled in Medicare.
}
\section{The Medicare Fee-For-Service Public Provider Enrollment}{
This dataset includes information on providers who are actively approved to
bill Medicare or have completed the 855O at the time the data was pulled
from the \strong{Provider Enrollment and Chain Ownership System} (PECOS).

These files are populated from PECOS and contain basic enrollment and
provider information, reassignment of benefits information and practice
location city, state and zip.

These files are not intended to be used as real-time reporting as the data
changes from day to day and the files are updated only on a quarterly basis.
If any information on these files needs to be updated, the provider needs
to contact their respective Medicare Administrative Contractor (MAC) to
have that information updated.

This data does not include information on opt-out providers. Information
is redacted where necessary to protect Medicare provider privacy.
\subsection{Data Update Frequency}{

Quarterly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Medicare Fee-For-Service Public Provider Enrollment API}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Medicare Fee-For-Service Public Provider Enrollment Data Dictionary}
\item \href{https://pecos.cms.hhs.gov/pecos}{Provider Enrollment and Chain Ownership System Site}
}
}
}

\examples{
\dontrun{
provider_mppe(npi = 1003026055)

provider_mppe(last = "phadke", first = "radhika")

provider_mppe(type_desc = "endocrinology")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |> purrr::map_dfr(provider_mppe)

# Data frame of NPIs
npi_df <- data.frame(npi = c(1003026055,
                             1316405939,
                             1720392988,
                             1518184605,
                             1922056829,
                             1083879860))
npi_df |>
tibble::deframe() |>
purrr::map_dfr(provider_mppe)

# Download First 1,000 Rows of Dataset =====================================
provider_mppe(full = TRUE)
}
}
