% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_mpstc.R
\name{provider_mpstc}
\alias{provider_mpstc}
\title{Search the Medicare Provider and Supplier Taxonomy Crosswalk API}
\usage{
provider_mpstc(txn_code = NULL, desc = NULL, clean_names = TRUE, full = FALSE)
}
\arguments{
\item{txn_code}{unique 10-character alphanumeric code that
designates a provider’s classification and specialization.}

\item{desc}{description of taxonomy classification or specialization.}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{full}{If true, downloads the first 1000 rows of data;
default is \code{FALSE}.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the search results.
}
\description{
A list of the type of providers and suppliers with the
proper taxonomy code eligible for Medicare programs.
}
\section{The Medicare Provider and Supplier Taxonomy Crosswalk}{
This dataset lists the providers and suppliers eligible
to enroll in Medicare programs with the proper healthcare
provider taxonomy code. This data includes the Medicare
speciality codes, if available, provider/supplier type
description, taxonomy code, and the taxonomy description.
This dataset is derived from information gathered from the
\strong{National Plan and Provider Enumerator System} (NPPES) and the
\strong{Provider Enrollment, Chain and Ownership System} (PECOS).
\subsection{Data Update Frequency}{

Weekly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}
}

\section{NUCC Taxonomy Codes}{
The Healthcare Provider Taxonomy codes are a HIPAA standard code
set named in the implementation specifications for some of the
ASC X12 standard HIPAA transactions. If the Taxonomy code is
required in order to properly pay or process a claim/encounter
information transaction, it is required to be reported. Thus,
reporting of the Healthcare Provider Taxonomy Code varies from
one health plan to another. The National Uniform Claim Committee
(NUCC) maintains this list of 10-digit codes. A provider must
select a code when applying for an NPI number.

Medicare does not use the taxonomy code for pricing a provider's
services. Medicare uses the provider information based on NPI,
locality and specialty (e.g., the specialty code in Medicare’s
database) to price appropriately.
}

\examples{
\dontrun{
provider_mpstc(txn_code = "2086S0102X")

provider_mpstc(desc = "surgery")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |> purrr::map_dfr(provider_nppes) |>
            dplyr::group_split(outcome) |>
            purrr::map_dfr(provider_unpack) |>
            dplyr::distinct(taxon_code) |>
            tibble::deframe() |>
            purrr::map_dfr(provider_mpstc)

# Download First 1,000 Rows of Dataset =====================================
provider_mpstc(full = TRUE)
}
}
\references{
\url{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-supplier-taxonomy-crosswalk}
\url{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/Find-Your-Taxonomy-Code}
\url{https://taxonomy.nucc.org/}
}
