% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_mrdd.R
\name{provider_mrdd}
\alias{provider_mrdd}
\title{Search the Medicare Revalidation Due Date API}
\usage{
provider_mrdd(
  npi = NULL,
  last = NULL,
  first = NULL,
  clean_names = TRUE,
  full = FALSE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{last}{Provider's last name}

\item{first}{Provider's first name}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{full}{If \code{TRUE}, downloads the entire dataset; default is \code{FALSE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on revalidation due dates for Medicare providers.
}
\section{Medicare Revalidation Due Date API}{
The Revalidation Due Date List dataset contains
revalidation due dates for Medicare providers who are due to
revalidate in the following six months. If a provider's due
date does not fall within the ensuing six months, the due
date is marked 'TBD'. In addition the dataset also includes
subfiles with reassignment information for a given provider
as well as due date listings for clinics and group practices
and their providers.
\subsection{Data Update Frequency}{

Monthly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-due-date-list}{Medicare Revalidation Due Date API}
}
}
}

\examples{
provider_mrdd(npi = 1184699621)

provider_mrdd(last = "Byrd", first = "Eric")

\dontrun{
# Unnamed List of NPIs

npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |> purrr::map_dfr(provider_mrdd)

# Download First 1,000 Records

provider_mrdd(full = TRUE)
}
}
