% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_promdci.R
\name{provider_promdci}
\alias{provider_promdci}
\title{Search the CMS Public Reporting of Missing Digital Contact Information API}
\usage{
provider_promdci(
  npi = NULL,
  last = NULL,
  first = NULL,
  clean_names = TRUE,
  full = FALSE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{last}{Provider's last name}

\item{first}{Provider's first name}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{full}{If true, downloads the first 1000 rows of data;
default is \code{FALSE}.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the search results.
}
\description{
Information on providers that are missing digital contact
information in NPPES.
}
\section{Public Reporting of Missing Digital Contact Information}{
In the May 2020 CMS Interoperability and Patient Access
final rule, CMS finalized the policy to publicly report the names
and NPIs of those providers who do not have digital contact
information included in the NPPES system (85 FR 25584). This
data includes the NPI and provider name of providers and
clinicians without digital contact information in NPPES. This data
is gathered from the NPPES for providers who are missing digital
contact information.
\subsection{Data Update Frequency}{

Quarterly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}
}

\examples{
\dontrun{
provider_promdci(npi = 1760485387)

provider_promdci(last = "Dewey", first = "Paul")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |> purrr::map_dfr(provider_promdci)

# Returns the First 1,000 Rows in the Dataset ============================
provider_promdci(full = TRUE)
}
}
\references{
\url{https://data.cms.gov/provider-compliance/public-reporting-of-missing-digital-contact-information}
}
