% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_rcgpr.R
\name{provider_rcgpr}
\alias{provider_rcgpr}
\title{Search the Medicare Revalidation Clinic Group Practice Reassignment API}
\usage{
provider_rcgpr(
  npi = NULL,
  last = NULL,
  first = NULL,
  clean_names = TRUE,
  full = FALSE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{last}{Last name of provider who is reassigning their
benefits or is an employee}

\item{first}{First name of provider who is reassigning their
benefits or is an employee}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{full}{If true, downloads the first 1000 rows of data;
default is \code{FALSE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on clinic group practice revalidation
for Medicare enrollment.
}
\section{Medicare Revalidation Clinic Group Practice Reassignment API}{
The Revalidation Clinic Group Practice Reassignment dataset
provides information between the physician and the group
practice they reassign their billing to. It also includes
individual employer association counts and the revalidation
dates for the individual physician as well as the clinic group
practice. This dataset is based on information gathered from the
Provider Enrollment, Chain and Ownership System (PECOS).
\subsection{Data Update Frequency}{

Monthly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-clinic-group-practice-reassignment}{Medicare Revalidation Clinic Group Practice Reassignment API}
}
}
}

\examples{
\dontrun{
provider_rcgpr(npi = 1427050418)

provider_rcgpr(last = "Denney", first = "James")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1871035477,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |> purrr::map_dfr(provider_rcgpr)

# Data frame of NPIs
npi_df <- data.frame(npi = c(1003026055,
                             1871035477,
                             1720392988,
                             1518184605,
                             1922056829,
                             1083879860))
npi_df |>
tibble::deframe() |>
purrr::map_dfr(provider_rcgpr)

# Download First 1,000 Rows of Dataset =====================================
provider_rcgpr(full = TRUE)
}
}
