% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provider_type_code}
\alias{provider_type_code}
\title{Provider Type Code Reference Table}
\format{
\subsection{\code{provider_type_code}}{

A data frame with 108 rows and 3 columns:
\describe{
\item{code}{State code}
\item{abbr}{State abbreviation}
\item{name}{State full name}
...
}
}
}
\source{
https://data.cms.gov/sites/default/files/2025-10/1e42b271-dfe0-4c64-8022-0c507993667e/PPEF_Data_Dictionary.pdf
}
\usage{
provider_type_code
}
\description{
Provider Type Code Reference Table
}
\keyword{internal}
