% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_unpack.R
\name{provider_unpack}
\alias{provider_unpack}
\title{Unpack NPPES NPI Registry Search Results}
\usage{
provider_unpack(df, clean_names = TRUE)
}
\arguments{
\item{df}{\code{\link[=tibble]{tibble()}} response from \code{\link[=provider_nppes]{provider_nppes()}}}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the unpacked list-columns of a search
performed with the \code{\link[=provider_nppes]{provider_nppes()}} function.
}
\description{
Unpack NPPES NPI Registry Search Results
}
\examples{
\dontrun{
# Single NPI ==============================================================
provider_nppes(npi = 1528060837) |>
provider_unpack()

# City, state, country ====================================================
provider_nppes(city = "Atlanta",
               state_abbr = "GA",
               country_abbr = "US") |>
provider_unpack()

# First name, city, state  ================================================
provider_nppes(first = "John",
               city = "Baltimore",
               state_abbr = "MD") |>
provider_unpack()
}
}
\references{
\url{https://npiregistry.cms.hhs.gov/api-page}
}
\seealso{
\code{\link[=provider_nppes]{provider_nppes()}}
}
