% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providers}
\alias{providers}
\title{Provider Enrollment in Medicare}
\usage{
providers(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  spec_code = NULL,
  spec_desc = NULL,
  state = NULL,
  org_name = NULL,
  multi = NULL
)
}
\arguments{
\item{npi}{\verb{<int>} 10-digit Individual National Provider Identifier}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control ID}

\item{enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{first, middle, last}{\verb{<chr>} Individual provider's name}

\item{spec_code}{\verb{<chr>} Enrollment specialty code}

\item{spec_desc}{\verb{<chr>} Enrollment specialty description}

\item{state}{\verb{<chr>} Enrollment state, full or abbreviation}

\item{org_name}{\verb{<chr>} Organization name}

\item{multi}{\verb{<chr>} Provider has multiple NPIs}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit NPI \cr
   \code{pac} \tab 10-digit PAC ID \cr
   \code{enid} \tab 15-digit provider enrollment ID \cr
   \code{spec_code} \tab Enrollment primary specialty type code \cr
   \code{spec_desc} \tab Enrollment specialty type description \cr
   \code{first} \tab Individual provider's first name \cr
   \code{middle} \tab Individual provider's middle name \cr
   \code{last} \tab Individual provider's last name \cr
   \code{state} \tab Enrollment state \cr
   \code{org_name} \tab Organizational provider's name \cr
}
}
\description{
Access enrollment level data on individual and organizational providers that are actively approved to bill Medicare.
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Provider Enrollment API}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Provider Enrollment Data Dictionary}
}
}

\examples{
providers()

providers(spec_code = "14")

providers(enid = "I20040309000221")

providers(npi = 1417918293)

providers(pac = 2860305554)

providers(state = "AK")

}
