% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpp.R
\name{qpp_experience}
\alias{qpp_experience}
\title{Search the CMS Quality Payment Program Experience API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
qpp_experience(
  year,
  npi = NULL,
  state = NULL,
  specialty = NULL,
  part_type = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{year}

\item{npi}{NPI of the Provider}

\item{state}{State where the provider is enrolled}

\item{specialty}{Type of enrollment - ASC to Hospital OR IFED}

\item{part_type}{Organizational name of the enrolled provider}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on participation and performance in the Merit-based
Incentive Payment System (MIPS) and Advanced Alternative Payment Models
(APMs) tracks.
}
\details{
The Quality Payment Program (QPP) Experience dataset provides
participation and performance information in the Merit-based Incentive
Payment System (MIPS) during each performance year. They cover
eligibility and participation, performance categories, and final score
and payment adjustments. The dataset provides additional details at the
TIN/NPI level on what was published in the previous performance year.
You can sort the data by variables like clinician type, practice size,
scores, and payment adjustments.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/quality-of-care/quality-payment-program-experience}{Quality Payment Program Experience}
}
}
}
\note{
Update Frequency: \strong{Annually}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
qpp_experience(year = 2020, state = "GA")
\dontshow{\}) # examplesIf}
}
