% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-quality.R
\name{quality}
\alias{quality}
\alias{metrics}
\title{Quality Payment Program}
\usage{
quality(
  year = NULL,
  npi = NULL,
  state = NULL,
  size = NULL,
  specialty = NULL,
  years = NULL,
  patients = NULL,
  services = NULL,
  charges = NULL,
  final_score = NULL,
  adjustment = NULL,
  count = FALSE
)

metrics(year = NULL)
}
\arguments{
\item{year}{\verb{<int>} A vector of years; for \code{quality()}, 2013-2024; for
\code{metrics()} 2018-2025}

\item{npi}{\verb{<int>} National Provider Identifier. Multiple rows for the same
NPI indicate that an individual clinician has reassigned billing rights to
multiple TINs and was identified as a MIPS eligible clinician under
multiple TIN/NPI combinations.}

\item{state}{\verb{<chr>} The practice state of the TIN associated with the
clinician.}

\item{size}{\verb{<int>} Number of clinicians associated with the TIN through
Medicare Part B claims for the performance year.}

\item{specialty}{\verb{<chr>} Derived from the specialty codes in Medicare Part B
claims.}

\item{years}{\verb{<int>} Number of years since NPI's first approved enrollment
date across all enrollments in PECOS.}

\item{patients}{\verb{<int>} Number of Medicare patients who received covered
professional services during MIPS eligibility determination period.}

\item{services}{\verb{<int>} Number of covered professional services provided to
Medicare Part B patients with a service date during MIPS eligibility
determination period.}

\item{charges}{\verb{<int>} Allowed charges under the PFS on Medicare Part B
claims with a service date during MIPS eligibility determination period.}

\item{final_score}{\verb{<int>} The MIPS final score attributed to the clinician
(identified by TIN/NPI combination).}

\item{adjustment}{\verb{<dbl>} Determined by comparing the \code{final_score} to
performance thresholds and scaling to ensure budget neutrality.
\itemize{
\item The \strong{Maximum negative} adjustment is \verb{-9\%}. (\code{final_score} = 0 - 18.75)
\item A \strong{negative} adjustment is between \verb{-9\%} and\verb{0\%}. (\code{final_score} = 18.76 - 74.99)
\item A \strong{neutral} adjustment is \verb{0\%}. (\code{final_score} = 75)
\item A \strong{positive} adjustment is greater than \verb{0\%}. (\code{final_score} = 75.01 - 100)
}}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
The Quality Payment Program (QPP) Experience dataset provides participation
and performance information in the Merit-based Incentive Payment System
(MIPS) during each performance year. They cover eligibility and
participation, performance categories, and final score and payment
adjustments.
}
\details{
The dataset provides additional details at the TIN/NPI level on what was
published in the previous performance year. You can sort the data by
variables like clinician type, practice size, scores, and payment
adjustments.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
quality(year = c(2021, 2024), state = "GA", count = TRUE)

quality(npi = c(1003026055, 1316939655))

metrics()
\dontshow{\}) # examplesIf}
}
