% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-quality.R
\name{quality_metrics}
\alias{quality_metrics}
\alias{quality}
\title{Quality Payment Program}
\usage{
quality_metrics(year)

quality(
  npi = NULL,
  state = NULL,
  size = NULL,
  specialty = NULL,
  years = NULL,
  patients = NULL,
  charges = NULL,
  services = NULL,
  final_score = NULL,
  adjustment = NULL,
  count = FALSE
)
}
\arguments{
\item{year}{\verb{<int>} A vector of years from 2018 to 2025}

\item{npi}{description}

\item{state}{description}

\item{size}{description}

\item{specialty}{description}

\item{years}{description}

\item{patients}{description}

\item{charges}{description}

\item{services}{description}

\item{final_score}{description}

\item{adjustment}{description}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
The Quality Payment Program (QPP) Experience dataset provides participation
and performance information in the Merit-based Incentive Payment System
(MIPS) during each performance year. They cover eligibility and
participation, performance categories, and final score and payment
adjustments.
}
\details{
The dataset provides additional details at the TIN/NPI level on what was
published in the previous performance year. You can sort the data by
variables like clinician type, practice size, scores, and payment
adjustments.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
quality(count = TRUE)
quality(count = TRUE, state = "GA")

quality() |> str()


quality_metrics(2018:2025)
\dontshow{\}) # examplesIf}
}
