% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_pay2.R
\name{quality_pay2}
\alias{quality_pay2}
\title{Quality Payment Program REWRITE}
\usage{
quality_pay2(
  year,
  npi = NULL,
  state = NULL,
  specialty = NULL,
  type = NULL,
  tidy = FALSE,
  nest = FALSE,
  eligibility = FALSE,
  ...
)
}
\arguments{
\item{year}{\verb{<int>} // \strong{required} QPP performance year, in \code{YYYY} format.
Run \code{\link[=qpp_years]{qpp_years()}} to return a vector of the years currently available.}

\item{npi}{\verb{<int>} 10-digit Individual National Provider Identifier
assigned to the clinician when they enrolled in Medicare. Multiple rows for
the same NPI indicate multiple TIN/NPI combinations.}

\item{state}{\verb{<chr>} State or US territory code location of the TIN
associated with the clinician.}

\item{specialty}{\verb{<chr>} Specialty corresponding to the type of
service that the clinician submitted most on their Medicare Part B claims
for this TIN/NPI combination.
\itemize{
\item Nurse Practitioner
\item Internal Medicine
\item Physician Assistant
\item Family Practice
\item Emergency Medicine
\item Diagnostic Radiology
\item Anesthesiology
\item Neurology
\item Cardiology
}}

\item{type}{\verb{<chr>} Participation type; level at which the
performance data was collected, submitted or reported for the final score
attributed to the clinician; drives most of the data returned.
\itemize{
\item \code{"Group"}
\item \code{"Individual"}
\item \code{"MIPS APM"}
}}

\item{tidy}{\verb{<lgl>} // \strong{default:} \code{TRUE} Tidy output}

\item{nest}{\verb{<lgl>} // \strong{default:} \code{TRUE} Nest \code{status} & \code{measures}}

\item{eligibility}{\verb{<lgl>} // \strong{default:} \code{TRUE} Append results
from \code{\link[=quality_eligibility]{quality_eligibility()}}}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Quality Payment Program REWRITE
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
quality_pay2(year = "2020", npi = 1144544834)

quality_pay2(year = "2022", npi = 1043477615)
\dontshow{\}) # examplesIf}
}
