% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_payment.R
\name{quality_payment_}
\alias{quality_payment_}
\title{Parallelized \code{\link[=quality_payment]{quality_payment()}}}
\usage{
quality_payment_(year = qpp_years(), ...)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Year data was reported, in \code{YYYY}
format. Run \code{\link[=qpp_years]{qpp_years()}} to return a vector of the years currently available.}

\item{...}{Pass arguments to \code{\link[=quality_payment]{quality_payment()}}.}
}
\description{
Parallelized \code{\link[=quality_payment]{quality_payment()}}
}
