% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpp.R
\name{quality_payment_years}
\alias{quality_payment_years}
\title{Check the current years available for the Quality Payments API}
\usage{
quality_payment_years()
}
\value{
integer vector of years available
}
\description{
Check the current years available for the Quality Payments API
}
\examples{
quality_payment_years()
}
