% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-reassignment.R
\name{reassignments}
\alias{reassignments}
\title{Reassignment of Benefits}
\usage{
reassignments(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  last = NULL,
  state = NULL,
  specialty = NULL,
  employers = NULL,
  employees = NULL,
  org_name = NULL,
  org_pac = NULL,
  org_enid = NULL,
  org_state = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{first, last}{\verb{<chr>} Provider's name}

\item{state}{\verb{<chr>} Enrollment state abbreviation}

\item{specialty}{\verb{<chr>} Enrollment specialty}

\item{employers, employees}{\verb{<int>} Enrollment specialty}

\item{org_name}{\verb{<chr>} Legal business name}

\item{org_pac}{\verb{<chr>} PECOS Associate Control ID}

\item{org_enid}{\verb{<chr>} Medicare Enrollment ID}

\item{org_state}{\verb{<chr>} Enrollment state abbreviation}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Returns information about:
\itemize{
\item Individual providers who are reassigning benefits or are an employee of
\item Organizational/Group providers who are receiving reassignment of benefits from or are the employer of the individual provider
}

It provides information regarding the physician and the group practice they
reassign their billing to, including individual employer association counts.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
reassignments(count = TRUE)

reassignments(count = TRUE,
              employers = greater_than(50, equal = TRUE))

reassignments(org_enid = "I20070209000135")

reassignments(pac = 9830437441)

reassignments(org_pac = 3173525888)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list}{API: Medicare Revalidation Reassignment List}
}
}
