% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rehab.R
\name{rehab}
\alias{rehab}
\title{Inpatient Rehabilitation Facilities}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/7t8x-u3ir}{API: Inpatient Rehabilitation Facility - General Information}
}
}
\usage{
rehab(
  ccn = NULL,
  name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  count = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} desc}

\item{name}{\verb{<chr>} desc}

\item{city, state, zip, county}{\verb{<chr>} desc}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
This dataset shows characteristics of the inpatient
rehabilitation facilities that are shown on Inpatient Rehabilitation
Facility Compare.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
rehab(count = TRUE)
rehab(state = "GA")
\dontshow{\}) # examplesIf}
}
