% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{results_cli}
\alias{results_cli}
\title{results_cli ------------------------------------------------------------}
\usage{
results_cli(apiname, url, results)
}
\arguments{
\item{apiname}{Name of API}

\item{url}{API url landing page}
}
\description{
results_cli ------------------------------------------------------------
}
