% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-revocations.R
\name{revocations}
\alias{revocations}
\title{Revoked Providers & Suppliers}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revoked-medicare-providers-and-suppliers}{API: Revoked Medicare Providers and Suppliers}
}
}
\usage{
revocations(
  npi = NULL,
  enid = NULL,
  first = NULL,
  last = NULL,
  org_name = NULL,
  multi = NULL,
  state = NULL,
  prov_desc = NULL,
  reason = NULL,
  year_start = NULL,
  year_end = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{first, last}{\verb{<chr>} Individual provider name}

\item{org_name}{\verb{<chr>} Organization name}

\item{multi}{\verb{<lgl>} Provider has multiple NPIs}

\item{state}{\verb{<chr>} Enrollment state abbreviation}

\item{prov_desc}{\verb{<chr>} Provider type enrollment description}

\item{reason}{\verb{<chr>} Reason for revocation}

\item{year_start, year_end}{\verb{<int>} year of revocation/expiration}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\description{
Information on providers and suppliers currently revoked from the Medicare program.
}
\details{
The Revoked Medicare Providers and Suppliers dataset contains information on
providers and suppliers who are revoked and under a current re-enrollment bar.
This dataset includes provider and supplier names, NPIs, revocation
authorities pursuant to \href{https://www.ecfr.gov/current/title-42/chapter-IV/subchapter-B/part-424/subpart-P/section-424.535}{42 CFR § 424.535}, revocation effective dates, and
re-enrollment bar expiration dates. This dataset is based on information
gathered from the \emph{Provider Enrollment, Chain and Ownership System (PECOS)}.
\subsection{Reasons}{
\itemize{
\item § 424.535 \emph{Revocation of enrollment in the Medicare program.}
\enumerate{
\item Noncompliance
\itemize{
\item Business Location Not Properly Licensed
\item Enrollment Requirements
\item Not Professionally Licensed
\item Provider/Supplier Type Requirements Not Met
\itemize{
\item DME Standards Not Met
\item IDTF Standards Not Met
}
}
\item Provider Or Supplier Conduct
\itemize{
\item Exclusion
\item Other
}
\item Felonies
\item False Or Misleading Information
\item On-Site Review
\item Grounds related to provider and supplier screening requirements
\item Misuse of billing number
\item Abuse Of Billing Privileges
\itemize{
\item Pattern Or Practice
}
\item Failure To Report
\item Failure To Document
\item Initial reserve operating funds
\item Other Program Termination
- Medicaid
- Other Federal Health Care Program
\item Prescribing Authority
\item Improper Prescribing Practices
\item False Claims Act (FCA)
\item \emph{Reserved}
\item Debt Referred To Treasury
\item Revoked Under Different Name, Numerical Identifier Or Business Identity
\item Affiliation That Poses An Undue Risk
\item Billing from non-compliant location
\item Abusive ordering, certifying, referring, or prescribing of Part A or B services, items or drugs
\item Patient Harm
\item Violation Of Provider And Supplier Standards
}
}

Source: \href{https://www.ecfr.gov/current/title-42/chapter-IV/subchapter-B/part-424/subpart-P/section-424.535}{Code of Federal Regulations}

(a) Reasons for revocation.  CMS may revoke a currently enrolled provide
}
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
revocations(count = TRUE)

revocations(org_name = not_blank(), count = TRUE)

revocations(org_name = starts("B"), count = TRUE)

revocations(prov_desc = contains("CARDIO"), state = excludes(c("GA", "OH")))
\dontshow{\}) # examplesIf}
}
