% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-owners.R
\name{rhc_owner}
\alias{rhc_owner}
\title{Rural Health Clinic Owners}
\usage{
rhc_owner(
  org_enid = NULL,
  org_pac = NULL,
  org_name = NULL,
  pac = NULL,
  owner = NULL,
  dba = NULL,
  percent = NULL,
  role = NULL,
  entity = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  title = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{org_enid}{\verb{<chr>} National Provider Identifier}

\item{org_pac}{\verb{<chr>} Provider's name}

\item{org_name}{\verb{<chr>} Provider's name}

\item{pac}{\verb{<chr>} Provider's name}

\item{owner}{\verb{<chr>} Provider's name}

\item{dba}{\verb{<chr>} Provider's name}

\item{percent}{\verb{<chr>} Provider's name}

\item{role}{\verb{<chr>} Provider's name}

\item{entity}{\verb{<chr>} Provider's name}

\item{first, middle, last}{\verb{<chr>} Provider's name}

\item{title}{\verb{<chr>} Provider's name}

\item{address, city, state, zip}{\verb{<chr>} Provider's name}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Providers with pending Medicare enrollment applications.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
rhc_owner(count = TRUE)

rhc_owner(state = c("GA", "FL"))
\dontshow{\}) # examplesIf}
}
