% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{ror}
\alias{ror}
\title{Calculate lagged rate of return}
\usage{
ror(df, col, n = 1L)
}
\arguments{
\item{df}{data frame}

\item{col}{numeric column}

\item{n}{values to offset}
}
\description{
Calculate lagged rate of return
}
\keyword{internal}
