% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sfcc}
\alias{sfcc}
\title{Concatenate Vectors}
\usage{
sfcc(...)
}
\arguments{
\item{...}{Any number of vectors, coerced to \verb{<character>} vector, if necessary}
}
\value{
concatenated \verb{<character>} vector
}
\description{
Concatenate Vectors
}
\examples{
sfcc(LETTERS, "A")

}
