% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-spending.R
\name{spending}
\alias{spending}
\alias{spending2}
\title{Medicare Spending Per Beneficiary - Hospital}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/rrqw-56er}{API: Medicare Spending Per Beneficiary - Hospital}
}

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}
}
\usage{
spending(
  ccn = NULL,
  name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  score = NULL,
  count = FALSE
)

spending2(ccn = NULL, name = NULL, state = NULL, claim = NULL, count = FALSE)
}
\arguments{
\item{ccn}{\verb{<chr>} desc}

\item{name}{\verb{<chr>} desc}

\item{city, state, zip, county}{\verb{<chr>} desc}

\item{state}{\verb{<chr>} desc}

\item{score}{\verb{<dbl>} desc}

\item{count}{\verb{<lgl>} Return the total row count}

\item{claim}{\verb{<chr>} desc}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.

A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
The Medicare Spending Per Beneficiary (MSPB) Measure shows
whether Medicare spends more, less, or about the same for an episode of
care (episode) at a specific hospital compared to all hospitals nationally.
An MSPB episode includes Medicare Part A and Part B payments for services
provided by hospitals and other healthcare providers the 3 days prior to,
during, and 30 days following a patient's inpatient stay. This measure
evaluates hospitals' costs compared to the costs of the national median (or
midpoint) hospital. This measure takes into account important factors like
patient age and health status (risk adjustment) and geographic payment
differences (payment-standardization).

Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
spending(count = TRUE)
spending(state = "GA")
\dontshow{\}) # examplesIf}
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
spending2(count = TRUE)
spending2(name = starts("SGMC"), state = "GA")
\dontshow{\}) # examplesIf}
}
