% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_stats}
\alias{summary_stats}
\title{Returns a tidytable of summary stats}
\usage{
summary_stats(
  df,
  condition = NULL,
  group_vars = NULL,
  summary_vars = NULL,
  arr = NULL
)
}
\arguments{
\item{df}{data frame}

\item{condition}{filter condition, i.e. \code{patient == "new"}}

\item{group_vars}{variables to group by, i.e. \code{c(specialty, state, hcpcs, cost)}}

\item{summary_vars}{variables to summarise, i.e. \code{c(min, max, mode, range)}}

\item{arr}{column to arrange data by, i.e. \code{cost}}
}
\value{
A \code{tidytable} containing the summary stats
}
\description{
Returns a tidy table of summary stats
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
download_datasets(specialty = "vascular surgery") |>
tidytable::slice_head(n = 10) |>
summary_stats(condition = patient == "new",
              group_vars = c(specialty, state, hcpcs, cost),
              summary_vars = c(min, max, mode, range),
              arr = cost)
\dontshow{\}) # examplesIf}
}
