% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-supplier.R
\name{supplier}
\alias{supplier}
\title{Medical Equipment Suppliers}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}
}
\usage{
supplier(
  id = NULL,
  par = NULL,
  org_dba = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  specialty = NULL,
  prov_type = NULL,
  supplies = NULL,
  count = FALSE
)
}
\arguments{
\item{id}{\verb{<chr>} provider identifier}

\item{par}{\verb{<lgl>} description}

\item{org_dba}{\verb{<chr>} desc}

\item{org_name}{\verb{<chr>} desc}

\item{city, state, zip}{\verb{<chr>} desc}

\item{specialty}{description}

\item{prov_type}{description}

\item{supplies}{\verb{<int>} desc}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
supplier(count = TRUE)
supplier(state = "GA")
\dontshow{\}) # examplesIf}
}
