% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy_codes}
\alias{taxonomy_codes}
\title{Provider Taxonomies}
\source{
\url{https://www.nucc.org/index.php/code-sets-mainmenu-41/provider-taxonomy-mainmenu-40/csv-mainmenu-57}
}
\usage{
taxonomy_codes(shape = c("wide", "long"))
}
\arguments{
\item{shape}{shape of the data frame returned, 'wide' or 'long'}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:
}
\description{
\code{taxonomy_codes()} returns a \code{\link[=tibble]{tibble()}} of the current Health Care Provider
Taxonomy code set

\strong{Update Frequency}: \emph{Biannually}
}
\section{Taxonomy Codes}{

The Health Care Provider Taxonomy code set is a collection of unique
alphanumeric codes, ten characters in length. They contain no embedded logic.

The code set Levels are organized to allow for drilling down to the
provider's most specific level of specialization, with three distinct "Levels":
}

\section{\strong{Level I} \emph{Provider Grouping}}{

A major grouping of service(s) or occupation(s) of health care providers.

\emph{Examples:} Allopathic & Osteopathic Physicians, Dental Providers, Hospitals
}

\section{\strong{Level II} \emph{Classification}}{

A more specific service or occupation related to the Provider Grouping.

For example, the classification for Allopathic & Osteopathic Physicians is
based upon the General Specialty Certificates as issued by the appropriate
national boards.

The following boards will, however, have their general certificates appear as
Level III Areas of specialization strictly due to display limitations of the
code set for Boards that have multiple general certificates:

\emph{Medical Genetics, Preventive Medicine, Psychiatry & Neurology, Radiology,}
\emph{Surgery, Otolaryngology, Pathology}
}

\section{\strong{Level III} \emph{Area of Specialization}}{

A more specialized area of the Classification in which a provider chooses to
practice or make services available.

For example, the area of specialization for Allopathic & Osteopathic
Physicians is based upon the Subspecialty Certificates as issued by the
appropriate national boards.
}

\section{\strong{Categories} \emph{(Level 0)}}{

The code set includes three specialty categories:

\strong{Group (of Individuals)}:
\enumerate{
\item Multi-Specialty
\item Single Specialty
}

\strong{Individuals}:
\enumerate{
\item Allopathic & Osteopathic Physicians
\item Behavioral Health and Social Service Providers
\item Chiropractic Providers
\item Dental Providers
\item Dietary and Nutritional Service Providers
\item Emergency Medical Service Providers
\item Eye and Vision Service Providers
\item Nursing Service Providers
\item Nursing Service Related Providers
\item Other Service Providers
\item Pharmacy Service Providers
\item Physician Assistants and Advanced Practice Nursing Providers
\item Podiatric Medicine and Surgery Service Providers
\item Respiratory, Developmental, Rehabilitative and Restorative Service Providers
\item Speech, Language and Hearing Service Providers
\item Student, Health Care
\item Technologists, Technicians, and Other Technical Service Providers
}

\strong{Non-Individuals}:
\enumerate{
\item Agencies
\item Ambulatory Health Care Facilities
\item Hospital Units
\item Hospitals
\item Laboratories
\item Managed Care Organizations
\item Nursing and Custodial Care Facilities
\item Other Service Providers
\item Residential Treatment Facilities
\item Respite Care Facilities
\item Suppliers
\item Transportation Services
}
}

\section{Display Name}{

The display name is a combination of the code name and the Level in
which the code is nested, which more precisely identifies the code.

For example, Addiction Medicine is a physician subspecialty in Anesthesiology,
Family Medicine, Internal Medicine, Preventive Medicine, and Psychiatry &
Neurology.

"Addiction Medicine" does not identify the specialty of
the physician, but the display name of "Addiction Medicine (Internal
Medicine) Physician" clearly does.

In another example, "Radiology" could be confused with several
taxonomies, but "Radiology Chiropractor" more accurately specifies the provider.
}

\section{Description}{

\itemize{
\item \code{taxonomy_code}: Provider Taxonomy Code
\item \code{taxonomy_category}: Indicates whether Taxonomy is Individual or Non-Individual, i.e., a group taxonomy
\item \code{taxonomy_grouping}: Level I, Provider Grouping
\item \code{taxonomy_classification}: Level II, Classification
\item \code{taxonomy_specialization}: Level III, Area of Specialization
\item \code{taxonomy_display_name}: Consumer-friendly taxonomy name, made of the code name and the Level in which the code is nested.
\item \code{taxonomy_definition}: Definition of Taxonomy
\item \code{version}: Three digit version of the code set. The first two digits indicate the year and the third digit indicates either the first release of the year ("0") or the second release of the year ("1").
\item \code{release_date}: Date the version of the code set was released
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
taxonomy_codes()
\dontshow{\}) # examplesIf}
}
