% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_crosswalk.R
\name{taxonomy_crosswalk}
\alias{taxonomy_crosswalk}
\title{Taxonomy to Medicare Specialty Crosswalk}
\usage{
taxonomy_crosswalk(
  taxonomy_code = NULL,
  taxonomy_description = NULL,
  medicare_specialty_code = NULL,
  medicare_specialty_type = NULL,
  tidy = TRUE
)
}
\arguments{
\item{taxonomy_code}{Taxonomy code}

\item{taxonomy_description}{Taxonomy code description}

\item{medicare_specialty_code}{Medicare specialty code}

\item{medicare_specialty_type}{Medicare specialty code description}

\item{tidy}{Tidy output; default is \code{TRUE}}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{taxonomy_crosswalk()} allows you to search the types of providers and
suppliers eligible for Medicare programs by taxonomy code or Medicare
specialty type code.
\subsection{Taxonomy Codes}{

The \strong{Healthcare Provider Taxonomy Code Set} is a hierarchical HIPAA
standard code set designed to categorize the type, classification, and
specialization of health care providers. It consists of two sections:
\enumerate{
\item Individuals and Groups of Individuals
\item Non-Individuals
}

When applying for an NPI, a provider must report the taxonomy that most
closely describes their type/classification/specialization. In some
situations, a provider may need to report more than one taxonomy but must
indicate one of them as the primary. The codes selected may not be the same
as categorizations used by Medicare for enrollment.
}
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-supplier-taxonomy-crosswalk}{Provider and Supplier Taxonomy Crosswalk}
\item \href{https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-methodology}{Taxonomy Crosswalk Methodology}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/Find-Your-Taxonomy-Code}{Find Your Taxonomy Code}
}
}

\section{Update Frequency}{
 \strong{Weekly}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
taxonomy_crosswalk(medicare_specialty_code = "B4[14]")
taxonomy_crosswalk(medicare_specialty_type = "Rehabilitation Agency")
taxonomy_crosswalk(taxonomy_code = "2086S0102X")
\dontshow{\}) # examplesIf}
}
