% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyup}
\alias{tidyup}
\title{Tidy a Data Frame}
\usage{
tidyup(
  df,
  dtype = NULL,
  dt = "date",
  yn = NULL,
  int = NULL,
  dbl = NULL,
  chr = NULL,
  up = NULL,
  cred = NULL,
  zip = NULL,
  lgl = NULL,
  cma = NULL
)
}
\arguments{
\item{df}{data frame}

\item{dtype}{\code{mdy} or \code{ymd}}

\item{dt}{convert to date, default is 'date'}

\item{yn}{convert to logical}

\item{int}{convert to integer}

\item{dbl}{convert to double}

\item{chr}{convert to character}

\item{up}{convert to upper case}

\item{cred}{remove periods}

\item{zip}{normalize zip code}

\item{lgl}{convert to logical}

\item{cma}{remove commas}
}
\value{
tidy data frame
}
\description{
Tidy a Data Frame
}
\keyword{internal}
