% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-revocations.R
\name{transparency}
\alias{transparency}
\title{Hospital Transparency Enforcement}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-price-transparency-enforcement-activities-and-outcomes}{API: Hospital Price Transparency Enforcement Activities and Outcomes}
}
}
\usage{
transparency(
  name = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  action = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{name}{\verb{<chr>} Hospital name}

\item{address}{\verb{<chr>} Hospital address}

\item{city}{\verb{<chr>} Hospital city}

\item{state}{\verb{<chr>} Hospital state}

\item{action}{\verb{<enum>} Action taken by CMS following a Compliance Review (see Details)
\itemize{
\item \code{"met"} = Met Requirements
\item \code{"admin"} = Administrative Closure
\item \code{"warn"} = Warning Notice
\item \code{"cap"} = CAP Request
\item \code{"closure"} = Closure Notice
\item \code{"cmp"} = CMP Notice
\item \code{"appeal"} = Appealed
}}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Information on enforcement actions as a result of CMS' assessment of a
hospital's compliance with the Hospital Price Transparency regulations.
\subsection{Overview}{

The \strong{Hospital Price Transparency Enforcement Activities and Outcomes} dataset
contains information related to enforcement actions taken by CMS following a
compliance review of a hospital's obligation to establish, update and make
public a list of the hospital's standard charges for items and services
provided by the hospital, in accordance with regulation (45 CFR 180).

This data set includes the name of each hospital or hospital location, the
hospital or hospital location address, the outcome or action following a
CMS compliance review and the date of the outcome or action taken.
\subsection{Actions:}{
\itemize{
\item \strong{Met Requirements:} The hospital was reviewed by CMS and no deficiencies
were cited. This category includes hospitals that were found to be in
compliance upon CMS' first review.
\item \strong{Administrative Closure:} Refers to the termination of compliance
review activities due to reasons such as, but not limited to, a hospital
has ceased operations, closed its physical location, is not subject to the
HPT requirements, or has already been deemed to meet the requirements.
\item \strong{Warning Notice:} The hospital was reviewed by CMS and deficiencies
were cited in a warning notice issued to the hospital.
\item \strong{CAP Request:} The hospital was reviewed by CMS and deficiencies were
cited in a request for a corrective action plan (CAP) issued to the hospital.
\item \strong{Closure Notice:} CMS determined that the hospital corrected previously
identified deficiencies and issued a closure notice to the hospital.
\item \strong{CMP Notice:} CMS issued a civil monetary penalty (CMP) to the hospital
if the hospital failed to respond to CMS' request to submit a corrective
action plan or comply with the requirements of a corrective action plan.
CMP \href{https://www.cms.gov/priorities/key-initiatives/hospital-price-transparency/enforcement-actions}{notices are publicized}.
\item \strong{Appealed:} The hospital filed an appeal of CMS' decision to issue a
civil monetary penalty.
}
}

}
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
transparency(count = TRUE)
transparency(count = TRUE, action = "met")
transparency(state = "GA", city = "Valdosta")
\dontshow{\}) # examplesIf}
}
