% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilization.R
\name{utilization_}
\alias{utilization_}
\title{Parallelized \code{\link[=utilization]{utilization()}}}
\usage{
utilization_(year = util_years(), ...)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Year data was reported, in \code{YYYY}
format. Run \code{\link[=util_years]{util_years()}} to return a vector of the years currently available.}

\item{...}{Pass arguments to \code{\link[=utilization]{utilization()}}.}
}
\description{
Parallelized \code{\link[=utilization]{utilization()}}
}
