% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-veteran.R
\name{veteran}
\alias{veteran}
\title{Veterans Health Administration Hospitals}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/uyx4-5s7f}{API: Veterans Health Administration Provider Level Data}
}
}
\usage{
veteran(
  ccn = NULL,
  name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  rating = NULL,
  count = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} desc}

\item{name}{\verb{<chr>} desc}

\item{city, state, zip, county}{\verb{<chr>} desc}

\item{rating}{\verb{<int>} desc}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of all VHA hospitals. The list includes addresses and phone numbers.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
veteran(count = TRUE)
veteran(state = "GA")
\dontshow{\}) # examplesIf}
}
