% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{years_df}
\alias{years_df}
\title{Calculate number of years since today's date}
\usage{
years_df(df, date_col)
}
\arguments{
\item{df}{data frame}

\item{date_col}{date column}
}
\value{
number of years since today's date
}
\description{
Calculate number of years since today's date
}
\examples{
dplyr::tibble(date = lubridate::today() - 366) |>
years_df(date_col = date)
}
