% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{years_floor}
\alias{years_floor}
\title{Years between two dates, rounded down to nearest whole number}
\usage{
years_floor(from, to)
}
\arguments{
\item{from}{\verb{<date>} Start date}

\item{to}{\verb{<date>} End date}
}
\value{
\verb{<int>} vector; number of years between \code{from} and \code{to}, rounded
down to nearest whole number
}
\description{
Years between two dates, rounded down to nearest whole number
}
\examples{
years_floor(
   as.Date("2020-01-01"),
   as.Date("2020-01-01") + 2057)

}
