% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpp.R
\name{years_qualpay}
\alias{years_qualpay}
\title{Check the current years available for the Quality Payments API}
\usage{
years_qualpay()
}
\description{
Check the current years available for the Quality Payments API
}
