% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{years_vec}
\alias{years_vec}
\title{Calculate number of years since today's date}
\usage{
years_vec(date)
}
\arguments{
\item{date}{date column}
}
\value{
number of years since today's date
}
\description{
Calculate number of years since today's date
}
\examples{
dplyr::tibble(date = lubridate::today() - 366) |>
dplyr::mutate(years_passed = years_vec(date = date))
}
