% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_counties}
\alias{add_counties}
\title{Add county name, FIPs, and geometry to data frame with zip codes}
\usage{
add_counties(df, zcol)
}
\arguments{
\item{df}{data frame}

\item{zcol}{column containing zip codes}
}
\description{
Add county name, FIPs, and geometry to data frame with zip codes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppes(npi = 1962026229) |>
add_counties(zip)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
