% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_counties}
\alias{add_counties}
\title{Add county name, FIPs, and geometry to data frame with zip codes}
\usage{
add_counties(
  df,
  statecol,
  zipcol,
  add_fips = FALSE,
  add_geo = FALSE,
  as_sf = FALSE
)
}
\arguments{
\item{df}{data frame}

\item{statecol}{bare column name column containing state abbreviations}

\item{zipcol}{bare column name containing zip codes}

\item{add_fips}{add county FIPS code column, default is \code{FALSE}}

\item{add_geo}{add county geometry column, default is \code{FALSE}}

\item{as_sf}{convert tibble to an \code{{sf}} object, default is \code{FALSE}}
}
\description{
Add county name, FIPs, and geometry to data frame with zip codes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppes(npi = 1720098791) |>
      add_counties(statecol = state,
                   zipcol   = zip,
                   add_fips = TRUE,
                   add_geo  = TRUE,
                   as_sf    = TRUE)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
