% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_counties.R
\name{add_counties}
\alias{add_counties}
\title{Add county name, FIPs, and geometry to data frame with zip codes}
\usage{
add_counties(
  df,
  statecol,
  zipcol,
  add_fips = FALSE,
  add_geo = FALSE,
  as_sf = FALSE
)
}
\arguments{
\item{df}{data frame}

\item{statecol}{bare column name column containing state abbreviations}

\item{zipcol}{bare column name containing zip codes}

\item{add_fips}{add county FIPS code column, default is \code{FALSE}}

\item{add_geo}{add county geometry column, default is \code{FALSE}}

\item{as_sf}{convert tibble to an \code{{sf}} object, default is \code{FALSE}}
}
\description{
Add county name, FIPs, and geometry to data frame with zip codes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Example data frame containing state abbreviation and zip code
ex <- dplyr::tibble(state = "GA",
                    zip = "31605")
ex

# Adds county name and latitude/longitude
ex |> add_counties(state, zip)

# Adds county FIPS
ex |> add_counties(state,
                   zip,
                   add_fips = TRUE)

# Adds county `geometry` column,
# based on county FIPS column
ex |> add_counties(state,
                   zip,
                   add_fips = TRUE,
                   add_geo  = TRUE)

# Converts data frame to an `sf` object
ex |> add_counties(state,
                   zip,
                   add_fips = TRUE,
                   add_geo  = TRUE,
                   as_sf    = TRUE)
\dontshow{\}) # examplesIf}
}
