% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addl_phone_numbers.R
\name{addl_phone_numbers}
\alias{addl_phone_numbers}
\title{Search the CMS Physician - Additional Phone Numbers API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
addl_phone_numbers(
  npi = NULL,
  ind_pac_id = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  org_pac_id = NULL,
  offset = 0,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{npi}{Unique clinician ID assigned by NPPES}

\item{ind_pac_id}{Unique individual clinician ID assigned by PECOS}

\item{first_name}{Individual clinician first name}

\item{middle_name}{Individual clinician middle name}

\item{last_name}{Individual clinician last name}

\item{city}{Group or individual's city}

\item{state}{Group or individual's state}

\item{zip}{Group or individual's ZIP code (9 digits when available)}

\item{org_pac_id}{Unique group ID assigned by PECOS to the group}

\item{offset}{offset; API pagination}

\item{clean_names}{Clean column names with {janitor}'s \code{clean_names()}
function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Dataset of additional phone numbers when clinicians have more
than one phone number at a single practice address.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/phys-phon}{Physician - Additional Phone Numbers}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
addl_phone_numbers(npi = 1407263999)
addl_phone_numbers(ind_pac_id = "0042100190")
addl_phone_numbers(org_pac_id = 6608028899)
\dontrun{
addl_phone_numbers(city = "Atlanta")
addl_phone_numbers(zip = 303421606)
}
}
