% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-affiliation.R
\name{affiliations}
\alias{affiliations}
\title{Provider-Facility Affiliations}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}
}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  suffix = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} Individual National Provider Identifier}

\item{pac}{\verb{<chr>} Individual PECOS Associate Control ID}

\item{first, middle, last, suffix}{\verb{<chr>} Individual provider's name}

\item{facility_type}{\verb{<enum>} facility type:
\itemize{
\item \code{esrd} = Dialysis facility
\item \code{hha} = Home health agency
\item \code{hospice} = Hospice
\item \code{hospital} = Hospital
\item \code{irf} = Inpatient rehabilitation facility
\item \code{ltch} = Long-term care hospital
\item \code{nurse} = Nursing home
\item \code{snf} = Skilled nursing facility
}}

\item{facility_ccn}{\verb{<chr>} CCN of \code{facility_type} column's
facility \strong{or} of a \strong{unit} within the hospital where the individual
provider provides services.}

\item{parent_ccn}{\verb{<int>} CCN of the \strong{primary} hospital containing the
unit where the individual provider provides services.}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
affiliations(middle = "", count = TRUE)

affiliations(facility_ccn = "33Z302")

affiliations(parent_ccn = 331302)

affiliations(facility_ccn = 331302)
\dontshow{\}) # examplesIf}
}
