% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider Facility Affiliations}
\usage{
affiliations(
  id = list(npi = NULL, pac = NULL, ccn = NULL),
  name = list(first = NULL, middle = NULL, last = NULL, suff = NULL),
  facility_type = NULL,
  ...
)
}
\arguments{
\item{id}{\verb{<list>} List of parameters that uniquely identify a provider, any
of the following: \code{npi}, \code{pac}, \code{ccn}.}

\item{name}{\verb{<list>} Individual provider's name(s), any of the following:
\code{first}, \code{middle}, \code{last}, \code{suffix}}

\item{facility_type}{\verb{<chr>} type of facility:
\itemize{
\item \code{"hp"} Hospital
\item \code{"lt"} Long-Term Care Hospital
\item \code{"nh"} Nursing Home
\item \code{"irf"} Inpatient Rehabilitation Facility
\item \code{"hha"} Home Health Agency
\item \code{"snf"} Skilled Nursing Facility
\item \code{"hs"} Hospice
\item \code{"df"} Dialysis Facility
}}

\item{...}{Empty dots}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=affiliations]{affiliations()}} allows the user access to data concerning
providers' facility affiliations
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
affiliations(id = list(ccn = 670055))
\dontshow{\}) # examplesIf}
}
