% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  suffix = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL
)
}
\arguments{
\item{npi}{\verb{<int>} Unique clinician ID assigned by NPPES}

\item{pac}{\verb{<chr>} Unique individual clinician ID assigned by PECOS}

\item{first}{\verb{<chr>} Individual clinician first name}

\item{middle}{\verb{<chr>} Individual clinician middle name}

\item{last}{\verb{<chr>} Individual clinician last name}

\item{suffix}{\verb{<chr>} Individual clinician suffix}

\item{facility_type}{\verb{<chr>} type of facility:
\itemize{
\item \code{"hp"} Hospital
\item \code{"lt"} Long-Term Care Hospital
\item \code{"nh"} Nursing Home
\item \code{"irf"} Inpatient Rehabilitation Facility
\item \code{"hha"} Home Health Agency
\item \code{"snf"} Skilled Nursing Facility
\item \code{"hs"} Hospice
\item \code{"df"} Dialysis Facility
}}

\item{facility_ccn}{\verb{<chr>} Medicare CCN of facility type or unit within hospital where an individual clinician provides service}

\item{parent_ccn}{\verb{<chr>} Medicare CCN of the primary hospital where individual clinician provides service, should the clinician provide services in a unit within the hospital}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=affiliations]{affiliations()}} allows the user access to data concerning
providers' facility affiliations
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}
}

\examples{
affiliations(facility_ccn = "33Z302")

affiliations(parent_ccn = 331302)

affiliations(pac = 7810891009)

affiliations(npi = 1003026055)

affiliations(first = "KIM")

}
