% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  type = NULL,
  ccn_unit = NULL,
  ccn_parent = NULL,
  tidy = TRUE,
  ...
)
}
\arguments{
\item{npi}{\verb{<chr>} Unique 10-digit National Provider Identifier number issued
by CMS to US healthcare providers through NPPES.}

\item{pac}{\verb{<int>} Unique 10-digit Provider Associate-level Control ID (PAC),
assigned to each individual or organization in PECOS. The PAC ID links all
entity-level information (e.g., tax identification numbers and
organizational names) and may be associated with multiple enrollment IDs if
the individual or organization enrolled multiple times under different
circumstances.}

\item{first, middle, last}{\verb{<chr>} Individual provider's name(s)}

\item{type}{\verb{<chr>} Type of facility, one of the following:
\itemize{
\item \code{"hp"} = Hospital
\item \code{"lt"} = Long-term care hospital
\item \code{"nh"} = Nursing home
\item \code{"irf"} = Inpatient rehabilitation facility
\item \code{"hha"} = Home health agency
\item \code{"snf"} = Skilled nursing facility
\item \code{"hs"} = Hospice
\item \code{"df"} = Dialysis facility
}}

\item{ccn_unit}{\verb{<chr>} 6-digit CCN of facility or unit within hospital
where an individual provider provides service.}

\item{ccn_parent}{\verb{<int>} 6-digit CCN of a sub-unit's
primary hospital, should the provider provide services in said unit.}

\item{tidy}{\verb{<lgl>} Tidy output; \strong{default} is \code{TRUE}}

\item{...}{Empty dots}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=affiliations]{affiliations()}} allows the user access to data concerning
providers' facility affiliations
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}

\emph{Update Frequency:} \strong{Monthly}
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
affiliations(parent_ccn = 670055)
\dontshow{\}) # examplesIf}
}
