% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider-Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  suffix = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL
)
}
\arguments{
\item{npi}{\verb{<int>} Individual clinician ID, assigned by \emph{NPPES}}

\item{pac}{\verb{<chr>} Individual clinician ID, assigned by \emph{PECOS}}

\item{first, middle, last, suffix}{\verb{<chr>} Individual clinician's name}

\item{facility_type}{\verb{<chr>} Type of facility:\tabular{ll}{
   \strong{abbr} \tab \strong{full} \cr
   \code{hp} \tab Hospital \cr
   \code{lt} \tab Long-term care hospital \cr
   \code{nh} \tab Nursing home \cr
   \code{irf} \tab Inpatient rehabilitation facility \cr
   \code{hha} \tab Home health agency \cr
   \code{snf} \tab Skilled nursing facility \cr
   \code{hs} \tab Hospice \cr
   \code{df} \tab Dialysis facility \cr
}}

\item{facility_ccn}{\verb{<chr>} Medicare CCN of \code{facility_type} column's
facility \emph{or} of a \emph{unit} within the hospital where the individual
clinician provides services.}

\item{parent_ccn}{\verb{<chr>} Medicare CCN of \emph{primary} hospital where the
individual clinician provides services in a unit within said hospital.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=affiliations]{affiliations()}} allows the user access to data concerning
providers' facility affiliations
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}
}

\examples{
affiliations()

affiliations(facility_ccn = "33Z302")

affiliations(parent_ccn = 331302)

affiliations(pac = 7810891009)

affiliations(npi = 1003026055)

affiliations(first = "KIM")

}
