% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider-Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  suffix = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL
)
}
\arguments{
\item{npi}{\verb{<int>} Individual National Provider Identifier}

\item{pac}{\verb{<chr>} Individual PECOS Associate Control ID}

\item{first, middle, last, suffix}{\verb{<chr>} Individual provider's name}

\item{facility_type}{\verb{<chr>} facility type abbreviation:\tabular{ll}{
   \strong{ABBR} \tab \strong{FULL} \cr
   \code{hp} \tab Hospital \cr
   \code{lt} \tab Long-term care hospital \cr
   \code{nh} \tab Nursing home \cr
   \code{irf} \tab Inpatient rehabilitation facility \cr
   \code{hha} \tab Home health agency \cr
   \code{snf} \tab Skilled nursing facility \cr
   \code{hs} \tab Hospice \cr
   \code{df} \tab Dialysis facility \cr
}}

\item{facility_ccn}{\verb{<chr>} CCN of \code{facility_type} column's
facility \emph{or} of a \emph{unit} within the hospital where the individual
provider provides services.}

\item{parent_ccn}{\verb{<chr>} CCN of the \emph{primary} hospital containing the
unit where the individual provider provides services.}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
affiliations()

affiliations(pac = 7810891009)

affiliations(npi = 1003026055)

affiliations(first = "KIM")

affiliations(facility_ccn = c("33Z302", 331302))

}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}
}
