% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac_id_ind = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  offset = 0L,
  tidy = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{pac_id_ind}{< \emph{integer} > 10-digit individual provider associate level variable
Links all entity-level information and may be associated with multiple
enrollment IDs if the individual or organization enrolled multiple times}

\item{first, middle, last}{< \emph{character} > Individual provider's first, middle,
or last name}

\item{facility_type}{< \emph{character} > Category of facility:
\itemize{
\item \code{"Hospital"} or \code{"hp"}
\item \code{"Long-term care hospital"} or \code{"ltch"}
\item \code{"Nursing home"} or \code{"nh"}
\item \code{"Inpatient rehabilitation facility"} or \code{"irf"}
\item \code{"Home health agency"} or \code{"hha"}
\item \code{"Skilled nursing facility"} or \code{"snf"}
\item \code{"Hospice"} or \code{"hs"}
\item \code{"Dialysis facility"} or \code{"df"}
}}

\item{facility_ccn}{< \emph{character} > 6-digit CMS Certification Number of
facility or unit within hospital where an individual provider provides service}

\item{parent_ccn}{< \emph{integer} > 6-digit CMS Certification Number of a
sub-unit's primary hospital, should the provider provide services in said unit}

\item{offset}{< \emph{integer} > offset; API pagination}

\item{tidy}{< \emph{boolean} > Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{affiliations()} allows you to access information concerning providers'
facility affiliations

Links:
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}

\emph{Update Frequency:} \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
affiliations(parent_ccn = 670055)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=clinicians]{clinicians()}}, \code{\link[=providers]{providers()}}, \code{\link[=hospitals]{hospitals()}}
}
