% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac_id = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  offset = 0L,
  tidy = TRUE
)
}
\arguments{
\item{npi}{Unique clinician ID assigned by NPPES}

\item{pac_id}{Unique individual clinician ID assigned by PECOS}

\item{first_name}{Individual clinician first name}

\item{middle_name}{Individual clinician middle name}

\item{last_name}{Individual clinician last name}

\item{facility_type}{Facilities can fall into the following type categories:
\itemize{
\item Hospital
\item Long-term Care Hospital (LTCH)
\item Nursing Home
\item Inpatient Rehabilitation Facility (IRF)
\item Home Health Agency (HHA)
\item Skilled Nursing Facility (SNF)
\item Hospice
\item Dialysis Facility
}}

\item{facility_ccn}{alphanumeric; Medicare CCN (CMS Certification Number) of
facility type or unit within hospital where an individual clinician
provides service. The CCN replaced the terms \emph{Medicare Provider Number},
\emph{Medicare Identification Number} and \emph{OSCAR Number} and is used to verify
Medicare/Medicaid certification for survey and certification,
assessment-related activities and communications.}

\item{parent_ccn}{numeric; CCN of the primary hospital where an
individual clinician provides service, should the clinician provide
services in a unit within the hospital.}

\item{offset}{offset; API pagination}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{affiliations()} allows you to access information concerning providers' facility affiliations
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}

\emph{Update Frequency:} \strong{Monthly}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
affiliations(parent_ccn = 670055)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=clinicians]{clinicians()}}, \code{\link[=providers]{providers()}}, \code{\link[=hospitals]{hospitals()}}
}
