% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\name{affiliations}
\alias{affiliations}
\title{Provider Facility Affiliations}
\usage{
affiliations(
  npi = NULL,
  pac = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  facility_type = NULL,
  facility_ccn = NULL,
  parent_ccn = NULL,
  offset = 0L,
  tidy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit Individual National Provider Identifier}

\item{pac}{< \emph{integer} > 10-digit Individual PECOS Associate Control ID}

\item{first, middle, last}{< \emph{character} > Individual Provider's name}

\item{facility_type}{< \emph{character} >
\itemize{
\item \code{"Hospital"} or \code{"hp"}
\item \code{"Long-term care hospital"} or \code{"ltch"}
\item \code{"Nursing home"} or \code{"nh"}
\item \code{"Inpatient rehabilitation facility"} or \code{"irf"}
\item \code{"Home health agency"} or \code{"hha"}
\item \code{"Skilled nursing facility"} or \code{"snf"}
\item \code{"Hospice"} or \code{"hs"}
\item \code{"Dialysis facility"} or \code{"df"}
}}

\item{facility_ccn}{< \emph{character} > 6-digit CMS Certification Number of
facility or unit within hospital where an individual provider provides service}

\item{parent_ccn}{< \emph{integer} > 6-digit CMS Certification Number of a sub-unit's
primary hospital, should the provider provide services in said unit}

\item{offset}{< \emph{integer} > // \strong{default:} \code{0L} API pagination}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit NPI \cr
   \code{pac} \tab 10-digit individual PAC ID \cr
   \code{first} \tab Individual provider's first name \cr
   \code{middle} \tab Individual provider's middle name \cr
   \code{last} \tab Individual provider's last name \cr
   \code{suffix} \tab Individual provider's suffix \cr
   \code{facility_type} \tab Category of facility \cr
   \code{facility_ccn} \tab Facility's 6-digit CCN \cr
   \code{parent_ccn} \tab Primary facility's 6-digit CCN \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=affiliations]{affiliations()}} allows the user access to data concerning providers'
facility affiliations
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{Physician Facility Affiliations}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/SurveyCertificationGenInfo/Downloads/Survey-and-Cert-Letter-16-09.pdf}{Certification Number (CCN) State Codes}
}

\emph{Update Frequency:} \strong{Monthly}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
affiliations(parent_ccn = 670055)
\dontshow{\}) # examplesIf}
}
