% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beneficiaries.R
\name{beneficiaries}
\alias{beneficiaries}
\title{Beneficiary Enrollment in Medicare}
\usage{
beneficiaries(
  year = NULL,
  period = NULL,
  level = NULL,
  state = NULL,
  county = NULL,
  fips = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{< \emph{integer} > Calendar year of Medicare enrollment; current years
can be checked with:
\itemize{
\item \code{bene_years("year")}: Years available for all 12 months
\item \code{bene_years("month")}: Years available for individual months
}}

\item{period}{< \emph{character} > Time frame of Medicare enrollment; options
are \code{"Year"}, \code{Month}, or any individual month name}

\item{level}{< \emph{character} > Geographic level of data; options are
\code{"National"}, \code{"State"}, or \code{"County"}}

\item{state}{< \emph{character} > Full state name or abbreviation of
beneficiary residence}

\item{county}{< \emph{character} > County of beneficiary residence}

\item{fips}{< \emph{character} > FIPS code of beneficiary residence}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{year} \tab Enrollment calendar year \cr
   \code{period} \tab Enrollment time frame \cr
   \code{level} \tab Geographic level of data \cr
   \code{state} \tab Beneficiary residence state \cr
   \code{state_name} \tab Beneficiary residence state name \cr
   \code{county} \tab Beneficiary residence county \cr
   \code{fips} \tab Beneficiary residence FIPS code \cr
   \code{bene_total} \tab Total Medicare beneficiaries \cr
   \code{bene_orig} \tab Original Medicare beneficiaries \cr
   \code{bene_ma_oth} \tab Medicare Advantage beneficiaries \cr
   \code{bene_total_aged} \tab Aged beneficiaries \cr
   \code{bene_aged_esrd} \tab Aged beneficiaries with ESRD \cr
   \code{bene_aged_no_esrd} \tab Aged beneficiaries without ESRD \cr
   \code{bene_total_dsb} \tab Disabled beneficiaries \cr
   \code{bene_dsb_esrd} \tab Disabled with ESRD and ESRD-only beneficiaries \cr
   \code{bene_dsb_no_esrd} \tab Disabled beneficiaries without ESRD \cr
   \code{bene_total_ab} \tab Total Medicare Part A and B beneficiaries \cr
   \code{bene_ab_orig} \tab Original Medicare Part A and B beneficiaries \cr
   \code{bene_ab_ma_oth} \tab Medicare Advantage Part A and B beneficiaries \cr
   \code{bene_total_rx} \tab Total Medicare Part D beneficiaries \cr
   \code{bene_rx_pdp} \tab Medicare Prescription Drug Plan beneficiaries \cr
   \code{bene_rx_mapd} \tab Medicare Advantage Prescription Drug Plan beneficiaries \cr
   \code{bene_rx_lis_elig} \tab Part D Low Income Subsidy Eligible (Full) Beneficiaries \cr
   \code{bene_rx_lis_full} \tab Part D Low Income Subsidy Full Beneficiaries \cr
   \code{bene_rx_lis_part} \tab Part D Low Income Subsidy Partial Beneficiaries \cr
   \code{bene_rx_lis_no} \tab Part D No Low Income Subsidy Beneficiaries \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

\code{\link[=beneficiaries]{beneficiaries()}} allows the user access current data on enrolled Medicare
beneficiaries.
}
\section{Medicare Monthly Enrollment}{

Current monthly information on the number of Medicare beneficiaries with
hospital/medical coverage and prescription drug coverage, available for
several geographic areas including national, state and county.

The hospital/medical coverage data can be broken down further by health care
delivery (Original Medicare versus Medicare Advantage and Other Health Plans)
and the prescription drug coverage data can be examined by those enrolled in
stand-alone Prescription Drug Plans and those enrolled in Medicare Advantage
Prescription Drug plans.

The dataset includes enrollee counts on a \emph{rolling 12 month basis}.
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/summary-statistics-on-beneficiary-enrollment/medicare-and-medicaid-reports/medicare-monthly-enrollment}{Medicare Monthly Enrollment}
}

\emph{Update Frequency:} \strong{Monthly}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
beneficiaries(year   = 2022,
              period = "Year",
              level  = "County",
              county = "Autauga")

beneficiaries(year   = 2022,
              period = "July",
              state  = "Georgia")

beneficiaries(level = "State",
              fips  = "10")
\dontshow{\}) # examplesIf}
}
