% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beneficiary_enrollment.R
\name{beneficiary_enrollment}
\alias{beneficiary_enrollment}
\title{Search the Medicare Monthly Enrollment API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
beneficiary_enrollment(
  year = 2021,
  month = "Year",
  geo_level = c("National", "State", "County"),
  state_abb = NULL,
  state = NULL,
  county = NULL,
  fips = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{year}{Calendar year of Medicare enrollment; current options are
\code{2017 - 2022}}

\item{month}{Timeframe of Medicare enrollment; options are \code{Year} or any
month within the 12-month timespan of the month in the dataset's version
name (listed \href{https://data.cms.gov/summary-statistics-on-beneficiary-enrollment/medicare-and-medicaid-reports/medicare-monthly-enrollment/api-docs}{here})}

\item{geo_level}{Geographic level of data; options are \code{National}, \code{State},
and \code{County}}

\item{state_abb}{Two-letter state abbreviation of beneficiary residence}

\item{state}{Full state name of beneficiary residence}

\item{county}{County of beneficiary residence}

\item{fips}{FIPS code of beneficiary residence}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Current information on the number of Medicare
beneficiaries with hospital/medical coverage and prescription
drug coverage, available for several geographical areas.
}
\details{
The Medicare Monthly Enrollment data provides current
monthly information on the number of Medicare beneficiaries
with hospital/medical coverage and prescription drug coverage,
available for several geographic areas including national,
state/territory, and county. The hospital/medical coverage
data can be broken down further by health care delivery
(Original Medicare versus Medicare Advantage and Other
Health Plans) and the prescription drug coverage data can
be examined by those enrolled in stand-alone Prescription
Drug Plans and those enrolled in Medicare Advantage
Prescription Drug plans. The dataset includes enrollee
counts on a \emph{rolling 12 month basis} and also provides
information on yearly trends. The dataset is based on
information gathered from CMS administrative enrollment
data for beneficiaries enrolled in the Medicare program
available from the CMS Chronic Conditions Data Warehouse.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/summary-statistics-on-beneficiary-enrollment/medicare-and-medicaid-reports/medicare-monthly-enrollment}{Medicare Monthly Enrollment}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
beneficiary_enrollment(year = 2018,
                       month = "Year",
                       geo_level = "County",
                       state_abb = "AL",
                       county = "Autauga")

beneficiary_enrollment(year = 2021,
                       geo_level = "County",
                       fips = "01001")

beneficiary_enrollment(year = 2022,
                       month = "July",
                       geo_level = "State",
                       state = "Georgia")

beneficiary_enrollment(year = 2017,
                       geo_level = "State",
                       fips = "10")
}
