% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_service.R
\name{by_service}
\alias{by_service}
\title{Search the Medicare Physician & Other Practitioners API by \strong{Provider and Service}}
\usage{
by_service(
  year,
  npi = NULL,
  last_name = NULL,
  first_name = NULL,
  organization_name = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  hcpcs_code = NULL,
  drug = NULL,
  pos = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{integer (\emph{required}); Year in \code{YYYY} format. Run helper function
\code{by_service_years()} to return a vector of the years
currently available.}

\item{npi}{National Provider Identifier for the rendering provider on the claim.}

\item{last_name}{Individual provider's last name.}

\item{first_name}{Individual provider's first name.}

\item{organization_name}{Organization name.}

\item{credential}{Individual provider's credentials.}

\item{gender}{Individual provider's gender.}

\item{entype}{Provider entity type.}

\item{city}{City where provider is located.}

\item{state}{State where provider is located.}

\item{zip}{Provider’s zip code.}

\item{fips}{Provider's state FIPS code.}

\item{ruca}{Rural-Urban Commuting Area Code (RUCA); a Census tract-based
classification scheme that utilizes the standard Bureau of Census
Urbanized Area and Urban Cluster definitions in combination with work
commuting information to characterize all of the nation's Census tracts
regarding their rural and urban status and relationships. The Referring
Provider ZIP code was cross walked to the United States Department of
Agriculture (USDA) 2010 Rural-Urban Commuting Area Codes.}

\item{country}{Country where provider is located.}

\item{specialty}{Provider specialty code reported on the largest number of
claims submitted.}

\item{par}{Identifies whether the provider participates in Medicare
and/or accepts assignment of Medicare allowed amounts. The value will
be \code{Y} for any provider that had at least one claim identifying the
provider as participating in Medicare or accepting assignment of
Medicare allowed amounts within HCPCS code and place of service. A
non-participating provider may elect to accept Medicare allowed amounts
for some services and not accept Medicare allowed amounts for other
services.}

\item{hcpcs_code}{HCPCS code used to identify the specific medical service
furnished by the provider. HCPCS codes include two levels. Level I codes
are the Current Procedural Terminology (CPT) codes that are maintained
by the American Medical Association and Level II codes are created by
CMS to identify products, supplies and services not covered by the CPT
codes (such as ambulance services).}

\item{drug}{Identifies whether the HCPCS code for the specific service
furnished by the provider is a HCPCS listed on the Medicare Part B Drug
Average Sales Price (ASP) File. Please visit the ASP drug pricing page
for additional information.}

\item{pos}{Identifies whether the place of service submitted on the claims
is a facility (\code{F}) or non-facility (\code{O}). Non-facility is generally an
office setting; however other entities are included in non-facility.}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on services and procedures provided to
Original Medicare (fee-for-service) Part B (Medical Insurance)
beneficiaries by physicians and other healthcare professionals;
aggregated by provider and service.
}
\details{
The \strong{Provider and Service} dataset provides information on
services and procedures provided to Medicare (fee-for-service) Part B
beneficiaries by physicians and other healthcare professionals.

The data is based on information gathered from CMS administrative claims
data for Part B beneficiaries available from the CMS Chronic Conditions
Data Warehouse.

The spending and utilization data in the Physician and Other
Practitioners by Provider and Service Dataset are aggregated
to the following:
\enumerate{
\item the NPI for the performing provider,
\item the Healthcare Common Procedure Coding System (HCPCS) code, and
\item the place of service (either facility or non-facility).
}

There can be multiple records for a given NPI based on the number of
distinct HCPCS codes that were billed and where the services were
provided. Data have been aggregated based on the place of service
because separate fee schedules apply depending on whether the place
of service submitted on the claim is facility or non-facility.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}{Medicare Physician & Other Practitioners: by Provider and Service API}
}

\emph{Update Frequency:} \strong{Annually}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
physician_by_service(npi = 1003000126)
physician_by_service(year = 2019, last_name = "Enkeshafi")
\dontshow{\}) # examplesIf}
}
