% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculations}
\alias{calculations}
\alias{change}
\alias{years_df}
\title{Calculations for Providers' stats}
\usage{
change(df, cols, digits = 3)

years_df(df, date_col)
}
\arguments{
\item{df}{data frame}

\item{cols}{numeric columns}

\item{digits}{Number of digits to round to, default is 3}

\item{date_col}{date column}
}
\value{
A \code{tibble}

number of years since today's date
}
\description{
Functions for calculating lagged changes and elapsed duration
between dates.
}
\examples{
df <- dplyr::tibble(year = 2015:2020,
                    charges = sample(1000:2000, size = 6),
                    payment = sample(1000:2000, size = 6))

# Calculate the lagged absolute/relative change
# and the cumulative sum of both:
change(df, c(charges, payment))

# Calculate the number of years between dates:
dplyr::tibble(date = lubridate::today() - 366) |>
years_df(date_col = date)
}
\keyword{internal}
