% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_multiple.R
\name{cc_multiple}
\alias{cc_multiple}
\title{Search the Medicare Multiple Chronic Conditions API}
\format{
\subsection{In addition to the searchable columns:}{

\describe{
\item{mcc}{To classify MCC for each Medicare beneficiary, the 21 chronic conditions are counted and grouped into four categories (0-1, 2-3, 4-5 and 6 or more).}
\item{prevalence}{Prevalence estimates are calculated by taking the beneficiaries within the MCC category divided by the total number of beneficiaries in the fee-for-service population, expressed as a percentage.}
\item{tot_std_pymt_percap}{Medicare standardized spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita). Standardized payments are presented to allow for comparisons across geographic areas in health care use among beneficiaries.}
\item{tot_pymt_percap}{Medicare spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita).}
\item{hosp_readmsn_rate}{Hospital readmissions are expressed as a percentage of all admissions. A 30-day readmission is defined as an admission to an acute care hospital for any cause within 30 days of discharge from an acute care hospital. Except when the patient died during the stay, each inpatient stay is classified as an index admission, a readmission, or both.}
\item{er_visits_per_1k}{Emergency department visits are presented as the number of visits per 1,000 beneficiaries. ED visits include visits where the beneficiary was released from the outpatient setting and where the beneficiary was admitted to an inpatient setting.}
}
}
}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
cc_multiple(
  year,
  level = NULL,
  sublevel = NULL,
  fips = NULL,
  age_group = NULL,
  demographic = NULL,
  subdemo = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{integer, YYYY, calendar year of Medicare enrollment. 2007-2018 data is
currently available.}

\item{level}{Geographic level of data; options are \code{National}, \code{State},
and \code{County}}

\item{sublevel}{The state and/or county where the Medicare beneficiary
resides. The values include the 50 United States, District of Columbia,
Puerto Rico or U.S. Virgin Islands. Data aggregated at the National level
are identified by "National'.}

\item{fips}{FIPS state and/or county code where the Medicare beneficiary
resides. This column will be blank for data aggregated at the National
level or for Puerto Rico and Virgin Islands.}

\item{age_group}{Identifies the age level of the population that the data has
been aggregated. A value of \code{All} indicates the data in the row represents
all Fee-for-Service Medicare Beneficiaries. A value of \verb{<65} or \verb{65+}
indicates that the data is aggregated by the age of the Medicare
Beneficiaries at the end of the calendar year.}

\item{demographic}{Identifies the demographic level of the population that the
data has been aggregated. A value of \code{All} indicates the data in the row
represents all Fee-for-Service Medicare beneficiaries. A value of \code{Sex}
indicates that the data has been aggregated by the Medicare beneficiary's
sex. A value of \code{Race} indicates that the data has been aggregated by the
Medicare beneficiary's race. A value of \verb{Dual Status} indicates that the
data has been aggregated by the Medicare beneficiary's dual eligibility
status.}

\item{subdemo}{For \code{Sex}, a beneficiary’s sex is classified
as Male or Female and is identified using information from the CMS
enrollment database. For \code{Race}, the race/ethnicity classifications are:
Non-Hispanic White, Black or African American, Asian/Pacific Islander,
Hispanic, and American Indian/Alaska Native. For \verb{Dual Status},
beneficiaries can be classified as 'Medicare and Medicaid' or 'Medicare Only'.
Beneficiaries enrolled in both Medicare and Medicaid are known as “dual
eligibles.” Medicare beneficiaries are classified as dual eligibles if in
any month in the given calendar year they were receiving full or partial
Medicaid benefits.}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on prevalence, use and spending by count of select
chronic conditions among Original Medicare (or fee-for-service)
beneficiaries.
}
\details{
The Multiple Chronic Conditions dataset provides information on the
number of chronic conditions among Original Medicare beneficiaries. The
dataset contains prevalence, use and spending organized by geography and
the count of chronic conditions from the set of select 21 chronic
conditions. The count of conditions is grouped into four categories (0-1,
2-3, 4-5 and 6 or more).
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/medicare-chronic-conditions/multiple-chronic-conditions}{Medicare Multiple Chronic Conditions}
}
}
}
\note{
Update Frequency: \strong{Annually}
}
\examples{
\dontrun{
cc_multiple(year = 2018, level  = "State", sublevel = "California")
cc_multiple(year = 2007, level = "National", demographic = "Race")
}
}
