% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_year}
\alias{change_year}
\title{Calculate lagged values by column}
\usage{
change_year(df, col, by)
}
\arguments{
\item{df}{data frame}

\item{col}{column of numeric values to calculate lag}

\item{by}{column to calculate lag by}
}
\value{
A \code{tibble}
}
\description{
Calculate lagged values by column
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
by_provider_years() |>
map(\(x) by_provider(year = x, npi = 1023076643)) |>
list_rbind() |>
select(year, tot_payment) |>
change_year(tot_payment, year)
\dontshow{\}) # examplesIf}
}
