% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{chg}
\alias{chg}
\title{Calculate lagged change}
\usage{
chg(x, n = 1L, fill_na = 0L)
}
\arguments{
\item{x}{numeric vector}

\item{n}{values to offset}

\item{fill_na}{fill value for any NAs; default is 0}
}
\value{
numeric vector
}
\description{
Calculate lagged change
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dplyr::tibble(year = 2015:2020,
              pay = 1000:1005) |>
dplyr::mutate(change = chg(pay))
\dontshow{\}) # examplesIf}
}
\keyword{internal}
