% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{chg}
\alias{chg}
\title{Calculate absolute change}
\usage{
chg(x, n = 1L)
}
\arguments{
\item{x}{numeric vector}

\item{n}{values to offset}
}
\value{
numeric vector
}
\description{
Calculate absolute change
}
\examples{
dplyr::tibble(year = 2015:2020,
              pay = 1000:1005) |>
dplyr::mutate(change = chg(pay, year)
}
