% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronic_conditions.R
\name{chronic_conditions}
\alias{chronic_conditions}
\alias{cc_multiple}
\alias{cc_specific}
\title{Chronic Conditions}
\format{
\describe{
\item{prevalence}{Prevalence estimates are calculated by taking the beneficiaries within the MCC category divided by the total number of beneficiaries in the fee-for-service population, expressed as a percentage.}
\item{tot_std_pymt_percap}{Medicare standardized spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita). Standardized payments are presented to allow for comparisons across geographic areas in health care use among beneficiaries.}
\item{tot_pymt_percap}{Medicare spending includes total Medicare payments for all covered services in Parts A and B and is presented per beneficiary (i.e. per capita).}
\item{hosp_readmsn_rate}{Hospital readmissions are expressed as a percentage of all admissions. A 30-day readmission is defined as an admission to an acute care hospital for any cause within 30 days of discharge from an acute care hospital. Except when the patient died during the stay, each inpatient stay is classified as an index admission, a readmission, or both.}
\item{er_visits_per_1k}{Emergency department visits are presented as the number of visits per 1,000 beneficiaries. ED visits include visits where the beneficiary was released from the outpatient setting and where the beneficiary was admitted to an inpatient setting.}
}
}
\usage{
cc_multiple(
  year,
  level = NULL,
  sublevel = NULL,
  fips = NULL,
  age = NULL,
  demo = NULL,
  subdemo = NULL,
  mcc = NULL,
  tidy = TRUE
)

cc_specific(
  year,
  condition = NULL,
  sublevel = NULL,
  level = NULL,
  fips = NULL,
  age = NULL,
  demo = NULL,
  subdemo = NULL,
  tidy = TRUE
)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Calendar year of Medicare
enrollment, in \code{YYYY} format. Run \code{\link[=cc_years]{cc_years()}} to return a vector of
currently available years.}

\item{level}{< \emph{character} > Geographic level of aggregation
\itemize{
\item \code{"National"}
\item \code{"State"}
\item \code{"County"}
}}

\item{sublevel}{< \emph{character} > Beneficiary's state or county}

\item{fips}{< \emph{character} > Beneficiary's state or county FIPS code}

\item{age}{< \emph{character} > Age group level of aggregation
\itemize{
\item \code{"All"}
\item \code{"<65"}
\item \code{"65+"}
}}

\item{demo, subdemo}{< \emph{character} > Demographic level of aggregation
\itemize{
\item \strong{\code{"All"}}
\item \strong{\code{"Sex"}}: \code{"Male"}, \code{"Female"}
\item \strong{\code{"Race"}}: \code{"non-Hispanic White"}, \code{"non-Hispanic Black"}, \code{"Asian Pacific Islander"}, \code{"Hispanic"}, \code{"Native American"}
\item \strong{\code{"Dual Status"}}: \code{"Medicare and Medicaid"}, \code{"Medicare Only"}
}}

\item{mcc}{< \emph{character} > Number of chronic conditions
\itemize{
\item \code{"0-1"}
\item \code{"2-3"}
\item \code{"4-5"}
\item \code{"6+"}
}}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{condition}{< \emph{character} > Chronic condition for which the prevalence
and utilization is compiled}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions allow the user access to data concerning chronic conditions
among Original Medicare (or fee-for-service) beneficiaries.
}
\section{\code{\link[=cc_specific]{cc_specific()}}}{


The \strong{Specific Chronic Conditions} dataset provides information on
prevalence, use and spending organized by geography and 21 distinct chronic
conditions among Original Medicare beneficiaries:
\enumerate{
\item Alcohol Abuse
\item Alzheimer’s Disease/Dementia
\item Arthritis
\item Asthma
\item Atrial Fibrillation
\item Autism Spectrum Disorders
\item Cancer
\item Chronic Kidney Disease
\item COPD
\item Depression
\item Diabetes
\item Drug Abuse/Substance Abuse
\item Heart Failure
\item Hepatitis (Chronic Viral B & C)
\item HIV/AIDS
\item Hyperlipidemia
\item Hypertension
\item Ischemic Heart Disease
\item Osteoporosis
\item Schizophrenia and Other Psychotic Disorders
\item Stroke
}
}

\section{\code{\link[=cc_multiple]{cc_multiple()}}}{


The \strong{Multiple Chronic Conditions} dataset provides information on
prevalence, use and spending organized by geography and the count of chronic
conditions from the set of select 21 chronic conditions. The count of
conditions is grouped into four categories:
\enumerate{
\item \strong{0-1}
\item \strong{2-3}
\item \strong{4-5}
\item \strong{6 or More}
}
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/medicare-chronic-conditions/multiple-chronic-conditions}{Medicare Multiple Chronic Conditions}
\item \href{https://data.cms.gov/medicare-chronic-conditions/specific-chronic-conditions}{Medicare Specific Chronic Conditions}
}

\emph{Update Frequency:} \strong{Annually}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cc_specific(year = 2018, level = "State", sublevel = "CA")
cc_multiple(year = 2018, level = "State", sublevel = "California")

cc_specific(year = 2007, level = "National", demo = "Race")
cc_multiple(year = 2007, level = "National", demo = "Race")
\dontshow{\}) # examplesIf}
}
