% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-clia.R
\name{clia}
\alias{clia}
\title{Clinical Laboratories}
\usage{
clia(
  name = NULL,
  ccn = NULL,
  certification = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  status = NULL,
  active = FALSE,
  count = FALSE
)
}
\arguments{
\item{name}{\verb{<chr>} Provider or clinical laboratory's name}

\item{ccn}{\verb{<chr>} 10-character CLIA number}

\item{certification}{\verb{<chr>} CLIA certificate type:
\itemize{
\item \code{"waiver"}
\item \code{"ppm"}
\item \code{"registration"}
\item \code{"compliance"}
\item \code{"accreditation"}
}}

\item{city}{\verb{<chr>} City}

\item{state}{\verb{<chr>} State}

\item{zip}{\verb{<chr>} Zip code}

\item{status}{\verb{<chr>} \code{A} (Compliant) or \code{B} (Non-Compliant)}

\item{active}{\verb{<lgl>} Return only active providers}

\item{count}{\verb{<lgl>} Return the dataset's total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Clinical laboratories including demographics and the type of testing
services the facility provides.
}
\section{CLIA}{


CMS regulates all laboratory testing (except research) performed on humans
in the U.S. through the \strong{Clinical Laboratory Improvement Amendments}.
In total, \strong{CLIA} covers approximately 320,000 laboratory entities.

Although all clinical laboratories must be properly certified to receive
Medicare or Medicaid payments, CLIA has no direct Medicare or Medicaid
program responsibilities.
}

\section{Certification}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories}{API: Provider of Services File - Clinical Laboratories}
\item \href{https://www.cms.gov/medicare/quality/clinical-laboratory-improvement-amendments}{CMS: CLIA}
\item \href{https://www.cdc.gov/clia/php/about/index.html}{CDC: CLIA}
\item \href{https://qcor.cms.gov/main.jsp}{CMS: QCOR}
\item \href{https://www.cdc.gov/labs/clia-certificates/index.html}{CLIA: Certificates}
\item \href{https://www.cms.gov/files/document/clia-certificate-fee-schedule-updated-06/7/2024.pdf}{CLIA: Certificate Fee Schedule}
\item \href{https://www.cms.gov/files/document/clia-cert-quick-start-guide.pdf}{CLIA: Certification Guide}
}
}

\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
clia(count = TRUE)
clia()
clia(ccn = provider:::cdc_labs$ccn)
clia(
  certification = "accreditation",
  city = "Valdosta",
  state = "GA"
)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories}{API: Provider of Services File - Clinical Laboratories}
\item \href{https://www.cms.gov/medicare/quality/clinical-laboratory-improvement-amendments}{CMS: CLIA}
\item \href{https://www.cdc.gov/clia/php/about/index.html}{CDC: CLIA}
\item \href{https://qcor.cms.gov/main.jsp}{CMS: QCOR}
\item \href{https://www.cdc.gov/labs/clia-certificates/index.html}{CLIA: Certificates}
\item \href{https://www.cms.gov/files/document/clia-certificate-fee-schedule-updated-06/7/2024.pdf}{CLIA: Certificate Fee Schedule}
\item \href{https://www.cms.gov/files/document/clia-cert-quick-start-guide.pdf}{CLIA: Certification Guide}
}
}
