% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicians.R
\name{clinicians}
\alias{clinicians}
\title{Clinician Demographics}
\usage{
clinicians(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  suffix = NULL,
  gender = NULL,
  credential = NULL,
  specialty = NULL,
  school = NULL,
  year = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  facility_name = NULL,
  facility_pac = NULL
)
}
\arguments{
\item{npi}{\verb{<int>} Individual National Provider Identifier}

\item{pac}{\verb{<int>} Individual PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Individual Medicare Enrollment ID}

\item{first, middle, last, suffix}{\verb{<chr>} Individual provider's name}

\item{gender}{\verb{<chr>} Individual provider's gender; \code{"F"} (Female), \code{"M"} (Male), or \code{"U"} (Unknown)}

\item{credential}{\verb{<chr>} Individual provider's credential, i.e. \code{"MD"}, \code{"OD"}}

\item{specialty}{\verb{<chr>} Individual provider’s primary medical specialty}

\item{school}{\verb{<chr>} Individual provider’s medical school}

\item{year}{\verb{<int>} Individual provider’s graduation year}

\item{city, state, zip}{\verb{<chr>} Facility's city, state, zip}

\item{facility_name}{\verb{<chr>} Facility associated with Provider}

\item{facility_pac}{\verb{<int>} Facility's PECOS Associate Control ID}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Demographics of doctors and clinicians listed in the Provider Data Catalog (PDC)

The Doctors and Clinicians national downloadable file is organized such that
each line is unique at the clinician/enrollment record/group/address level.
Clinicians with multiple Medicare enrollment records and/or single enrollments
linking to multiple practice locations are listed on multiple lines.
}
\examples{
clinicians()

clinicians(enid = "I20081002000549")

clinicians(first = "ETAN")

clinicians(city = starts_with("At"), state = "GA", year = 2020)

}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/mj5m-pzi6}{National Downloadable File}
\item \href{https://data.cms.gov/provider-data/sites/default/files/data_dictionaries/physician/DOC_Data_Dictionary.pdf}{Provider Data Catalog (PDC) Data Dictionary}
}
}
