% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicians.R
\name{clinicians}
\alias{clinicians}
\title{Clinicians Enrolled in Medicare}
\usage{
clinicians(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  suffix = NULL,
  gender = NULL,
  credential = NULL,
  grad_school = NULL,
  grad_year = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  specialty = NULL,
  facility_name = NULL,
  facility_pac = NULL
)
}
\arguments{
\item{npi}{\verb{<int>} Provider's NPI}

\item{pac}{\verb{<int>} Provider's PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Provider's Medicare Enrollment ID}

\item{first, middle, last, suffix}{\verb{<chr>} Provider's name}

\item{gender}{\verb{<chr>} Provider's gender; \code{"F"} (Female), \code{"M"} (Male), or \code{"U"} (Unknown)}

\item{credential}{\verb{<chr>} Provider’s credential, i.e. "MD"}

\item{grad_school}{\verb{<chr>} Medical school provider graduated from}

\item{grad_year}{\verb{<int>} Provider’s graduation year; YYYY}

\item{city, state, zip}{\verb{<chr>} Provider's city, state, zip}

\item{specialty}{\verb{<chr>} Provider’s primary medical specialty}

\item{facility_name}{\verb{<chr>} Facility associated with Provider}

\item{facility_pac}{\verb{<int>} Facility's PECOS Associate Control ID}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit individual NPI \cr
   \code{pac} \tab 10-digit individual PAC ID \cr
   \code{enid} \tab 15-digit individual enrollment ID \cr
   \code{first} \tab Provider's first name \cr
   \code{middle} \tab Provider's middle name \cr
   \code{last} \tab Provider's last name \cr
   \code{suffix} \tab Provider's name suffix \cr
   \code{gender} \tab Provider's gender \cr
   \code{credential} \tab Provider's credential \cr
   \code{school} \tab Provider's medical school \cr
   \code{grad_year} \tab Provider's graduation year \cr
   \code{specialty} \tab Provider's primary specialty \cr
   \code{specialty_sec} \tab Provider's secondary specialty \cr
   \code{facility_name} \tab Facility associated with provider \cr
   \code{pac_org} \tab Facility's 10-digit PAC ID \cr
   \code{members} \tab Number of providers associated with facility's PAC ID \cr
   \code{address} \tab Provider's street address \cr
   \code{city} \tab Provider's city \cr
   \code{state} \tab Provider's state \cr
   \code{zip} \tab Provider's zip code \cr
   \code{phone} \tab Provider's phone number \cr
   \code{telehealth} \tab Indicates if provider offers telehealth services \cr
   \code{assign_ind} \tab Indicates if provider accepts Medicare assignment \cr
   \code{assign_org} \tab Indicates if facility accepts Medicare assignment \cr
}
}
\description{
Access information about providers enrolled in Medicare,
including the medical school that they attended and the year they graduated
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/mj5m-pzi6}{National Downloadable File}
\item \href{https://data.cms.gov/provider-data/sites/default/files/data_dictionaries/physician/DOC_Data_Dictionary.pdf}{Provider Data Catalog (PDC) Data Dictionary}
}
}

\examples{
clinicians()

clinicians(enid = "I20081002000549")

clinicians(first = "ETAN")

}
