% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicians.R
\name{clinicians}
\alias{clinicians}
\title{Doctors and Clinicians National Downloadable File}
\usage{
clinicians(
  npi = NULL,
  pac_id = NULL,
  enroll_id = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  gender = NULL,
  school = NULL,
  grad_year = NULL,
  specialty = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  offset = 0L,
  tidy = TRUE
)
}
\arguments{
\item{npi}{Unique clinician ID assigned by NPPES}

\item{pac_id}{Unique individual clinician ID assigned by PECOS}

\item{enroll_id}{Unique ID for the clinician enrollment that is the source
for the data in the observation}

\item{first_name}{Individual clinician first name}

\item{middle_name}{Individual clinician middle name}

\item{last_name}{Individual clinician last name}

\item{gender}{Individual clinician gender}

\item{school}{Individual clinician’s medical school}

\item{grad_year}{Individual clinician’s medical school graduation year}

\item{specialty}{Primary medical specialty reported by the individual
clinician in the selected enrollment}

\item{city}{Group or individual's city}

\item{state}{Group or individual's state}

\item{zip}{Group or individual's ZIP code}

\item{offset}{offset; API pagination}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{clinicians()} . The Doctors and Clinicians national downloadable file is organized
such that each line is unique at the clinician/enrollment
record/group/address level. Clinicians with multiple Medicare enrollment
records and/or single enrollments linking to multiple practice locations
are listed on multiple lines.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/mj5m-pzi6}{National Downloadable File}
\item \href{https://data.cms.gov/provider-data/sites/default/files/data_dictionaries/physician/DOC_Data_Dictionary.pdf}{Provider Data Catalog (PDC) Data Dictionary}
}

\emph{Update Frequency:} \strong{Monthly}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clinicians(enroll_id = "I20081002000549")
clinicians(school = "NEW YORK UNIVERSITY SCHOOL OF MEDICINE")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=hospitals]{hospitals()}}, \code{\link[=providers]{providers()}}, \code{\link[=affiliations]{affiliations()}}
}
