% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicians.R
\name{clinicians}
\alias{clinicians}
\title{Clinicians Enrolled in Medicare}
\usage{
clinicians(
  npi = NULL,
  pac_id_ind = NULL,
  enroll_id_ind = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  gender = NULL,
  school = NULL,
  grad_year = NULL,
  specialty = NULL,
  facility_name = NULL,
  pac_id_org = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  offset = 0L,
  tidy = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{pac_id_ind}{< \emph{integer} > 10-digit individual provider associate-level
control identifier}

\item{enroll_id_ind}{< \emph{character} > 15-digit individual provider Medicare
enrollment identifier; begins with capital "I"}

\item{first, middle, last}{< \emph{character} > Individual provider's first,
middle, or last name}

\item{gender}{< \emph{character} > Individual provider's gender; \code{"F"} (Female)
or \code{"M"} (Male)}

\item{school}{< \emph{character} > Individual provider’s medical school}

\item{grad_year}{< \emph{integer} > Individual provider’s graduation year}

\item{specialty}{< \emph{character} > Individual provider’s primary medical
specialty reported in the selected enrollment}

\item{facility_name}{< \emph{character} > Name of facility associated with the
individual provider}

\item{pac_id_org}{< \emph{integer} > 10-digit organizational/group provider
associate-level control identifier}

\item{city}{< \emph{character} > Provider's city}

\item{state}{< \emph{character} > Provider's state}

\item{zip}{< \emph{character} > Provider's ZIP code}

\item{offset}{< \emph{integer} > offset; API pagination}

\item{tidy}{< \emph{boolean} > Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit individual NPI \cr
   \code{pac_id_ind} \tab 10-digit individual PAC ID \cr
   \code{enroll_id_ind} \tab 15-digit individual enrollment ID \cr
   \code{first} \tab Provider's first name \cr
   \code{middle} \tab Provider's middle name \cr
   \code{last} \tab Provider's last name \cr
   \code{suffix} \tab Provider's name suffix \cr
   \code{gender} \tab Provider's gender \cr
   \code{credential} \tab Provider's credential \cr
   \code{school} \tab Provider's medical school \cr
   \code{grad_year} \tab Provider's graduation year \cr
   \code{specialty} \tab Provider's primary specialty \cr
   \code{specialty_sec} \tab Provider's secondary specialty \cr
   \code{facility_name} \tab Facility associated with provider \cr
   \code{pac_id_org} \tab Facility's 10-digit PAC ID \cr
   \code{members} \tab Number of providers associated with facility's PAC ID \cr
   \code{address} \tab Provider's street address \cr
   \code{city} \tab Provider's city \cr
   \code{state} \tab Provider's state \cr
   \code{zip} \tab Provider's zip code \cr
   \code{phone} \tab Provider's phone number \cr
   \code{telehealth} \tab Indicates if provider offers telehealth services \cr
   \code{assign_ind} \tab Indicates if provider accepts Medicare assignment \cr
   \code{assign_group} \tab Indicates if facility accepts Medicare assignment \cr
}
}
\description{
\code{clinicians()} allows you to access information about providers enrolled in
Medicare, including the medical school that they attended and the year they graduated

Links:
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/mj5m-pzi6}{National Downloadable File}
\item \href{https://data.cms.gov/provider-data/sites/default/files/data_dictionaries/physician/DOC_Data_Dictionary.pdf}{Provider Data Catalog (PDC) Data Dictionary}
}

\emph{Update Frequency:} \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clinicians(enroll_id = "I20081002000549")
clinicians(school = "NEW YORK UNIVERSITY SCHOOL OF MEDICINE")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=hospitals]{hospitals()}}, \code{\link[=providers]{providers()}}, \code{\link[=affiliations]{affiliations()}}
}
