% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicians.R
\name{clinicians}
\alias{clinicians}
\title{Clinicians Enrolled in Medicare}
\usage{
clinicians(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  gender = NULL,
  credential = NULL,
  school = NULL,
  grad_year = NULL,
  specialty = NULL,
  facility_name = NULL,
  pac_org = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  offset = 0L,
  tidy = TRUE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit Individual National Provider Identifier}

\item{pac}{< \emph{integer} > 10-digit Individual PECOS Associate Control ID}

\item{enid}{< \emph{character} > 15-digit Individual Medicare Enrollment ID}

\item{first, middle, last}{< \emph{character} > Individual provider's name}

\item{gender}{< \emph{character} > Individual provider's gender; \code{"F"} (Female)
or \code{"M"} (Male)}

\item{credential}{Individual provider’s credential}

\item{school}{< \emph{character} > Individual provider’s medical school}

\item{grad_year}{< \emph{integer} > Individual provider’s graduation year}

\item{specialty}{< \emph{character} > Individual provider’s primary medical
specialty reported in the selected enrollment}

\item{facility_name}{< \emph{character} > Name of facility associated with the
individual provider}

\item{pac_org}{< \emph{integer} > 10-digit Organizational PECOS
Associate Control ID}

\item{city}{< \emph{character} > Provider's city}

\item{state}{< \emph{character} > Provider's state}

\item{zip}{< \emph{character} > Provider's ZIP code}

\item{offset}{< \emph{integer} > // \strong{default:} \code{0L} API pagination}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit individual NPI \cr
   \code{pac} \tab 10-digit individual PAC ID \cr
   \code{enid} \tab 15-digit individual enrollment ID \cr
   \code{first} \tab Provider's first name \cr
   \code{middle} \tab Provider's middle name \cr
   \code{last} \tab Provider's last name \cr
   \code{suffix} \tab Provider's name suffix \cr
   \code{gender} \tab Provider's gender \cr
   \code{credential} \tab Provider's credential \cr
   \code{school} \tab Provider's medical school \cr
   \code{grad_year} \tab Provider's graduation year \cr
   \code{specialty} \tab Provider's primary specialty \cr
   \code{specialty_sec} \tab Provider's secondary specialty \cr
   \code{facility_name} \tab Facility associated with provider \cr
   \code{pac_org} \tab Facility's 10-digit PAC ID \cr
   \code{members} \tab Number of providers associated with facility's PAC ID \cr
   \code{address} \tab Provider's street address \cr
   \code{city} \tab Provider's city \cr
   \code{state} \tab Provider's state \cr
   \code{zip} \tab Provider's zip code \cr
   \code{phone} \tab Provider's phone number \cr
   \code{telehealth} \tab Indicates if provider offers telehealth services \cr
   \code{assign_ind} \tab Indicates if provider accepts Medicare assignment \cr
   \code{assign_org} \tab Indicates if facility accepts Medicare assignment \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=clinicians]{clinicians()}} allows you to access information about providers enrolled in
Medicare, including the medical school that they attended and the year they graduated

\emph{Update Frequency:} \strong{Monthly}
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/mj5m-pzi6}{National Downloadable File}
\item \href{https://data.cms.gov/provider-data/sites/default/files/data_dictionaries/physician/DOC_Data_Dictionary.pdf}{Provider Data Catalog (PDC) Data Dictionary}
}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clinicians(enid = "I20081002000549") # enid not working
clinicians(school = "NEW YORK UNIVERSITY SCHOOL OF MEDICINE")
\dontshow{\}) # examplesIf}
}
