% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicians.R
\name{clinicians}
\alias{clinicians}
\title{Clinicians Enrolled in Medicare}
\usage{
clinicians(
  npi = NULL,
  pac_id_ind = NULL,
  enroll_id_ind = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  gender = NULL,
  school = NULL,
  grad_year = NULL,
  specialty = NULL,
  facility_name = NULL,
  pac_id_org = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  offset = 0L,
  tidy = TRUE
)
}
\arguments{
\item{npi}{Unique clinician ID assigned by NPPES}

\item{pac_id_ind}{Unique individual clinician ID assigned by PECOS}

\item{enroll_id_ind}{Unique ID for the clinician enrollment that is the source
for the data in the observation}

\item{first, middle, last}{Individual clinician's first, middle, or last name}

\item{gender}{Individual clinician gender; \code{"F"} (Female) or \code{"M"} (Male)}

\item{school}{Individual clinician’s medical school}

\item{grad_year}{Individual clinician’s medical school graduation year}

\item{specialty}{Primary medical specialty reported by the individual
clinician in the selected enrollment}

\item{facility_name}{Name of facility associated with the clinician}

\item{pac_id_org}{description}

\item{city}{Group or individual's city}

\item{state}{Group or individual's state}

\item{zip}{Group or individual's ZIP code}

\item{offset}{offset; API pagination}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{clinicians()} allows you to access information about providers
enrolled in Medicare, including the medical school that they attended ant the
year they graduated
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/mj5m-pzi6}{National Downloadable File}
\item \href{https://data.cms.gov/provider-data/sites/default/files/data_dictionaries/physician/DOC_Data_Dictionary.pdf}{Provider Data Catalog (PDC) Data Dictionary}
}

\emph{Update Frequency:} \strong{Monthly}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
clinicians(enroll_id = "I20081002000549")
clinicians(school = "NEW YORK UNIVERSITY SCHOOL OF MEDICINE")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=hospitals]{hospitals()}}, \code{\link[=providers]{providers()}}, \code{\link[=affiliations]{affiliations()}}
}
