% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\alias{compare_hcpcs}
\alias{compare_conditions}
\title{Compare Providers to State and National Benchmarks}
\usage{
compare_hcpcs(df, ...)

compare_conditions(df, pivot = FALSE)
}
\arguments{
\item{df}{< \emph{tbl_df} > // \strong{required} \code{\link[=tibble]{tibble()}} returned from \code{utilization(type = "provider")}}

\item{...}{For future use.

\code{\link[=tibble]{tibble()}} returned from \code{utilization(type = "service")}}

\item{pivot}{< \emph{boolean} > // \strong{default:} \code{FALSE} Pivot output}
}
\value{
A \link[tibble:tibble-package]{tibble} containing:
\itemize{
\item \code{compare_hcpcs()}
\item \code{compare_conditions()}
}
}
\description{
\itemize{
\item \code{compare_hcpcs()} allows the user to compare a provider's yearly HCPCS
utilization data to state and national averages
\item \code{compare_conditions()} allows the user to compare the average yearly
prevalence of chronic conditions among a provider's patient mix to state and
national averages
}
}
\section{Functions}{
\itemize{
\item \code{compare_hcpcs()}: performance detail

\item \code{compare_conditions()}: chronic condition prevalence

}}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

compare_hcpcs(utilization(year = 2018, type = "service", npi = 1023076643))

map_dfr(util_years(), ~utilization(year = .x,
                                   npi = 1023076643,
                                   type = "service")) |>
compare_hcpcs()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

compare_conditions(utilization(year = 2018, type = "provider", npi = 1023076643))

map_dfr(util_years(), ~utilization(year = .x,
                                   npi = 1023076643,
                                   type = "provider")) |>
compare_conditions()
\dontshow{\}) # examplesIf}
}
