% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_hcpcs}
\alias{compare_hcpcs}
\title{Compare Yearly HCPCS Utilization Data}
\usage{
compare_hcpcs(df)
}
\arguments{
\item{df}{data frame returned by \code{by_service()}}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the results.
}
\description{
\code{compare_hcpcs()} allows you to compare yearly HCPCS utilization by provider,
state and national averages
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
by_service_years() |>
map(\(x) by_service(year = x, npi = 1023076643)) |>
list_rbind() |>
compare_hcpcs()
\dontshow{\}) # examplesIf}
}
