% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_long}
\alias{display_long}
\title{Pivot data frame to long format for easy printing}
\usage{
display_long(df, cols = dplyr::everything())
}
\arguments{
\item{df}{data frame}

\item{cols}{columns to pivot long, default is \code{\link[dplyr:reexports]{dplyr::everything()}}}
}
\description{
Pivot data frame to long format for easy printing
}
\keyword{internal}
