% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{distros_main}
\alias{distros_main}
\title{Main CMS Distributions}
\usage{
distros_main(datasets = api_names("main"))
}
\arguments{
\item{datasets}{character vector of dataset names}
}
\value{
description
}
\description{
Main CMS Distributions
}
\keyword{internal}
