% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-esrd.R
\name{esrd}
\alias{esrd}
\title{Dialysis Facilities}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}
}
\usage{
esrd(
  ccn = NULL,
  facility_name = NULL,
  stars = NULL,
  network = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  status = NULL,
  chain_owned = NULL,
  chain_name = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} Individual National Provider Identifier}

\item{facility_name}{\verb{<chr>} facility type}

\item{stars}{\verb{<int>} 1 - 5}

\item{network}{\verb{<int>} 1 - 18}

\item{address, city, state, zip, county}{\verb{<chr>} Individual provider's name}

\item{status}{\verb{<enum>} Non-profit/profit}

\item{chain_owned}{\verb{<lgl>} CCN of the \strong{primary} hospital containing}

\item{chain_name}{\verb{<chr>} facility type}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
esrd(count = TRUE)
esrd()
esrd(stars = "1.0")
esrd(network = 18)
\dontshow{\}) # examplesIf}
}
