% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-cms-fqhc.R
\name{fqhc_owner}
\alias{fqhc_owner}
\title{FQHC Owners}
\usage{
fqhc_owner(npi = NULL, first = NULL, last = NULL, count = FALSE, set = FALSE)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{first, last}{\verb{<chr>} Provider's name}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Providers with pending Medicare enrollment applications.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
fqhc_owner(count = TRUE)

fqhc_owner() |> str()
\dontshow{\}) # examplesIf}
}
