% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospital_enrollment.R
\name{hospital_enrollment}
\alias{hospital_enrollment}
\title{Search the CMS Hospital Enrollments API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
hospital_enrollment(
  npi = NULL,
  facility_ccn = NULL,
  enroll_id = NULL,
  enroll_state = NULL,
  specialty_code = NULL,
  pac_id_org = NULL,
  org_name = NULL,
  dba_name = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{Hospital’s National Provider Identifier (NPI)}

\item{facility_ccn}{Hospital’s CMS Certification Number (CCN), formerly called an
OSCAR Number}

\item{enroll_id}{Hospital’s enrollment ID}

\item{enroll_state}{Hospital’s enrollment state}

\item{specialty_code}{Enrollment application and specialty type code}

\item{pac_id_org}{Hospital’s PECOS Associate Control (PAC) ID}

\item{org_name}{Hospital’s legal business name}

\item{dba_name}{Hospital’s doing-business-as name}

\item{city}{City of the hospital’s practice location address}

\item{state}{State of the hospital’s practice location address.}

\item{zipcode}{Zip code of the hospital’s practice location address}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on hospitals currently enrolled in Medicare.
}
\details{
The Hospital Enrollments dataset provides enrollment information of
all Hospitals currently enrolled in Medicare. This data includes
information on the Hospital's sub-group type, legal business name, doing
business as name, organization type and address.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{Hospital Enrollments}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
hospital_enrollment(npi = 1689653487)
hospital_enrollment(facility_ccn = "440058")
hospital_enrollment(pac_id_org = 6103733050)
hospital_enrollment(city = "Atlanta")
hospital_enrollment(zipcode = 117771928)
\dontshow{\}) # examplesIf}
}
