% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals.R
\name{hospitals}
\alias{hospitals}
\title{Hospitals Enrolled in Medicare}
\usage{
hospitals(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  enid_state = NULL,
  specialty = NULL,
  org_name = NULL,
  dba_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  designation = NULL,
  multi = NULL,
  reh = NULL,
  subgroup = list(acute = NULL, drug = NULL, child = NULL, general = NULL, long = NULL,
    short = NULL, psych = NULL, rehab = NULL, swing = NULL, psych_unit = NULL, rehab_unit
    = NULL, specialty = NULL, other = NULL)
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{ccn}{\verb{<int>} CMS Certification Number}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{enid_state}{\verb{<chr>} Enrollment State}

\item{specialty}{\verb{<chr>} Part A Provider Specialty:
\itemize{
\item \code{"00-09"}: Hospital
\item \code{"00-24"}: REH (Rural Emergency Hospital)
\item \code{"00-85"}: CAH (Critical Access Hospital)
}}

\item{org_name}{\verb{<chr>} Legal Business Name}

\item{dba_name}{\verb{<chr>} Doing-Business-As Name}

\item{city}{\verb{<chr>} Practice Location City}

\item{state}{\verb{<chr>} Practice Location State}

\item{zip}{\verb{<chr>} Practice Location Zip Code}

\item{designation}{\verb{<chr>} IRS designation; \code{"Proprietor"}/\code{"Non-Profit"}}

\item{multi}{\verb{<lgl>} Hospital has more than one NPI}

\item{reh}{\verb{<lgl>} Former Hospital/CAH now a Rural Emergency Hospital}

\item{subgroup}{\verb{<list>} Hospital’s subgroup/unit:
\itemize{
\item \code{acute}: Acute Care
\item \code{drug}: Alcohol/Drug Treatment
\item \code{child}: Children's Hospital
\item \code{general}: General Hospital
\item \code{long}: Long-Term Care
\item \code{short}: Short-Term Care
\item \code{psych}: Psychiatric
\item \code{rehab}: Rehabilitation
\item \code{swing}: Swing-Bed Approved
\item \code{psych_unit}: Psychiatric Unit
\item \code{rehab_unit}: Rehabilitation Unit
\item \code{specialty}: Specialty Hospital
\item \code{other}: Unlisted on CMS form
}}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Hospitals currently enrolled in Medicare. Data includes the hospital's
sub-group types, legal business name, doing-business-as name, organization
type and address.
}
\examples{
hospitals()
hospitals(pac = 6103733050)
hospitals(state = "GA", reh = TRUE)
hospitals(city = "Atlanta", state = "GA", subgroup = list(acute = FALSE))
hospitals(state = "GA", subgroup = list(psych = TRUE))
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{Hospital Enrollments API}
}
}
