% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-hospitals.R
\name{hospitals}
\alias{hospitals}
\alias{subgroups}
\title{Hospitals Enrolled in Medicare}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{API: Hospital Enrollments}
}
}
\usage{
hospitals(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  enid_state = NULL,
  org_name = NULL,
  dba_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  multi = NULL,
  org_type = NULL,
  provider_type = NULL,
  location_type = NULL,
  subgroup = subgroups(),
  count = FALSE
)

subgroups(
  acute = NULL,
  drug = NULL,
  child = NULL,
  general = NULL,
  long = NULL,
  short = NULL,
  psych = NULL,
  rehab = NULL,
  swing = NULL,
  psych_unit = NULL,
  rehab_unit = NULL,
  specialty = NULL,
  other = NULL
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{ccn}{\verb{<int>} CMS Certification Number}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid, enid_state}{\verb{<chr>} Medicare Enrollment ID, Enrollment state}

\item{org_name}{\verb{<chr>} Legal business name}

\item{dba_name}{\verb{<chr>} Doing-business-as name}

\item{city, state, zip}{\verb{<chr>} Location city, state, zip}

\item{multi}{\verb{<lgl>} Does hospital have more than one NPI?}

\item{org_type}{\verb{<enum>} \code{P}: Proprietary, \code{N}: Non-Profit, \code{D}: Unknown}

\item{provider_type}{\verb{<enum>} Provider type; \code{hospital}: Part A Hospital,
\code{reh}: Rural Emergency Hospital, or \code{cah}: Critical Access Hospital}

\item{location_type}{\verb{<enum>} Location type; \code{primary}, \code{psych}, \code{rehab},
\code{extension}, \code{other}}

\item{subgroup}{\verb{<subgroups>} Hospital’s subgroup/unit. See \code{\link[=subgroups]{subgroups()}}.}

\item{count}{\verb{<lgl>} Return the total row count}

\item{acute}{\verb{<lgl>} Acute Care}

\item{drug}{\verb{<lgl>} Alcohol/Drug Treatment}

\item{child}{\verb{<lgl>} Children's Hospital}

\item{general}{\verb{<lgl>} General Hospital}

\item{long}{\verb{<lgl>} Long-Term Care}

\item{short}{\verb{<lgl>} Short-Term Care}

\item{psych}{\verb{<lgl>} Psychiatric}

\item{rehab}{\verb{<lgl>} Rehabilitation}

\item{swing}{\verb{<lgl>} Swing-Bed Approved}

\item{psych_unit}{\verb{<lgl>} Psychiatric Unit}

\item{rehab_unit}{\verb{<lgl>} Rehabilitation Unit}

\item{specialty}{\verb{<lgl>} Specialty Hospital}

\item{other}{\verb{<lgl>} Unlisted on CMS form}
}
\value{
A \link[tibble:tibble-package]{tibble}

A \verb{<subgroups>} object
}
\description{
Hospitals currently enrolled in Medicare. Data includes the hospital's
sub-group types, legal business name, doing-business-as name, organization
type and address.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
hospitals(count = TRUE)
hospitals(state = "GA", provider_type = "reh")
hospitals(
  city = "Atlanta",
  state = "GA",
  subgroup = subgroups(
    acute = FALSE,
    psych = TRUE
  )
)
\dontshow{\}) # examplesIf}
subgroups(acute = TRUE, rehab = TRUE)
}
