% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-hospitals.R
\name{hospitals}
\alias{hospitals}
\title{Hospitals Enrolled in Medicare}
\usage{
hospitals(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  enid_state = NULL,
  specialty = NULL,
  org_name = NULL,
  dba_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  designation = NULL,
  multi = NULL,
  reh = NULL,
  subgroup = subgroups(),
  count = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{ccn}{\verb{<int>} CMS Certification Number}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{enid_state}{\verb{<chr>} Enrollment State}

\item{specialty}{\verb{<chr>}
\itemize{
\item \code{"00-09"}: Hospital
\item \code{"00-24"}: Rural Emergency Hospital
\item \code{"00-85"}: Critical Access Hospital
}}

\item{org_name}{\verb{<chr>} Legal Business Name}

\item{dba_name}{\verb{<chr>} Doing-Business-As Name}

\item{city}{\verb{<chr>} Practice Location City}

\item{state}{\verb{<chr>} Practice Location State}

\item{zip}{\verb{<chr>} Practice Location Zip Code}

\item{designation}{\verb{<chr>} \code{"Proprietor"}/\code{"Non-Profit"}}

\item{multi}{\verb{<lgl>} Hospital has more than one NPI}

\item{reh}{\verb{<lgl>} Former Hospital/CAH now a Rural Emergency Hospital}

\item{subgroup}{\verb{<subgroups>} Hospital’s subgroup/unit. See \code{\link[=subgroups]{subgroups()}}.}

\item{count}{\verb{<lgl>} Return the dataset's total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Hospitals currently enrolled in Medicare. Data includes the hospital's
sub-group types, legal business name, doing-business-as name, organization
type and address.
}
\examples{
hospitals(count = TRUE)
hospitals(pac = 6103733050)
hospitals(state = "GA", reh = TRUE)
hospitals(city = "Atlanta", state = "GA", subgroup = subgroups(acute = FALSE))
hospitals(state = "GA", subgroup = subgroups(psych = TRUE))
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{Hospital Enrollments API}
}
}
