% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-hospitals.R
\name{hospitals}
\alias{hospitals}
\title{Hospitals Enrolled in Medicare}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{API: Hospital Enrollments}
}
}
\usage{
hospitals(
  npi = NULL,
  ccn = NULL,
  pac = NULL,
  enid = NULL,
  enid_state = NULL,
  org_name = NULL,
  dba_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  multi = NULL,
  status = NULL,
  org_type = NULL,
  prov_type = NULL,
  loc_type = NULL,
  subgroup = subgroups(),
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{ccn}{\verb{<int>} CMS Certification Number}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid, enid_state}{\verb{<chr>} Medicare Enrollment ID, Enrollment state}

\item{org_name}{\verb{<chr>} Legal business name}

\item{dba_name}{\verb{<chr>} Doing-business-as name}

\item{city, state, zip}{\verb{<chr>} Location city, state, zip}

\item{multi}{\verb{<lgl>} Does hospital have more than one NPI?}

\item{status}{\verb{<enum>} Organization status
\itemize{
\item \code{P} = Proprietary
\item \code{N} = Non-Profit
\item \code{D} = Unknown
}}

\item{org_type}{\verb{<enum>} Organization structure type
\itemize{
\item \code{corp} = Corporation
\item \code{other} = Other
\item \code{llc} = LLC
\item \code{part} = Partnership
\item \code{sole} = Sole Proprietor
}}

\item{prov_type}{\verb{<enum>} Provider type;
\itemize{
\item \code{hospital} = Part A Hospital
\item \code{reh} = Rural Emergency Hospital
\item \code{cah} = Critical Access Hospital
}}

\item{loc_type}{\verb{<enum>} Practice location type
\itemize{
\item \code{main} = Main/Primary Hospital Location
\item \code{psych} = Hospital Psychiatric Unit
\item \code{rehab} = Hospital Rehabilitation Unit
\item \code{swing} = Hospital Swing-Bed Unit
\item \code{ext} = Opt Extension Site
\item \code{other} = Other Hospital Practice Location
}}

\item{subgroup}{\verb{<subgroups>} Hospital’s subgroup/unit. See \code{\link[=subgroups]{subgroups()}}.}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Hospitals currently enrolled in Medicare. Data includes the hospital's
sub-group types, legal business name, doing-business-as name, organization
type and address.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
hospitals(count = TRUE)
hospitals2(count = TRUE)

hospitals(prov_type = "reh")

x <- hospitals(
  city = "Atlanta",
  state = "GA",
  subgroup = subgroups(
    acute = FALSE,
    psych = TRUE))
x
hospitals2(ccn = x$ccn)

x <- hospitals2()
x
hospitals(ccn = x$ccn)
\dontshow{\}) # examplesIf}
}
