% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals.R
\name{hospitals}
\alias{hospitals}
\title{Hospitals Enrolled in Medicare}
\usage{
hospitals(
  npi = NULL,
  facility_ccn = NULL,
  enroll_id_org = NULL,
  enroll_state = NULL,
  specialty_code = NULL,
  pac_id_org = NULL,
  organization = NULL,
  doing_business_as = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  tidy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{facility_ccn}{< \emph{character} > 6-digit CMS Certification Number of hospital}

\item{enroll_id_org}{< \emph{character} > 15-digit organizational provider Medicare
enrollment identifier; begins with capital "O"}

\item{enroll_state}{< \emph{character} > Hospital’s enrollment state}

\item{specialty_code}{< \emph{character} > Enrollment specialty type code}

\item{pac_id_org}{< \emph{integer} > 10-digit organizational/group provider
associate-level control identifier}

\item{organization}{< \emph{character} > Hospital’s legal business name}

\item{doing_business_as}{< \emph{character} > Hospital’s doing-business-as name}

\item{city}{< \emph{character} > City of the hospital’s practice location}

\item{state}{< \emph{character} > State of the hospital’s practice location}

\item{zip}{< \emph{character} > Zip code of the hospital’s practice location}

\item{tidy}{< \emph{boolean} > Tidy output; default is \code{TRUE}.}

\item{na.rm}{< \emph{boolean} > Remove empty rows and columns; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{hospitals()} allows you to search for information on all hospitals
currently enrolled in Medicare. Data returned includes the hospital's
sub-group types, legal business name, doing-business-as name, organization
type and address.

Links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{Hospital Enrollments}
}

\emph{Update Frequency:} \strong{Monthly}
}
\examples{
hospitals(pac_id_org = 6103733050)
}
\seealso{
\code{\link[=clinicians]{clinicians()}}, \code{\link[=providers]{providers()}}, \code{\link[=affiliations]{affiliations()}}
}
