% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals.R
\name{hospitals}
\alias{hospitals}
\title{Hospitals Enrolled in Medicare}
\usage{
hospitals(
  npi = NULL,
  facility_ccn = NULL,
  enroll_id_org = NULL,
  enroll_state = NULL,
  specialty_code = NULL,
  pac_org = NULL,
  organization = NULL,
  doing_business_as = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  proprietary_nonprofit = NULL,
  multiple_npis = NULL,
  general = NULL,
  acute_care = NULL,
  alcohol_drug = NULL,
  childrens = NULL,
  long_term = NULL,
  psychiatric = NULL,
  rehabilitation = NULL,
  short_term = NULL,
  swing_bed = NULL,
  psych_unit = NULL,
  rehab_unit = NULL,
  specialty_hospital = NULL,
  other = NULL,
  reh_conversion = NULL,
  tidy = TRUE,
  pivot = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{facility_ccn}{< \emph{character} > 6-digit CMS Certification Number of hospital}

\item{enroll_id_org}{< \emph{character} > 15-digit organizational provider Medicare
enrollment identifier; begins with capital "O"}

\item{enroll_state}{< \emph{character} > Hospital’s enrollment state}

\item{specialty_code}{< \emph{character} > Enrollment specialty type code}

\item{pac_org}{< \emph{integer} > 10-digit organizational/group provider
associate-level control identifier}

\item{organization}{< \emph{character} > Hospital’s legal business name}

\item{doing_business_as}{< \emph{character} > Hospital’s doing-business-as name}

\item{city}{< \emph{character} > City of the hospital’s practice location}

\item{state}{< \emph{character} > State of the hospital’s practice location}

\item{zip}{< \emph{character} > Zip code of the hospital’s practice location}

\item{proprietary_nonprofit}{< \emph{character} > \code{"P"} if hospital registered as
proprietor with the IRS; \code{"N"} if registered as non-profit}

\item{multiple_npis}{< \emph{boolean} > Indicates hospital has more than one NPI}

\item{general}{< \emph{boolean} > Indicates hospital’s subgroup/unit is General}

\item{acute_care}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Acute Care}

\item{alcohol_drug}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Alcohol/Drug}

\item{childrens}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Children's Hospital}

\item{long_term}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Long-Term}

\item{psychiatric}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Psychiatric}

\item{rehabilitation}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Rehabilitation}

\item{short_term}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Short-Term}

\item{swing_bed}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Swing-Bed Approved}

\item{psych_unit}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Psychiatric Unit}

\item{rehab_unit}{< \emph{boolean} > Indicates hospital’s subgroup/unit is
Rehabilitation Unit}

\item{specialty_hospital}{< \emph{boolean} > Indicates hospital’s subgroup/unit
is Specialty Hospital}

\item{other}{< \emph{boolean} > Indicates hospital’s subgroup/unit is not listed
on the CMS form}

\item{reh_conversion}{< \emph{boolean} > Indicates a former Hospital or Critical
Access Hospital that converted to a Rural Emergency Hospital}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{pivot}{< \emph{boolean} > // \strong{default:} \code{TRUE} Pivot output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{hospitals()} allows you to search for information on all hospitals
currently enrolled in Medicare. Data returned includes the hospital's
sub-group types, legal business name, doing-business-as name, organization
type and address.

Links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{Hospital Enrollments}
}

\emph{Update Frequency:} \strong{Monthly}
}
\examples{
hospitals(pac_org = 6103733050)

hospitals(state = "GA", reh_conversion = TRUE)
}
\seealso{
\code{\link[=clinicians]{clinicians()}}, \code{\link[=providers]{providers()}}, \code{\link[=affiliations]{affiliations()}}
}
