% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitals.R
\name{hospitals}
\alias{hospitals}
\title{Hospitals Enrolled in Medicare}
\usage{
hospitals(
  npi = NULL,
  facility_ccn = NULL,
  enid_org = NULL,
  enroll_state = NULL,
  pac_org = NULL,
  specialty_code = NULL,
  organization = NULL,
  dba = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  registration = NULL,
  multi = NULL,
  gen = NULL,
  acute = NULL,
  alc_drug = NULL,
  child = NULL,
  long = NULL,
  short = NULL,
  psych = NULL,
  rehab = NULL,
  swing = NULL,
  psych_unit = NULL,
  rehab_unit = NULL,
  spec = NULL,
  other = NULL,
  reh = NULL,
  tidy = TRUE,
  pivot = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{facility_ccn}{< \emph{character} > 6-digit CMS Certification Number}

\item{enid_org}{< \emph{character} > 15-digit organizational provider Medicare
enrollment identifier}

\item{enroll_state}{< \emph{character} > Hospital’s enrollment state}

\item{pac_org}{< \emph{integer} > 10-digit organizational/group provider
associate-level control identifier}

\item{specialty_code}{< \emph{character} > Medicare Part A Provider specialty code:
\itemize{
\item \code{"00-00"}: RELIGIOUS NON-MEDICAL HEALTH CARE INSTITUTION (RNHCI)
\item \code{"00-01"}: COMMUNITY MENTAL HEALTH CENTER
\item \code{"00-02"}: COMPREHENSIVE OUTPATIENT REHABILITATION FACILITY
\item \code{"00-03"}: END-STAGE RENAL DISEASE FACILITY (ESRD)
\item \code{"00-04"}: FEDERALLY QUALIFIED HEALTH CENTER (FQHC)
\item \code{"00-05"}: HISTOCOMPATIBILITY LABORATORY
\item \code{"00-06"}: HOME HEALTH AGENCY
\item \code{"00-08"}: HOSPICE
\item \code{"00-09"}: HOSPITAL
\item \code{"00-10"}: INDIAN HEALTH SERVICES FACILITY
\item \code{"00-13"}: ORGAN PROCUREMENT ORGANIZATION (OPO)
\item \code{"00-14"}: OUTPATIENT PHYSICAL THERAPY/OCCUPATIONAL THERAPY/SPEECH PATHOLOGY SERVICES
\item \code{"00-17"}: RURAL HEALTH CLINIC (RHC)
\item \code{"00-18"}: SKILLED NURSING FACILITY (SNF)
\item \code{"00-19"}: OTHER
\item \code{"00-24"}: RURAL EMERGENCY HOSPITAL
\item \code{"00-85"}: CRITICAL ACCESS HOSPITAL (CAH)
}}

\item{organization}{< \emph{character} > Hospital’s legal business name}

\item{dba}{< \emph{character} > Hospital’s doing-business-as name}

\item{city}{< \emph{character} > City of the hospital’s practice location}

\item{state}{< \emph{character} > State of the hospital’s practice location}

\item{zip}{< \emph{character} > Zip code of the hospital’s practice location}

\item{registration}{< \emph{character} > Hospital's IRS designation:
\itemize{
\item \code{"P"}: Registered as \strong{Proprietor}
\item \code{"N"}: Registered as \strong{Non-Profit}
}}

\item{multi}{< \emph{boolean} > Indicates hospital has more than one NPI}

\item{gen, acute, alc_drug, child, long, short, psych, rehab, swing, psych_unit, rehab_unit, spec, other}{< \emph{boolean} >
Indicates hospital’s subgroup/unit designation:
\itemize{
\item \code{acute}: Acute Care
\item \code{alc_drug}: Alcohol/Drug
\item \code{child}: Children's Hospital
\item \code{gen}: General
\item \code{long}: Long-Term
\item \code{short}: Short-Term
\item \code{psych}: Psychiatric
\item \code{rehab}: Rehabilitation
\item \code{swing}: Swing-Bed Approved
\item \code{psych_unit}: Psychiatric Unit
\item \code{rehab_unit}: Rehabilitation Unit
\item \code{spec}: Specialty Hospital
\item \code{other}: Not listed on CMS form
}}

\item{reh}{< \emph{boolean} > Indicates a former Hospital or Critical Access
Hospital (CAH) that converted to a Rural Emergency Hospital (REH)}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{pivot}{< \emph{boolean} > // \strong{default:} \code{TRUE} Pivot output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab Hospital's 10-digit National Provider Identifier \cr
   \code{pac_org} \tab Hospital’s 10-digit PECOS Associate Control ID \cr
   \code{enid_org} \tab Hospital’s 15-digit Enrollment ID \cr
   \code{enid_state} \tab Hospital’s Enrollment State \cr
   \code{facility_ccn} \tab Hospital’s CMS Certification Number \cr
   \code{organization} \tab Hospital’s Legal Business Name \cr
   \code{doing_business_as} \tab Hospital’s Doing-Business-As Name \cr
   \code{specialty_code} \tab Hospital’s Medicare Part A Provider Specialty Code \cr
   \code{specialty} \tab Specialty Code Description \cr
   \code{incorporation_date} \tab Date the Business was Incorporated \cr
   \code{incorporation_state} \tab State in which the Business was Incorporated \cr
   \code{structure} \tab Hospital’s Organization Structure Type \cr
   \code{address} \tab Hospital’s Practice Location Address \cr
   \code{city} \tab Hospital’s Practice Location City \cr
   \code{state} \tab Hospital’s Practice Location State \cr
   \code{zip} \tab Hospital’s Practice Location Zip Code \cr
   \code{location_type} \tab Practice Location Type \cr
   \code{reh_conversion_date} \tab Date the Hospital/CAH converted to an REH \cr
   \code{associated_ccns} \tab CCNs associated with the Hospital/CAH that converted to an REH \cr
   \code{registration} \tab Hospital’s IRS Registration Status; \strong{Proprietary} or \strong{Non-Profit} \cr
   \code{designations} \tab Hospital's Subgroup/Unit Designations, Multiple NPIs, REH Conversion \cr
}
}
\description{
\code{hospitals()} allows you to search for information on all hospitals
currently enrolled in Medicare. Data returned includes the hospital's
sub-group types, legal business name, doing-business-as name, organization
type and address.

Links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/hospital-enrollments}{Hospital Enrollments}
}

\emph{Update Frequency:} \strong{Monthly}
}
\examples{
hospitals(pac_org = 6103733050)

hospitals(state = "GA", reh = TRUE)
}
\seealso{
\code{\link[=clinicians]{clinicians()}}, \code{\link[=providers]{providers()}}, \code{\link[=affiliations]{affiliations()}}
}
