% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-hospitals.R
\name{hospitals2}
\alias{hospitals2}
\title{Hospitals 2}
\usage{
hospitals2(
  ccn = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  hosp_type = NULL,
  ownership = NULL,
  count = FALSE
)
}
\arguments{
\item{county}{\verb{<chr>}}

\item{hosp_type}{\verb{<chr>}}

\item{ownership}{\verb{<chr>}}
}
\description{
Hospitals 2
}
\keyword{internal}
