% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrsa.R
\name{hrsa}
\alias{hrsa}
\alias{hrsa_open}
\alias{hrsa_layers}
\alias{hrsa_items}
\alias{hrsa_fields}
\title{HRSA Facilities}
\usage{
hrsa_open()

hrsa_layers(ms)

hrsa_items(ms)

hrsa_fields(facility, fl)
}
\arguments{
\item{ms}{\verb{<MapServer>} String indicating what to retrieve.}

\item{facility}{Facility type}

\item{fl}{\verb{<FeatureLayer>}}
}
\value{
A list of endpoints.
}
\description{
HRSA Facilities
}
\examples{
ms <- hrsa_open()
ms
hrsa_items(ms)

fl <- hrsa_layers(ms)

hrsa_fields("asc", fl)
}
