% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-in_outpatient.R
\name{inpatient}
\alias{inpatient}
\alias{outpatient}
\title{Inpatient Hospitals}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/uyx4-5s7f}{API: Medicare Inpatient Hospitals - by Provider}
}

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/uyx4-5s7f}{API: Medicare Outpatient Hospitals - by Provider and Service}
}
}
\usage{
inpatient(
  year = NULL,
  ccn = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  patients = NULL,
  discharges = NULL,
  charge = NULL,
  payment = NULL,
  avg_age = NULL,
  avg_risk = NULL,
  dual = NULL,
  ndual = NULL,
  count = FALSE
)

outpatient(
  year = NULL,
  ccn = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  patients = NULL,
  services = NULL,
  charge = NULL,
  allowed = NULL,
  payment = NULL,
  outliers = NULL,
  count = FALSE
)
}
\arguments{
\item{year}{\verb{<int>} Year data was reported}

\item{ccn}{\verb{<int>} The CMS Certification Number (CCN) of the provider billing
for outpatient hospital services.}

\item{org_name}{\verb{<chr>} Organizational provider's name}

\item{city, state, zip}{\verb{<chr>} The provider's city and state, as reported in
NPPES.}

\item{patients}{\verb{<int>} The number of Medicare fee-for-service beneficiaries
receiving outpatient hospital services.}

\item{discharges}{\verb{<int>} The number of discharges billed by the provider for
inpatient hospital services.}

\item{charge}{\verb{<int>} The provider's average estimated submitted charge for
services covered by Medicare for the Ambulatory Payment Classification
(APC).}

\item{payment}{\verb{<dbl>} The average of total regular payments the provider
receives directly from Medicare. It excludes special outlier payments which
is reported in a separate column.}

\item{avg_age}{\verb{<dbl>} The average age of Medicare enrollees who used a
covered health care or medical service from the provider.}

\item{avg_risk}{\verb{<dbl>} Average Hierarchical Condition Category (HCC) risk
score of beneficiaries}

\item{dual}{\verb{<int>} Number of Medicare beneficiaries qualified to receive
Medicare and Medicaid benefits. Beneficiaries are classified as Medicare
and Medicaid entitlement if in any month in the given calendar year they
were receiving full or partial Medicaid benefits.}

\item{ndual}{\verb{<int>} Number of Medicare beneficiaries qualified to receive
Medicare only benefits. Beneficiaries are classified as Medicare only
entitlement if they received zero months of any Medicaid benefits (full or
partial) in the given calendar year.}

\item{count}{\verb{<lgl>} Return the total row count}

\item{services}{\verb{<int>} The number of primary HCPCS services billed by the
provider for outpatient hospital services.}

\item{allowed}{\verb{<dbl>} The average of total regular payments the provider
receives for the APC. It includes both Medicare direct provider payments as
well as beneficiaries’ co-payment and deductible payments. It excludes
special outlier payments which is reported in a separate column.}

\item{outliers}{\verb{<int>} The number of comprehensive APC services with outlier
payments.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.

A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\subsection{Outlier Payments}{
}
}
\details{
OPPS APC payment amounts are based on the average costs for a set of
services. In the event that a hospital’s costs for these services exceed a
given threshold tied to the average APC payment, CMS must issue an outlier
payment to the hospital to that service to compensate for the costly
provision of service.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
inpatient(count = TRUE)
inpatient(state = "GA", city = "Valdosta")
\dontshow{\}) # examplesIf}
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
outpatient(count = TRUE)
outpatient(state = "GA", city = "Valdosta")
\dontshow{\}) # examplesIf}
}
