% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laboratories.R
\name{laboratories}
\alias{laboratories}
\title{Clinical Laboratories}
\usage{
laboratories(
  name = NULL,
  clia = NULL,
  certificate = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  active = FALSE,
  tidy = TRUE,
  na.rm = TRUE,
  pivot = TRUE,
  ...
)
}
\arguments{
\item{name}{\verb{<chr>} Provider or clinical laboratory's name}

\item{clia}{\verb{<chr>} 10-character CLIA number}

\item{certificate}{\verb{<chr>} CLIA certificate type:
\itemize{
\item \code{"waiver"}
\item \code{"ppm"}
\item \code{"registration"}
\item \code{"compliance"}
\item \code{"accreditation"}
}}

\item{city}{\verb{<chr>} City}

\item{state}{\verb{<chr>} State}

\item{zip}{\verb{<chr>} Zip code}

\item{active}{\verb{<lgl>} // \strong{default:} \code{FALSE} Return only active providers}

\item{tidy}{\verb{<lgl>} // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{\verb{<lgl>} // \strong{default:} \code{TRUE} Remove empty rows and columns}

\item{pivot}{\verb{<lgl>} // \strong{default:} \code{TRUE} Pivot output}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Access information on clinical laboratories including demographics and
the type of testing services the facility provides.
}
\section{Clinical Laboratory Improvement Amendments (CLIA)}{


CMS regulates all laboratory testing (except research) performed on humans
in the U.S. through the Clinical Laboratory Improvement Amendments (CLIA).
In total, CLIA covers approximately 320,000 laboratory entities.

The Division of Clinical Laboratory Improvement & Quality, within the Quality,
Safety & Oversight Group, under the Center for Clinical Standards and Quality
(CCSQ) has the responsibility for implementing the CLIA Program.

Although all clinical laboratories must be properly certified to receive
Medicare or Medicaid payments, CLIA has no direct Medicare or Medicaid
program responsibilities.
}

\section{CLIA Certificates}{


There are five CLIA certificate types all of which are effective for a
period of two years. They are as follows, in order of increasing complexity:
\enumerate{
\item Certificate of \strong{Waiver}: Issued to a laboratory to perform only waived
tests; does not waive the lab from all CLIA requirements. Waived tests are
laboratory tests that are simple to perform. Routine inspections are not
conducted for waiver labs, although 2\% are visited each year to ensure
quality laboratory testing.
\item Certificate for \strong{Provider-Performed Microscopy Procedures} (PPM): Issued
to a laboratory in which a physician, midlevel practitioner or dentist
performs limited tests that require microscopic examination. PPM tests are
considered moderate complexity. Waived tests can also be performed under this
certificate type. There are no routine inspections conducted for PPM labs.
\item Certificate of \strong{Registration}: Initially issued to a laboratory that has
applied for a Certificate of Compliance or Accreditation, enabling the lab to
conduct moderate/high complexity testing until the survey is performed and
the laboratory is found to be in CLIA compliance. Includes PPM and waived
testing.
\item Certificate of \strong{Compliance}: Allows the laboratory to conduct
moderate/high complexity testing and is issued after an inspection finds the
lab to be in compliance with all applicable CLIA requirements. Includes PPM
and waived testing.
\item Certificate of \strong{Accreditation}: Exactly the same as the Certificate of
Compliance, except that the laboratory must be accredited by one of the
following CMS-approved accreditation organizations:
}
\itemize{
\item \href{https://a2la.org/}{American Association for Laboratory Accreditation} (A2LA)
\item \href{https://www.aabb.org/}{Association for the Advancement of Blood & Biotherapies} (AABB)
\item \href{https://osteopathic.org/}{American Osteopathic Association} (AOA)
\item \href{https://www.ashi-hla.org/}{American Society for Histocompatibility and Immunogenetics} (ASHI)
\item \href{https://www.cap.org/}{College of American Pathologists} (CAP)
\item \href{https://www.cola.org/}{Commission on Office Laboratory Accreditation} (COLA)
\item \href{https://www.jointcommission.org/}{The Joint Commission} (JCAHO)
}
}

\section{Resources}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories}{Provider of Services File - Clinical Laboratories}
\item \href{https://www.cms.gov/medicare/quality/clinical-laboratory-improvement-amendments}{CMS.gov CLIA}
\item \href{https://www.cdc.gov/clia/php/about/index.html}{CDC.gov CLIA}
\item \href{https://www.fda.gov/medical-devices/ivd-regulatory-assistance/public-databases}{FDA CLIA Databases}
\item \href{https://www.cdc.gov/labs/clia-certificates/index.html}{CDC.gov CLIA Certificates}
\item \href{https://qcor.cms.gov/main.jsp}{CMS QCOR - S&C's Quality, Certification and Oversight Reports}
}

\emph{Update Frequency:} \strong{Quarterly}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Artic Envestigations Program Laboratory, Anchorage, AK
laboratories(clia = "02D0873639")

# Dengue Laboratory, San Juan, PR
laboratories(clia = "40D0869394")

# CDC/CGH/DGHA International Laboratory, Atlanta, GA
laboratories(clia = "11D1061576")

# Infectious Diseases Laboratory, Atlanta, GA
laboratories(clia = "11D0668319")

# National Center for Environmental Health, Division of Laboratory Science, Atlanta, GA
laboratories(clia = "11D0668290")

# Vector-Borne Diseases Laboratory, Fort Collins, CO
laboratories(clia = "06D0880233")

# Wiregrass Georgia Tech College Student Health Center, Valdosta, GA
laboratories(clia = "11D2306220")

laboratories(clia = "11D0265516")

laboratories(certificate = "ppm", city = "Valdosta", state = "GA", active = TRUE)
\dontshow{\}) # examplesIf}
}
