% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laboratories.R
\name{laboratories}
\alias{laboratories}
\title{Clinical Laboratories}
\usage{
laboratories(
  name = NULL,
  clia = NULL,
  certificate = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  tidy = TRUE,
  pivot = TRUE
)
}
\arguments{
\item{name}{Provider or clinical laboratory's name}

\item{clia}{10-character CLIA number}

\item{certificate}{CLIA certificate type:
\itemize{
\item \code{"waiver"}
\item \code{"compliance"}
\item \code{"accreditation"}
\item \code{"ppm"} (Provider-Performed Microscopy Procedures)
\item \code{"registration"}
}}

\item{city}{City}

\item{state}{State}

\item{zip}{Zip code}

\item{tidy}{Tidy output; default is \code{TRUE}.}

\item{pivot}{Pivot output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{laboratories()} allows you to search for information on clinical laboratories
including demographics and the type of testing services the facility provides.
\subsection{Clinical Laboratory Improvement Amendments (CLIA)}{

CMS regulates all laboratory testing (except research) performed on humans
in the U.S. through the Clinical Laboratory Improvement Amendments (CLIA).
In total, CLIA covers approximately 320,000 laboratory entities. The Division
of Clinical Laboratory Improvement & Quality, within the Quality, Safety &
Oversight Group, under the Center for Clinical Standards and Quality (CCSQ)
has the responsibility for implementing the CLIA Program. Although all
clinical laboratories must be properly certified to receive Medicare or
Medicaid payments, CLIA has no direct Medicare or Medicaid program
responsibilities.
}

\subsection{CLIA Certificate types}{

There are five different CLIA certificate types which all are effective for a
period of two years:

Certificate of \strong{Waiver}
Issued to a laboratory to perform only waived tests; does not waive the lab
from all CLIA requirements. Waived tests are laboratory tests that are simple
and easy to perform. Routine inspections are not conducted for waiver labs,
though 2\% are visited each year to ensure quality laboratory testing.

Certificate for \strong{Provider-Performed Microscopy Procedures} (PPM)
Issued to a laboratory in which a physician, midlevel practitioner or dentist
performs limited tests that require microscopic examination. PPM tests are
considered moderate complexity. Waived tests can also be performed under this
certificate type. There are no routine inspections conducted for PPM labs.

Certificate of \strong{Registration}
Initially issued to a laboratory that has applied for a Certificate of
Compliance or a Certificate of Accreditation, enabling the lab to conduct
moderate/high complexity testing until the survey is performed and the
laboratory found to be in compliance with the CLIA regulations. Includes PPM
and waived testing.

Certificate of \strong{Compliance}
Allows the laboratory to conduct moderate/high complexity testing and is
issued after an inspection finds the lab to be in compliance with all
applicable CLIA requirements. Includes PPM and waived testing.

Certificate of \strong{Accreditation}
Exactly the same as the Certificate of Compliance, except that the laboratory
must be accredited by one of the following CMS-approved accreditation
organizations:
\itemize{
\item American Association for Laboratory Accreditation (AALA)
\item American Association of Blood Banks (AABB)
\item American Osteopathic Association (AOA)
\item American Society for Histocompatibility and Immunogenetics (ASHI)
\item College of American Pathologists (CAP)
\item Commission on Office Laboratory Accreditation (COLA)
\item the Joint Commission (JCAHO)
}
}

\subsection{Links:}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/hospitals-and-other-facilities/provider-of-services-file-clinical-laboratories}{Provider of Services File - Clinical Laboratories}
}

\emph{Update Frequency:} \strong{Quarterly}
}
}
\examples{
laboratories(clia = "11D0265516")
}
