% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-longterm.R
\name{long_term}
\alias{long_term}
\title{Long-Term Care Hospitals}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/azum-44iv}{API: Long-Term Care Hospital - General Information}
}
}
\usage{
long_term(
  ccn = NULL,
  name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  own_type = NULL,
  beds = NULL,
  count = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} The CMS Certification Number (CCN) is used to identify the facility listed}

\item{name}{\verb{<chr>} Name of the facility}

\item{city, state, zip, county}{\verb{<chr>} Facility’s city, state, zip, county}

\item{own_type}{Facility’s ownership type: For Profit, Non-profit, Government, Physician, Tribal}

\item{beds}{\verb{<int>} The total number of beds in the facility}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of long-term care hospitals with information such as
address, phone number, ownership data and more.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
long_term(count = TRUE)
long_term(state = "GA")
\dontshow{\}) # examplesIf}
}
