% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-modifiers.R
\name{modifier}
\alias{modifier}
\alias{excludes}
\alias{contains}
\alias{not}
\alias{between}
\alias{greater_than}
\alias{less_than}
\alias{starts_with}
\alias{ends_with}
\title{A variety of different query operators}
\source{
\href{https://jsonapi.org/format/#query-parameters}{JSON-API: Query Parameters}
}
\usage{
excludes(...)

contains(x)

not(x)

between(...)

greater_than(x, or_equal = FALSE)

less_than(x, or_equal = FALSE)

starts_with(x)

ends_with(x)
}
\arguments{
\item{...}{input}

\item{x}{input}

\item{or_equal}{\verb{<lgl>} append \code{=}}
}
\value{
An S7 \verb{<Modifier>} object.
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query conditions,
in the \href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API} format.
}
\examples{
list(
   # !x \%in\% c("AL", "AK", "AZ")
   `excludes(c("AL", "AK", "AZ"))` = excludes(c("AL", "AK", "AZ")),
   # endsWith(x, "bar")
   `ends_with("bar")` = ends_with("bar"),
   # startsWith(x, "foo")
   `starts_with("foo")` = starts_with("foo"),
   # x < 1000
   `less_than(1000)` = less_than(1000),
   # x <= 1000
   `less_than(0.125, or_equal = TRUE)` = less_than(0.125, or_equal = TRUE),
   # x > 1000
   `greater_than(1000)` = greater_than(1000),
   # x >= 1000
   `greater_than(0.125, or_equal = TRUE)` = greater_than(0.125, or_equal = TRUE),
   # x > 0.125 & x < 2
   `between(0.125, 2)` = between(0.125, 2),
   # grepl("baz", x)
   `contains("baz")` = contains("baz"),
   # x != "zzz"
   `not("zzz")` = not("zzz")
 )
}
