% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-modifiers.R
\name{modifier}
\alias{modifier}
\alias{excludes}
\alias{contains}
\alias{not}
\alias{between}
\alias{greater_than}
\alias{less_than}
\alias{starts_with}
\alias{ends_with}
\title{A variety of different query operators}
\source{
\href{https://jsonapi.org/format/#query-parameters}{JSON-API: Query Parameters}
}
\usage{
excludes(...)

contains(x)

not(x)

between(...)

greater_than(x, or_equal = FALSE)

less_than(x, or_equal = FALSE)

starts_with(x)

ends_with(x)
}
\arguments{
\item{...}{parameter input}

\item{x}{parameter input}

\item{or_equal}{\verb{<lgl>} append \code{=} to \code{greater_than()}/\code{less_than()}}
}
\value{
An S7 \verb{<Modifier>} object.
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query conditions,
in the \href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API} format.
}
\examples{
params()
x <- excludes(state.abb[1:5])
x
operator(x)
value(x)
contains("baz")
not(1000)
between(0.125, 2)
greater_than(1000)
greater_than(0.125, or_equal = TRUE)
less_than(1000)
less_than(0.125, or_equal = TRUE)
starts_with("foo")
ends_with("bar")
}
