% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{modifier}
\alias{modifier}
\alias{excludes}
\alias{between}
\alias{contains}
\alias{not}
\alias{not_blank}
\alias{is_blank}
\alias{greater}
\alias{less}
\alias{starts}
\alias{ends}
\title{A variety of different query operators}
\source{
\href{https://jsonapi.org/format/#query-parameters}{JSON-API: Query Parameters}
}
\usage{
excludes(x)

between(x, y)

contains(x)

not(x)

not_blank()

is_blank()

greater(x, equal = FALSE)

less(x, equal = FALSE)

starts(x)

ends(x)
}
\arguments{
\item{x}{input}

\item{y}{input}

\item{equal}{\verb{<lgl>} append \code{=} to \code{less()} or \code{greater()}}
}
\value{
An S7 \verb{<Modifier>} object.
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query
conditions, in the
\href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API}
format.
}
\examples{
list(
   `excludes(c("AL", "AK", "AZ"))` = excludes(c("AL", "AK", "AZ")),
   `ends("bar")` = ends("bar"),
   `starts("foo")` = starts("foo"),
   `less(1000)` = less(1000),
   `less(0.125, equal = TRUE)` = less(0.125, equal = TRUE),
   `greater(1000)` = greater(1000),
   `greater(0.125, equal = TRUE)` = greater(0.125, equal = TRUE),
   `between(0.125, 2)` = between(0.125, 2),
   `contains("baz")` = contains("baz"),
   `not("zzz")` = not("zzz"),
   `not_blank()` = not_blank(),
   `is_blank()` = is_blank()
 )
}
