% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-modifiers.R
\name{modifier}
\alias{modifier}
\alias{excludes}
\alias{between}
\alias{contains}
\alias{not}
\alias{not_blank}
\alias{is_blank}
\alias{greater_than}
\alias{less_than}
\alias{starts_with}
\alias{ends_with}
\title{A variety of different query operators}
\source{
\href{https://jsonapi.org/format/#query-parameters}{JSON-API: Query Parameters}
}
\usage{
excludes(...)

between(...)

contains(x)

not(x)

not_blank()

is_blank()

greater_than(x, equal = FALSE)

less_than(x, equal = FALSE)

starts_with(x)

ends_with(x)
}
\arguments{
\item{...}{input}

\item{x}{input}

\item{equal}{\verb{<lgl>} append \code{=} to \code{less_than()} or \code{greater_than()}}
}
\value{
An S7 \verb{<Modifier>} object.
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query
conditions, in the
\href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API}
format.
}
\examples{
list(
   `excludes("AL", "AK", "AZ")` = excludes("AL", "AK", "AZ"),
   `ends_with("bar")` = ends_with("bar"),
   `starts_with("foo")` = starts_with("foo"),
   `less_than(1000)` = less_than(1000),
   `less_than(0.125, equal = TRUE)` = less_than(0.125, equal = TRUE),
   `greater_than(1000)` = greater_than(1000),
   `greater_than(0.125, equal = TRUE)` = greater_than(0.125, equal = TRUE),
   `between(0.125, 2)` = between(0.125, 2),
   `contains("baz")` = contains("baz"),
   `not("zzz")` = not("zzz"),
   `not_blank()` = not_blank(),
   `is_blank()` = is_blank()
 )
}
