% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{new_modifier}
\alias{new_modifier}
\alias{is_modifier}
\alias{modifier}
\alias{greater_than}
\alias{less_than}
\alias{between}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{equals}
\alias{any_of}
\title{Query Modifiers}
\usage{
new_modifier(operator, value)

is_modifier(x)

greater_than(x, or_equal = FALSE)

less_than(x, or_equal = FALSE)

between(x, y, negate = FALSE)

starts_with(x)

ends_with(x)

contains(x)

equals(x, negate = FALSE)

any_of(x, negate = FALSE)
}
\arguments{
\item{operator}{input}

\item{value}{input}

\item{x, y}{input}

\item{or_equal}{\verb{<lgl>} append \code{=}}

\item{negate}{\verb{<lgl>} prepend \code{NOT}}
}
\value{
An object of class \verb{<modifier>}
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query conditions,
in the \href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API} format.
}
\examples{
greater_than(1000)
greater_than(0.125, or_equal = TRUE)
less_than(1000)
less_than(0.125, or_equal = TRUE)
between(1000, 1100)
between(0.125, 2, negate = TRUE)
starts_with("foo")
ends_with("bar")
contains("baz")
equals(1000)
equals(1000, negate = TRUE)
any_of(state.abb[10:15])
any_of(state.abb[10:15], negate = TRUE)
}
\keyword{internal}
