% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-modifiers.R
\name{modifier}
\alias{modifier}
\alias{excludes}
\alias{contains}
\alias{not}
\alias{between}
\alias{greater_than}
\alias{less_than}
\alias{starts_with}
\alias{ends_with}
\title{A variety of different query operators}
\source{
\href{https://jsonapi.org/format/#query-parameters}{JSON-API: Query Parameters}
}
\usage{
excludes(...)

contains(x)

not(x)

between(...)

greater_than(x, or_equal = FALSE)

less_than(x, or_equal = FALSE)

starts_with(x)

ends_with(x)
}
\arguments{
\item{...}{input}

\item{x}{input}

\item{or_equal}{\verb{<lgl>} append \code{=}}
}
\value{
An S7 \verb{<Modifier>} object.
}
\description{
Helpers for use in constructing conditions in queries.
}
\details{
Query modifiers are a small DSL for use in constructing query conditions,
in the \href{https://www.drupal.org/docs/core-modules-and-themes/core-modules/jsonapi-module/filtering}{JSON-API} format.
}
\examples{
params(
   state = excludes(state.abb[1:5]),
   title = ends_with("bar"),
   name = starts_with("foo"),
   n = less_than(1000),
   avg = less_than(0.125, or_equal = TRUE),
   rank = greater_than(1000),
   score = greater_than(0.125, or_equal = TRUE),
   interval = between(0.125, 2),
   category = contains("baz"),
   type = not("standard")
 )
}
