% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes.R
\name{nppes}
\alias{nppes}
\title{National Registry of All Health Care Providers}
\usage{
nppes(
  npi = NULL,
  entype = NULL,
  first_name = NULL,
  last_name = NULL,
  org_name = NULL,
  purpose_name = NULL,
  taxonomy_desc = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  country = NULL,
  limit = 1200,
  skip = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI).}

\item{entype}{Entity type. Choices are either \code{I} for an Individual provider
or \code{O} for an Organizational provider. Cannot be the only criteria entered.}

\item{first_name}{Individual provider's first name. Trailing wildcard
entries are permitted requiring at least two characters to be entered
(e.g. \verb{jo*} ).}

\item{last_name}{Individual provider's last name. Trailing wildcard
entries are permitted requiring at least two characters to be entered
(e.g. \verb{jo*} ).}

\item{org_name}{Healthcare organization's name (NPI-2). Trailing wildcard entries are
permitted requiring at least two characters to be entered. All types of
Organization Names (LBN, DBA, Former LBN, Other Name) associated with an
NPI are examined for matching contents, therefore, the results might
contain an organization name different from the one entered in the
Organization Name criterion.}

\item{purpose_name}{Refers to whether the name information entered pertains
to an Authorized Official's name or a Provider's name. When not specified,
the results will search against a provider's first and last name. AO will
only search against Authorized Official names. While PROVIDER will only
search against Provider name. Valid values are: \code{AO} and \code{Provider.}}

\item{taxonomy_desc}{Search for providers by their taxonomy by entering the
taxonomy description.}

\item{city}{City associated with the provider's address. To search for a
Military Address, enter either \code{APO} or \code{FPO}.}

\item{state}{State abbreviation associated with the provider's address. If
this field is used, at least one other field, besides the \code{entype} and
\code{country}, must be populated. Valid values for state abbreviations:
\href{https://npiregistry.cms.hhs.gov/help-api/state}{State Abbreviations}.}

\item{zipcode}{The Postal Code associated with the provider's address
identified in Address Purpose. If you enter a 5 digit postal code, it
will match any appropriate 9 digit (zip+4) codes in the data. Trailing
wildcard entries are permitted requiring at least two characters to be
entered (e.g., \verb{21*}).}

\item{country}{Country abbreviation associated with the provider's
address. This field \strong{can} be used as the only input criterion, as long
as the value selected \emph{is not} \strong{US} (United States). Valid values for
country abbreviations:
\href{https://npiregistry.cms.hhs.gov/help-api/country}{Country Abbreviations}.}

\item{limit}{Maximum number of results to return; default is 1200.}

\item{skip}{Number of results to skip after searching
the previous number; set in \code{limit}.}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{nppes()} allows you to search the National Plan and Provider Enumeration
System (NPPES) NPI Registry's public API. Th Registry is a free directory of
all active National Provider Identifier (NPI) records.
\subsection{National Provider Identifier (NPI)}{

Healthcare providers acquire their unique 10-digit NPIs to identify
themselves in a standard way throughout their industry. Once CMS supplies
an NPI, they publish the parts of the NPI record that have public relevance,
including the provider’s name, taxonomy and practice address.
\subsection{Links}{
\itemize{
\item \href{https://npiregistry.cms.hhs.gov/api-page}{NPPES NPI Registry API Documentation}
\item \href{https://npiregistry.cms.hhs.gov/demo-api}{NPPES NPI Registry API Demo}
}

\emph{Update Frequency:} \strong{Weekly}
}

}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nppes_full(npi = 1528060837)
\dontshow{\}) # examplesIf}
}
