% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-nursing.R
\name{nursing}
\alias{nursing}
\title{Nursing Homes}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}
}
\usage{
nursing(
  ccn = NULL,
  name = NULL,
  org_dba = NULL,
  org_name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  beds = NULL,
  rating = NULL,
  count = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} provider identifier}

\item{name}{\verb{<chr>} description}

\item{org_dba}{\verb{<chr>} desc}

\item{org_name}{\verb{<chr>} desc}

\item{city, state, zip, county}{\verb{<chr>} desc}

\item{beds}{\verb{<int>} description}

\item{rating}{\verb{<int>} description}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
nursing(count = TRUE)
nursing(state = "GA")
\dontshow{\}) # examplesIf}
}
