% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_payments.R
\name{open_payments}
\alias{open_payments}
\title{Open Payments Program}
\usage{
open_payments(
  year,
  npi = NULL,
  covered_type = NULL,
  first = NULL,
  last = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  teaching_hospital = NULL,
  payer = NULL,
  payer_id = NULL,
  pay_form = NULL,
  pay_nature = NULL,
  offset = 0L,
  tidy = TRUE,
  pivot = FALSE
)
}
\arguments{
\item{year}{Program reporting year. Run the helper function
\code{open_payments_years()} to return a vector of currently
available years.}

\item{npi}{Covered recipient's National Provider Identifier.}

\item{covered_type}{Type of covered recipient, e.g.,
\itemize{
\item \code{"Physician"}
\item \code{"Non-Physician Practitioner"}
\item \code{"Teaching Hospital"}
}}

\item{first, last}{Covered recipient's first and/or last name}

\item{city}{Covered recipient's city}

\item{state}{Covered recipient's state abbreviation}

\item{zip}{Covered recipient's zip code}

\item{teaching_hospital}{Name of teaching hospital, e.g.
\verb{Vanderbilt University Medical Center}}

\item{payer}{Paying entity's name; examples:
\itemize{
\item \code{"Pharmacosmos Therapeutics Inc."}
\item \code{"Getinge USA Sales, LLC"}
\item \code{"Agiliti Health, Inc."}
\item \code{"OrthoScan, Inc."}
}}

\item{payer_id}{Paying entity's unique Open Payments ID}

\item{pay_form}{Form of payment, examples:
\itemize{
\item \code{"Stock option"}
\item \code{"Cash or cash equivalent"}
\item \code{"In-kind items and services"}
}}

\item{pay_nature}{Nature of payment or transfer of value; examples:
\itemize{
\item \code{"Royalty or License"}
\item \code{"Charitable Contribution"}
\item \code{"Current or prospective ownership or investment interest"}
\item \code{"Food and Beverage"}
}}

\item{offset}{offset; API pagination}

\item{tidy}{Tidy output; default is \code{TRUE}.}

\item{pivot}{Pivot output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=open_payments]{open_payments()}} allows you to search CMS' Open Payments
Program API.
}
\details{
The Open Payments program is a national disclosure program that
promotes a more transparent and accountable health care system. Open
Payments houses a publicly accessible database of payments that reporting
entities, including drug and medical device companies, make to covered
recipients like physicians. Please note that CMS does not comment on
what relationships may be beneficial or potential conflicts of interest.
CMS publishes the data attested to by reporting entities. The data is
open to individual interpretation.

Open Payments is a national transparency program that collects and
publishes information about financial relationships between drug and
medical device companies (referred to as "reporting entities") and
certain health care providers (referred to as "covered recipients").
These relationships may involve payments to providers for things
including but not limited to research, meals, travel, gifts or speaking
fees.

The purpose of the program is to provide the public with a more
transparent health care system. All information available on the Open
Payments database is open to personal interpretation and if there are
questions about what the data means, patients and their advocates should
speak directly to the health care provider for a better understanding.
\subsection{Links}{
\itemize{
\item \href{https://www.cms.gov/priorities/key-initiatives/open-payments}{What is the Open Payments Program?}
}

\emph{Update Frequency:} \strong{Yearly}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
open_payments(year = 2021, npi = 1043218118)
open_payments(pay_nature = "Royalty or License")
open_payments(pay_form = "Stock option")
open_payments(payer = "Adaptive Biotechnologies Corporation")
open_payments(teaching_hospital = "Nyu Langone Hospitals")
\dontshow{\}) # examplesIf}
}
