% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_payments.R
\name{open_payments}
\alias{open_payments}
\title{Search CMS' Open Payments Program API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
open_payments(
  recipient_npi = NULL,
  recipient_type = NULL,
  recipient_id = NULL,
  recipient_first_name = NULL,
  recipient_last_name = NULL,
  recipient_city = NULL,
  recipient_state = NULL,
  recipient_zip_code = NULL,
  teaching_hospital_name = NULL,
  manufacturer_gpo_name = NULL,
  manufacturer_gpo_id = NULL,
  form_of_payment = NULL,
  nature_of_payment = NULL,
  year = 2021,
  offset = 0,
  clean_names = TRUE,
  nest = FALSE
)
}
\arguments{
\item{recipient_npi}{10-digit National Provider Identifier (NPI).}

\item{recipient_type}{e.g.,
\itemize{
\item \verb{Covered Recipient Physician}
\item \verb{Covered Recipient Non-Physician Practitioner}
\item \verb{Covered Recipient Teaching Hospital}
}}

\item{recipient_id}{Payment recipient's unique Open Payments ID}

\item{recipient_first_name}{Recipient's first name}

\item{recipient_last_name}{Recipient's last name}

\item{recipient_city}{City}

\item{recipient_state}{State, abbreviation}

\item{recipient_zip_code}{Zip code}

\item{teaching_hospital_name}{Name of teaching hospital, e.g.
\verb{Vanderbilt University Medical Center}}

\item{manufacturer_gpo_name}{Paying entity's name. Examples:
\itemize{
\item \verb{Pharmacosmos Therapeutics Inc.}
\item \verb{Getinge USA Sales, LLC}
\item \verb{Agiliti Health, Inc.}
\item \verb{OrthoScan, Inc.}
}}

\item{manufacturer_gpo_id}{Paying entity's unique Open Payments ID}

\item{form_of_payment}{Form of payment, examples:
\itemize{
\item \verb{Stock option}
\item \verb{Cash or cash equivalent}
\item \verb{In-kind items and services}
}}

\item{nature_of_payment}{Nature of payment or transfer of value, examples:
\itemize{
\item \verb{Royalty or License}
\item \verb{Charitable Contribution}
\item \verb{Current or prospective ownership or investment interest}
\item \verb{Food and Beverage}
}}

\item{year}{Reporting year, 2015-2021, default is \code{2021}}

\item{offset}{offset; API pagination}

\item{clean_names}{Convert column names to snakecase; default is \code{TRUE}.}

\item{nest}{Nest related columns together; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=open_payments]{open_payments()}} allows you to search CMS' Open Payments
Program API.
}
\details{
The Open Payments program is a national disclosure program that
promotes a more transparent and accountable health care system. Open
Payments houses a publicly accessible database of payments that reporting
entities, including drug and medical device companies, make to covered
recipients like physicians. Please note that CMS does not comment on
what relationships may be beneficial or potential conflicts of interest.
CMS publishes the data attested to by reporting entities. The data is
open to individual interpretation.

Open Payments is a national transparency program that collects and
publishes information about financial relationships between drug and
medical device companies (referred to as "reporting entities") and
certain health care providers (referred to as "covered recipients").
These relationships may involve payments to providers for things
including but not limited to research, meals, travel, gifts or speaking
fees.

The purpose of the program is to provide the public with a more
transparent health care system. All information available on the Open
Payments database is open to personal interpretation and if there are
questions about what the data means, patients and their advocates should
speak directly to the health care provider for a better understanding.
\subsection{Links}{
\itemize{
\item \href{https://www.cms.gov/OpenPayments}{What is the Open Payments Program?}
\item \href{https://www.cms.gov/OpenPaymentsData.cms.gov}{OpenPaymentsData.cms.gov}
\item \href{https://openpaymentsdata.cms.gov/dataset/0380bbeb-aea1-58b6-b708-829f92a48202}{2021 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/a08c4b30-5cf3-4948-ad40-36f404619019}{2020 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/4e54dd6c-30f8-4f86-86a7-3c109a89528e}{2019 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/f003634c-c103-568f-876c-73017fa83be0}{2018 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/74e3a32c-e8f8-595c-899c-35f9bffa828f}{2017 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/4c774e90-7f9e-5d19-b168-ff9be1e69034}{2016 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/e657f6f0-7abb-5e82-8b42-23bff09f0763}{2015 General Payment Data}
}
}
}
\note{
Update Frequency: \strong{Yearly}
}
\examples{
\dontrun{
open_payments(recipient_npi = 1043218118, year = 2021)
open_payments(nature_of_payment = "Royalty or License")
open_payments(form_of_payment = "Stock option")
open_payments(manufacturer_gpo_name = "Adaptive Biotechnologies Corporation")
open_payments(teaching_hospital_name = "Nyu Langone Hospitals")
}
}
