% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_payments.R
\name{open_payments}
\alias{open_payments}
\title{Search CMS' Open Payments Program API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
open_payments(
  covered_recipient_type = NULL,
  covered_recipient_profile_id = NULL,
  covered_recipient_npi = NULL,
  covered_recipient_first_name = NULL,
  covered_recipient_middle_name = NULL,
  covered_recipient_last_name = NULL,
  recipient_city = NULL,
  recipient_state = NULL,
  recipient_zip_code = NULL,
  applicable_manufacturer_or_applicable_gpo_making_payment_id = NULL,
  year = 2021,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{covered_recipient_type}{"Covered Recipient Physician"}

\item{covered_recipient_profile_id}{Payment recipient's unique Open Payments ID}

\item{covered_recipient_npi}{10-digit National Provider Identifier (NPI).}

\item{covered_recipient_first_name}{Recipient's first name}

\item{covered_recipient_middle_name}{Recipient's middle name}

\item{covered_recipient_last_name}{Recipient's last name}

\item{recipient_city}{City}

\item{recipient_state}{State}

\item{recipient_zip_code}{Zip code}

\item{applicable_manufacturer_or_applicable_gpo_making_payment_id}{Paying entity's unique Open Payments ID}

\item{year}{Reporting year, 2015-2021, default is 2021}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=open_payments]{open_payments()}} allows you to search CMS' Open Payments
Program API.
}
\details{
The Open Payments program is a national disclosure program that
promotes a more transparent and accountable health care system. Open
Payments houses a publicly accessible database of payments that reporting
entities, including drug and medical device companies, make to covered
recipients like physicians. Please note that CMS does not comment on
what relationships may be beneficial or potential conflicts of interest.
CMS publishes the data attested to by reporting entities. The data is
open to individual interpretation.

Open Payments is a national transparency program that collects and
publishes information about financial relationships between drug and
medical device companies (referred to as "reporting entities") and
certain health care providers (referred to as "covered recipients").
These relationships may involve payments to providers for things
including but not limited to research, meals, travel, gifts or speaking
fees.

The purpose of the program is to provide the public with a more
transparent health care system. All information available on the Open
Payments database is open to personal interpretation and if there are
questions about what the data means, patients and their advocates should
speak directly to the health care provider for a better understanding.
\subsection{Links}{
\itemize{
\item \href{https://www.cms.gov/OpenPayments}{What is the Open Payments Program?}
\item \href{https://www.cms.gov/OpenPaymentsData.cms.gov}{OpenPaymentsData.cms.gov}
\item \href{https://openpaymentsdata.cms.gov/dataset/0380bbeb-aea1-58b6-b708-829f92a48202}{2021 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/a08c4b30-5cf3-4948-ad40-36f404619019}{2020 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/4e54dd6c-30f8-4f86-86a7-3c109a89528e}{2019 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/f003634c-c103-568f-876c-73017fa83be0}{2018 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/74e3a32c-e8f8-595c-899c-35f9bffa828f}{2017 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/4c774e90-7f9e-5d19-b168-ff9be1e69034}{2016 General Payment Data}
\item \href{https://openpaymentsdata.cms.gov/dataset/e657f6f0-7abb-5e82-8b42-23bff09f0763}{2015 General Payment Data}
}
}
}
\note{
Update Frequency: \strong{Yearly}
}
\examples{
\dontrun{
open_payments(covered_recipient_npi = "1043218118")
}
}
