% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_payments.R
\name{open_payments}
\alias{open_payments}
\title{Provider Financial Relationships with Drug & Medical Device Companies}
\usage{
open_payments(
  year,
  npi = NULL,
  covered_type = NULL,
  first = NULL,
  last = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  teaching_hospital = NULL,
  payer = NULL,
  payer_id = NULL,
  pay_form = NULL,
  pay_nature = NULL,
  offset = 0L,
  tidy = TRUE,
  pivot = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Year data was reported, in \code{YYYY}
format. Run \code{\link[=open_years]{open_years()}} to return a vector of the years currently available.}

\item{npi}{< \emph{integer} > Covered recipient's 10-digit National Provider Identifier}

\item{covered_type}{< \emph{character} > Type of covered recipient:
\itemize{
\item \code{"Physician"}
\item \code{"Non-Physician Practitioner"}
\item \code{"Teaching Hospital"}
}}

\item{first, last}{< \emph{character} > Covered recipient's name}

\item{city}{< \emph{character} > Covered recipient's city}

\item{state}{< \emph{character} > Covered recipient's state abbreviation}

\item{zip}{< \emph{character} > Covered recipient's zip code}

\item{teaching_hospital}{< \emph{character} > Name of teaching hospital, e.g.
\itemize{
\item \code{"Vanderbilt University Medical Center"}
}}

\item{payer}{< \emph{character} > Paying entity's name, e.g.
\itemize{
\item \code{"Pharmacosmos Therapeutics Inc."}
\item \code{"Getinge USA Sales, LLC"}
\item \code{"Agiliti Health, Inc."}
\item \code{"OrthoScan, Inc."}
}}

\item{payer_id}{< \emph{integer} > Paying entity's unique Open Payments ID}

\item{pay_form}{< \emph{character} > Form of payment:
\itemize{
\item \code{"Stock option"}
\item \code{"Cash or cash equivalent"}
\item \code{"In-kind items and services"}
}}

\item{pay_nature}{< \emph{character} > Nature of payment or transfer of value:
\itemize{
\item \code{"Royalty or License"}
\item \code{"Charitable Contribution"}
\item \code{"Current or prospective ownership or investment interest"}
\item \code{"Food and Beverage"}
}}

\item{offset}{< \emph{integer} > // \strong{default:} \code{0L} API pagination}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{pivot}{< \emph{boolean} > // \strong{default:} \code{TRUE} Pivot output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=open_payments]{open_payments()}} allows the user access to CMS' Open Payments Program API

The \strong{Open Payments} program is a national disclosure program that collects and
publishes information about financial relationships between drug and medical
device companies (referred to as "reporting entities") and certain health
care providers (referred to as "covered recipients"). These relationships may
involve payments to providers for things including but not limited to
research, meals, travel, gifts or speaking fees.
}
\section{Terminology}{

\strong{Reporting Entities}: Applicable manufacturers or GPOs.

\strong{Applicable Group Purchasing Organizations} (GPOs) are entities that operate
in the United States and purchase, arrange for or negotiate the purchase of
covered drugs, devices, biologicals, or medical supplies for a group of
individuals or entities, but not solely for use by the entity itself.

\strong{Applicable Manufacturers} are entities that operate in the United States and
are (1) engaged in the production, preparation, propagation, compounding, or
conversion of a covered drug, device, biological, or medical supply, but not
if such covered drug, device, biological or medical supply is solely for use
by or within the entity itself or by the entity's own patients (this
definition does not include distributors or wholesalers (including, but not
limited to, re-packagers, re-labelers, and kit assemblers) that do not hold
title to any covered drug, device, biological or medical supply); or are (2)
entities under common ownership with an entity described in part (1) of this
definition, which provides assistance or support to such entities with
respect to the production, preparation, propagation, compounding, conversion,
marketing, promotion, sale, or distribution of a covered drug, device,
biological or medical supply.

\strong{Covered Recipients} are any physician, physician assistant, nurse
practitioner, clinical nurse specialist, certified registered nurse
anesthetist, or certified nurse-midwife who is not a bona fide employee of
the applicable manufacturer that is reporting the payment; or a teaching
hospital, which is any institution that received a payment.

\strong{Teaching Hospitals} are hospitals that receive payment for Medicare direct
graduate medical education (GME), IPPS indirect medical education (IME), or
psychiatric hospital IME programs.

\strong{Natures of Payment} are categories that must be used to describe why a
payment or other transfer of value was made. They are only applicable to
the “general” payment type, not research or ownership. The categories are:
\itemize{
\item Acquisitions (2021 - current)
\item Charitable contributions:
\itemize{
\item Compensation for services other than consulting
\item Compensation for serving as faculty or speaker for:
\itemize{
\item An accredited or certified continuing education program (2013 - 2020)
\item An unaccredited and non-certified continuing education program (2013 - 2020)
\item A medical education program (2021 - current)
}
}
\item Consulting fees
\item Current or prospective ownership or investment interest (prior to 2023)
\item Debt Forgiveness (2021 - current)
\item Education
\item Entertainment
\item Food and beverage
\item Gift
\item Grant
\item Honoraria
\item Long-term medical supply or device loan (2021 - current)
\item Royalty or license
\item Space rental or facility fees (Teaching Hospitals only)
\item Travel and lodging
}

\strong{Transfers of Value} are anything of value given by an applicable manufacturer
or applicable GPO to a covered recipient or physician owner/investor that
does not fall within one of the excluded categories in the rule.

\strong{Ownership and Investment Interests} include, but are not limited to:
\itemize{
\item Stock
\item Stock option(s) (not received as compensation, until they are exercised)
\item Partnership share(s)
\item Limited liability company membership(s)
\item Loans
\item Bonds
\item Financial instruments secured with an entity’s property or revenue
}

This may be direct or indirect and through debt, equity or other means.
}

\section{Links}{

\itemize{
\item \href{https://www.cms.gov/priorities/key-initiatives/open-payments}{What is the Open Payments Program?}
\item \href{https://www.cms.gov/OpenPayments/Resources}{Open Payments: General Resources}
}
}

\section{Update Frequency}{

Yearly
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
open_payments(year = 2021, npi = 1043218118)
open_payments(year = 2021, pay_nature = "Royalty or License")
open_payments(year = 2021, pay_form = "Stock option")
open_payments(year = 2021, payer = "Adaptive Biotechnologies Corporation")
open_payments(year = 2021, teaching_hospital = "Nyu Langone Hospitals")

# Use the years helper function to retrieve results for all avaliable years:
open_years() |>
map(\(x) open_payments(year = x, npi = 1043477615)) |>
list_rbind()
\dontshow{\}) # examplesIf}
}
