% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_out.R
\name{opt_out}
\alias{opt_out}
\title{Search the Medicare Opt Out Affidavits API}
\usage{
opt_out(
  first = NULL,
  last = NULL,
  npi = NULL,
  specialty = NULL,
  date_start = NULL,
  date_end = NULL,
  address = NULL,
  city = NULL,
  state_abb = NULL,
  zip = NULL,
  eligible = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{first}{First Name of the Opt Out Provider}

\item{last}{Last Name of the Opt Out Provider}

\item{npi}{National Provider Identifier (NPI) number of the Opt Out Provider}

\item{specialty}{Specialty of the Opt Out Provider}

\item{date_start}{Date from which the Provider's Opt Out Status is
effective}

\item{date_end}{Date on which the Provider's Opt Out Status ends}

\item{address}{Provider's Street Address}

\item{city}{Provider's City}

\item{state_abb}{Provider's State Abbreviation}

\item{zip}{Provider's Zip Code}

\item{eligible}{Flag indicating whether the Provider is eligible to
Order and Refer}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of practitioners who are currently opted out
of Medicare.
}
\details{
The Opt Out Affidavits dataset provides information on providers who have decided not to participate in Medicare. It contains the provider's NPI, specialty, address, and effective dates.
\subsection{Data Update Frequency}{

Monthly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits}{Medicare Opt Out Affidavits API}
}
}
}
\examples{
\dontrun{
opt_out(specialty = "Psychiatry")
opt_out(first = "David")
opt_out(npi = 1114974490)
opt_out(state_abb = "NY", eligible = "N")

opt_out(specialty = "Clinical\%20Psychologist")
opt_out(city = "Los\%20Angeles")

# Returns empty list i.e., provider is not in the database
opt_out(npi = 1326011057)
}
}
