% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_out.R
\name{opt_out}
\alias{opt_out}
\title{Search the Medicare Opt Out Affidavits API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
opt_out(
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  specialty = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  order_and_refer = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{National Provider Identifier (NPI) number of the Opt Out Provider}

\item{first_name}{First Name of the Opt Out Provider}

\item{last_name}{Last Name of the Opt Out Provider}

\item{specialty}{Specialty of the Opt Out Provider}

\item{address}{Provider's Street Address}

\item{city}{Provider's City}

\item{state}{Provider's State Abbreviation}

\item{zip}{Provider's Zip Code}

\item{order_and_refer}{Flag indicating whether the Provider is eligible to
Order and Refer}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of practitioners who are currently opted out
of Medicare.
}
\details{
The Opt Out Affidavits dataset provides information on providers
who have decided not to participate in Medicare. It contains the
provider's NPI, specialty, address, and effective dates.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits}{Medicare Opt Out Affidavits API}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
opt_out(specialty = "Psychiatry", zipcode = "07626")
opt_out(first = "David", last = "Smith")
opt_out(npi = 1114974490)
opt_out(city = "Los Angeles", address = "9201 W SUNSET BLVD")
opt_out(state = "NY", order_and_refer = FALSE)
opt_out(npi = 1326011057)
\dontshow{\}) # examplesIf}
}
