% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_out.R
\name{opt_out}
\alias{opt_out}
\title{Providers Opted Out of Medicare}
\usage{
opt_out(
  npi = NULL,
  first = NULL,
  last = NULL,
  specialty = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  order_refer = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{< \code{integer} > 10-digit Opt-out National Provider Identifier}

\item{first, last}{< \code{character} > Opt-out provider's name}

\item{specialty}{< \code{character} > Opt-out provider's specialty}

\item{address}{< \code{character} > Opt-out provider's address}

\item{city}{< \code{character} > Opt-out provider's city}

\item{state}{< \code{character} > Opt-out provider's state abbreviation}

\item{zip}{< \code{character} > Opt-out provider's zip code}

\item{order_refer}{< \code{boolean} > Indicates order and refer eligibility}

\item{tidy}{< \code{boolean} > // \strong{default:} \code{TRUE} Tidy output}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit NPI \cr
   \code{first} \tab Opt-out provider's first name \cr
   \code{last} \tab Opt-out provider's last name \cr
   \code{specialty} \tab Opt-out provider's specialty \cr
   \code{order_refer} \tab Indicates if the provider can order and refer \cr
   \code{optout_start_date} \tab Date that provider's Opt-Out period begins \cr
   \code{optout_end_date} \tab Date that provider's Opt-Out period ends \cr
   \code{last_updated} \tab Date information was last updated \cr
   \code{address} \tab Opt-out provider's street address \cr
   \code{city} \tab Opt-out provider's city \cr
   \code{state} \tab Opt-out provider's state \cr
   \code{zip} \tab Opt-out provider's zip code \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{opt_out()} allows the user to access information on providers who have
decided not to participate in Medicare.
}
\section{Opting Out}{


Providers who do not wish to enroll in the Medicare program may “opt-out”,
meaning neither they nor the beneficiary can bill Medicare for services rendered.

Instead, a private contract between provider and beneficiary is signed,
neither party is reimbursed by Medicare and the beneficiary pays
the provider out-of-pocket.

To opt out, a provider must:
\itemize{
\item Be of an \emph{eligible specialty} type
\item Submit an \emph{opt-out affidavit} to Medicare
\item Enter into a \emph{private contract} with their Medicare patients, reflecting
the agreement that they will pay out-of-pocket and that no one will submit
the bill to Medicare for reimbursement
}
}

\section{Opt-Out Periods}{


Opt-out periods last for two years and cannot be terminated early unless the
provider is opting out for the very first time and terminates the opt-out no
later than 90 days after the opt-out period's effective date.

Opt-out statuses are also effective for two years and automatically renew.
Providers that do not want to extend their opt-out status at the end of an
opt-out period may cancel by notifying all MACs an affidavit was filed at
least 30 days prior to the start of the next opt-out period.

If a provider retires, surrenders their license, or no longer wants to
participate in the Medicare program, they must officially withdraw within 90
days. DMEPOS suppliers must withdraw within 30 days.

Providers may \emph{NOT} opt-out if they intend to be a Medicare Advantage
(Part C) provider or furnish services covered by traditional Medicare
fee-for-service (Part B).

\emph{Update Frequency:} \strong{Monthly}
}

\examples{
opt_out(npi = 1043522824)

\dontshow{if ( interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# For opt-out providers eligible
# to order and refer, use [order_refer()]
# to look up their eligibility status:

opt_out(npi = 1043522824) |>
        pull(npi) |>
        map(\(x) order_refer(npi = x)) |>
        list_rbind()
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits}{Medicare Opt Out Affidavits API}
}
}
