% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-order_refer.R
\name{order_refer}
\alias{order_refer}
\title{Order and Refer Eligibility}
\usage{
order_refer(
  npi = NULL,
  first = NULL,
  last = NULL,
  part_b = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  hospice = NULL,
  count = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{first, last}{\verb{<chr>} Individual provider's first/last name}

\item{part_b, dme, hha, pmd, hospice}{\verb{<lgl>} Eligibility for:
\itemize{
\item \code{part_b}: Medicare Part B
\item \code{dme}: Durable Medical Equipment
\item \code{hha}: Home Health Agency
\item \code{pmd}: Power Mobility Devices
\item \code{hospice}: Hospice
}}

\item{count}{\verb{<lgl>} Return the dataset's total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Eligibility to order and refer within Medicare
}
\section{Criteria}{

\itemize{
\item \emph{Individual} NPI
\item Medicare enrollment with \emph{Approved} or \emph{Opt-Out} status
\item Eligible \emph{specialty} type
}
}

\section{Types}{

\itemize{
\item \emph{Ordering}: can order non-physician services for patients.
\item \emph{Referring/Certifying}: can request items/services Medicare may reimburse.
\item \emph{Opt-Out}: can enroll solely to order and refer.
}
}

\section{Services}{

\itemize{
\item \emph{Medicare Part B}: Clinical Labs, Imaging
\item \emph{Medicare Part A}: Home Health
\item \emph{DMEPOS}: Durable medical equipment, prosthetics, orthotics, & supplies
}
}

\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
order_refer(count = TRUE)
order_refer(npi = 1003026055)
order_refer(first = "Jennifer", last = "Smith")
order_refer(
  part_b = TRUE,
  dme = TRUE,
  hha = FALSE,
  pmd = TRUE,
  hospice = FALSE)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{API: Medicare Order and Referring}
\item \href{https://www.cms.gov/medicare/enrollment-renewal/providers-suppliers/chain-ownership-system-pecos/ordering-certifying}{CMS: Ordering & Certifying}
}
}
