% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_refer.R
\name{order_refer}
\alias{order_refer}
\title{Order and Referral Eligibility}
\usage{
order_refer(
  npi = NULL,
  first = NULL,
  last = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  hos = NULL,
  tidy = TRUE,
  pivot = TRUE,
  ...
)
}
\arguments{
\item{npi}{\verb{<chr>} Unique 10-digit National Provider Identifier number issued
by CMS to US healthcare providers through NPPES.}

\item{first, last}{\verb{<chr>} Individual provider's first/last name}

\item{partb, dme, hha, pmd, hos}{\verb{<lgl>} Whether a provider is eligible to
order and refer to:
\itemize{
\item \code{partb}: Medicare Part B
\item \code{dme}: Durable Medical Equipment
\item \code{hha}: Home Health Agency
\item \code{pmd}: Power Mobility Devices
\item \code{hos}: Hospice
}}

\item{tidy}{\verb{<lgl>} Tidy output; \strong{default} is \code{TRUE}}

\item{pivot}{\verb{<lgl>} Pivot output; \strong{default} is \code{TRUE}}

\item{...}{Empty dots}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab National Provider Identifier \cr
   \code{first} \tab Order and Referring Provider's First Name \cr
   \code{last} \tab Order and Referring Provider's Last Name \cr
   \code{eligible} \tab Services An Eligible Provider Can Order/Refer To \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{order_refer()} returns a provider's eligibility to order and refer within
Medicare to:
\itemize{
\item \strong{Part B}: Clinical Laboratory Services, Imaging Services
\item \strong{DME}: Durable Medical Equipment, Prosthetics, Orthotics, & Supplies (DMEPOS)
\item \strong{Part A}: Home Health Services
}

To be eligible, a provider must:
\itemize{
\item have an \emph{Individual} NPI
\item be enrolled in Medicare in either an \emph{Approved} or \emph{Opt-Out} status
\item be of an \emph{Eligible Specialty} type
}

\strong{Ordering Providers} can order non-physician services for patients.

\strong{Referring (or Certifying) Providers} can request items or services that
Medicare may reimburse on behalf of its beneficiaries.

\strong{Opt-Out Providers}: Providers who have opted out of Medicare may still
order and refer. They can also enroll solely to order and refer.
}
\examples{
order_refer(npi = 1003026055)

# Filter for certain privileges
order_refer(first = "Jennifer",
            last  = "Smith",
            partb = TRUE,
            hos   = FALSE,
            hha   = FALSE,
            pmd   = FALSE)

}
\references{
links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{Medicare Order and Referring API}
\item \href{https://www.cms.gov/medicare/enrollment-renewal/providers-suppliers/chain-ownership-system-pecos/ordering-certifying}{CMS.gov: Ordering & Certifying}
\item \href{https://data.cms.gov/resources/order-and-referring-methodology}{Order and Referring Methodology}
}

\emph{Update Frequency:} \strong{Twice Weekly}
}
