% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_refer.R
\name{order_refer}
\alias{order_refer}
\title{Search the Medicare Order and Referring API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
order_refer(
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{first_name}{Provider's first name}

\item{last_name}{Provider's last name}

\item{partb}{logical}

\item{dme}{logical}

\item{hha}{logical}

\item{pmd}{logical}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
All physicians and non-physician practitioners who are legally
eligible to order and refer in the Medicare program and who have current
enrollment records in Medicare.
}
\details{
The Medicare Order and Referring dataset provides information on
all physicians and non-physician practitioners, by their National
Provider Identifier (NPI), who are of a type/specialty that is legally
eligible to order and refer in the Medicare program and who have current
enrollment records in Medicare.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{Medicare Order and Referring API}
}
}
}
\note{
Update Frequency: \strong{Weekly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
order_refer(npi = 1003026055)

order_refer(last_name = "phadke",
            first_name = "radhika")

# Unnamed List of NPIs
c(1003026055, 1316405939, 1720392988,
  1518184605, 1922056829, 1083879860) |>
  purrr::map(order_refer) |>
  purrr::list_rbind()

# Data frame of NPIs
data.frame(npi = c(1003026055, 1316405939, 1720392988,
                   1518184605, 1922056829, 1083879860)) |>
                   dplyr::pull(npi) |>
                   purrr::map(order_refer) |>
                   purrr::list_rbind()
\dontshow{\}) # examplesIf}
}
