% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_refer.R
\name{order_refer}
\alias{order_refer}
\title{Ordering and Referring Eligibility for Medicare}
\usage{
order_refer(
  npi = NULL,
  first_name = NULL,
  last_name = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier}

\item{first_name}{Provider's first name}

\item{last_name}{Provider's last name}

\item{partb}{boolean; eligibility to order & refer to Medicare Part B}

\item{dme}{boolean; eligibility to order Durable Medical Equipment}

\item{hha}{boolean; eligibility to refer to a Home Health Agency}

\item{pmd}{boolean; eligibility to order Power Mobility Devices}

\item{tidy}{boolean; Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} object containing the search results.
}
\description{
\code{order_refer()} allows you to search for physicians and non-physician
practitioners who are of a type/specialty that is legally eligible to order
and refer to Part B (clinical laboratory and imaging), DME and Part A HHA
claims in the Medicare program.
\subsection{Ordering and Referring (or Certifying) Providers}{

\strong{Ordering providers} can order non-physician services for patients.
\strong{Referring providers} can request items or services Medicare may reimburse
on behalf of Medicare beneficiaries.

To qualify as an ordering and referring/certifying provider, a provider must:
\itemize{
\item have an \emph{individual} NPI
\item be enrolled in Medicare in either an \emph{approved} or \emph{opt-out} status
\item be of an \emph{eligible specialty} type
}

Providers currently enrolled in Medicare Part B can order and refer. A
provider that does not want to bill Medicare for their services can still
order and refer by opting out of Medicare or enrolling solely to order and
refer. Medicare coverage will then apply when ordering or referring:
\itemize{
\item Durable Medical Equipment, Prosthetics, Orthotics, and Supplies (DMEPOS)
\item Clinical Laboratory Services
\item Imaging Services
\item Home Health Services
}
}
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{Medicare Order and Referring API}
\item \href{https://www.cms.gov/medicare/provider-enrollment-and-certification/ordering-and-certifying}{CMS.gov: Ordering & Certifying}
\item \href{https://data.cms.gov/resources/order-and-referring-methodology}{Order and Referring Methodology}
}
}

\section{Update Frequency}{
 \strong{Twice Weekly}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Search by NPI
order_refer(npi = 1003026055)

# Search by name and/or filter for certain privileges
order_refer(last_name = "Smith", partb = FALSE, hha = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=providers]{providers()}}, \code{\link[=opt_out]{opt_out()}}, \code{\link[=pending]{pending()}}
}
