% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_refer.R
\name{order_refer}
\alias{order_refer}
\title{Order and Referral Eligibility}
\usage{
order_refer(
  npi = NULL,
  first = NULL,
  last = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  tidy = TRUE,
  pivot = TRUE,
  ...
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{first, last}{< \emph{character} > Individual provider's first/last name}

\item{partb, dme, hha, pmd}{< \emph{boolean} > Whether a provider is eligible to
order and refer to:
\itemize{
\item \code{partb}: Medicare Part B
\item \code{dme}: Durable Medical Equipment
\item \code{hha}: Home Health Agency
\item \code{pmd}: Power Mobility Devices
}}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{pivot}{< \emph{boolean} > // \strong{default:} \code{TRUE} Pivot output}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab National Provider Identifier \cr
   \code{first} \tab Order and Referring Provider's First Name \cr
   \code{last} \tab Order and Referring Provider's Last Name \cr
   \code{service} \tab Services An Eligible Provider Can Order/Refer To \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{order_refer()} returns a provider's eligibility to order and refer within
Medicare to:
\itemize{
\item \strong{Part B}: Clinical Laboratory Services, Imaging Services
\item \strong{DME}: Durable Medical Equipment, Prosthetics, Orthotics, & Supplies (DMEPOS)
\item \strong{Part A}: Home Health Services
}

To be eligible, a provider must:
\itemize{
\item have an \emph{Individual} NPI
\item be enrolled in Medicare in either an \emph{Approved} or \emph{Opt-Out} status
\item be of an \emph{Eligible Specialty} type
}

\strong{Ordering Providers} can order non-physician services for patients.

\strong{Referring (or Certifying) Providers} can request items or services that
Medicare may reimburse on behalf of its beneficiaries.

\strong{Opt-Out Providers}: Providers who have opted out of Medicare may still
order and refer. They can also enroll solely to order and refer.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
order_refer(npi = 1003026055)

# Filter for certain privileges
order_refer(last = "Smith", partb = FALSE, hha = TRUE)
\dontshow{\}) # examplesIf}
}
\references{
links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{Medicare Order and Referring API}
\item \href{https://www.cms.gov/medicare/enrollment-renewal/providers-suppliers/chain-ownership-system-pecos/ordering-certifying}{CMS.gov: Ordering & Certifying}
\item \href{https://data.cms.gov/resources/order-and-referring-methodology}{Order and Referring Methodology}
}

\emph{Update Frequency:} \strong{Twice Weekly}
}
