% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_refer.R
\name{order_refer}
\alias{order_refer}
\title{Search the Medicare Order and Referring API}
\usage{
order_refer(
  npi = NULL,
  last = NULL,
  first = NULL,
  partb = NULL,
  dme = NULL,
  hha = NULL,
  pmd = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{last}{Provider's last name}

\item{first}{Provider's first name}

\item{partb}{logical}

\item{dme}{logical}

\item{hha}{logical}

\item{pmd}{logical}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
All physicians and non-physician practitioners who are legally
eligible to order and refer in the Medicare program and who have current
enrollment records in Medicare.
}
\details{
The Medicare Order and Referring dataset provides information on
all physicians and non-physician practitioners, by their National
Provider Identifier (NPI), who are of a type/specialty that is legally
eligible to order and refer in the Medicare program and who have current
enrollment records in Medicare.
\subsection{Data Update Frequency}{

Weekly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{Medicare Order and Referring API}
}
}
}
\examples{
\dontrun{
order_refer(npi = 1003026055)

order_refer(last = "phadke",
            first = "radhika")

# Unnamed List of NPIs
npi_list <- c(1003026055,
              1316405939,
              1720392988,
              1518184605,
              1922056829,
              1083879860)

npi_list |>
purrr::map_dfr(order_refer)

# Data frame of NPIs
npi_df <- data.frame(npi = c(1003026055,
                             1316405939,
                             1720392988,
                             1518184605,
                             1922056829,
                             1083879860))
npi_df |>
tibble::deframe() |>
purrr::map_dfr(order_refer)
}
}
