% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpatient.R
\name{outpatient}
\alias{outpatient}
\title{Outpatient Hospitals Enrolled in Medicare}
\usage{
outpatient(
  year,
  ccn = NULL,
  organization = NULL,
  street = NULL,
  city = NULL,
  state = NULL,
  fips = NULL,
  zip = NULL,
  ruca = NULL,
  apc = NULL,
  tidy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Year data was reported, in \code{YYYY}
format. Run \code{\link[=out_years]{out_years()}} to return a vector of the years currently available.}

\item{ccn}{< \emph{integer} > 6-digit CMS Certification Number}

\item{organization}{< \emph{character} > Organization's name}

\item{street}{< \emph{character} > Street where provider is located}

\item{city}{< \emph{character} > City where provider is located}

\item{state}{< \emph{character} > State where provider is located}

\item{fips}{< \emph{character} > Provider's state FIPS code}

\item{zip}{< \emph{character} > Provider’s zip code}

\item{ruca}{< \emph{character} > Provider’s RUCA code}

\item{apc}{< \emph{character} > comprehensive ambulatory payment classification code}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}
}
\description{
Outpatient Hospitals Enrolled in Medicare
}
\note{
State/FIP values are restricted to 49 of the United States and the District
of Columbia where hospitals are subjected to Medicare's Outpatient
Prospective Payment System (OPPS). Maryland and US territory hospitals are
excluded.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
outpatient(year = 2021, ccn = "110122")

outpatient(year = 2021, state = "GA")
\dontshow{\}) # examplesIf}
}
