% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-owner.R
\name{owner}
\alias{owner}
\title{Owners}
\source{
Medicare
}
\usage{
owner(
  fac_type = NULL,
  fac_enid = NULL,
  fac_pac = NULL,
  fac_name = NULL,
  own_pac = NULL,
  own_org = NULL,
  own_dba = NULL,
  own_pct = NULL,
  own_role = NULL,
  own_entity = NULL,
  own_first = NULL,
  own_last = NULL,
  own_title = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  count = FALSE
)
}
\arguments{
\item{fac_type}{\verb{<enum>} Facility type; if NULL (default), will search all:
\itemize{
\item \code{HHA} = Home Health Agency
\item \code{RHC} = Rural Health Clinic
\item \code{FQHC} = Federally Qualified Health Clinic
\item \code{SNF} = Skilled Nursing Facility
\item \code{Hospice} = Hospice
\item \code{Hospital} = Hospital
}}

\item{own_pac}{\verb{<chr>} Provider's name}

\item{own_org}{\verb{<chr>} Provider's name}

\item{own_dba}{\verb{<chr>} Provider's name}

\item{own_pct}{\verb{<dbl>} Provider's name}

\item{own_role}{\verb{<chr>} Provider's name}

\item{own_entity}{\verb{<enum>} Provider's name}

\item{own_first, own_last}{\verb{<chr>} Provider's name}

\item{own_title}{\verb{<chr>} Provider's name}

\item{city, state, zip}{\verb{<chr>} Provider's name}

\item{count}{\verb{<lgl>} Return the total row count}

\item{enid}{\verb{<chr>} National Provider Identifier}

\item{pac}{\verb{<chr>} Provider's name}

\item{org_name}{\verb{<chr>} Provider's name}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Owners of facilities enrolled in Medicare.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
owner(count = TRUE)
owner(city = "Valdosta", state = "GA")
\dontshow{\}) # examplesIf}
}
