% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{pct}
\alias{pct}
\title{Calculate lagged percentage change}
\usage{
pct(x, n = 1L, fill_na = 0L)
}
\arguments{
\item{x}{numeric vector}

\item{n}{values to offset}

\item{fill_na}{fill value for any NAs; default is 0}
}
\description{
Calculate lagged percentage change
}
\keyword{internal}
