% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-pending.R
\name{pending}
\alias{pending}
\title{Pending Medicare Enrollments}
\source{
Medicare Pending Initial Logging and Tracking:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-physicians}{API: Physicians}
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-non-physicians}{API: Non-Physicians}
}
}
\usage{
pending(npi = NULL, first = NULL, last = NULL, count = FALSE, set = FALSE)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{first, last}{\verb{<chr>} Provider's name}

\item{count}{\verb{<lgl>} Return the dataset's total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Providers with pending Medicare enrollment applications.

The Pending Initial Logging and Tracking (L & T) dataset provides a list of
pending applications for both Physicians and Non-Physicians that have not
been processed by CMS contractors.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
pending(count = TRUE)
pending(first = "Victor", count = TRUE)
pending(first = starts_with("V"))
\dontshow{\}) # examplesIf}
}
