% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending.R
\name{pending}
\alias{pending}
\title{Pending Medicare Enrollments}
\usage{
pending(npi = NULL, first = NULL, last = NULL)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{first, last}{\verb{<chr>} Provider's name}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Providers with pending Medicare enrollment applications.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
pending(first = "John")
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-physicians}{Medicare Pending Initial Logging and Tracking Physicians API}
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-non-physicians}{Medicare Pending Initial Logging and Tracking Non-Physicians API}
}
}
