% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending.R
\name{pending}
\alias{pending}
\title{Pending Medicare Enrollment Applications}
\usage{
pending(type = "P", npi = NULL, first = NULL, last = NULL, tidy = TRUE)
}
\arguments{
\item{type}{< \code{character} > // \strong{default:} \code{"P"}

Physician (\code{P}) or Non-physician (\code{N})}

\item{npi}{< \code{integer} >

10-digit National Provider Identifier}

\item{first, last}{< \code{character} >

Provider's name}

\item{tidy}{< \code{boolean} > // \strong{default:} \code{TRUE}

Tidy output}
}
\value{
A \code{\link[=tibble]{tibble()}} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit individual NPI \cr
   \code{first} \tab Provider's first name \cr
   \code{last} \tab Provider's last name \cr
   \code{type} \tab Type of Provider \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

\code{\link[=pending]{pending()}} allows the user to search for providers with pending Medicare
enrollment applications.
}
\section{Update Frequency}{

\strong{QUARTERLY}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

pending(type = "P", first = "John")

pending(type = "N", last = "Smith")
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-physicians}{Medicare Pending Initial Logging and Tracking Physicians API}
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-non-physicians}{Medicare Pending Initial Logging and Tracking Non-Physicians API}
}
}
