% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending.R
\name{pending}
\alias{pending}
\title{Pending Medicare Enrollment Applications}
\usage{
pending(type, npi = NULL, first = NULL, last = NULL, tidy = TRUE)
}
\arguments{
\item{type}{< \emph{character} > // \strong{required} Physician ("\code{P}") or
Non-physician ("\code{N}")}

\item{npi}{< \emph{integer} > 10-digit National Provider Identifier}

\item{first, last}{< \emph{character} > Provider's name}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

\code{pending()} allows the user to search for providers with pending Medicare
enrollment applications.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pending(type = "N", last = "Smith")

pending(type = "P", first = "John")
\dontshow{\}) # examplesIf}
}
\references{
APIs:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-physicians}{Medicare Pending Initial Logging and Tracking Physicians API}
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-non-physicians}{Medicare Pending Initial Logging and Tracking Non-Physicians API}
}

\emph{Update Frequency:} \strong{Weekly}
}
