% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pending_applications.R
\name{pending_applications}
\alias{pending_applications}
\title{Providers with Pending Medicare Enrollment Applications}
\usage{
pending_applications(
  type,
  npi = NULL,
  last_name = NULL,
  first_name = NULL,
  tidy = TRUE
)
}
\arguments{
\item{type}{physician or non-physician}

\item{npi}{National Provider Identifier (NPI) number}

\item{last_name}{Last name of provider}

\item{first_name}{First name of provider}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{pending_applications()} A list of enrollment applications pending CMS contractor
review for physicians & non-physicians.
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-physicians}{Medicare Pending Initial Logging and Tracking Physicians API}
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/pending-initial-logging-and-tracking-non-physicians}{Medicare Pending Initial Logging and Tracking Non-Physicians API}
}
}

\section{Update Frequency}{
 \strong{Weekly}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pending_applications(last_name = "Smith", type = "non-physician")
pending_applications(first_name = "John", type = "physician")
\dontshow{\}) # examplesIf}
}
