% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physician_by_geography.R
\name{physician_by_geography}
\alias{physician_by_geography}
\title{Search the Medicare Physician & Other Practitioners API
by \strong{Geography and Service}}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
physician_by_geography(
  geo_level = NULL,
  fips = NULL,
  geo_desc = NULL,
  hcpcs_code = NULL,
  hcpcs_desc = NULL,
  hcpcs_drug = NULL,
  pos = NULL,
  year = 2020,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{geo_level}{Identifies the level of geography by which the data in the
row has been aggregated. A value of 'State' indicates the data in the
row is aggregated to a single state identified in the Rendering Provider
State column for a given HCPCS Code Level. A value of 'National'
indicates the data in the row is aggregated across all states for a
given HCPCS Code Level.}

\item{fips}{FIPS code of the referring provider state. This variable is
blank when reported at the national level.}

\item{geo_desc}{The state name where the provider is located, as reported
in NPPES. The values include the 50 United States, District of Columbia,
U.S. territories, Armed Forces areas, Unknown and Foreign Country. Data
aggregated at the National level are identified by the word 'National'.}

\item{hcpcs_code}{HCPCS code used to identify the specific medical service
furnished by the provider.}

\item{hcpcs_desc}{Description of the HCPCS code for the specific medical
service furnished by the provider. HCPCS descriptions associated with
CPT codes are consumer friendly descriptions provided by the AMA. CPT
Consumer Friendly Descriptors are lay synonyms for CPT descriptors that
are intended to help healthcare consumers who are not medical
professionals understand clinical procedures on bills and patient
portals. CPT Consumer Friendly Descriptors should not be used for
clinical coding or documentation. All other descriptions are CMS Level
II descriptions provided in long form. Due to variable length
restrictions, the CMS Level II descriptions have been truncated to
256 bytes. As a result, the same HCPCS description can be associated
with more than one HCPCS code.}

\item{hcpcs_drug}{Identifies whether the HCPCS code for the specific
service furnished by the provider is a HCPCS listed on the Medicare Part
B Drug Average Sales Price (ASP) File.}

\item{pos}{Identifies whether the place of service submitted on the claims
is a facility (value of ‘F’) or non-facility (value of ‘O’). Non-facility
is generally an office setting; however other entities are included
in non-facility.}

\item{year}{Year in YYYY format, between 2013-2020; default is 2020}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on services and procedures provided to
Original Medicare (fee-for-service) Part B (Medical Insurance)
beneficiaries by physicians and other healthcare professionals;
aggregated by geography and service.
}
\details{
The \strong{Geography and Service} dataset contains information on
utilization, payment (allowed amount and Medicare payment), and
submitted charges organized by HCPCS and place of service in the
national table and organized by provider state, HCPCS and place of
service in the state table. The national and state tables also include
a HCPCS drug indicator to identify whether the HCPCS product/service is
a drug as defined from the Medicare Part B Drug ASP list.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service}{Medicare Physician & Other Practitioners: by Geography and Service API}
}
}
}
\note{
Update Frequency: \strong{Annually}
}
\examples{
physician_by_geography(hcpcs_code = "0002A",
                       year = 2020)
\dontrun{
service <- purrr::map_dfr(as.character(2013:2020),
           ~physician_by_service(npi = 1003000126, year = .x))

procedures <- service |>
              dplyr::distinct(hcpcs_cd) |>
              tibble::deframe()

arg_cross <- purrr::cross_df(list(
             x = as.character(2013:2020),
             y = procedures))

# National Level
purrr::map2_dfr(arg_cross$x, arg_cross$y,
~physician_by_geography(geo_level = "National", year = .x, hcpcs_code = .y))

# State Level
purrr::map2_dfr(arg_cross$x, arg_cross$y,
~physician_by_geography(geo_level = "Georgia", year = .x, hcpcs_code = .y))
}
}
