% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescribers.R
\name{prescribers}
\alias{prescribers}
\alias{prescribers_}
\title{Prescriber Utilization & Demographics by Year}
\usage{
prescribers(
  year,
  type,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  brand_name = NULL,
  generic_name = NULL,
  level = NULL,
  opioid = NULL,
  opioidLA = NULL,
  antibiotic = NULL,
  antipsychotic = NULL,
  tidy = TRUE,
  nest = TRUE,
  na.rm = TRUE,
  ...
)

prescribers_(year = rx_years(), ...)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Year data was reported, in \code{YYYY}
format. Run \code{\link[=rx_years]{rx_years()}} to return a vector of the years currently available.}

\item{type}{< \emph{character} > // \strong{required} dataset to query, \code{"Provider"},
\code{"Drug"}, \code{"Geography"}}

\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{first, last, organization}{< \emph{character} > Individual/Organizational
prescriber's name}

\item{credential}{< \emph{character} > Individual prescriber's credentials}

\item{gender}{< \emph{character} > Individual prescriber's gender; \code{"F"} (Female),
\code{"M"} (Male)}

\item{entype}{< \emph{character} > Prescriber entity type; \code{"I"} (Individual),
\code{"O"} (Organization)}

\item{city}{< \emph{character} > City where prescriber is located}

\item{state}{< \emph{character} > State where prescriber is located}

\item{zip}{< \emph{character} > Prescriber’s zip code}

\item{fips}{< \emph{character} > Prescriber's state's FIPS code}

\item{ruca}{< \emph{character} > Prescriber’s RUCA code}

\item{country}{< \emph{character} > Country where prescriber is located}

\item{specialty}{< \emph{character} > Prescriber specialty code reported on the
largest number of claims submitted}

\item{brand_name}{< \emph{character} > Brand name (trademarked name) of the drug
filled, derived by linking the National Drug Codes (NDCs) from PDEs to a
drug information database.}

\item{generic_name}{< \emph{character} > USAN generic name of the drug filled (short
version); A term referring to the chemical ingredient of a drug rather than
the trademarked brand name under which the drug is sold, derived by linking
the National Drug Codes (NDCs) from PDEs to a drug information database.}

\item{level}{< \emph{character} > Geographic level by which the data will be
aggregated:
\itemize{
\item \code{"State"}: Data is aggregated for each state
\item \code{"National"}: Data is aggregated across all states for a given HCPCS Code
}}

\item{opioid}{< \emph{boolean} > \emph{type = 'Geography'}, \code{TRUE} returns Opioid drugs}

\item{opioidLA}{< \emph{boolean} > \emph{type = 'Geography'}, \code{TRUE} returns Long-acting Opioids}

\item{antibiotic}{< \emph{boolean} > \emph{type = 'Geography'}, \code{TRUE} returns antibiotics}

\item{antipsychotic}{< \emph{boolean} > \emph{type = 'Geography'}, \code{TRUE} returns antipsychotics}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{nest}{< \emph{boolean} > // \strong{default:} \code{TRUE} Nest output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}

\item{...}{Pass arguments to \code{\link[=prescribers]{prescribers()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=prescribers]{prescribers()}} allows access to information on prescription drugs provided
to Medicare beneficiaries enrolled in Part D (Prescription Drug Coverage),
by physicians and other health care providers; aggregated by provider, drug
and geography.

The Medicare Part D Prescribers Datasets contain information on prescription
drug events (PDEs) incurred by Medicare beneficiaries with a Part D
prescription drug plan. The Part D Prescribers Datasets are organized by
National Provider Identifier (NPI) and drug name and contains information
on drug utilization (claim counts and day supply) and total drug costs.
}
\section{By Provider}{

\strong{type =}\code{"Provider"}:

The Medicare Part D Prescribers by \strong{Provider} dataset summarizes for each
prescriber the total number of prescriptions that were dispensed, which
include original prescriptions and any refills, and the total drug cost.
}

\section{By Provider and Drug}{

\strong{type =}\code{"Drug"}:

The Medicare Part D Prescribers by \strong{Provider and Drug} dataset contains
the total number of prescription fills that were dispensed and the total
drug cost paid organized by prescribing National Provider Identifier (NPI),
drug brand name (if applicable) and drug generic name.
}

\section{By Geography and Drug}{

\strong{type =}\code{"Geography"}:

For each drug, the \strong{Geography and Drug} dataset includes the total number
of prescriptions that were dispensed, which include original prescriptions
and any refills, and the total drug cost.

The total drug cost includes the ingredient cost of the medication,
dispensing fees, sales tax, and any applicable administration fees and is
based on the amount paid by the Part D plan, Medicare beneficiary, government
subsidies, and any other third-party payers.
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider}{Medicare Part D Prescribers: by Provider}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-provider-and-drug}{Medicare Part D Prescribers: by Provider and Drug}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-part-d-prescribers/medicare-part-d-prescribers-by-geography-and-drug}{Medicare Part D Prescribers: by Geography and Drug}
\item \href{https://data.cms.gov/sites/default/files/2021-08/mup_dpr_ry21_20210819_technical_specifications.pdf}{Medicare Part D Prescribers Technical Specifications}
}

\emph{Update Frequency:} \strong{Annually}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
prescribers(year = 2020,
            type = 'Provider',
            npi = 1003000423)

prescribers(year = 2019,
            type = 'Drug',
            npi = 1003000126)

prescribers(year = 2021,
            type = 'Geography',
            brand_name = 'Clotrimazole-Betamethasone')

prescribers(year = 2017,
            type = 'Geography',
            level = 'National',
            brand_name = 'Paroxetine Hcl')

prescribers(year = 2017,
            type = 'Geography',
            opioid = TRUE)

# Use the years helper function to
# retrieve results for every year:
rx_years() |>
map(\(x) prescribers(year = x,
                     type = 'Provider',
                     npi = 1043477615)) |>
list_rbind()

# Parallelized version
prescribers_(type = 'Provider',
             npi = 1043477615)

prescribers_(type = 'Drug',
             npi = 1003000423)

prescribers_(type = 'Geography',
             level = 'National',
             generic_name = 'Mirabegron')
\dontshow{\}) # examplesIf}
}
