% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prov_npi_nppes.R
\name{prov_npi_nppes}
\alias{prov_npi_nppes}
\title{Search the NPPES by NPI}
\usage{
prov_npi_nppes(npi)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}
}
\value{
A tibble containing the NPI(s) searched for,
the date-time the search was performed, and
a list-column of the results.
}
\description{
Search the NPPES by NPI
}
\examples{
prov_npi_nppes(1528060837)
}
