% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_enrollment.R
\name{provider_enrollment}
\alias{provider_enrollment}
\title{Search the Medicare Fee-For-Service Public Provider Enrollment API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
provider_enrollment(
  npi = NULL,
  pecos_id = NULL,
  enroll_id = NULL,
  prov_type_code = NULL,
  prov_type_desc = NULL,
  city = NULL,
  state = NULL,
  zipcode = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  org_name = NULL,
  gender = NULL,
  enroll_id_reassign = NULL,
  enroll_id_receive = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{npi}{An NPI is a 10-digit unique numeric identifier that all providers
must obtain before enrolling in Medicare. It is assigned to health care
providers upon application through the National Plan and Provider
Enumeration System (NPPES).}

\item{pecos_id}{Provider associate level variable (PAC ID) from PECOS
database used to link across tables. A PAC ID is a 10-digit unique
numeric identifier that is assigned to each individual or organization
in PECOS. All entity-level information (e.g., tax identification numbers
and organizational names) is linked through the PAC ID. A PAC ID may be
associated with multiple Enrollment IDs if the individual or
organization enrolled multiple times under different circumstances.}

\item{enroll_id}{Provider enrollment ID from PECOS database used to link
across tables. An Enrollment ID is a 15-digit unique alphanumeric
identifier that is assigned to each new provider enrollment application.
All enrollment-level information (e.g., enrollment type, enrollment
state, provider specialty and reassignment of benefits) is linked
through the Enrollment ID.}

\item{prov_type_code}{Provider enrollment application and enrollment
specialty type. This field shows the provider’s primary specialty code.
For practitioners and DME suppliers, please see the Secondary Specialty
file for a list of secondary specialties (when applicable). Only about
20\% of practitioners and DME suppliers have at least one secondary
specialty.}

\item{prov_type_desc}{Provider enrollment application and enrollment
specialty type description}

\item{city}{Provider practice location city. Each practice location is
linked to an Enrollment ID. If any other provider reassigns benefits to
an Enrollment ID listed in the Practice Location file, the practice
location is also associated with that other provider.}

\item{state}{Provider enrollment state, abbreviated location. Providers
enroll at the state level, so one PAC ID may be associated with multiple
ENRLMT_IDs and multiple STATE_CD values.}

\item{zipcode}{Provider practice location zip code. Each practice location
is linked to an Enrollment ID. If any other provider reassigns benefits
to an Enrollment ID listed in the Practice Location file, the practice
location is also associated with that other provider.}

\item{first_name}{Individual provider first name}

\item{middle_name}{Individual provider middle name}

\item{last_name}{Individual provider last name}

\item{org_name}{Organizational provider name}

\item{gender}{Individual provider gender; F(female), M(male), 9(unknown)}

\item{enroll_id_reassign}{Enrollment ID of the provider reassigning benefits
to bill to Medicare. To find enrollment information for providers
reassigning their benefits to bill to Medicare, link
REASGN_BNFT_ENRLMT_ID to ENRLMT_ID in the Enrollment file.}

\item{enroll_id_receive}{Enrollment ID of the provider receiving
reassignment benefits to bill to Medicare. To find enrollment
information for providers receiving reassignment benefits to bill to
Medicare, link RCV_BNFT_ENRLMT_ID to ENRLMT_ID in the Enrollment file.}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on a point in time snapshot of enrollment
level data for providers actively enrolled in Medicare.
}
\details{
The Medicare Fee-For-Service Public Provider Enrollment dataset
includes information on providers who are actively approved to bill
Medicare or have completed the 855O at the time the data was pulled
from the Provider Enrollment and Chain Ownership System (PECOS). The
release of this provider enrollment data is not related to other
provider information releases such as Physician Compare or Data
Transparency.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Medicare Fee-For-Service Public Provider Enrollment API}
}
}
}
\note{
Update Frequency: \strong{Quarterly}
}
\examples{
provider_enrollment(npi = 1003026055)
provider_enrollment(last = "phadke", first = "radhika")
provider_enrollment(prov_type_desc = "endocrinology")
\dontrun{
# Won't filter, returns entire dataset
provider_enrollment(taxonomy_desc = "Trauma Surgery")
}
}
