% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_enrollment.R
\name{provider_enrollment}
\alias{provider_enrollment}
\title{Provider Enrollment in Medicare}
\usage{
provider_enrollment(
  npi = NULL,
  pac_id = NULL,
  enroll_id = NULL,
  specialty_code = NULL,
  specialty = NULL,
  state = NULL,
  first_name = NULL,
  middle_name = NULL,
  last_name = NULL,
  organization_name = NULL,
  gender = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{10-digit National Provider Identifier}

\item{pac_id}{10-digit Provider associate level variable. Links all
entity-level information and may be associated with multiple enrollment IDs
if the individual or organization enrolled multiple times.}

\item{enroll_id}{15-digit Provider enrollment ID. Assigned to each new
provider enrollment application. Links all enrollment-level information
(enrollment type, state, reassignment of benefits).}

\item{specialty_code}{Enrollment primary specialty type code.}

\item{specialty}{Enrollment specialty type description.}

\item{state}{Enrollment state abbreviation. Since Providers enroll at the state level,
a PAC ID can be associated with multiple enrollment IDs and states.}

\item{first_name}{Individual provider first name}

\item{middle_name}{Individual provider middle name}

\item{last_name}{Individual provider last name}

\item{organization_name}{Organizational provider name}

\item{gender}{Individual provider gender:
\itemize{
\item \code{F}: Female
\item \code{M}: Male
\item \code{9}: Unknown or Organizational provider
}}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
\link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{provider_enrollment()} allows you to access enrollment level
data on individual and organizational providers that are actively approved
to bill Medicare.
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Provider Enrollment API}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Provider Enrollment Data Dictionary}
}
}

\section{Update Frequency}{
 Quarterly
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
provider_enrollment(npi = 1417918293, specialty_code = "14-41")
provider_enrollment(first_name = "DEBRA", last_name = "FROMER")
provider_enrollment(organization_name = "Elizabethtown Community Hospital", state = "NY")
provider_enrollment(specialty = "PRACTITIONER - ENDOCRINOLOGY", state = "AK", gender = "F")
provider_enrollment(pac_id = 2860305554, gender = "9")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=order_refer]{order_refer()}}, \code{\link[=opt_out]{opt_out()}}, \code{\link[=pending_applications]{pending_applications()}}
}
