% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_icd10.R
\name{provider_icd10}
\alias{provider_icd10}
\title{Search the National Library of Medicine's ICD-10-CM API}
\usage{
provider_icd10(code = NULL, term = NULL, field = "both", limit = 500)
}
\arguments{
\item{code}{All or part of an ICD-10-CM code}

\item{term}{Associated term describing an ICD-10 code}

\item{field}{options are "code" or "both"; default is "both"}

\item{limit}{default is 700}
}
\value{
A \code{\link[=tibble]{tibble()}} containing the search results.
}
\description{
\link{provider_icd10} allows you to search the NLM's ICD-10-CM
API by code or associated term.
}
\section{The NIH's Clinical Table Search Service API for ICD-10-CM}{
ICD-10-CM (International Classification of Diseases, 10th Revision,
Clinical Modification) is a medical coding system for classifying
diagnoses and reasons for visits in U.S. health care settings.
\href{http://www.cdc.gov/nchs/icd/icd10cm.htm}{Learn more about ICD-10-CM.}
\subsection{Current Version}{

ICD-10-CM 2023
}

\subsection{Data Source}{

National Institute of Health/National Library of Medicine
}

\subsection{Links}{
\itemize{
\item \href{https://clinicaltables.nlm.nih.gov/apidoc/icd10cm/v3/doc.html}{NIH NLM Clinical Table Service ICD-10-CM API}
}
}
}

\examples{
# Returns the seven ICD-10-CM codes beginning with "A15".
provider_icd10(code = "A15", limit = 10)

# Returns the first 20 ICD-10-CM codes associated with tuberculosis.
provider_icd10(term = "tuber", limit = 20)

# Returns the two ICD-10-CM codes associated with pleurisy.
provider_icd10(term = "pleurisy")

# Returns first 500 "Z" codes (Z00.00 - Z46.9)
provider_icd10(code = "z", field = "code")

# Must specify `field` as "code" or it will search term, for example:
# Returns codes with the letter "Z" in their associated term
provider_icd10(code = "z", limit = 10)

# Returns the 95 codes beginning with "Z5-"
provider_icd10(code = "z5")
}
