% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_mooa.R
\name{provider_mooa}
\alias{provider_mooa}
\title{Search the Medicare Opt Out Affidavits API}
\usage{
provider_mooa(npi = NULL, last = NULL, first = NULL, clean_names = TRUE)
}
\arguments{
\item{npi}{10-digit National Provider Identifier (NPI)}

\item{last}{Provider's last name}

\item{first}{Provider's first name}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of practitioners who are currently
opted out of Medicare.
}
\section{Medicare Opt Out Affidavits Dataset}{
The Opt Out Affidavits dataset provides
information on providers who have decided not to
participate in Medicare. It contains the provider's NPI,
specialty, address, and effective dates.
\subsection{Data Update Frequency}{

Quarterly
}

\subsection{Data Source}{

Centers for Medicare & Medicaid Services
}

\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/opt-out-affidavits}{Medicare Opt Out Affidavits API}
}
}
}

\examples{
\dontrun{
# Search by NPI
provider_mooa(npi = 1114974490)

# Search by Last Name
provider_mooa(last = "Altchek")

# Returns empty list i.e., provider is not in the database
provider_mooa(1326011057)
}
}
