% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_unpack.R
\name{provider_unpack}
\alias{provider_unpack}
\title{Unpack NPI NPPES Search Results}
\usage{
provider_unpack(df)
}
\arguments{
\item{df}{data.frame or tbl_df, response from provider_nppes()}
}
\value{
A tibble containing the unpacked list-columns of a search
performed with the provider_nppes() function.
}
\description{
Unpack NPI NPPES Search Results
}
\examples{
\dontrun{
provider_nppes(1528060837) |>
provider_unpack()

provider_nppes(city = "Atlanta", state = "GA") |>
provider_unpack()

provider_nppes(first = "John", city = "Baltimore", state = "MD") |>
provider_unpack()
}

}
