% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_unpack.R
\name{provider_unpack}
\alias{provider_unpack}
\title{Unpack NPPES NPI Registry Search Results}
\usage{
provider_unpack(df, clean_names = TRUE)
}
\arguments{
\item{df}{\link[tibble:tibble-package]{tibble} response from \code{\link[=provider_nppes]{provider_nppes()}}}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the unpacked
list-columns of a search performed with the \code{\link[=provider_nppes]{provider_nppes()}} function.
}
\description{
\code{provider_unpack()} allows you to unpack the list-columns
within the response from \code{provider_nppes()}.
}
\examples{
\dontrun{
# Single NPI ==============================================================
provider_nppes(npi = 1528060837) |>
provider_unpack()

# City, state, country ====================================================
provider_nppes(city = "Atlanta",
               state = "GA",
               country = "US") |>
provider_unpack()

# First name, city, state  ================================================
provider_nppes(first = "John",
               city = "Baltimore",
               state = "MD") |>
provider_unpack()
}
}
\seealso{
\code{\link[=provider_nppes]{provider_nppes()}}
}
