% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-enrolled.R
\name{providers}
\alias{providers}
\title{Provider Enrollment in Medicare}
\usage{
providers(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  prov_type = NULL,
  prov_desc = NULL,
  state = NULL,
  org_name = NULL,
  multi = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{pac}{\verb{<chr>} PECOS Associate Control ID}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{first, middle, last}{\verb{<chr>} Individual provider's name}

\item{prov_type}{\verb{<chr>} Enrollment specialty code}

\item{prov_desc}{\verb{<chr>} Enrollment specialty description}

\item{state}{\verb{<chr>} Enrollment state, full or abbreviation}

\item{org_name}{\verb{<chr>} Organizational provider's name}

\item{multi}{\verb{<lgl>} Provider has multiple NPIs}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Enrollment data on individual and organizational providers that are
actively approved to bill Medicare.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
providers(count = TRUE)

providers(count = TRUE, org_name = not_blank())

providers(org_name = starts_with("Z"))
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{API: Medicare Provider Supplier Enrollment}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Provider Enrollment Data Dictionary}
}
}
