% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-providers.R
\name{providers}
\alias{providers}
\title{Provider Enrollment in Medicare}
\usage{
providers(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  spec = NULL,
  specialty = NULL,
  state = NULL,
  org_name = NULL,
  multi = NULL,
  count = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} 10-digit Individual National Provider Identifier}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control ID}

\item{enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{first, middle, last}{\verb{<chr>} Individual provider's name}

\item{spec}{\verb{<chr>} Enrollment specialty code}

\item{specialty}{\verb{<chr>} Enrollment specialty description}

\item{state}{\verb{<chr>} Enrollment state, full or abbreviation}

\item{org_name}{\verb{<chr>} Organization name}

\item{multi}{\verb{<lgl>} Provider has multiple NPIs}

\item{count}{\verb{<lgl>} Return the dataset's total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Enrollment data on individual and organizational providers that are
actively approved to bill Medicare.
}
\examples{
providers(count = TRUE)
providers(enid = "I20040309000221")
providers(npi = 1417918293)
providers(pac = 2860305554)
providers(state = "AK")
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{API: Medicare Provider Supplier Enrollment}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Provider Enrollment Data Dictionary}
}
}
