% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providers}
\alias{providers}
\title{Provider Enrollment in Medicare}
\usage{
providers(
  npi = NULL,
  pac_id = NULL,
  enroll_id = NULL,
  specialty_code = NULL,
  specialty = NULL,
  state = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  organization = NULL,
  gender = NULL,
  tidy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{pac_id}{< \emph{integer} > 10-digit provider associate level variable. Links
all entity-level information and may be associated with multiple enrollment
IDs if the individual or organization enrolled multiple times.}

\item{enroll_id}{< \emph{character} > 15-digit provider enrollment ID. Assigned to
each new provider enrollment application. Links all enrollment-level
information (enrollment type, state, reassignment of benefits).}

\item{specialty_code}{< \emph{character} > Enrollment primary specialty type code}

\item{specialty}{< \emph{character} > Enrollment specialty type description}

\item{state}{< \emph{character} > Enrollment state abbreviation/full name.
Providers enroll at the state level, so a PAC ID can be associated with
multiple enrollment IDs and states.}

\item{first, middle, last}{Individual provider's first/middle/last name}

\item{organization}{Organizational provider's name}

\item{gender}{Individual provider gender. Options are:
\itemize{
\item \code{"F"}: Female
\item \code{"M"}: Male
\item \code{"9"}: Unknown (or Organizational provider)
}}

\item{tidy}{< \emph{boolean} > Tidy output; default is \code{TRUE}.}

\item{na.rm}{< \emph{boolean} > Remove empty rows and columns; default is \code{TRUE}.}
}
\value{
\link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit NPI \cr
   \code{pac_id} \tab 10-digit PAC ID \cr
   \code{enroll_id} \tab 15-digit provider enrollment ID \cr
   \code{specialty_code} \tab Enrollment primary specialty type code \cr
   \code{specialty} \tab Enrollment specialty type description \cr
   \code{state} \tab Enrollment state \cr
   \code{first} \tab Individual provider's first name \cr
   \code{middle} \tab Individual provider's middle name \cr
   \code{last} \tab Individual provider's last name \cr
   \code{gender} \tab Individual provider's gender \cr
   \code{organization} \tab Organizational provider's name \cr
}
}
\description{
\code{providers()} allows you to access enrollment level
data on individual and organizational providers that are actively approved
to bill Medicare.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Provider Enrollment API}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Provider Enrollment Data Dictionary}
}

\emph{Update Frequency:} \strong{Quarterly}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
providers(npi = 1417918293, specialty_code = "14-41")
providers(pac_id = 2860305554, gender = "9")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=order_refer]{order_refer()}}, \code{\link[=opt_out]{opt_out()}}, \code{\link[=pending]{pending()}}
}
