% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{providers}
\alias{providers}
\title{Provider Enrollment in Medicare}
\usage{
providers(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  specialty_code = NULL,
  specialty_description = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  organization = NULL,
  state = NULL,
  gender = NULL,
  tidy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit Individual National Provider Identifier}

\item{pac}{< \emph{integer} > 10-digit PECOS Associate Control ID}

\item{enid}{< \emph{character} > 15-digit Medicare Enrollment ID}

\item{specialty_code}{< \emph{character} > Enrollment specialty code}

\item{specialty_description}{< \emph{character} > Enrollment specialty description}

\item{first, middle, last}{< \emph{character} > Individual provider's name}

\item{organization}{< \emph{character} > Organizational provider's name}

\item{state}{< \emph{character} > Enrollment state, full or abbreviation}

\item{gender}{< \emph{character} > Individual provider's gender: \code{"F"} (Female),
\code{"M"} (Male), \code{"9"} (Unknown/Organization)}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}
}
\value{
\link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{npi} \tab 10-digit NPI \cr
   \code{pac} \tab 10-digit PAC ID \cr
   \code{enid} \tab 15-digit provider enrollment ID \cr
   \code{specialty_code} \tab Enrollment primary specialty type code \cr
   \code{specialty_description} \tab Enrollment specialty type description \cr
   \code{first} \tab Individual provider's first name \cr
   \code{middle} \tab Individual provider's middle name \cr
   \code{last} \tab Individual provider's last name \cr
   \code{organization} \tab Organizational provider's name \cr
   \code{state} \tab Enrollment state \cr
   \code{gender} \tab Individual provider's gender \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{providers()} allows you to access enrollment level data on individual and
organizational providers that are actively approved to bill Medicare.
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-fee-for-service-public-provider-enrollment}{Provider Enrollment API}
\item \href{https://data.cms.gov/resources/medicare-fee-for-service-public-provider-enrollment-data-dictionary}{Provider Enrollment Data Dictionary}
}

\emph{Update Frequency:} \strong{Quarterly}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
providers(enid = "I20040309000221")
providers(npi = 1417918293, specialty_code = "14-41")
providers(pac = 2860305554, gender = "9")
\dontshow{\}) # examplesIf}
}
