% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych.R
\name{psych}
\alias{psych}
\title{Inpatient Psychiatric Facilities}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/q9vs-r7wp}{API: Inpatient Psychiatric Facility Quality Measure Data - by Facility}
}
}
\usage{
psych(
  ccn = NULL,
  name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  rating = NULL,
  count = FALSE
)
}
\arguments{
\item{ccn}{\verb{<chr>} desc}

\item{name}{\verb{<chr>} desc}

\item{city, state, zip, county}{\verb{<chr>} desc}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
This dataset includes provider-level data for quality measures
included under the IPFQR program, including HBIPS, SUB, TOB, Transition
Record (TR), Screening for Metabolic Disorders (SMD), FAPH, IMM,
Readmissions (READM), and Medication Continuation (MedCont, formerly known
as MedCoPsy). Psychiatric facilities that are eligible for the Inpatient
Psychiatric Facility Quality Reporting (IPFQR) program are required to meet
all program requirements, otherwise their Medicare payments may be reduced.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
psych(count = TRUE)
psych(state = "GA")
\dontshow{\}) # examplesIf}
}
