% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-quality.R
\name{quality}
\alias{quality}
\alias{metrics}
\title{Quality Payment Program}
\usage{
quality(
  npi = NULL,
  state = NULL,
  size = NULL,
  specialty = NULL,
  years = NULL,
  patients = NULL,
  charges = NULL,
  services = NULL,
  final_score = NULL,
  adjustment = NULL,
  count = FALSE
)

metrics(year = NULL)
}
\arguments{
\item{npi}{\verb{<int>} description}

\item{state}{description}

\item{size}{\verb{<int>} description}

\item{specialty}{description}

\item{years}{\verb{<int>} description}

\item{patients}{\verb{<int>} description}

\item{charges}{\verb{<int>} description}

\item{services}{\verb{<int>} description}

\item{final_score}{\verb{<int>} description}

\item{adjustment}{\verb{<int>} description}

\item{count}{\verb{<lgl>} Return the total row count}

\item{year}{\verb{<int>} A vector of years from 2018 to 2025}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
The Quality Payment Program (QPP) Experience dataset provides participation
and performance information in the Merit-based Incentive Payment System
(MIPS) during each performance year. They cover eligibility and
participation, performance categories, and final score and payment
adjustments.
}
\details{
The dataset provides additional details at the TIN/NPI level on what was
published in the previous performance year. You can sort the data by
variables like clinician type, practice size, scores, and payment
adjustments.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
quality(count = TRUE)

quality(count = TRUE, state = "GA")

quality(npi = 1043245657)

quality(npi = c(1003026055, 1316939655))

metrics()
\dontshow{\}) # examplesIf}
}
