% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpp.R
\name{quality_eligibility}
\alias{quality_eligibility}
\title{Search the CMS Quality Payment Program Eligibility API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
quality_eligibility(year, npi, tidy = TRUE)
}
\arguments{
\item{year}{year}

\item{npi}{The NPI assigned to the clinician when they enrolled in Medicare.}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Data pulled from across CMS that is used to create an
eligibility determination for a clinician. Using what CMS knows about a
clinician from their billing patterns and enrollments, eligibility is
"calculated" multiple times before and during the performance year.
}
\details{
The Quality Payment Program (QPP) Eligibility System pulls together
data from across the Centers for Medicare and Medicaid Services (CMS) to
create an eligibility determination for every clinician in the system.
Using what CMS knows about a clinician from their billing patterns and
enrollments, eligibility is "calculated" multiple times before and during
the performance year. Information can be obtained primarily by the
Clinician type. You can query the Clinician type by passing in an National
Provider Identifier, or NPI. This number is a unique 10-digit
identification number issued to health care providers in the United
States by CMS. The information contained in these endpoints includes
basic enrollment information, associated organizations, information
about those organizations, individual and group special status
information, and in the future, any available Alternative Payment Model
(APM) affiliations.
\subsection{Links}{
\itemize{
\item \href{https://cmsgov.github.io/qpp-eligibility-docs/}{QPP Eligibility API Documentation}
}
}
}
\note{
Update Frequency: \strong{Annually}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
quality_payment(year = 2020, state = "GA")
\dontshow{\}) # examplesIf}
}
