% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_eligibility.R
\name{quality_eligibility}
\alias{quality_eligibility}
\title{Quality Payment Program Eligibility}
\usage{
quality_eligibility(year, npi, tidy = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} QPP performance year, in \code{YYYY}format.
Run \code{\link[=qpp_years]{qpp_years()}} to return a vector of the years currently available.}

\item{npi}{< \emph{integer} > 10-digit Individual National Provider Identifier
assigned to the clinician when they enrolled in Medicare. Multiple rows for
the same NPI indicate multiple TIN/NPI combinations.}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{FALSE} Remove empty rows and columns}

\item{...}{For future use.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=quality_eligibility]{quality_eligibility()}} allows the user access to information on eligibility
in the Merit-based Incentive Payment System (MIPS) and Advanced Alternative
Payment Models (APMs) tracks.

Data pulled from across CMS is used to create an eligibility determination
for a clinician. Using what CMS knows about a clinician from their billing
patterns and enrollments, eligibility is "calculated" multiple times before
and during the performance year.
}
\section{Quality Payment Program (QPP) Eligibility}{

The QPP Eligibility System aggregates data from across CMS to create an
eligibility determination for every clinician in the system. Using what CMS
knows about a clinician from their billing patterns and enrollments,
eligibility is "calculated" multiple times before and during the performance
year.

The information contained in these endpoints includes basic enrollment
information, associated organizations, information about those organizations,
individual and group special status information, and in the future, any
available Alternative Payment Model (APM) affiliations.
}

\section{Types}{

\itemize{
\item \strong{Clinicians} represent healthcare providers and are referenced using a NPI.
\item \strong{Practices} represent a clinician or group of clinicians that assign their
billing rights to the same TIN. These are represented with a TIN, EIN, or
SSN, and querying by this number requires an authorization token.
\item \strong{Virtual Groups} represent a combination of two or more TINs with certain
characteristics, represented by a virtual group identifier, also requiring an
authorization token.
\item \strong{APM Entities} represent a group of practices which participate in an
APM, characterized by an APM Entity ID.
}
}

\section{Links}{

\itemize{
\item \href{https://cmsgov.github.io/qpp-eligibility-docs/}{QPP Eligibility API Documentation}
\item \href{https://qpp.cms.gov/api/eligibility/docs/?urls.primaryName=Eligibility\%2C\%20v6}{QPP Eligibility & MVP/CAHPS/Subgroups Registration Services (v6)}
}
}

\section{Update Frequency}{
 \strong{Annually}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
quality_eligibility(year = 2020, npi = 1144544834)
\dontshow{\}) # examplesIf}
}
