% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_metrics.R
\name{quality_metrics}
\alias{quality_metrics}
\title{Quality Payment Program}
\usage{
quality_metrics(year)
}
\arguments{
\item{year}{\verb{<int>} A vector of years from 2018 to 2025}
}
\description{
Quality Payment Program
}
