% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_payment.R
\name{quality_payment}
\alias{quality_payment}
\title{Quality Payment Program Experience}
\usage{
quality_payment(
  year,
  npi = NULL,
  state = NULL,
  specialty = NULL,
  type = NULL,
  tidy = TRUE,
  nest = TRUE,
  ...
)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} QPP performance year, in \code{YYYY}format.
Run \code{\link[=qpp_years]{qpp_years()}} to return a vector of the years currently available.}

\item{npi}{< \emph{integer} > 10-digit Individual National Provider Identifier
assigned to the clinician when they enrolled in Medicare. Multiple rows for
the same NPI indicate multiple TIN/NPI combinations.}

\item{state}{< \emph{character} > State or US territory code location of the TIN
associated with the clinician.}

\item{specialty}{< \emph{character} > Specialty corresponding to the type of
service that the clinician submitted most on their Medicare Part B claims
for this TIN/NPI combination.
\itemize{
\item Nurse Practitioner
\item Internal Medicine
\item Physician Assistant
\item Family Practice
\item Emergency Medicine
\item Diagnostic Radiology
\item Anesthesiology
\item Neurology
\item Cardiology
}}

\item{type}{< \emph{character} > Participation type; level at which the
performance data was collected, submitted or reported for the final score
attributed to the clinician; drives most of the data returned.
\itemize{
\item \code{"Group"}
\item \code{"Individual"}
\item \code{"MIPS APM"}
}}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{nest}{< \emph{boolean} > // \strong{default:} \code{TRUE} Nest \code{statuses} & \code{measures}}

\item{...}{For future use.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=quality_payment]{quality_payment()}}  allows the user access to information on participation
and performance in the Merit-based Incentive Payment System (MIPS) and
Advanced Alternative Payment Models (APMs) tracks.
}
\section{Quality Payment Program (QPP) Experience}{

The QPP dataset provides participation and performance information in the
Merit-based Incentive Payment System (MIPS) during each performance year.
They cover eligibility and participation, performance categories, final
score and payment adjustments. The dataset provides additional details at the
TIN/NPI level on what was published in the previous performance year.
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/quality-of-care/quality-payment-program-experience}{Quality Payment Program Experience}
}
}

\section{Update Frequency}{

\strong{Annually}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
quality_payment(year = 2020, npi = 1144544834)
\dontshow{\}) # examplesIf}
}
