% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassignments.R
\name{reassignments}
\alias{reassignments}
\title{Reassignment of Benefits}
\usage{
reassignments(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  last = NULL,
  state = NULL,
  specialty = NULL,
  org_name = NULL,
  org_pac = NULL,
  org_enid = NULL,
  org_state = NULL,
  count = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} 10-digit National Provider Identifier}

\item{pac}{\verb{<chr>} 10-digit PECOS Associate Control ID}

\item{enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{first, last}{\verb{<chr>} Provider's name}

\item{state}{\verb{<chr>} Enrollment state abbreviation}

\item{specialty}{\verb{<chr>} Enrollment specialty}

\item{org_name}{\verb{<chr>} Legal business name}

\item{org_pac}{\verb{<chr>} 10-digit PECOS Associate Control ID}

\item{org_enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{org_state}{\verb{<chr>} Enrollment state abbreviation}

\item{count}{\verb{<lgl>} Return the dataset's total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Returns information about:
\itemize{
\item Individual providers who are reassigning benefits or are an employee of
\item Organizational/Group providers who are receiving reassignment of benefits from or are the employer of the individual provider
}

It provides information regarding the physician and the group practice they
reassign their billing to, including individual employer association counts.
}
\examples{
reassignments(count = TRUE)
reassignments(org_enid = "I20070209000135")
reassignments(pac = 9830437441)
reassignments(org_pac = 3173525888)
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list}{API: Medicare Revalidation Reassignment List}
}
}
