% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassignments.R
\name{reassignments}
\alias{reassignments}
\title{Reassignment of Benefits}
\usage{
reassignments(
  npi = NULL,
  pac = NULL,
  enid = NULL,
  first = NULL,
  last = NULL,
  state = NULL,
  specialty = NULL,
  organization = NULL,
  pac_org = NULL,
  enid_org = NULL,
  state_org = NULL,
  record = NULL,
  tidy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > Individual provider's 10-digit National Provider Identifier}

\item{pac}{< \emph{integer} > Individual provider's 10-digit PECOS Associate Control ID}

\item{enid}{< \emph{character} > Individual provider's 15-digit Medicare Enrollment ID}

\item{first, last}{< \emph{character} > Individual provider's name}

\item{state}{< \emph{character} > Individual provider's enrollment state}

\item{specialty}{< \emph{character} > Individual provider's enrollment specialty}

\item{organization}{< \emph{character} > Organizational provider's legal business name}

\item{pac_org}{< \emph{integer} > Organizational provider's 10-digit PECOS Associate Control ID}

\item{enid_org}{< \emph{character} > Organizational provider's 15-digit Medicare Enrollment ID}

\item{state_org}{< \emph{character} > Organizational provider's enrollment state}

\item{record}{< \emph{character} > Identifies whether the record is for reassignment (\code{"R"}) or employment (\code{"E"})}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link[=reassignments]{reassignments()}} returns information about:
\itemize{
\item Individual providers who are reassigning benefits or are an employee of
\item Organizational/Group providers who are receiving reassignment of benefits from or are the employer of the individual provider
}

Links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list}{Medicare Revalidation Reassignment List API}
}

\emph{Update Frequency:} \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
reassignments(enid = "I20200929003184")
reassignments(pac = 9830437441)
reassignments(pac_org = 3173525888)
\dontshow{\}) # examplesIf}
}
