% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalidation_date.R
\name{revalidation_date}
\alias{revalidation_date}
\title{Search the Medicare Revalidation Due Date List API}
\usage{
revalidation_date(
  npi = NULL,
  enroll_id = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  state = NULL,
  enrollment_type = NULL,
  specialty_description = NULL,
  tidy = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{enroll_id}{< \emph{character} > 15-digit provider enrollment ID}

\item{first, last}{< \emph{character} > Individual provider's first/last name}

\item{organization}{< \emph{character} > Organizational provider's legal business name}

\item{state}{< \emph{character} > Enrollment state}

\item{enrollment_type}{< \emph{integer} > Provider enrollment type:
\itemize{
\item \code{1}: Part A
\item \code{2}: DME
\item \code{3}: Non-DME Part B
}}

\item{specialty_description}{< \emph{character} > Enrollment specialty}

\item{tidy}{Tidy output; default is \code{TRUE}.}

\item{na.rm}{< \emph{boolean} > Remove empty rows and columns; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on revalidation due dates for Medicare providers.
Medicare Providers must validate their enrollment record every three or
five years. CMS sets every Provider’s Revalidation due date at the end
of a month and posts the upcoming six to seven months of due dates
online. A due date of ‘TBD’ means that CMS has not set the due date yet.
These lists are refreshed every two months and two months’ worth of due
dates are appended to the list
}
\details{
The Revalidation Due Date List dataset contains revalidation due
dates for Medicare providers who are due to revalidate in the following
six months. If a provider's due date does not fall within the ensuing
six months, the due date is marked 'TBD'. In addition the dataset also
includes subfiles with reassignment information for a given provider
as well as due date listings for clinics and group practices and
their providers.

Links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-due-date-list}{Medicare Revalidation Due Date API}
}

\emph{Update Frequency:} \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
revalidation_date(enroll_id = "I20031110000070")
revalidation_date(enroll_id = "O20110620000324")
revalidation_date(state = "FL", enrollment_type = 3, specialty = "General Practice")
\dontshow{\}) # examplesIf}
}
