% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalidation_date.R
\name{revalidation_date}
\alias{revalidation_date}
\title{Search the Medicare Revalidation Due Date API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
revalidation_date(
  enroll_id = NULL,
  npi = NULL,
  first = NULL,
  last = NULL,
  org_name = NULL,
  enroll_state = NULL,
  enroll_type = NULL,
  prov_type = NULL,
  enroll_spec = NULL,
  reval_date = NULL,
  adjust_date = NULL,
  indiv_reassign = NULL,
  benefits_reassign = NULL,
  version = "Nov",
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{enroll_id}{Enrollment ID}

\item{npi}{National Provider Identifier (NPI)}

\item{first}{First name of individual provider}

\item{last}{Last name of individual provider}

\item{org_name}{Legal business name of organizational provider}

\item{enroll_state}{Enrollment state}

\item{enroll_type}{Provider type code (\code{1} if Part A; \code{2} if DME;
\code{3} if Non-DME Part B)}

\item{prov_type}{Provider type description}

\item{enroll_spec}{Enrollment specialty}

\item{reval_date}{Previously assigned revalidation due date (blank if not
previously assigned a due date)}

\item{adjust_date}{Next revalidation due date (blank if not yet assigned)}

\item{indiv_reassign}{Number of individual enrollment associations that are
reassigning benefits to or are employed by the organizational provider}

\item{benefits_reassign}{Number of organizational enrollment associations
to which the individual provider reassigns benefits or is employed by}

\item{version}{dataset version; current possible values are "Feb" - "Nov"}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on revalidation due dates for Medicare providers.
Medicare Providers must validate their enrollment record every three or
five years. CMS sets every Provider’s Revalidation due date at the end
of a month and posts the upcoming six to seven months of due dates
online. A due date of ‘TBD’ means that CMS has not set the due date yet.
These lists are refreshed every two months and two months’ worth of due
dates are appended to the list
}
\details{
The Revalidation Due Date List dataset contains revalidation due
dates for Medicare providers who are due to revalidate in the following
six months. If a provider's due date does not fall within the ensuing
six months, the due date is marked 'TBD'. In addition the dataset also
includes subfiles with reassignment information for a given provider
as well as due date listings for clinics and group practices and
their providers.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-due-date-list}{Medicare Revalidation Due Date API}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
revalidation_date(enroll_id = "I20031110000070", npi = 1184699621)

revalidation_date(first = "Eric", last = "Byrd")

revalidation_date(enroll_state = "FL", enroll_type = "3",
                  enroll_spec = "General Practice")
\dontrun{
npi_list <- rep(c(1003026055, 1316405939,
                  1720392988, 1518184605), each = 8)
months <- rep(c("Nov", "Oct", "Sep", "Aug",
                "Jul", "Jun", "Apr", "Feb"), times = 4)
purrr::map2_dfr(npi_list, months, ~revalidation_date(npi = .x,
                                                     version = .y))
}
}
