% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalidation_group.R
\name{revalidation_group}
\alias{revalidation_group}
\title{Search the Medicare Revalidation Clinic Group Practice Reassignment API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
revalidation_group(
  npi = NULL,
  enroll_id = NULL,
  first_name = NULL,
  last_name = NULL,
  state = NULL,
  specialty = NULL,
  pac_id_group = NULL,
  enroll_id_group = NULL,
  business_name = NULL,
  state_group = NULL,
  record_type = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{NPI of provider who is reassigning their benefits or is an
employee}

\item{enroll_id}{Enrollment ID of provider reassigning their benefits or is an employee}

\item{first_name}{First name of provider who is reassigning their benefits
or is an employee}

\item{last_name}{Last name of provider who is reassigning their benefits or
is an employee}

\item{state}{Enrollment state of provider who is reassigning their
benefits or is an employee}

\item{specialty}{Enrollment specialty of the provider who is
reassigning their benefits or is an employee}

\item{pac_id_group}{PAC ID of provider who is receiving reassignment or is
the employer}

\item{enroll_id_group}{Enrollment ID of provider who is receiving
reassignment or is the employer}

\item{business_name}{Legal business name of provider who is receiving
reassignment or is the employer}

\item{state_group}{Enrollment state of provider who is receiving
reassignment or is the employer}

\item{record_type}{Identifies whether the record is for a reassignment
(\code{Reassignment}) or employment (\verb{Physician Assistant})}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Information on clinic group practice revalidation for Medicare
enrollment.
}
\details{
The Revalidation Clinic Group Practice Reassignment dataset
provides information between the physician and the group practice they
reassign their billing to. It also includes individual employer
association counts and the revalidation dates for the individual
physician as well as the clinic group practice.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-clinic-group-practice-reassignment}{Medicare Revalidation Clinic Group Practice Reassignment API}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
revalidation_group(enroll_id = "I20200929003184",
                   npi = 1962026229,
                   first_name = "Rashadda",
                   last_name = "Wong",
                   state = "CT",
                   specialty = "Physician Assistant")

revalidation_group(pac_id_group = 9436483807,
                   enroll_id_group = "O20190619002165",
                   business_name = "1st Call Urgent Care",
                   state_group = "FL",
                   record_type = "Reassignment")
\dontshow{\}) # examplesIf}
}
