% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalidation_reassign.R
\name{revalidation_reassign}
\alias{revalidation_reassign}
\title{Search the Medicare Revalidation Reassignment List API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
revalidation_reassign(
  npi = NULL,
  enroll_id = NULL,
  first_name = NULL,
  last_name = NULL,
  state = NULL,
  specialty = NULL,
  pac_id_group = NULL,
  enroll_id_group = NULL,
  business_name = NULL,
  state_group = NULL,
  record_type = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{NPI of provider who is reassigning their benefits or is an
employee}

\item{enroll_id}{Enrollment ID of provider reassigning their benefits or is an employee}

\item{first_name}{First name of provider who is reassigning their benefits
or is an employee}

\item{last_name}{Last name of provider who is reassigning their benefits or
is an employee}

\item{state}{Enrollment state of provider who is reassigning their
benefits or is an employee}

\item{specialty}{Enrollment specialty of the provider who is
reassigning their benefits or is an employee}

\item{pac_id_group}{PAC ID of provider who is receiving reassignment or is
the employer. Providers enroll at the state level, so one PAC ID may be
associated with multiple Enrollment IDs.}

\item{enroll_id_group}{Enrollment ID of provider who is receiving
reassignment or is the employer}

\item{business_name}{Legal business name of provider who is receiving
reassignment or is the employer}

\item{state_group}{Enrollment state of provider who is receiving
reassignment or is the employer}

\item{record_type}{Identifies whether the record is for a reassignment
(\code{Reassignment}) or employment (\verb{Physician Assistant})}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Reassignments of Providers who are due for Revalidation.
}
\details{
The Revalidation Reassignment List dataset provides information on
reassignments of providers who are due for revalidation.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list}{Medicare Revalidation Reassignment List API}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
revalidation_reassign(enroll_id = "I20200929003184",
                   npi = 1962026229,
                   first_name = "Rashadda",
                   last_name = "Wong",
                   state = "CT",
                   specialty = "Physician Assistant")

revalidation_reassign(pac_id_group = 9436483807,
                   enroll_id_group = "O20190619002165",
                   business_name = "1st Call Urgent Care",
                   state_group = "FL",
                   record_type = "Reassignment")
\dontshow{\}) # examplesIf}
}
