% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalidation_reassign.R
\name{revalidation_reassign}
\alias{revalidation_reassign}
\title{Search the Medicare Revalidation Reassignment List API}
\usage{
revalidation_reassign(
  npi = NULL,
  pac_id_ind = NULL,
  enroll_id_ind = NULL,
  first = NULL,
  last = NULL,
  state_ind = NULL,
  specialty = NULL,
  pac_id_org = NULL,
  enroll_id_org = NULL,
  state_org = NULL,
  business_name = NULL,
  record_type = NULL,
  tidy = TRUE
)
}
\arguments{
\item{npi}{NPI of provider who is reassigning their benefits or is an
employee}

\item{pac_id_ind}{PAC ID of provider reassigning their benefits or
is an employee}

\item{enroll_id_ind}{Enrollment ID of provider reassigning their benefits or
is an employee}

\item{first, last}{First and/or last name of provider who is reassigning their benefits
or is an employee}

\item{state_ind}{Enrollment state of provider who is reassigning their
benefits or is an employee}

\item{specialty}{Enrollment specialty of the provider who is
reassigning their benefits or is an employee}

\item{pac_id_org}{PAC ID of provider who is receiving reassignment or is
the employer.}

\item{enroll_id_org}{Enrollment ID of provider who is receiving
reassignment or is the employer}

\item{state_org}{Enrollment state of provider who is receiving
reassignment or is the employer}

\item{business_name}{Legal business name of provider who is receiving
reassignment or is the employer}

\item{record_type}{Identifies whether the record is for a reassignment
(\code{Reassignment}) or employment (\verb{Physician Assistant})}

\item{tidy}{Tidy output; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Reassignments of Providers who are due for Revalidation.
}
\details{
The Revalidation Reassignment List dataset provides information on
reassignments of providers who are due for revalidation.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list}{Medicare Revalidation Reassignment List API}
}

\emph{Update Frequency:} \strong{Monthly}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
revalidation_reassign(enroll_id_ind = "I20200929003184")
revalidation_reassign(pac_id_ind = 9830437441)
revalidation_reassign(pac_id_org = 3173525888)
\dontshow{\}) # examplesIf}
}
