% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revalidation_reassign.R
\name{revalidation_reassign}
\alias{revalidation_reassign}
\title{Search the Medicare Revalidation Reassignment List API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
revalidation_reassign(
  group_pac_id = NULL,
  group_enroll_id = NULL,
  group_bus_name = NULL,
  group_state = NULL,
  record_type = NULL,
  ind_enroll_id = NULL,
  ind_npi = NULL,
  ind_first = NULL,
  ind_last = NULL,
  ind_state = NULL,
  ind_specialty = NULL,
  month = "Latest",
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{group_pac_id}{PAC ID of provider who is receiving reassignment or is
the employer}

\item{group_enroll_id}{Enrollment ID of provider who is receiving
reassignment or is the employer}

\item{group_bus_name}{Legal business name of provider who is receiving
reassignment or is the employer}

\item{group_state}{Enrollment state of provider who is receiving
reassignment or is the employer}

\item{record_type}{Identifies whether the record is for a reassignment
(\code{Reassignment}) or employment (\verb{Physician Assistant})}

\item{ind_enroll_id}{Enrollment ID of provider who is reassigning their
benefits or is an employee}

\item{ind_npi}{NPI of provider who is reassigning their benefits or is an
employee}

\item{ind_first}{First name of provider who is reassigning their benefits
or is an employee}

\item{ind_last}{Last name of provider who is reassigning their benefits or
is an employee}

\item{ind_state}{Enrollment state of provider who is reassigning their
benefits or is an employee}

\item{ind_specialty}{Enrollment specialty of the provider who is
reassigning their benefits or is an employee}

\item{month}{dataset version, \code{Latest} is default; possible months are
\code{Jan}, \code{Feb}, \code{Apr}, \code{Jun}, \code{Jul}, \code{Aug}, \code{Sep}, \code{Oct}, \code{Nov}, \code{Dec}}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Reassignments of Providers who are due for Revalidation.
}
\details{
The Revalidation Reassignment List dataset provides information on
reassignments of providers who are due for revalidation.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revalidation-reassignment-list}{Medicare Revalidation Reassignment List API}
}
}
}
\note{
Update Frequency: \strong{Monthly}
}
\examples{
revalidation_reassign(ind_enroll_id = "I20200929003184",
                   ind_npi = 1962026229,
                   ind_first = "Rashadda",
                   ind_last = "Wong",
                   ind_state = "CT",
                   ind_specialty = "Physician Assistant")

revalidation_reassign(group_pac_id = 9436483807,
                   group_enroll_id = "O20190619002165",
                   group_bus_name = "1st Call Urgent Care",
                   group_state = "FL",
                   record_type = "Reassignment")
\dontrun{
npi_list <- rep(c(1326063900,
                  1588191308,
                  1639176803,
                  1326057050), each = 9)

months <- rep(c("Dec",
                "Nov",
                "Oct",
                "Sep",
                "Aug",
                "Jul",
                "Jun",
                "Apr",
                "Feb"),
                times = 4)

purrr::map2_dfr(npi_list,
months, ~revalidation_reassign(ind_npi = .x, month = .y))
}
}
