% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-revocations.R
\name{revocations}
\alias{revocations}
\title{Revocations of Medicare Enrollments}
\usage{
revocations(
  npi = NULL,
  enid = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  org_name = NULL,
  multi = NULL,
  state = NULL,
  specialty = NULL,
  reason = NULL,
  count = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} 10-digit Individual National Provider Identifier}

\item{enid}{\verb{<chr>} 15-digit Medicare Enrollment ID}

\item{first, middle, last}{\verb{<chr>} Individual provider's name}

\item{org_name}{\verb{<chr>} Organization name}

\item{multi}{\verb{<lgl>} Provider has multiple NPIs}

\item{state}{\verb{<chr>} Enrollment state, full or abbreviation}

\item{specialty}{\verb{<chr>} Enrollment specialty description}

\item{reason}{\verb{<chr>} reason for revocation}

\item{count}{\verb{<lgl>} Return the dataset's total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Eligibility to order and refer within Medicare
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
revocations(count = TRUE)
revocations(state = "GA")
revocations(specialty = contains("CARDIO"))
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{API: Medicare Order and Referring}
\item \href{https://www.cms.gov/medicare/enrollment-renewal/providers-suppliers/chain-ownership-system-pecos/ordering-certifying}{CMS: Ordering & Certifying}
}
}
