% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-revoked.R
\name{revocations}
\alias{revocations}
\title{Revoked Providers & Suppliers}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/revoked-medicare-providers-and-suppliers}{API: Revoked Medicare Providers and Suppliers}
}
}
\usage{
revocations(
  npi = NULL,
  enid = NULL,
  first = NULL,
  middle = NULL,
  last = NULL,
  org_name = NULL,
  multi = NULL,
  state = NULL,
  specialty = NULL,
  reason = NULL,
  count = FALSE,
  set = FALSE
)
}
\arguments{
\item{npi}{\verb{<int>} National Provider Identifier}

\item{enid}{\verb{<chr>} Medicare Enrollment ID}

\item{first, middle, last}{\verb{<chr>} Individual provider's name}

\item{org_name}{\verb{<chr>} Organization name}

\item{multi}{\verb{<lgl>} Provider has multiple NPIs}

\item{state}{\verb{<chr>} Enrollment state abbreviation}

\item{specialty}{\verb{<chr>} Enrollment specialty description}

\item{reason}{\verb{<chr>} Reason for revocation}

\item{count}{\verb{<lgl>} Return the total row count}

\item{set}{\verb{<lgl>} Return the entire dataset}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Information on providers and suppliers currently revoked from the Medicare program.
\subsection{Details}{

The Revoked Medicare Providers and Suppliers dataset contains information on
providers and suppliers who are revoked and under a current re-enrollment bar.
This dataset includes provider and supplier names, NPIs, revocation
authorities pursuant to \href{https://www.ecfr.gov/current/title-42/chapter-IV/subchapter-B/part-424/subpart-P/section-424.535}{42 CFR § 424.535}, revocation effective dates, and
re-enrollment bar expiration dates. This dataset is based on information
gathered from the \emph{Provider Enrollment, Chain and Ownership System (PECOS)}.
}
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
revocations(count = TRUE)

revocations(
  count = TRUE,
  org_name = not_blank())

revocations(org_name = starts("B"))

revocations(
  specialty = contains("CARDIO"),
  state = excludes(c("GA", "OH")))
\dontshow{\}) # examplesIf}
}
