% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{ror}
\alias{ror}
\title{Calculate lagged rate of return}
\usage{
ror(df, col, n = 1L)
}
\arguments{
\item{df}{numeric vector}

\item{col}{column}

\item{n}{values to offset}
}
\value{
numeric vector
}
\description{
Calculate lagged rate of return
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dplyr::tibble(year = 2021:2023,
              pay = c(2000, 2200, 1980)) |>
ror(pay)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
