% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndc.R
\name{rxnorm}
\alias{rxnorm}
\title{RxNorm API}
\usage{
rxnorm(ndc, ...)
}
\arguments{
\item{ndc}{< \emph{character} > // \strong{required} 10- to 11-digit National Drug Code}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
RxNorm API
}
\keyword{internal}
