% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-spending.R
\name{spending}
\alias{spending}
\alias{spending2}
\title{Medicare Spending Per Beneficiary - Hospital}
\source{
\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}

\itemize{
\item \href{https://data.cms.gov/provider-data/dataset/27ea-46a8}{API: Physician Facility Affiliations}
}
}
\usage{
spending(
  ccn = NULL,
  name = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  county = NULL,
  score = NULL,
  count = FALSE
)

spending2(ccn = NULL, name = NULL, state = NULL, claim = NULL, count = FALSE)
}
\arguments{
\item{ccn}{\verb{<chr>} desc}

\item{name}{\verb{<chr>} desc}

\item{city, state, zip, county}{\verb{<chr>} desc}

\item{state}{\verb{<chr>} desc}

\item{score}{\verb{<dbl>} desc}

\item{count}{\verb{<lgl>} Return the total row count}

\item{claim}{\verb{<chr>} desc}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.

A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Access information concerning individual providers'
affiliations with organizations/facilities.

Access information concerning individual providers'
affiliations with organizations/facilities.
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
spending2(count = TRUE)
spending2(name = starts("SGMC"), state = "GA")
\dontshow{\}) # examplesIf}
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
spending2(count = TRUE)
spending2(state = "GA")
\dontshow{\}) # examplesIf}
}
