% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-hospitals.R
\name{subgroups}
\alias{subgroups}
\title{Hospital Subgroups}
\usage{
subgroups(
  acute = NULL,
  drug = NULL,
  child = NULL,
  general = NULL,
  long = NULL,
  short = NULL,
  psych = NULL,
  rehab = NULL,
  swing = NULL,
  psych_unit = NULL,
  rehab_unit = NULL,
  specialty = NULL,
  other = NULL
)
}
\arguments{
\item{acute}{\verb{<lgl>} Acute/Short Term Care Hospital}

\item{drug}{\verb{<lgl>} Alcohol/Drug Treatment}

\item{child}{\verb{<lgl>} Children's Hospital}

\item{general}{\verb{<lgl>} General Hospital}

\item{long}{\verb{<lgl>} Long-Term Care}

\item{short}{\verb{<lgl>} Short-Term Care}

\item{psych}{\verb{<lgl>} Psychiatric}

\item{rehab}{\verb{<lgl>} Rehabilitation}

\item{swing}{\verb{<lgl>} Swing-Bed Approved}

\item{psych_unit}{\verb{<lgl>} Psychiatric Unit}

\item{rehab_unit}{\verb{<lgl>} Rehabilitation Unit}

\item{specialty}{\verb{<lgl>} Specialty Hospital}

\item{other}{\verb{<lgl>} Unlisted on CMS form}
}
\value{
A \verb{<subgroups>} object
}
\description{
Hospital Subgroups
}
\examples{
subgroups(acute = TRUE, rehab = TRUE)

subgroups()

}
\keyword{internal}
