% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_crosswalk.R
\name{taxonomy_crosswalk}
\alias{taxonomy_crosswalk}
\title{Taxonomy Code // Medicare Specialty}
\usage{
taxonomy_crosswalk(
  taxonomy_code = NULL,
  taxonomy_description = NULL,
  specialty_code = NULL,
  specialty_description = NULL,
  keyword_search = NULL,
  tidy = TRUE
)
}
\arguments{
\item{taxonomy_code}{< \emph{character} > 10-digit taxonomy code}

\item{taxonomy_description}{< \emph{character} > Provider's taxonomy description}

\item{specialty_code}{< \emph{character} > Medicare specialty code}

\item{specialty_description}{< \emph{character} > Medicare provider/supplier type}

\item{keyword_search}{< \emph{character} > Search term to use for quick full-text search.}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{specialty_code} \tab Code that corresponds to the Medicare specialty \cr
   \code{specialty_description} \tab Description of the Medicare provider/Supplier Type \cr
   \code{taxonomy_code} \tab Provider's taxonomy code \cr
   \code{taxonomy_description} \tab Description of the taxonomy code \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{taxonomy_crosswalk()} allows you to search the types of providers and
suppliers eligible for Medicare programs by taxonomy code or Medicare
specialty type code.
\subsection{Taxonomy Codes}{

The \strong{Healthcare Provider Taxonomy Code Set} is a hierarchical HIPAA
standard code set designed to categorize the type, classification, and
specialization of health care providers. It consists of two sections:
\enumerate{
\item Individuals and Groups of Individuals
\item Non-Individuals
}

When applying for an NPI, a provider must report the taxonomy that most
closely describes their type/classification/specialization. In some
situations, a provider may need to report more than one taxonomy but must
indicate one of them as the primary. The codes selected may not be the same
as categorizations used by Medicare for enrollment.

Links:
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-supplier-taxonomy-crosswalk}{Provider and Supplier Taxonomy Crosswalk}
\item \href{https://data.cms.gov/resources/medicare-provider-and-supplier-taxonomy-crosswalk-methodology}{Taxonomy Crosswalk Methodology}
\item \href{https://www.cms.gov/medicare/enrollment-renewal/providers-suppliers/health-care-taxonomy}{Find Your Taxonomy Code}
}

\emph{Update Frequency:} \strong{Weekly}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
taxonomy_crosswalk(keyword_search = "B4")
taxonomy_crosswalk(keyword_search = "Histocompatibility")
taxonomy_crosswalk(specialty_description = "Rehabilitation Agency")
taxonomy_crosswalk(taxonomy_code = "2086S0102X")
\dontshow{\}) # examplesIf}
}
