% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_crosswalk.R
\name{taxonomy_crosswalk}
\alias{taxonomy_crosswalk}
\title{Search the Medicare Provider and Supplier Taxonomy Crosswalk API}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
taxonomy_crosswalk(
  specialty_code = NULL,
  specialty_desc = NULL,
  taxonomy_code = NULL,
  taxonomy_desc = NULL,
  clean_names = TRUE,
  lowercase = TRUE
)
}
\arguments{
\item{specialty_code}{Code that corresponds to the listed Medicare specialty}

\item{specialty_desc}{Description of the Medicare Provider/Supplier Type}

\item{taxonomy_code}{The taxonomy codes the providers use}

\item{taxonomy_desc}{The description of the taxonomy that the providers use}

\item{clean_names}{Clean column names with {janitor}'s
\code{clean_names()} function; default is \code{TRUE}.}

\item{lowercase}{Convert column names to lowercase; default is \code{TRUE}.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of the type of providers and suppliers with the
proper taxonomy code eligible for medicare programs.
}
\details{
The Medicare Provider and Supplier Taxonomy Crosswalk dataset
lists the providers and suppliers eligible to enroll in Medicare
programs with the proper healthcare provider taxonomy code. This data
includes the Medicare speciality codes, if available, provider/supplier
type description, taxonomy code, and the taxonomy description. This
dataset is derived from information gathered from National Plan and
Provider Enumerator System (NPPES) and Provider Enrollment, Chain and
Ownership System (PECOS).
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/medicare-provider-and-supplier-taxonomy-crosswalk}{Medicare Provider and Supplier Taxonomy Crosswalk API}
\item \href{https://www.cms.gov/Medicare/Provider-Enrollment-and-Certification/Find-Your-Taxonomy-Code}{Medicare Find Your Taxonomy Code}
}
}
}
\note{
Update Frequency: \strong{Weekly}
}
\examples{
taxonomy_crosswalk(specialty_code = "B4[14]")

taxonomy_crosswalk(specialty_desc = "Rehabilitation Agency")

taxonomy_crosswalk(taxonomy_code = "2086S0102X")

taxonomy_crosswalk(taxonomy_desc = "Agencies/Hospice Care Community Based")
}
