% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyup}
\alias{tidyup}
\title{Tidy a Data Frame}
\usage{
tidyup(
  df,
  dtype = NULL,
  dt = "date",
  yn = NULL,
  int = NULL,
  dbl = NULL,
  chr = NULL,
  up = NULL,
  cred = NULL
)
}
\arguments{
\item{df}{data frame}

\item{dtype}{\code{mdy} or \code{ymd}}

\item{dt}{cols to convert to date}

\item{yn}{cols to convert to logical}

\item{int}{cols to convert to integer}

\item{dbl}{cols to convert to double}

\item{chr}{cols to convert to character}

\item{up}{cols to convert to upper case}

\item{cred}{cols to remove periods from}
}
\value{
tidy data frame
}
\description{
Tidy a Data Frame
}
\keyword{internal}
