% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms-transparency.R
\name{transparency}
\alias{transparency}
\title{Hospital Price Transparency}
\usage{
transparency(
  name = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  action = NULL,
  count = FALSE
)
}
\arguments{
\item{name}{\verb{<chr>} Hospital name}

\item{address}{\verb{<chr>} Hospital address}

\item{city}{\verb{<chr>} Hospital city}

\item{state}{\verb{<chr>} Hospital state}

\item{action}{\verb{<chr>} Action taken by CMS following a Compliance Review}

\item{count}{\verb{<lgl>} Return the dataset's total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Eligibility to order and refer within Medicare
}
\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
transparency(count = TRUE)
transparency(state = "GA", city = "Valdosta")
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item \href{https://data.cms.gov/provider-characteristics/medicare-provider-supplier-enrollment/order-and-referring}{API: Medicare Order and Referring}
\item \href{https://www.cms.gov/medicare/enrollment-renewal/providers-suppliers/chain-ownership-system-pecos/ordering-certifying}{CMS: Ordering & Certifying}
}
}
