% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-utilization.R
\name{utilization}
\alias{utilization}
\alias{services}
\alias{geography}
\title{Provider Utilization by Year}
\usage{
utilization(
  year = NULL,
  npi = NULL,
  first = NULL,
  last = NULL,
  cred = NULL,
  entity = NULL,
  specialty = NULL,
  par = NULL,
  hcpcs = NULL,
  patients = NULL,
  services = NULL,
  charge = NULL,
  allowed = NULL,
  payment = NULL,
  avg_age = NULL,
  avg_risk = NULL,
  dual = NULL,
  ndual = NULL,
  city = NULL,
  state = NULL,
  count = FALSE
)

services(
  year = NULL,
  npi = NULL,
  entity = NULL,
  first = NULL,
  last = NULL,
  cred = NULL,
  specialty = NULL,
  par = NULL,
  hcpcs = NULL,
  drug = NULL,
  pos = NULL,
  patients = NULL,
  services = NULL,
  charge = NULL,
  allowed = NULL,
  payment = NULL,
  count = FALSE
)

geography(
  year = NULL,
  level = NULL,
  state = NULL,
  hcpcs = NULL,
  drug = NULL,
  pos = NULL,
  providers = NULL,
  patients = NULL,
  services = NULL,
  charge = NULL,
  allowed = NULL,
  payment = NULL,
  count = FALSE
)
}
\arguments{
\item{year}{\verb{<int>} Year data was reported}

\item{npi}{\verb{<int>} 10-digit national provider identifier}

\item{first, last}{\verb{<chr>} Individual/Organizational provider's name}

\item{cred}{\verb{<chr>} Individual provider's credentials}

\item{entity}{\verb{<chr>} Type of entity reported in NPPES. \code{I} identifies
individual providers and \code{O} identifies those registered as organizations.}

\item{specialty}{\verb{<chr>} Provider specialty reported on the largest number of
claims submitted}

\item{par}{\verb{<lgl>} Identifies a provider \emph{with at least one claim}
identifying them as \emph{participating} in Medicare or accepting assignment of
Medicare allowed amounts within HCPCS code and place of service. A
\emph{non-participating} provider is one that may elect to accept Medicare
allowed amounts for some services and not accept Medicare allowed amounts
for other services.}

\item{hcpcs}{\verb{<int/chr>} Total number of unique HCPCS codes}

\item{patients}{\verb{<int>} Total Medicare beneficiaries receiving services from
the provider}

\item{services}{\verb{<int>} Total provider services}

\item{charge}{\verb{<int>} The total charges that the provider submitted for all
services}

\item{allowed}{\verb{<dbl>} The Medicare allowed amount for all provider services.
This figure is the sum of the amount Medicare pays, the deductible and
coinsurance amounts that the beneficiary is responsible for paying, and any
amounts that a third party is responsible for paying.}

\item{payment}{\verb{<dbl>} Total amount that Medicare paid after deductible and
coinsurance amounts have been deducted for all the provider's line item
services.}

\item{avg_age}{\verb{<dbl>} Average age of beneficiaries. Beneficiary age is
calculated at the end of the calendar year or at the time of death}

\item{avg_risk}{\verb{<dbl>} Average Hierarchical Condition Category (HCC) risk
score of beneficiaries}

\item{dual}{\verb{<int>} Number of Medicare beneficiaries qualified to receive
Medicare and Medicaid benefits. Beneficiaries are classified as Medicare
and Medicaid entitlement if in any month in the given calendar year they
were receiving full or partial Medicaid benefits.}

\item{ndual}{\verb{<int>} Number of Medicare beneficiaries qualified to receive
Medicare only benefits. Beneficiaries are classified as Medicare only
entitlement if they received zero months of any Medicaid benefits (full or
partial) in the given calendar year.}

\item{city, state}{\verb{<chr>} The provider's city and state, as reported in NPPES.}

\item{count}{\verb{<lgl>} Return the total row count}

\item{drug}{\verb{<lgl>} Total number of unique HCPCS codes}

\item{pos}{\verb{<chr>} Total number of unique HCPCS codes}

\item{level}{\verb{<chr>} National or State}

\item{providers}{\verb{<int>} Total}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
Access information on services and procedures provided to Original
Medicare (fee-for-service) Part B beneficiaries by physicians and other
healthcare professionals; aggregated by provider, service and geography.

The \strong{Provider} dataset allows the user access to data such as
services and procedures performed; charges submitted and payment received;
and beneficiary demographic and health characteristics for providers
treating Original Medicare (fee-for-service) Part B beneficiaries,
aggregated by year.
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider}{Medicare Physician & Other Practitioners: by Provider API}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}{Medicare Physician & Other Practitioners: by Provider and Service API}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service}{Medicare Physician & Other Practitioners: by Geography and Service API}
}
}

\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
utilization(count = TRUE)
services(count = TRUE)
geography(count = TRUE)

utilization(npi = 1003000423)

services(npi = 1003000423)

geography(
  hcpcs = c("Q0091", "G0101", "99213", "99212", "99203", "81002", "76830"),
  pos = "O",
  state = c("National", "Ohio"))
\dontshow{\}) # examplesIf}
}
