% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-utilization.R
\name{utilization}
\alias{utilization}
\title{Provider Utilization by Year}
\usage{
utilization(
  year = NULL,
  npi = NULL,
  first = NULL,
  last = NULL,
  cred = NULL,
  entity = NULL,
  address = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  specialty = NULL,
  participating = NULL,
  hcpcs = NULL,
  patients = NULL,
  services = NULL,
  charges = NULL,
  allowed = NULL,
  payment = NULL,
  avg_age = NULL,
  avg_risk = NULL,
  dual = NULL,
  ndual = NULL,
  count = FALSE
)
}
\arguments{
\item{year}{\verb{<int>} Year data was reported}

\item{npi}{\verb{<int>} 10-digit national provider identifier}

\item{first, last}{\verb{<chr>} Individual/Organizational provider's name}

\item{cred}{\verb{<chr>} Individual provider's credentials}

\item{entity}{\verb{<chr>} Type of entity reported in NPPES. An entity code of \code{I}
identifies providers registered as individuals and an entity type code of
\code{O} identifies providers registered as organizations.}

\item{address, city, state, zip}{The provider's street address, city, state and
zip code, as reported in NPPES.}

\item{specialty}{\verb{<chr>} Provider specialty reported on the largest number of
claims submitted}

\item{participating}{\verb{<lgl>} Identifies whether the provider participates in
Medicare and/or accepts assignment of Medicare allowed amounts. The value
will be \code{Y} for any provider that had at least one claim identifying the
provider as participating in Medicare or accepting assignment of Medicare
allowed amounts within HCPCS code and place of service. A non-participating
provider may elect to accept Medicare allowed amounts for some services and
not accept Medicare allowed amounts for other services.}

\item{hcpcs}{\verb{<int>} Total number of unique HCPCS codes}

\item{patients}{\verb{<int>} Total Medicare beneficiaries receiving services from
the provider}

\item{services}{\verb{<int>} Total provider services}

\item{charges}{\verb{<int>} The total charges that the provider submitted for all
services}

\item{allowed}{\verb{<dbl>} The Medicare allowed amount for all provider services.
This figure is the sum of the amount Medicare pays, the deductible and
coinsurance amounts that the beneficiary is responsible for paying, and any
amounts that a third party is responsible for paying.}

\item{payment}{\verb{<dbl>} Total amount that Medicare paid after deductible and
coinsurance amounts have been deducted for all the provider's line item
services.}

\item{avg_age}{\verb{<dbl>} Average age of beneficiaries. Beneficiary age is
calculated at the end of the calendar year or at the time of death}

\item{avg_risk}{\verb{<dbl>} Average Hierarchical Condition Category (HCC) risk
score of beneficiaries}

\item{dual}{\verb{<int>} Number of Medicare beneficiaries qualified to receive
Medicare and Medicaid benefits. Beneficiaries are classified as Medicare
and Medicaid entitlement if in any month in the given calendar year they
were receiving full or partial Medicaid benefits.}

\item{ndual}{\verb{<int>} Number of Medicare beneficiaries qualified to receive
Medicare only benefits. Beneficiaries are classified as Medicare only
entitlement if they received zero months of any Medicaid benefits (full or
partial) in the given calendar year.}

\item{count}{\verb{<lgl>} Return the total row count}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Access information on services and procedures provided to Original
Medicare (fee-for-service) Part B beneficiaries by physicians and other
healthcare professionals; aggregated by provider, service and geography.

The \strong{Provider} dataset allows the user access to data such as
services and procedures performed; charges submitted and payment received;
and beneficiary demographic and health characteristics for providers
treating Original Medicare (fee-for-service) Part B beneficiaries,
aggregated by year.
}
\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider}{Medicare Physician & Other Practitioners: by Provider API}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}{Medicare Physician & Other Practitioners: by Provider and Service API}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service}{Medicare Physician & Other Practitioners: by Geography and Service API}
}
}

\examples{
\dontshow{if (httr2::is_online()) withAutoprint(\{ # examplesIf}
utilization(count = TRUE)

utilization(npi = 1003000423)
\dontshow{\}) # examplesIf}
}
