% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilization.R
\name{utilization}
\alias{utilization}
\alias{by_provider}
\alias{by_service}
\alias{by_geography}
\title{Utilization Statistics for Medicare Part B Providers}
\usage{
by_provider(
  year,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  tidy = TRUE
)

by_service(
  year,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  hcpcs_code = NULL,
  drug = NULL,
  pos = NULL,
  tidy = TRUE,
  rbcs = TRUE
)

by_geography(
  year,
  state = NULL,
  hcpcs_code = NULL,
  pos = NULL,
  level = NULL,
  fips = NULL,
  hcpcs_desc = NULL,
  drug = NULL,
  tidy = TRUE,
  rbcs = TRUE
)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Year in \code{YYYY} format. Run helper
function \code{prac_years()} to return a vector of the years currently available.}

\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{first, last}{< \emph{character} > Individual provider's name}

\item{organization}{< \emph{character} > Organizational provider's name}

\item{credential}{< \emph{character} > Individual provider's credentials}

\item{gender}{< \emph{character} > Individual provider's gender; \code{"F"} (Female),
\code{"M"} (Male)}

\item{entype}{< \emph{character} > Provider entity type; \code{"I"} (Individual),
\code{"O"} (Organization)}

\item{city}{< \emph{character} > City where provider is located}

\item{state}{< \emph{character} > State where the provider is located, as reported
in NPPES. Values include the 50 United States, District of Columbia,
U.S. territories, Armed Forces areas, Unknown and Foreign Country. Data
aggregated at the National level are identified by the word 'National'.}

\item{zip}{< \emph{character} > Provider’s zip code}

\item{fips}{< \emph{character} > FIPS code of the referring provider state. This variable is
blank when reported at the national level.}

\item{ruca}{< \emph{character} > Rural-Urban Commuting Area Code (RUCA):
- \strong{Metro Area Core}
- \code{"1"}: Primary flow within Urbanized Area (UA)
- \code{"1.1"}: Secondary flow 30-50\% to larger UA
- \strong{Metro Area High Commuting}
- \code{"2"}: Primary flow 30\% or more to UA
- \code{"2.1"}: Secondary flow 30-50\% to larger UA
- \strong{Metro Area Low Commuting}
- \code{3}: Primary flow 10-30\% to UA
- \strong{Micro Area Core}
- \code{4}: Primary flow within large Urban Cluster (10k - 49k)
- \code{4.1}: Secondary flow 30-50\% to UA
- \strong{Micro High Commuting}
- \code{5}: Primary flow 30\% or more to large UC
- \code{5.1}: Secondary flow 30-50\% to UA
- \strong{Micro Low Commuting}
- \code{6}: Primary flow 10-30\% to large UC
- \strong{Small Town Core}
- \code{7}: Primary flow within small UC (2.5k - 9.9k)
- \code{7.1}: Secondary flow 30-50\% to UA
- \code{7.2}: Secondary flow 30-50\% to large UC
- \strong{Small Town High Commuting}
- \code{8}: Primary flow 30\% or more to small UC
- \code{8.1}: Secondary flow 30-50\% to UA
- \code{8.2}: Secondary flow 30-50\% to large UC
- \strong{Small Town Low Commuting}
- \code{9}: Primary flow 10-30\% to small UC
- \strong{Rural Areas}
- \code{10}: Primary flow to tract outside a UA or UC
- \code{10.1}: Secondary flow 30-50\% to UA
- \code{10.2}: Secondary flow 30-50\% to large UC
- \code{10.3}: Secondary flow 30-50\% to small UC
- \code{99}: Zero population and no rural-urban identifier information}

\item{country}{< \emph{character} > Country where provider is located.}

\item{specialty}{< \emph{character} > Provider specialty code reported on the largest number of
claims submitted.}

\item{par}{Identifies whether the provider participates in Medicare
and/or accepts assignment of Medicare allowed amounts. The value will
be \code{Y} for any provider that had at least one claim identifying the
provider as participating in Medicare or accepting assignment of
Medicare allowed amounts within HCPCS code and place of service. A
non-participating provider may elect to accept Medicare allowed amounts
for some services and not accept Medicare allowed amounts for other
services.}

\item{tidy}{< \emph{boolean} > Tidy output; default is \code{TRUE}}

\item{hcpcs_code}{< \emph{character} > HCPCS code used to identify the specific medical service
furnished by the provider.}

\item{drug}{< \emph{boolean} > Flag that identifies whether the HCPCS code for the specific
service furnished by the provider is a HCPCS listed on the Medicare Part
B Drug Average Sales Price (ASP) File.}

\item{pos}{< \emph{character} > Identifies whether the place of service submitted on the claims is a facility (\code{"F"}) or non-facility (\code{"O"}). Non-facility is generally an office setting; however other entities are included in non-facility.}

\item{rbcs}{< \emph{boolean} > Add Restructured BETOS Classifications to HCPCS
codes; default is \code{TRUE}}

\item{level}{< \emph{character} > Level of geography by which the data will be aggregated. \code{State}
indicates the data is aggregated to a single state identified in the Rendering Provider
State column. \code{National} indicates the data is aggregated across all states for a
given HCPCS Code Level.}

\item{hcpcs_desc}{< \emph{character} > Description of the HCPCS code for the specific medical
service furnished by the provider.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.

A \link[tibble:tibble-package]{tibble} containing the search results.

A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
These functions allow the user access to information on services and
procedures provided to Original Medicare (fee-for-service) Part B
beneficiaries by physicians and other healthcare professionals; aggregated
by provider, service and geography.
}
\section{\code{by_provider()}}{

The \strong{Provider and Service} dataset allows the user access to data such as
services and procedures performed; charges submitted and payment received;
and beneficiary demographic and health characteristics for providers
treating Original Medicare (fee-for-service) Part B beneficiaries,
aggregated by year.
}

\section{\code{by_service()}}{

The \strong{Provider and Service} dataset is aggregated by the following:
\enumerate{
\item the NPI for the performing provider,
\item the Healthcare Common Procedure Coding System (HCPCS) code, and
\item the place of service (either facility or non-facility).
}

There can be multiple records for a given NPI based on the number of
distinct HCPCS codes that were billed and where the services were
provided. Data have been aggregated based on the place of service
because separate fee schedules apply depending on whether the place
of service submitted on the claim is facility or non-facility.
}

\section{\code{by_geography()}}{

The \strong{Geography and Service} subset contains information on utilization,
allowed amount, Medicare payment, and submitted charges organized nationally
and state-wide by HCPCS code and place of service.

Links:
\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider}{Medicare Physician & Other Practitioners: by Provider API}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}{Medicare Physician & Other Practitioners: by Provider and Service API}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service}{Medicare Physician & Other Practitioners: by Geography and Service API}
}

\emph{Update Frequency:} \strong{Annually}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
by_provider(year = 2020, npi = 1003000423)

# Use the years helper function to retrieve results for every year:
prac_years() |>
map(\(x) by_provider(year = x, npi = 1043477615)) |>
list_rbind()

by_service(year = 2019, npi = 1003000126)

by_geography(year = 2020, hcpcs_code = "0002A")
\dontshow{\}) # examplesIf}
}
