% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilization.R
\name{utilization}
\alias{utilization}
\alias{by_provider}
\alias{by_service}
\alias{by_geography}
\title{Provider Utilization & Demographics}
\usage{
by_provider(
  year,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  tidy = TRUE,
  nest = TRUE,
  detailed = FALSE,
  na.rm = TRUE
)

by_service(
  year,
  npi = NULL,
  first = NULL,
  last = NULL,
  organization = NULL,
  credential = NULL,
  gender = NULL,
  entype = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  fips = NULL,
  ruca = NULL,
  country = NULL,
  specialty = NULL,
  par = NULL,
  hcpcs_code = NULL,
  drug = NULL,
  pos = NULL,
  tidy = TRUE,
  rbcs = TRUE,
  na.rm = TRUE
)

by_geography(
  year,
  state = NULL,
  hcpcs_code = NULL,
  pos = NULL,
  level = NULL,
  fips = NULL,
  drug = NULL,
  tidy = TRUE,
  rbcs = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{year}{< \emph{integer} > // \strong{required} Year data was reported, in \code{YYYY}
format. Run \code{\link[=pop_years]{pop_years()}} to return a vector of the years currently available.}

\item{npi}{< \emph{integer} > 10-digit national provider identifier}

\item{first, last, organization}{< \emph{character} > Individual/Organizational
provider's name}

\item{credential}{< \emph{character} > Individual provider's credentials}

\item{gender}{< \emph{character} > Individual provider's gender; \code{"F"} (Female),
\code{"M"} (Male)}

\item{entype}{< \emph{character} > Provider entity type; \code{"I"} (Individual),
\code{"O"} (Organization)}

\item{city}{< \emph{character} > City where provider is located}

\item{state}{< \emph{character} > State where provider is located}

\item{zip}{< \emph{character} > Provider’s zip code}

\item{fips}{< \emph{character} > Provider's state FIPS code}

\item{ruca}{< \emph{character} > Provider’s RUCA code}

\item{country}{< \emph{character} > Country where provider is located}

\item{specialty}{< \emph{character} > Provider specialty code reported on the
largest number of claims submitted}

\item{par}{< \emph{boolean} > Identifies whether the provider participates in
Medicare and/or accepts assignment of Medicare allowed amounts}

\item{tidy}{< \emph{boolean} > // \strong{default:} \code{TRUE} Tidy output}

\item{nest}{< \emph{boolean} > // \strong{default:} \code{TRUE} Nest \code{demographics} and \code{conditions}}

\item{detailed}{< \emph{boolean} > // \strong{default:} \code{FALSE} Include \code{detailed} column}

\item{na.rm}{< \emph{boolean} > // \strong{default:} \code{TRUE} Remove empty rows and columns}

\item{hcpcs_code}{< \emph{character} > HCPCS code used to identify the specific
medical service furnished by the provider}

\item{drug}{< \emph{boolean} > Identifies whether the HCPCS code is listed in the
Medicare Part B Drug Average Sales Price (ASP) File}

\item{pos}{< \emph{character} > Identifies whether the Place of Service (POS)
submitted on the claims is a:
\itemize{
\item Facility (\code{"F"}): Hospital, Skilled Nursing Facility, etc.
\item Non-facility (\code{"O"}): Office, Home, etc.
}}

\item{rbcs}{< \emph{boolean} > // \strong{default:} \code{TRUE} Add Restructured BETOS
Classifications to HCPCS codes}

\item{level}{< \emph{character} > Geographic level by which the data will be
aggregated:
\itemize{
\item \code{"State"}: Data is aggregated for each state
\item \code{National}: Data is aggregated across all states for a given HCPCS Code
}}
}
\value{
\code{by_provider()} returns a \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{year} \tab Year data was reported \cr
   \code{npi} \tab 10-digit national provider identifier \cr
   \code{entity_type} \tab Provider's entity/enumeration type \cr
   \code{first} \tab Individual provider's first name \cr
   \code{middle} \tab Individual provider's middle name \cr
   \code{last} \tab Individual provider's last name/Organization's name \cr
   \code{gender} \tab Individual provider's gender \cr
   \code{credential} \tab Individual provider's credential \cr
   \code{specialty} \tab Provider's specialty \cr
   \code{address} \tab Provider's street address \cr
   \code{city} \tab Provider's city \cr
   \code{state} \tab Provider's state \cr
   \code{zip} \tab Provider's zip code \cr
   \code{fips} \tab Provider's state's FIPS code \cr
   \code{ruca} \tab Provider's RUCA code \cr
   \code{country} \tab Provider's country \cr
   \code{par} \tab Indicates if provider participates in and/or accepts Medicare assignment \cr
   \code{tot_hcpcs} \tab Total number of \emph{unique} HCPCS codes submitted \cr
   \code{tot_benes} \tab Total Medicare beneficiaries receiving services from the provider \cr
   \code{tot_srvcs} \tab Total services provided to beneficiaries by the provider \cr
   \code{tot_charges} \tab Total charges that the provider submitted for all services \cr
   \code{tot_allowed} \tab Total allowed amount for all services; the sum of Medicare reimbursement, deductible, coinsurance, etc. \cr
   \code{tot_payment} \tab Total amount Medicare paid, less any deductible and coinsurance \cr
   \code{tot_std_pymt} \tab Total amount Medicare paid, standardized to account for geographic differences \cr
   \code{hcc_risk_avg} \tab Average Hierarchical Condition Category (HCC) risk score of beneficiaries \cr
   \code{hcpcs_detailed} \tab Nested list columns \code{drug} & \code{medical}, offering further detail about the HCPCS codes submitted \cr
   \code{demographics} \tab Nested list column containing demographic data about the beneficiaries seen by the provider \cr
   \code{conditions} \tab Nested list column containing data about beneficiaries' chronic conditions \cr
}


\code{by_service()} returns a \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{year} \tab Year data was reported \cr
   \code{npi} \tab 10-digit national provider identifier \cr
   \code{level} \tab Data aggregation level, will always be "Provider" here \cr
   \code{entity_type} \tab Provider's entity/enumeration type \cr
   \code{first} \tab Individual provider's first name \cr
   \code{middle} \tab Individual provider's middle name \cr
   \code{last} \tab Individual provider's last name/Organization's name \cr
   \code{gender} \tab Individual provider's gender \cr
   \code{credential} \tab Individual provider's credential \cr
   \code{specialty} \tab Provider's specialty \cr
   \code{address} \tab Provider's street address \cr
   \code{city} \tab Provider's city \cr
   \code{state} \tab Provider's state \cr
   \code{zip} \tab Provider's zip code \cr
   \code{fips} \tab Provider's state's FIPS code \cr
   \code{ruca} \tab Provider's RUCA code \cr
   \code{country} \tab Provider's country \cr
   \code{par} \tab Indicates if the provider participates in and/or accepts Medicare assignment \cr
   \code{hcpcs_code} \tab HCPCS code used to identify the specific medical service furnished by the provider \cr
   \code{hcpcs_desc} \tab Consumer Friendly Description of the HCPCS code \cr
   \code{category} \tab Restructured BETOS Classification System Category \cr
   \code{subcategory} \tab Restructured BETOS Classification System Subcategory \cr
   \code{family} \tab Restructured BETOS Classification System Family \cr
   \code{procedure} \tab Restructured BETOS Classification System Major Procedure Indicator \cr
   \code{drug} \tab Indicates if the HCPCS code is listed in the Medicare Part B Drug Average Sales Price (ASP) File \cr
   \code{pos} \tab Identifies the Place of Service (POS) submitted on the claim \cr
   \code{tot_benes} \tab Distinct number of Medicare beneficiaries for each HCPCS code/POS combination \cr
   \code{tot_srvcs} \tab Number of services provided for each HCPCS code/POS combination \cr
   \code{tot_day} \tab Number of distinct Medicare beneficiary/per day services for each HCPCS code/POS combination \cr
   \code{avg_charge} \tab Average charge that the provider submitted for the service \cr
   \code{avg_allowed} \tab Average of the Medicare allowed amount for the service \cr
   \code{avg_payment} \tab Average amount Medicare paid, less any deductible and coinsurance \cr
   \code{avg_std_pymt} \tab Average amount Medicare paid, standardized to account for geographic differences \cr
}


\code{by_geography()} returns a \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Field} \tab \strong{Description} \cr
   \code{year} \tab Year data was reported \cr
   \code{level} \tab Data aggregation level, either "National" or "State" \cr
   \code{state} \tab Provider's state \cr
   \code{fips} \tab Provider's state's FIPS code \cr
   \code{hcpcs_code} \tab HCPCS code used to identify the specific medical service furnished \cr
   \code{hcpcs_desc} \tab Consumer Friendly Description of the HCPCS code \cr
   \code{category} \tab Restructured BETOS Classification System Category \cr
   \code{subcategory} \tab Restructured BETOS Classification System Subcategory \cr
   \code{family} \tab Restructured BETOS Classification System Family \cr
   \code{procedure} \tab Restructured BETOS Classification System Major Procedure Indicator \cr
   \code{drug} \tab Indicates if the HCPCS code is listed in the Medicare Part B Drug Average Sales Price (ASP) File \cr
   \code{pos} \tab Identifies the Place of Service (POS) submitted on the claim \cr
   \code{tot_provs} \tab Number of providers that submitted each HCPCS code/POS combination \cr
   \code{tot_benes} \tab Distinct number of Medicare beneficiaries for each HCPCS code/POS combination \cr
   \code{tot_srvcs} \tab Number of services provided for each HCPCS code/POS combination \cr
   \code{tot_day} \tab Number of distinct Medicare beneficiary/per day services for each HCPCS code/POS combination \cr
   \code{avg_charge} \tab Average charge submitted for the service \cr
   \code{avg_allowed} \tab Average of the Medicare allowed amount for the service \cr
   \code{avg_payment} \tab Average amount Medicare paid, less any deductible and coinsurance \cr
   \code{avg_std_pymt} \tab Average amount Medicare paid, standardized to account for geographic differences \cr
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions allow the user access to information on services and
procedures provided to Original Medicare (fee-for-service) Part B
beneficiaries by physicians and other healthcare professionals; aggregated
by provider, service and geography.
}
\section{\code{by_provider()}}{


The \strong{Provider} dataset allows the user access to data such as
services and procedures performed; charges submitted and payment received;
and beneficiary demographic and health characteristics for providers
treating Original Medicare (fee-for-service) Part B beneficiaries,
aggregated by year.
}

\section{\code{by_service()}}{


The \strong{Provider and Service} dataset is aggregated by:
\enumerate{
\item Rendering provider's NPI
\item Healthcare Common Procedure Coding System (HCPCS) code
\item Place of Service (Facility or Non-facility)
}

There can be multiple records for a given NPI based on the number of
distinct HCPCS codes that were billed and where the services were
provided. Data have been aggregated based on the place of service
because separate fee schedules apply depending on whether the place
of service submitted on the claim is facility or non-facility.
}

\section{\code{by_geography()}}{


The \strong{Geography and Service} dataset contains information on utilization,
allowed amount, Medicare payment, and submitted charges organized nationally
and state-wide by HCPCS code and place of service.
}

\section{Rural-Urban Commuting Area Codes (RUCA)}{

\itemize{
\item \strong{Metro Area Core}
\itemize{
\item \code{"1"}: Primary flow within Urbanized Area (UA)
\item \code{"1.1"}: Secondary flow 30-50\% to larger UA
}
\item \strong{Metro Area High Commuting}
\itemize{
\item \code{"2"}: Primary flow 30\% or more to UA
\item \code{"2.1"}: Secondary flow 30-50\% to larger UA
}
\item \strong{Metro Area Low Commuting}
\itemize{
\item \code{"3"}: Primary flow 10-30\% to UA
}
\item \strong{Micro Area Core}
\itemize{
\item \code{"4"}: Primary flow within large Urban Cluster (10k - 49k)
\item \code{"4.1"}: Secondary flow 30-50\% to UA
}
\item \strong{Micro High Commuting}
\itemize{
\item \code{"5"}: Primary flow 30\% or more to large UC
\item \code{"5.1"}: Secondary flow 30-50\% to UA
}
\item \strong{Micro Low Commuting}
\itemize{
\item \code{"6"}: Primary flow 10-30\% to large UC
}
\item \strong{Small Town Core}
\itemize{
\item \code{"7"}: Primary flow within small UC (2.5k - 9.9k)
\item \code{"7.1"}: Secondary flow 30-50\% to UA
\item \code{"7.2"}: Secondary flow 30-50\% to large UC
}
\item \strong{Small Town High Commuting}
\itemize{
\item \code{"8"}: Primary flow 30\% or more to small UC
\item \code{"8.1"}: Secondary flow 30-50\% to UA
\item \code{"8.2"}: Secondary flow 30-50\% to large UC
}
\item \strong{Small Town Low Commuting}
\itemize{
\item \code{"9"}: Primary flow 10-30\% to small UC
}
\item \strong{Rural Areas}
\itemize{
\item \code{"10"}: Primary flow to tract outside a UA or UC
\item \code{"10.1"}: Secondary flow 30-50\% to UA
\item \code{"10.2"}: Secondary flow 30-50\% to large UC
\item \code{"10.3"}: Secondary flow 30-50\% to small UC
}
\item \code{"99"}: Zero population and no rural-urban identifier information
}
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider}{Medicare Physician & Other Practitioners: by Provider API}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-provider-and-service}{Medicare Physician & Other Practitioners: by Provider and Service API}
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/medicare-physician-other-practitioners/medicare-physician-other-practitioners-by-geography-and-service}{Medicare Physician & Other Practitioners: by Geography and Service API}
}

\emph{Update Frequency:} \strong{Annually}
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
by_provider(year = 2020, npi = 1003000423)

by_service(year = 2019, npi = 1003000126)

by_geography(year = 2020, hcpcs_code = "0002A")

# Use the years helper function to retrieve results for every year:
pop_years() |>
map(\(x) by_provider(year = x, npi = 1043477615)) |>
list_rbind()
\dontshow{\}) # examplesIf}
}
