% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{years}
\alias{years}
\alias{open_years}
\alias{pop_years}
\alias{cc_years}
\alias{qpp_years}
\alias{bene_years}
\title{Years Currently Searchable for APIs}
\usage{
open_years()

pop_years()

cc_years()

qpp_years()

bene_years(period = c("year", "month"))
}
\arguments{
\item{period}{One of \code{"year"} or \code{"month"}}
}
\value{
integer vector of years available to search
}
\description{
Years Currently Searchable for APIs
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bene_years("year")
bene_years("month")
open_years()
pop_years()
cc_years()
qpp_years()
\dontshow{\}) # examplesIf}
}
\keyword{internal}
