% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beneficiaries.R, R/years.R
\name{beneficiaries_years}
\alias{beneficiaries_years}
\alias{years}
\alias{open_years}
\alias{prac_years}
\alias{cc_years}
\alias{quality_years}
\title{Years Currently Searchable for APIs}
\usage{
beneficiaries_years(period = c("year", "month"))

open_years()

prac_years()

cc_years()

quality_years()
}
\arguments{
\item{period}{"year", "month"}
}
\value{
integer vector of years available to search
}
\description{
Years Currently Searchable for APIs
}
\examples{
beneficiaries_years("year")
beneficiaries_years("month")
open_years()
prac_years()
cc_years()
quality_years()
}
\seealso{
\code{\link[=open_payments]{open_payments()}}, \code{\link[=by_geography]{by_geography()}}, \code{\link[=by_provider]{by_provider()}}, \code{\link[=by_service]{by_service()}}, \code{\link[=quality_payment]{quality_payment()}}, \code{\link[=quality_eligibility]{quality_eligibility()}}, \code{\link[=quality_stats]{quality_stats()}}
}
