% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years.R
\name{years}
\alias{years}
\alias{open_years}
\alias{prac_years}
\alias{cc_years}
\alias{quality_years}
\alias{bene_years}
\title{Years Currently Searchable for APIs}
\usage{
open_years()

prac_years()

cc_years()

quality_years()

bene_years(period = c("year", "month"))
}
\arguments{
\item{period}{One of \code{"year"} or \code{"month"}}
}
\value{
integer vector of years available to search
}
\description{
Years Currently Searchable for APIs
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bene_years("year")
bene_years("month")
open_years()
prac_years()
cc_years()
quality_years()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=open_payments]{open_payments()}}, \code{\link[=by_geography]{by_geography()}}, \code{\link[=by_provider]{by_provider()}}, \code{\link[=by_service]{by_service()}}, \code{\link[=quality_payment]{quality_payment()}}, \code{\link[=quality_eligibility]{quality_eligibility()}}, \code{\link[=quality_stats]{quality_stats()}}, \code{\link[=beneficiaries]{beneficiaries()}}
}
