% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_geography.R, R/by_provider.R,
%   R/by_service.R, R/cc_multiple.R, R/cc_specific.R, R/open_payments.R, R/qpp.R
\name{by_geography_years}
\alias{by_geography_years}
\alias{by_provider_years}
\alias{by_service_years}
\alias{cc_multiple_years}
\alias{cc_specific_years}
\alias{open_payments_years}
\alias{quality_payment_years}
\title{Current years available for the Geography and Service API}
\usage{
by_geography_years()

by_provider_years()

by_service_years()

cc_multiple_years()

cc_specific_years()

open_payments_years()

quality_payment_years()
}
\value{
integer vector of years available

integer vector of years available

integer vector of years available

integer vector of years available

integer vector of years available

integer vector of years available

integer vector of years available
}
\description{
Current years available for the Geography and Service API

Current years available for the Provider API

Current years available for the Provider and Service API

Check the current years available for the Multiple Chronic Conditions API

Check the current years available for the Specific Chronic Conditions API

Check the current years available for the Open Payments API

Check the current years available for the Quality Payments API
}
\examples{
by_geography_years()
by_provider_years()
by_service_years()
cc_multiple_years()
cc_specific_years()
open_payments_years()
quality_payment_years()
}
