% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_multiple.R, R/cc_specific.R,
%   R/open_payments.R, R/physician_by_geography.R, R/physician_by_provider.R,
%   R/physician_by_service.R, R/qpp.R
\name{cc_multiple_years}
\alias{cc_multiple_years}
\alias{cc_specific_years}
\alias{open_payments_years}
\alias{physician_by_geography_years}
\alias{physician_by_provider_years}
\alias{physician_by_service_years}
\alias{quality_payment_years}
\title{Check the current years available for the Multiple Chronic Conditions API}
\usage{
cc_multiple_years()

cc_specific_years()

open_payments_years()

physician_by_geography_years()

physician_by_provider_years()

physician_by_service_years()

quality_payment_years()
}
\value{
integer vector of years available

integer vector of years available

integer vector of years available

integer vector of years available

integer vector of years available

integer vector of years available

integer vector of years available
}
\description{
Check the current years available for the Multiple Chronic Conditions API

Check the current years available for the Specific Chronic Conditions API

Check the current years available for the Open Payments API

Check the current years available for the Physician & Other Practitioners by Geography and Service API

Check the current years available for the Physician & Other Practitioners by Provider API

Check the current years available for the Physician & Other Practitioners by Provider and Service API

Check the current years available for the Quality Payments API
}
\examples{
cc_multiple_years()
cc_specific_years()
open_payments_years()
physician_by_geography_years()
physician_by_provider_years()
physician_by_service_years()
quality_payment_years()
}
